Cooked.Attack

Documentation

newtype DatumHijackingLbl

Constructors

DatumHijackingLbl Address

redirectScriptOutputTweak

Arguments

:: (MonadTweak m, Is k A_Traversal, Show (DatumType a), ToData (DatumType a))

=> Optic' k is TxSkel (ConcreteOutput (TypedValidator a) TxSkelOutDatum Value (Versioned Script))

-> (ConcreteOutput (TypedValidator a) TxSkelOutDatum Value (Versioned Script) -> Maybe (TypedValidator a))
Return Just the new validator, or Nothing if you want to leave this output unchanged.

-> (Integer -> Bool)
The redirection described by the previous argument might apply to more than one of the script outputs of the transaction. Use this predicate to select which of the redirectable script outputs to actually redirect. We count the redirectable script outputs from the left to the right, starting with zero.

-> m [ConcreteOutput (TypedValidator a) TxSkelOutDatum Value (Versioned Script)]

Redirect script outputs from one validator to another validator of the same type. Returns the list of outputs it redirected (as they were before the modification), in the order in which they occurred on the original transaction.

Something like txSkelOutsL % traversed % txSkelOutOwnerTypeP (Pl.TypedValidator a)@ might be useful to construct the optics used by this tweak.

datumHijackingAttack

Arguments

:: forall a m. (MonadTweak m, Show (DatumType a), PrettyCooked (DatumType a), ToData (DatumType a), Typeable (DatumType a), Typeable a)

=> (ConcreteOutput (TypedValidator a) TxSkelOutDatum Value (Versioned Script) -> Bool)
Predicate to select outputs to steal, depending on the intended recipient, the datum, and the value.

-> (Integer -> Bool)
The selection predicate may match more than one output, restrict to the i-th of the output(s) (counting from the left, starting at zero) chosen by the selection predicate with this predicate.

-> m [ConcreteOutput (TypedValidator a) TxSkelOutDatum Value (Versioned Script)]

A datum hijacking attack, simplified: This attack tries to substitute a different recipient on PaysScript constraints, but leaves the datum as it is. That is, it tests for careless uses of something like txInfoOutputs in places where something like getContinuingOutputs should be used. If this attack goes through, however, a "proper" datum hijacking attack that modifies the datum in a way that (the relevant part of) the toBuiltinData-translation stays the same will also work.

A DatumHijackingLbl with the hash of the "thief" validator is added to the labels of the TxSkel using addLabel.

This attack returns the list of outputs it redirected, in the order in which they occurred on the original transaction. If no output is redirected, this attack fails.

data DupTokenLbl

Constructors

DupTokenLbl

dupTokenAttack

Arguments

:: MonadTweak m

=> (AssetClass -> Integer -> Integer)
A function describing how the amount of tokens specified by a Mints constraint should be changed, depending on the asset class and the amount specified by the constraint. The given function f should probably satisfy f ac i > i for all ac and i, i.e. it should increase the minted amount. If it does *not* increase the minted amount, the amount will be left unchanged.

-> Wallet
The wallet of the attacker. Any additional tokens that are minted by the modified transaction but were not minted by the original transaction are paid to this wallet.

-> m Value

A token duplication attack increases values in Mints-constraints of a TxSkel according to some conditions, and pays the extra minted value to a given recipient wallet. This adds a DupTokenLbl to the labels of the transaction using addLabel. Returns the Value by which the minted value was increased.

data DoubleSatLbl

Constructors

DoubleSatLbl

type DoubleSatDelta = (Map TxOutRef TxSkelRedeemer, [TxSkelOut], TxSkelMints)

A triple of transaction inputs, transaction outputs, and minted value. This is what we can add to the transaction in order to try a double satisfaction attack.

data DoubleSatSplitMode

How to combine extra constraints added for different reasons in the doubleSatAttack. See the comments there.

Constructors

TryCombinations

AllSeparate

doubleSatAttack

Arguments

:: (MonadTweak m, Is k A_Fold)

=> Optic' k is TxSkel a
Each focus of this optic is a potential reason to add some extra constraints.

As an example, one could go through the paysScript outputs for validators of type t with the following traversal:

txSkelOutsL % taversed % txSkelOutputToTypedValidatorP @t

-> (a -> m [DoubleSatDelta])
Which inputs, outputs, and mints to add, for each of the foci. There might be different options for each focus, that's why the return value is a list.

Continuing the example, for each of the focused paysScript outputs, you might want to try adding some spendsScript inputs to the transaction. Since it might be interesting to try different redeemers on these extra spendsScript inputs, you can just provide a list of all the options you want to try adding for a given paysScript that's already on the transaction.

###################################

ATTENTION: If you modify the state while computing these lists, the behaviour of the doubleSatAttack might be strange: Any modification of the state that happens on any call to this function will be applied to all returned transactions. For example, if you awaitTime in any of these computations, the doubleSatAttack will wait for all returned transactions.

TODO: Make this interface safer, for example by using (some kind of) an UtxoState argument.

###################################

-> Wallet
The wallet of the attacker, where any surplus is paid to.

In the example, the extra value in the added spendsScript constraints will be paid to the attacker.

-> DoubleSatSplitMode
Since there are potentially multiple triples of inputs, outputs, and mints produced for each of the foci, the question is whether (and how) to combine additions that were triggered by different foci.

In the example: Let's say that the unmodified transaction had three focused paysScript outputs (let's denote them by a, b, and c), and that you want to try 2, 3, and 5 additional spendsScript inputs for each of them, respectively (let's call those a1, a2, and b1, b2, b3, and c1, c2, c3, c4, c5).

If you want to try each additional spendsScript on its own modified transaction, use AllSeparate. Thus, there'll be 2+3+5=10 modified transactions. Namely, for each element of the list

[a1, a2, b1, b2, b3, c1, c2, c3, c4, c5]  ,

you'll get one modified transaction that includes that additional input.

If you want to try combining all options from one focus with all options from all of the other foci, use TryCombinations. Then, there'll be (2+1)*(3+1)*(5+1)=72 possible combinations, if you include all of the combinations where at most three (one for each focus) extra constraints are added. One of these combinations is of course the one where nothing is added, and that one is omitted, which brings the grand total down to 71 modified transactions, the additional spendsScript inputs of which are given by the following list:

[ -- one additional input (the 10 cases from above)
  [a1],
  [a2],
  ...
  [c4],
  [c5],

  -- two additional inputs, from different foci (2*3 + 2*5 + 3*5 = 31 cases)
  [a1, b1],
  [a1, b2],
  ...
  [b3, c4],
  [b3, c5],

  -- three additional inputs, one from each focus (2*3*5 = 30 cases)
  [a1, b1, c1],
  [a1, b1, c2],
  ...
  [a1, b3, c4],
  [a1, b3, c5]
]

So you see that this attack can branch quite wildly. Use with caution!

-> m ()

Double satisfacion attack. See the comment above for what such an attack is about conceptually.

This attack consists in adding some extra constraints to a transaction, and hoping that the additional minting policies or validator scripts thereby involved are fooled by what's already present on the transaction. Any extra value contained in new inputs to the transaction is then paid to the attacker.

newtype AddTokenLbl

Constructors

AddTokenLbl TokenName

addTokenAttack

Arguments

:: MonadTweak m

=> (Versioned MintingPolicy -> [(TokenName, NonZero Integer)])
For each policy that occurs in some Mints constraint, return a list of token names together with how many tokens with that name should be minted.

For each of the elements of the returned list, one modified transaction with the additional tokens will be generated. (This means for example that, if there were three minting policies on the original transaction, and the lists returned for each of them have n,m, and o elements, respectively, there'll be n*m*o modified transactions.)

The redeemer will be the one that's already being used on the transaction.

-> Wallet
The wallet of the attacker. Any extra tokens will be paid to this wallet.

-> m Value

This attack adds extra tokens, depending on the minting policy. It is different from the dupTokenAttack in that it does not merely try to increase the amount of tokens minted: It tries to mint tokens of asset classes that were not necessarily present on the unmodified transaction.

This attack adds an AddTokenLbl with the token name of the additional minted token(s). It returns additional value minted.