Cooked.Attack.Common

The type of attacks

type Attack = MockChainSt -> TxSkel -> [TxSkel]

The type of attacks that operate on a single transaction. The idea is that in every MockChainState, there are number of possible modifications to a transaction described by the given TxSkel. Return [] if the modification does not apply. Use these attacks with the modalities from MonadModalMockChain.

Constructing Attacks that return at most one modified transaction

mkAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify.

-> (a -> Maybe a)
The modification to apply; return Nothing if you want to leave the given focus as it is.

-> Attack

The simplest way to make an attack from an optic: Try to apply a given function of type a -> Maybe a to all foci of an optic, modifying all foci that return Just. This attack returns a singleton list of the modified TxSkel if at least one modification was successful, i.e. if at least one of the foci evaluates to Just; otherwise it returns [].

mkSelectAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify.

-> (a -> Maybe a)
The modification to apply; return Nothing if you want to leave the given focus as it is.

-> (Integer -> Bool)
Maybe the modification applies to (i.e. returns Just on) more than one focus. Use this function to select the foci to modify by the order in which they are traversed. If you set this function to const True, you should probably use mkAttack, because that the semantic of that function.

-> Attack

Traverse all foci of the given optic from the left to the right, while counting the foci to which the modification successfully applies, and modify only those modifiable foci whose index satisfies a given predicate. (This is useful for example if you have many identical outputs on a transaction but you only want to modify a few of them.) This attack returns a singleton list of the modified TxSkel if at least one modification was successful, i.e. if at least one focus was modified; otherwise it returns [].

Constructing Attacks that return zero or more modified transactions

mkSplittingAttack

Arguments

:: forall a b k is. Is k A_Traversal

=> SplitStrategy

-> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify

-> (MockChainSt -> a -> [(a, b)])
Function that returns possible modifications of the current focus, together with some piece of extra information (labels)

-> ([b] -> TxSkel -> [TxSkel])
function to look at a modified TxSkel and the list of labels of all modifications applied to it, in order to apply optional further modifications. The order of labels is not specified (yet?)

-> Attack

This attack generates potentially very many modified TxSkels, or potentially very few. For each focus, a list of possible modifications, each with an associated label, is computed. The SplitStrategy argument determines how to combine modifications and labels.

By way of example, consider

mkSplittingAttack strategy optic f g st skel

where optic has three foci x,y,z in skel. Assume that for each of these foci, the lists of possible modifications together with their labels are given by

f st x = [(x1, 'a'), (x2, 'b')]

f st y = [(y1, 'c'), (y2, 'd'), (y3, 'e')]

f st z = []

The strategy OneChange passes to g all possible modifications of the intial TxSkel obtained by applying exactly one of the modifications. In the example, this would mean 2 + 3 + 0 = 5 modified TxSkels, namely the ones corresponding to the following lists of (un)modified foci, together with their (singleton) lists of labels:

[ ([x1, y,  z], "a"),
  ([x2, y,  z], "b"),
  ([x,  y1, z], "c"),
  ([x,  y2, z], "d"),
  ([x,  y3, z], "e")
]

The strategy AllCombinations passes to g all modifications that change at least focus. In the example, there are (2+1) * (3+1) * (0+1) = 12 combinations of applied and non-applied modifications, but one of them is the one that leaves everything unchanged, which leaves us with the following 11:

[ ([x,  y1, z], "c"),
  ([x,  y2, z], "d"),
  ([x,  y3, z], "e"),
  ([x1, y,  z], "a"),
  ([x1, y1, z], "ac"),
  ([x1, y2, z], "ad"),
  ([x1, y3, z], "ae"),
  ([x2, y,  z], "b"),
  ([x2, y1, z], "bc"),
  ([x2, y2, z], "bd"),
  ([x2, y3, z], "be")
]

oneChange :: [(a, [(a, b)])] -> [([a], [b])]

allCombinations :: [(a, [(a, b)])] -> [([a], [b])]

data SplitStrategy

Constructors

OneChange

AllCombinations

mkAccumLAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify

-> (acc -> a -> (a, acc))
function that describes the modification of the accumulator and the current focus.

-> acc
initial accumulator

-> (MockChainSt -> TxSkel -> acc -> [TxSkel])
function to decide whether the traversal modified the TxSkel in the desired way. This will typically look like this: > state skel acc -> if someTest state skel acc > then computeFinalModifications state skel acc > else []

-> Attack

A very general attack: Traverse all foci of the optic from the left to the right, collecting an accumulator while also (optionally) modifying the TxSkel. At the end, look at the modified TxSkel and the accumulator to decide whether the traversal was a success, or whether we should return [].

Helpers to interact with MockChainSt

utxosSuchThatMcst :: FromData a => MockChainSt -> Address -> UtxoPredicate a -> [(SpendableOut, Maybe a)]

Like utxosSuchThat, but with the MockChainState as an explicit argument

scriptUtxosSuchThatMcst :: FromData (DatumType a) => MockChainSt -> TypedValidator a -> (DatumType a -> Value -> Bool) -> [(SpendableOut, DatumType a)]

Like scriptUtxosSuchThat, but with the MockChainState as an explicit argument

General helpers

addLabel :: LabelConstrs x => x -> TxSkel -> TxSkel

Add a label to a TxSkel. If there is already a pre-existing label, the given label will be added, forming a pair (newlabel, oldlabel).