Cooked.Attack.DoubleSat

Documentation

data DSSplitMode

How to combine extra constraints added for different reasons in the doubleSatAttack. See the comments there.

Constructors

TryCombinations

AllSeparate

doubleSatAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Each focus of this optic is a potential reason to add some extra constraints.

As an example, one could go through the PaysScript constraints for validators of type t with the following traversal:

paysScriptConstraintsT % paysScriptConstraintTypeP @t

-> (MockChainSt -> a -> [Constraints])
Which constraints to add, for each of the foci. There might be different options for each focus, that's why the return value is a list.

Continuing the example, for each of the focused PaysScript constraints, you might want to try adding some SpendsScript constraints to the transaction. Since it might be interesting to try different redeemers on these extra SpendsScript constraints, you can just provide a list of all the options you want to try adding for a given PaysScript that's already on the transaction.

-> Wallet
The wallet of the attacker, where any surplus is paid to.

In the example, the extra value in the added SpendsScript constraints will be paid to the attacker.

-> DSSplitMode
Since there are potentially multiple Constraints produced for each of the foci, the question is whether (and how) to combine additions that were triggered by different foci.

In the example: Let's say that the unmodified transaction had three focused PaysScript constraints (let's denote them by a, b, and c), and that you want to try 2, 3, and 5 additional SpendsScript constraints for each of them, respectively (let's call those a1, a2, and b1, b2, b3, and c1, c2, c3, c4, c5).

If you want to try each additional SpendsScript on its own modified transaction, use AllSeparate. Thus, there'll be 2+3+5=10 modified transactions. Namely, for each element of the list

[a1, a2, b1, b2, b3, c1, c2, c3, c4, c5]  ,

you'll get one modified transaction that includes that constraint.

If you want to try combining all options from one focus with all options from all of the other foci, use TryCombinations. Then, there'll be (2+1)*(3+1)*(5+1)=72 possible combinations, if you include all of the combinations where at most three (one for each focus) extra constraints are added. One of these combinations is of course the one where nothing is added, and that one is omitted, which brings the grand total down to 71 modified transactions, the additional SpendsScript constraints of which are given by the following list:

[ -- one additional constraint (the 10 cases from above)
  [a1],
  [a2],
  ...
  [c4],
  [c5],

  -- two additional constraints, from different foci (2*3 + 2*5 + 3*5 = 31 cases)
  [a1, b1],
  [a1, b2],
  ...
  [b3, c4],
  [b3, c5],

  -- three additional constraints, one from each focus (2*3*5 = 30 cases)
  [a1, b1, c1],
  [a1, b1, c2],
  ...
  [a1, b3, c4],
  [a1, b3, c5]
]

So you see that this attack can branch quite wildly. Use with caution!

-> Tweak ()

Double satisfacion attack. See the comment above for what such an attack is about conceptually.

This attack consists in adding some extra constraints to a transaction, and hoping that the additional minting policies or validator scripts thereby involved are fooled by what's already present on the transaction. Any extra value contained in new inputs to the transaction is then paid to the attacker.

data DoubleSatLbl

Constructors

DoubleSatLbl