Cooked.Attack.OutPermutations

Documentation

data PermutOutAttackMode

Constructors

KeepIdentity (Maybe Int)

OmitIdentity (Maybe Int)

permutOutAttack :: PermutOutAttackMode -> Attack

Modify transactions by changing the ordering of output constraints. If the PermutAttackMode is

KeepIdentity (Just n), the unmodified transaction is included in the list of modified transactions and only the first n outputs are permuted,

KeepIdentity Nothing, the unmodified transaction is included and all outputs are permuted. Use this with care; there might be a lot of permutations!

OmitIdentity (Just n), the unmodified transaction is not included in the list of modified transactions and only the first n outputs are permuted,

OmitIdentity Nothing, the unmodified transaction is not included and all outputs are permuted. Use this with care; there might be a lot of permutations!

(In particular, this is clever enough to only generate distinct permutations)

tryOutPermutations :: Maybe Int -> Attack -> Attack

For each modified TxSkel returned by the given attack, try all TxSkels obtained by permuting the output constraints. The optional integer argument is a bound on how long the prefix of the list of outputs to be permuted should be (use Nothing with care!).

distinctPermutations :: Eq a => [a] -> [[a]]

nonIdentityPermutations :: Eq a => [a] -> [[a]]