Cooked.Attack.PeerTampering

Description

This module provides an attack to replace a peer with another in every possible locations in a skeleton. If the peers have different levels of privileges, this can uncover vulnerabilities.

Peer tampering params

data PeerTamperingParams effs

Parameters of the peer tampering attack

Constructors

PeerTamperingParams

Fields

ptpBranching :: Branching

The branching policy to use when several peers are targeted

ptpChanges :: Sem effs (PubKeyHash, [PubKeyHash])

The peer to replace, associated with a list of replacing peers

purePeerTamperingParams :: Member NonDet effs => Branching -> [(PubKeyHash, [PubKeyHash])] -> PeerTamperingParams effs

A pure variant of PeerTamperingParams

singlePeerTamperingParams :: (ToPubKeyHash existing, ToPubKeyHash new) => existing -> new -> PeerTamperingParams effs

Peer tampering params transforming a single user into another

balancingPeerTamperingParams :: (ToPubKeyHash new, Members '[Tweak, NonDet] effs) => new -> PeerTamperingParams effs

Peer tampering params transforming the balancing user into another

Peer tampering label

newtype PeerTamperingLabel

A label added to a TxSkel on which a tweak tampering a peer has been applied. The label contains all the peers that have been replaced, as they were before being replaced.

Constructors

PeerTamperingLabel [PubKeyHash]

Peer tampering attack

peerTamperingAttack :: Members '[Tweak, NonDet] effs => PeerTamperingParams effs -> Sem effs [PubKeyHash]

Attempts to change the given peer into other peers everywhere in a TxSkel in an attempt to uncover permission breaches.