Cooked.Attack.TamperDatum

Documentation

tamperDatumAttack

Arguments

:: forall a. PaysScriptConstrs a

=> (DatumType a -> Maybe (DatumType a))
Use this function to return Just the changed datum, if you want to perform a change, and Nothing, if you want to leave it as-is. All datums on PaysScript constraints not paying to a validator of type a are never touched.

-> Attack

An attack that tries to change the datum on PaysScript constraints with a prescribed tampering function, that only applies to datums of a certain type.

data TamperDatumLbl

Constructors

TamperDatumLbl