Cooked.Attack.Tweak.AddConstraints

Description

Some attacks that add constraints to or remove constraints from transactions.

Documentation

addConstraintsTweak :: Constraints -> Tweak Constraints

Add constraints to a transaction. This tweak uses addMiscConstraintsTweak to add MiscConstraints in a conflict-avoiding and minimal manner, so read also the comment at that function.

If there are contradictions between the constraints that are already present and the additional constraints, this tweak fails.

OutConstraints are always added at the end of the list of output constraints, because some contracts rely on the ordering of the (intial segment of) the list of transaction outputs. If that's not to your need, you might be interested in permutOutTweak.

The returned Constraints will be the constraints that were *actually* added to the transaction, which might differ from your specified constraints (but only in MiscConstraints, so on the left-hand side of :=>:). In every case, this tweak ensures that every constraint you specified will be included in the constraints of the modified transaction.

Adding and removing MiscConstraints from transactions

addMiscConstraintTweak :: MiscConstraint -> Tweak (Maybe MiscConstraint)

This tweak adds MiscConstraints to a transaction, in a way that avoids conflicts and also potentially simplifies the constraints. In particular:

Adding a SpendsPK for an UTxO which is already consumed by the unmodified transaction will leave the transaction unmodified.

Adding a SpendsScript for an UTxO that is already consumed by the unmodified transaction will lead to failure, unless the SpendsScript constraint on the unmodified transaction uses exactly the same validator, redeemer, and datum. In that case, the tweak will leave the transaction unmodified.

Mints constraints are added without any additional checks.

if any constraint concerning time (i.e. Before, After, or ValidateIn) is added, the resulting transaction will only have one ValidateIn constraint, specifying as the validity time range the intersection of all the pre-existing time ranges on the transaction and the new time range.

if a SignedBy constraint is added, the resulting transaction will similarly have only one SignedBy constraint, which holds all of the signers required by the unmodified transaction, together with the new signers.

The returned MiscConstraint is either Just the constraint that was added to the transaction, or Nothing, if your constraint was already present on the unmodified transaction.

removeMiscConstraintsTweak :: (MiscConstraint -> Maybe a) -> Tweak [a]

This tweak removes some MiscConstraints from a transaction. Exactly those constraints that return Just something are removed from the transaction.

The tweak returns a list of all the Just values that were returned by the removed constraints.

partitionMaybe :: (a -> Maybe b) -> [a] -> ([b], [a])

Partition a list into two lists, where the first output list contains all values that were returned by Just, and the second records all elements of the input list that returned Nothing. The order of elements is preserved in both output lists.

addSpendsScriptTweak :: forall a. SpendsConstrs a => TypedValidator a -> RedeemerType a -> SpendableOut -> Tweak (Maybe MiscConstraint)

this tweak ensures that a certain SpendsScript constraint is present on a transaction.

Adding a SpendsScript for an UTxO that is already consumed by the unmodified transaction will lead to failure, unless the SpendsScript constraint on the unmodified transaction uses exactly the same validator, redeemer, and datum. In that case, the tweak will leave the transaction unmodified.

This tweak returns Just the added constraint iff the transaction was modified.

addSpendsPKTweak :: SpendableOut -> Tweak (Maybe MiscConstraint)

This tweak ensures that a certain SpendsPK constraint is present on a transaction.

This tweak returns Just the added constraint iff the transaction was modified.

addMintsTweak :: MintsConstrs a => Maybe a -> [MintingPolicy] -> Value -> Tweak (Maybe MiscConstraint)

This tweak adds a Mints constraint.

This tweak always returns Just the added constraint. Its return type is in Maybe in order to be homogeneous with addSpendsScriptTweak, addSpendsPKTweak, addValidateInTweak, and addSignedByTweak.

removeTimeConstraintsTweak :: Tweak POSIXTimeRange

This tweak removes all time constraints (i.e. Before, After, and ValidateIn) from a transaction.

Returns the validity time range of the unmodified transaction.

addValidateInTweak :: POSIXTimeRange -> Tweak (Maybe MiscConstraint)

This tweak restricts the validity time range of a transaction to the intersection of the given range and its current validity time range.

If this tweak performs any change at all, it also "cleans up" the time constraints (i.e. Before, After, and ValidateIn), so that the resulting transaction has only one ValidateIn constraint.

This tweak returns Just the added constraint iff the transaction was modified.

removeSignedByTweak :: Tweak [PubKeyHash]

This tweak removes all SignedBy constraints from a transaction. It returns a list of the signers it removed.

addSignedByTweak :: [PubKeyHash] -> Tweak (Maybe MiscConstraint)

This tweak adds signers to a transaction. It also combines all SignedBy constraints of the transaction into one, if it performs any change at all.

This tweak returns Just the added constraint iff the transaction was modified.

Adding and Removing OutConstraints

addOutConstraintTweak :: OutConstraint -> Tweak ()

Add an OutConstraint to a transaction. The additional constraint will be added to the end of the list of output constraints. This is because some contracts rely on the shape of (the initial segment of) the list of transaction outputs. Try permutOutTweak if you want to change the ordering of output constraints.

removeOutConstraintsTweak :: (OutConstraint -> Maybe a) -> Tweak [a]

This tweak removes some OutConstraints from a transaction. Exactly those constraints that return Just something are removed from the transaction.

The tweak returns a list of all the Just values that were returned by the removed constraints.