Cooked.Attack.Tweak.Common

Description

This module defines Tweaks which are the fundamental building blocks of our "domain specific language" for attacks.

The type of tweaks

newtype Tweak a

A tweak is a function that, depending on the MockChainState, looks at a transaction and computes a list of modified transactions, together with some additional values.

Our intuition (and also the language of the comments pertaining to tweaks) is that a tweak

fails if it returns []

modifies a transaction, where the unmodified transaction is the name we give to the input TxSkel, and each of the TxSkels in the output list is a modified transaction

returns the value(s) in the snd component of the pairs in the output list. This is reflected by the Monad instance for Tweaks.

Constructors

Tweak

Fields

getTweak :: MockChainSt -> TxSkel -> [(TxSkel, a)]

data UntypedTweak where

Internal wrapper type for compatibility with the LTL modalities. You'll probably never work with this type if you want to build and use tweaks or peeks.

Constructors

UntypedTweak :: Tweak a -> UntypedTweak

Simple tweaks

failingTweak :: Tweak a

The never-applicable tweak.

doNothingTweak :: Tweak ()

The tweak that always applies and leaves the transaction unchanged.

mcstTweak :: Tweak MockChainSt

The "tweak" that returns the current MockChainSt. This does *not* modify the transaction.

viewTweak :: Is k A_Getter => Optic' k is TxSkel a -> Tweak a

The "tweak" that obtains some value from the TxSkel. This does *not* modify the transaction.

setTweak :: Is k A_Setter => Optic' k is TxSkel a -> a -> Tweak ()

The tweak that sets a certain value in the TxSkel.

overTweak :: Is k A_Setter => Optic' k is TxSkel a -> (a -> a) -> Tweak ()

The tweak that modifies a certain value in the TxSkel.

Composing tweaks

tryTweak :: Tweak a -> Tweak (Maybe a)

Turn a potentially failing tweak into an always successful tweak, that just leaves the TxSkel unmodified if the original tweak would have failed.

In cases where the original tweak would have failed, this returns Nothing. If the original tweak would have been applicable, this is signalled by wrapping the original tweak's return value in a Just.

getTxSkel :: Tweak TxSkel

This "tweak" returns the transaction as it has been modified by now. Use this as a kind of savepoint, if you want execute some sequence of tweaks of which you're not sure that they will lead to the right result. You can restore the savepoint after that sequence with setTxSkel.

setTxSkel :: TxSkel -> Tweak ()

See the comment at getTxSkel.

Applying Tweaks to Constraints

applyToConstraints :: Tweak a -> MockChainSt -> Constraints -> [Constraints]

Most tweaks do something interesting only to the Constraints of a transaction. This function extracts that action on Constraints from an tweak.

Constructing Tweaks from Optics

mkTweak

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify.

-> (MockChainSt -> a -> Maybe a)
The modification to apply; return Nothing if you want to leave the given focus as it is.

-> Tweak [a]

Probably the most common way to make an tweak from an optic: Try to apply a given function of type MockChainSt -> a -> Maybe a to all foci of an optic, modifying all foci that return Just ....

If at least one focus was modified, this tweak returns a list of the foci that were modified, in the order in which they occured on the original transaction (and with no modification applied to them).

If no focus was modified, or if nothing was in focus, the tweak fails.

mkSelectTweak

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify.

-> (MockChainSt -> a -> Maybe a)
The modification to apply; return Nothing if you want to leave the given focus as it is.

-> (Integer -> Bool)
Maybe more than one of the foci is modifiable (i.e. the modification returns Just). Use this predicate to selectively apply the modification only to some of the modifiable foci: This tweak counts the modifiable foci in the order of the traversal, starting with 0, and only applies the modification to the i-th modifiable focus if i satisfies the predicate.

-> Tweak [a]

Sometimes mkTweak modifies too many foci. For example, there might be many identical outputs on a transaction but you only want to modify a few of them.

This is the problem solved by this tweak. It traverses all foci of the given optic from the left to the right, while counting the foci to which the modification successfully applies, but modifies the i-th modifiable focus only if i satisfies a given predicate.

The return value and failure behaviour of this tweak is similar to the one of mkTweak: It returns a list of the modified foci, as they were before the modification, and fails if nothing was modified.

mkAccumLTweak

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify

-> (MockChainSt -> acc -> a -> (a, acc))
function that describes the modification of the accumulator and the current focus.

-> acc
Initial accumulator.

-> Tweak acc

A very simple, but flexible way to build an tweak from an optic: Traverse all foci of the optic from the left to the right, collecting an accumulator while also modifying the foci.

This tweak never fails, and returns the accumulator, even if no modifications were made to the TxSkel. So, it's up to the caller of this function to look at the accumulator and decide how to proceed.

Helpers to interact with MockChainSt

utxosSuchThatMcst :: FromData a => MockChainSt -> Address -> UtxoPredicate a -> [(SpendableOut, Maybe a)]

Like utxosSuchThat, but with the MockChainState as an explicit argument

scriptUtxosSuchThatMcst :: FromData (DatumType a) => MockChainSt -> TypedValidator a -> (DatumType a -> Value -> Bool) -> [(SpendableOut, DatumType a)]

Like scriptUtxosSuchThat, but with the MockChainState as an explicit argument

Some more simple tweaks

changeValueTweak

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel Value

-> (Value -> Value)
A function describing how a focused value should change. Return the modified value.

-> Tweak [Value]

Change some Values in a TxSkel. Returns a list of the increments by which all of the focused values were changed.

This tweak never fails.

addLabelTweak :: LabelConstrs x => x -> Tweak ()

Add a label to a TxSkel. If there is already a pre-existing label, the given label will be added, forming a pair (newlabel, oldlabel).