Cooked.Conversion.ToHash

Description

This module provides an interface for plutus elements that can be hashed. This is mostly used to provide alias for hashes when pretty printing those elements.

Documentation

class ToHash a where

Methods

toHash :: a -> BuiltinByteString