Cooked.Currencies

Description

This module introduces currencies (namely the quick values and the permanent values) which make it convenient to manipulate assets that are already supposed to exist when running a mock chain. For example, a market maker would exchange Ada against other assets. Yet, when writing traces for such a contract we would need to define a minting policy for those tokens, which is tedious. Moreover, we often want wallets to have some of such tokens from the start (see Cooked.InitialDistribution).

The quick prefixed functions provide access to tokens from the const (const True) minting policy. That is, these can be minted and burnt at will, at any point in time.

The permanent prefixed functions provide access to tokens from the const (const False) minting policy. That is, these cannot ever be minted or burnt and must be present in an initial distribution.

Documentation

quickTokenName :: String -> TokenName

Token name of a quick asset class; prefixes the name with a q to make it easy to distinguish between quick and permanent tokens.

quickAssetClass :: String -> AssetClass

Quick asset class from a token name

quickValue :: String -> Integer -> Value

Constructor for quick values from token name and amount

permanentTokenName :: String -> TokenName

Token name of a permanent asset class

permanentAssetClass :: String -> AssetClass

Permanent asset class from a token name

permanentValue :: String -> Integer -> Value

Constructor for Permanent values from token name and amount

quickCurrencyPolicy :: MintingPolicy

quickCurrencyPolicyV3 :: Versioned MintingPolicy

quickCurrencySymbol :: CurrencySymbol

permanentCurrencyPolicy :: MintingPolicy

permanentCurrencyPolicyV3 :: Versioned MintingPolicy

permanentCurrencySymbol :: CurrencySymbol

currencySymbolFromLanguageAndMP :: Language -> MintingPolicy -> CurrencySymbol

Takes a minting policy and a language version and returns the associated currency symbol