Cooked.MockChain.AutoReferenceScripts

Description

This module provides a function to ensure that each redeemer used in a skeleton is attached a reference input with the right reference script when it exists in the index.

Documentation

toTxSkelWithReferenceScripts :: MonadBlockChain m => TxSkel -> m TxSkel

Goes through the various parts of the skeleton where a redeemer can appear, and attempt to attach a reference input to each of them, following the rules from updateRedeemer