Cooked.MockChain.Balancing

Description

This module handles auto-balancing of transaction skeleton. This includes computation of fees and collaterals because their computation cannot be separated from the balancing.

Documentation

type Body = TxBody ConwayEra

A transaction body

data ExtendedTxSkel

A TxSkel with extra pieces of information produced during balancing

Constructors

ExtendedTxSkel

Fields

eSkel :: TxSkel

The skeleton itself

eFee :: Fee

The fee associated with this skeleton

eMCollaterals :: Maybe Collaterals

The optional collaterals associated with this skeleton

eBody :: Body

The Cardano body generated from this skeleton

balanceTxSkel :: Members '[MockChainRead, MockChainLog, Error MockChainError, Error ToCardanoError, Fail] effs => TxSkel -> Sem effs ExtendedTxSkel

This is the main entry point of our balancing mechanism. This function takes a skeleton and returns a (possibly) balanced skeleton alongside the associated fee, collateral inputs and return collateral user, which might be empty when no script is involved in the transaction. The options from the skeleton control whether it should be balanced, and how to compute its associated elements.

getMinAndMaxFee :: Members '[MockChainRead] effs => Integer -> Sem effs (Fee, Fee)

This computes the minimum and maximum possible fee a transaction can cost based on the current protocol parameters and its number of scripts. In the Dijsktra era, this will be modified with new protocol parameters. See https://github.com/IntersectMBO/cardano-ledger/blob/master/docs/adr/2024-08-14_009-refscripts-fee-change.md for more information

estimateTxSkelFee :: Members '[MockChainRead, Error MockChainError, Error ToCardanoError, Fail] effs => TxSkel -> Fee -> Maybe Collaterals -> Sem effs (Fee, Body)

Estimates the required fee for a given skeleton with a given initial fee and collaterals