Cooked.MockChain.BlockChain

BlockChain Monad

class MonadFail m => MonadBlockChainWithoutValidation m where

Methods

allUtxos :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs

datumFromHash :: DatumHash -> m (Maybe (Datum, Doc ()))

Returns the datum with the given hash, along with its pretty-printed representation, or Nothing if there is none

txOutByRef :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

ownPaymentPubKeyHash :: m PubKeyHash

Returns the hash of our own public key. When running in the Plutus.Contract.Contract monad, this is a proxy to ownPubKey; when running in mock mode, the return value can be controlled with signingWith: the head of the non-empty list will be considered as the "ownPubkey".

currentSlot :: m Slot

Returns the current slot number

currentTime :: m POSIXTime

Returns the current time

awaitSlot :: Slot -> m Slot

Either waits until the given slot or returns the current slot. Note that that it might not wait for anything if the current slot is larger than the argument.

awaitTime :: POSIXTime -> m POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

Example: if starting time is 0 and slot length is 3s, then `awaitTime 4` waits until slot 2 and returns the value `POSIXTime 5`.

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates and balances a transaction from a skeleton, then attemps to validate such transaction. A balanced transaction is such that inputs + mints == outputs + fees. To balance a transaction, we need access to the current UTxO state to choose which inputs to add in case the output-side of the balancing equation is bigger.

The TxSkel receives a TxOpts record with a number of options to customize how validation works.

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

txOutV2fromV1 :: TxOut -> TxOut

filterUtxos :: (o1 -> Maybe o2) -> [(TxOutRef, o1)] -> [(TxOutRef, o2)]

Helper function to filter the output of allUtxos and utxosFromCardanoTx

pkUtxosMaybeDatum :: MonadBlockChainWithoutValidation m => PubKeyHash -> m [(TxOutRef, PKOutputMaybeDatum)]

Return all UTxOs belonging to a particular pubkey, no matter their datum or value.

pkUtxos :: MonadBlockChainWithoutValidation m => PubKeyHash -> m [(TxOutRef, PKOutput)]

Return all UTxOs belonging to a particular pubkey that have no datum on them.

allUtxosWithDatums :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

Like allUtxos, but on every OutputDatumHash, try to resolve the complete datum from the state

filteredUtxos :: MonadBlockChainWithoutValidation m => (TxOut -> Maybe output) -> m [(TxOutRef, output)]

filteredUtxosWithDatums :: MonadBlockChainWithoutValidation m => (TxOut -> Maybe output) -> m [(TxOutRef, output)]

Like filteredUtxos, but will all resolvable datum hashes resolved. This means that the outputs are presented differently from how a script would see them; the information on whether there are inline datums or datum hashes is lost.

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

Slot and Time Management

Slots are integers that monotonically increase and model the passage of time. By looking at the current slot, a validator gets to know that it is being executed within a certain window of wall-clock time. Things can get annoying pretty fast when trying to mock traces and trying to exercise certain branches of certain validators; make sure you also read the docs on autoSlotIncrease to be able to simulate sending transactions in parallel.

waitNSlots :: MonadBlockChainWithoutValidation m => Integer -> m Slot

waitNMilliSeconds :: MonadBlockChainWithoutValidation m => DiffMilliSeconds -> m POSIXTime

Deriving further MonadBlockChain instances

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTrans.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the boilerplate of defining all the methods of the class yourself.

Constructors

AsTrans

Fields

getTrans :: t m a