Cooked.MockChain.Common

Description

This module exposes some type aliases common to our MockChain library

Type aliases

type Fee = Integer

An alias for Integers used as fees

type CollateralIns = Set TxOutRef

An alias for sets of utxos used as collateral inputs

type Collaterals = (CollateralIns, Maybe TxSkelOut)

An alias for optional pairs of collateral inputs and optional return collateral output

type Utxo = (TxOutRef, TxSkelOut)

An alias for an output and its reference

type Utxos = [Utxo]

An alias for lists of Utxo