Cooked.MockChain.Direct

Description

This module provides a direct (as opposed to Staged) implementation of the MonadBlockChain specification. This rely on the emulator from cardano-node-emulator for transaction validation, although we have our own internal state. This choice might be revised in the future.

Direct Emulation

The MockChainT monad provides a direct emulator; that is, it gives us a simple way to run a full validation process directly, without relying on a deployed node. While simulated, the validation is performed by the cardano-ledger code, thus ensuring similar results on the real chain.

A MockChain:

stores and updates a MockChainState

returns a UtxoState when run

emits entries in a MockChainBook

data MockChainBook where

This represents elements that can be emitted throughout a MockChain run. These elements are either log entries corresponding to internal events worth logging, or aliases for hashables corresponding to elements users wishes to be properly displayed when printed with PrettyCooked

Constructors

MockChainBook

Fields

:: { mcbJournal :: [MockChainLogEntry]

Log entries generated by cooked-validators

, mcbAliases :: Map BuiltinByteString String

Aliases stored by the user

} -> MockChainBook

newtype MockChainT m a

A MockChainT builds up a stack of monads on top of a given monad m to reflect the requirements of the simulation. It writes a MockChainBook, updates and reads from a MockChainState and throws possible MockChainErrors.

Constructors

MockChainT

Fields

unMockChain :: ExceptT MockChainError (StateT MockChainState (WriterT MockChainBook m)) a

type MockChain = MockChainT Identity

Our MockChain naturally instantiate the inner monad with Identity

combineMockChainT :: (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

Combines two MockChainT together

MockChain return types

data MockChainReturn a where

The returned type when running a MockChainT. This is both a reorganizing and filtering of the natural returned type ((Either MockChainError a, MockChainState), MockChainBook), which is much easier to query.

Constructors

MockChainReturn

Fields

:: { mcrValue :: Either MockChainError a

The value returned by the computation, or an error

, mcrOutputs :: Map TxOutRef (TxSkelOut, Bool)

The outputs at the end of the run

, mcrUtxoState :: UtxoState

The UtxoState at the end of the run

, mcrJournal :: [MockChainLogEntry]

The final journal emitted during the run

, mcrAliases :: Map BuiltinByteString String

The map of aliases defined during the run

} -> MockChainReturn a

type RawMockChainReturn a = ((Either MockChainError a, MockChainState), MockChainBook)

Raw return type of running a MockChainT

unRawMockChainReturn :: RawMockChainReturn a -> MockChainReturn a

Building a MockChainReturn from a RawMockChainReturn

MockChain configurations

data MockChainConf a b where

Configuration to run a MockChainT

Constructors

MockChainConf

Fields

:: { mccInitialState :: MockChainState

The initial state from which to run the MockChainT

, mccInitialDistribution :: InitialDistribution

The initial payments to issue in the run

, mccFunOnResult :: RawMockChainReturn a -> b

The function to apply on the result of the run

} -> MockChainConf a b

initDistConf :: InitialDistribution -> MockChainConf a (MockChainReturn a)

A configuration with a default initial state, a given distribution, returning a MockChainReturn

mockChainStateConf :: MockChainState -> MockChainConf a MockChainState

A configuration with a given initial MockChainState, a default initial distribution, returning the final MockChainState

MockChain runs

runMockChainTFromConf :: Monad m => MockChainConf a b -> MockChainT m a -> m b

Runs a MockChainT using a certain configuration

runMockChainFromConf :: MockChainConf a b -> MockChain a -> b

Runs a MockChain using a certain configuration

runMockChainTFromInitDist :: Monad m => InitialDistribution -> MockChainT m a -> m (MockChainReturn a)

Runs a MockChainT from an initial InitialDistribution

runMockChainFromInitDist :: InitialDistribution -> MockChain a -> MockChainReturn a

See runMockChainTFromInitDist

runMockChainT :: Monad m => MockChainT m a -> m (MockChainReturn a)

Uses runMockChainTFromInitDist with a default InitialDistribution

runMockChain :: MockChain a -> MockChainReturn a

Uses runMockChainFromInitDist with a default InitialDistribution

Direct Interpretation of Operations