Cooked.MockChain.GenerateTx

Description

This module exposes translation functions to transform a TxSkel into a signed transaction

Documentation

txSignersAndBodyToCardanoTx :: [Wallet] -> TxBody ConwayEra -> Tx ConwayEra

Generates a Cardano transaction and signs it

txSkelToCardanoTx :: MonadBlockChainBalancing m => TxSkel -> Integer -> Maybe (Set TxOutRef, Wallet) -> m (Tx ConwayEra)

Generates a full Cardano transaction from a skeleton, fees and collaterals