Cooked.MockChain.GenerateTx.Anchor

Description

Transforming TxSkelAnchor into its Cardano counterpart

Documentation

toCardanoAnchor :: TxSkelAnchor -> Anchor

This function transforms a TxSkelAnchor into its Cardano counterpart. If the provided anchor does not provde a resolved page, it will be unsafely fetched online, so use at your own discretion.