Cooked.MockChain.GenerateTx.Body

Description

This modules exposes entry points to convert a TxSkel into a fully fledged transaction body

Documentation

txSkelToTxBody :: MonadBlockChainBalancing m => TxSkel -> Fee -> Collaterals -> m (TxBody ConwayEra)

Generates a transaction body from a TxSkel and associated fee and collateral information. This transaction body accounts for the actual execution units of each of the scripts involved in the skeleton.

txBodyContentToTxBody :: MonadBlockChainBalancing m => TxBodyContent BuildTx ConwayEra -> m (TxBody ConwayEra)

Generates a transaction body from a body content

txSkelToTxBodyContent :: MonadBlockChainBalancing m => TxSkel -> Fee -> Collaterals -> m (TxBodyContent BuildTx ConwayEra)

Generates a body content from a skeleton

txSkelToIndex :: MonadBlockChainBalancing m => TxSkel -> Collaterals -> m (UTxO ConwayEra)

Generates an index with utxos known to a TxSkel

txSignatoriesAndBodyToCardanoTx :: [TxSkelSignatory] -> TxBody ConwayEra -> Tx ConwayEra

Generates a Cardano transaction and signs it

txSkelToCardanoTx :: MonadBlockChainBalancing m => TxSkel -> Fee -> Collaterals -> m (Tx ConwayEra)

Generates a full Cardano transaction from a skeleton, fees and collaterals