Cooked.MockChain.GenerateTx.Collateral

Description

This module exposes the generation of transaction collaterals, which consist of a collateral amount, collateral inputs and return collateral

Documentation

toCollateralTriplet :: MonadBlockChainBalancing m => Fee -> Collaterals -> m (TxInsCollateral ConwayEra, TxTotalCollateral ConwayEra, TxReturnCollateral CtxTx ConwayEra)

Computes the collateral triplet from the fees and the collateral inputs in the context. What we call a collateral triplet is composed of: * The set of collateral inputs * The total collateral paid by the transaction in case of phase 2 failure * An output returning excess collateral value when collaterals are used These quantity should satisfy the equation (in terms of their values): collateral inputs = total collateral + return collateral