Cooked.MockChain.GenerateTx.Common

Documentation

data GenerateTxError

Errors that can arise during transaction generation

Constructors

ToCardanoError String ToCardanoError

TxBodyError String TxBodyError

GenerateTxErrorGeneral String

type TxGen context a = ReaderT context (Either GenerateTxError) a

The domain in which transaction parts are generated.

class Transform a b where

Methods

transform :: a -> b

throwOnLookup :: Ord k => String -> k -> Map k a -> TxGen context a

Looks up a key in a map. Throws a GenerateTxErrorGeneral error with a given message when the key is absent, returns the associated value otherwise.

throwOnString :: String -> TxGen context a

Throws a general error from a String.

throwOnToCardanoErrorOrApply :: String -> (a -> b) -> Either ToCardanoError a -> TxGen context b

Lifts a ToCardanoError with an associated error message, or apply a function if a value exists.

throwOnToCardanoError :: String -> Either ToCardanoError a -> TxGen context a

Lifts a ToCardanoError with an associated error message, or leaves the value unchanged if it exists.

liftTxGen :: Transform context' context => TxGen context a -> TxGen context' a

Lifts a computation from a smaller context