Cooked.MockChain.GenerateTx.Credential

Description

This module exposes the generation of various kinds of credentials

Documentation

toRewardAccount :: Member (Error ToCardanoError) effs => Credential -> Sem effs RewardAccount

Translates a given credential to a reward account.

toCardanoCredential :: (Member (Error ToCardanoError) effs, SerialiseAsRawBytes (Hash key)) => AsType key -> (Hash key -> KeyHash kr) -> Credential -> Sem effs (Credential kr)

Converts an Credential to a Cardano Credential of the expected kind

toStakeCredential :: Member (Error ToCardanoError) effs => Credential -> Sem effs (Credential 'Staking)

Translates a credential into a Cardano stake credential

deserialiseFromBuiltinByteString :: (Member (Error ToCardanoError) effs, SerialiseAsRawBytes a) => AsType a -> BuiltinByteString -> Sem effs a

Converts an PubKeyHash to any kind of key

toScriptHash :: Member (Error ToCardanoError) effs => ScriptHash -> Sem effs ScriptHash

Converts a plutus script hash into a cardano ledger script hash

toKeyHash :: (Member (Error ToCardanoError) effs, SerialiseAsRawBytes (Hash key)) => AsType key -> (Hash key -> KeyHash kr) -> PubKeyHash -> Sem effs (KeyHash kr)

Converts a plutus pkhash into a certain cardano ledger hash

toDRepCredential :: Member (Error ToCardanoError) effs => Credential -> Sem effs (Credential 'DRepRole)

Translates a credential into a Cardano drep credential

toStakePoolKeyHash :: Member (Error ToCardanoError) effs => PubKeyHash -> Sem effs (KeyHash 'StakePool)

Converts an PubKeyHash into a cardano ledger stake pool key hash

toColdCredential :: Member (Error ToCardanoError) effs => Credential -> Sem effs (Credential 'ColdCommitteeRole)

Translates a credential into a Cardano cold committee credential

toHotCredential :: Member (Error ToCardanoError) effs => Credential -> Sem effs (Credential 'HotCommitteeRole)

Translates a credential into a Cardano hot committee credential

toVRFVerKeyHash :: Member (Error ToCardanoError) effs => PubKeyHash -> Sem effs (VRFVerKeyHash a)

Converts an PubKeyHash into a cardano ledger VRFVerKeyHash