Cooked.MockChain.GenerateTx.ReferenceInputs

Description

This module allows the generation of Cardano reference inputs

Documentation

toInsReference :: MonadBlockChainBalancing m => TxSkel -> m (TxInsReference BuildTx ConwayEra)

Takes a TxSkel and generates the associated TxInsReference from its content. These reference inputs can be found in two places, either in direct reference inputs txSkelInsReference or scattered in the various redeemers of the transaction, which can be gathered with txSkelInsReferenceInRedeemers.