Cooked.MockChain.Instances

Description

This module exposes concrete instances to run a mockchain. There are 4 of them :

DirectMockChain exposes the minimal set of effects required to run a mockchain, without the ability to branch or modify runs. Use this only if you specifically want to disallow Ltl modifications.

StagedMockChain exposes all the primitives required to run a mockchain, with the addition of branching and Ltl modifications using tweaks. This should be the environement to use in 99% of the cases.

ExtendedStagedMockChain exposes the same primitives as StagedMockChain, with an additional custom effect that can both be used in the main thread and in the associated tweaks. This allows a mockchain run to depend on arbitrary additional effects (if multiple effects are needed, this single effect can be instantiated to a Bundle wrapping up those effects).

FullMockChain exposes all the effects used to process a mockchain run, including intermediate hidden in the other instances. This should only be used when explicitly executing internal primitives of cooked, such as balancing, is required.

Direct, simple mockchain instance

type DirectEffs = '[MockChainWrite, MockChainRead, MockChainMisc, Fail]

The most direct stack of effects to run a mockchain

type DirectMockChain a = Sem DirectEffs a

A mockchain computation built on top of the DirectEffs stack of effects

Staged mockchain instance with all effects

type FullTweakEffs = '[MockChainMisc, MockChainRead, Fail, Error ToCardanoError, Error MockChainError, State MockChainState, MockChainLog, Writer MockChainJournal]

A stack of effects aimed at being used as modifications for a FullMockChain computation

type FullTweak a = TypedTweak FullTweakEffs a

A tweak computation based on the FullTweakEffs stack of effects

type FullEffs = '[ModifyGlobally (UntypedTweak FullTweakEffs), MockChainWrite, ModifyLocally (UntypedTweak FullTweakEffs), State [Ltl (UntypedTweak FullTweakEffs)], MockChainMisc, MockChainRead, Fail, Error ToCardanoError, Error MockChainError, State MockChainState, MockChainLog, Writer MockChainJournal, NonDet]

A stack of effects which allows everything allowed by StagedEffs with the addition of all the lower level effects required to interpret it.

type FullMockChain a = Sem FullEffs a

A mockchain computation built on top of the FullEffs stack of effects

Staged mockchain instance with minimal effects

type StagedTweakEffs = ExtendedStagedTweakEffs (Bundle '[])

A stack of effects aimed at being used as modifications for a StagedMockChain computation

type StagedTweak a = TypedTweak StagedTweakEffs a

A tweak computation based on the StagedTweakEffs stack of effects

type StagedEffs = ExtendedStagedEffs (Bundle '[])

A stack of effects which allows everything allowed by DirectEffs with the addition of branching and Ltl modification with tweaks living in StagedTweakEffs

type StagedMockChain a = Sem StagedEffs a

A mockchain computation built on top of the StagedEffs stack of effects

Staged mockchain instance with minimal effects and a custom effect

class InterpretAlone eff where

The class of effects that can be interpreted on their own on top of an arbitrary stack of effects

Methods

runInterpretAlone :: Sem (eff : effs) a -> Sem effs a

type ExtendedStagedTweakEffs extraEff = '[extraEff, MockChainMisc, MockChainRead, Fail]

A stack of effects aimed at being used as modifications for a StagedMockChain computation

type ExtendedStagedTweak extraEff a = TypedTweak (ExtendedStagedTweakEffs extraEff) a

A tweak computation based on the ExtendedStagedTweakEffs stack of effects

type ExtendedStagedEffs extraEff = '[ModifyGlobally (UntypedTweak (ExtendedStagedTweakEffs extraEff)), MockChainWrite, extraEff, MockChainMisc, MockChainRead, Fail, NonDet]

A stack of effects which allows everything allowed by DirectEffs with the addition of branching and Ltl modification with tweaks living in ExtendedStagedTweakEffs

type ExtendedStagedMockChain extraEff a = Sem (ExtendedStagedEffs extraEff) a

A mockchain computation built on top of the ExtendedStagedEffs stack of effects

Orphan instances

RunnableMockChain DirectEffs

Instance details

Methods

runMockChain :: MockChainState -> Sem DirectEffs a -> [RawMockChainReturn a]

RunnableMockChain FullEffs

Instance details

Methods

runMockChain :: MockChainState -> Sem FullEffs a -> [RawMockChainReturn a]

InterpretAlone extraEff => RunnableMockChain (ExtendedStagedEffs extraEff)

Instance details

Methods

runMockChain :: MockChainState -> Sem (ExtendedStagedEffs extraEff) a -> [RawMockChainReturn a]