Cooked.MockChain.Journal

Description

This module exposes the various events emitted during a mockchain run.

Documentation

data MockChainJournal where

This represents the writable elements that can be emitted throughout a mockchain run.

Constructors

MockChainJournal

Fields

:: { mcbLog :: [MockChainLogEntry]

Log entries generated by cooked-validators

, mcbAliases :: Map BuiltinByteString String

Aliases stored by the user

, mcbNotes :: [PrettyCookedOpts -> DocCooked]

Notes taken by the user, parameterized by some pretty cooked options, to get a better display at the end of the run

, mcbAssertions :: [(String, Bool)]

Assertions gathered during the run, alongside their associated error messages to display in case of failure

} -> MockChainJournal

fromLogEntry :: MockChainLogEntry -> MockChainJournal

Build a MockChainJournal from a single log entry

fromAlias :: String -> BuiltinByteString -> MockChainJournal

Build a MockChainJournal from a single alias

fromNote :: (PrettyCookedOpts -> DocCooked) -> MockChainJournal

Build a MockChainJournal from a single note

fromAssert :: String -> Bool -> MockChainJournal

Build a MockChainJournal from a single assertion and error message