Cooked.MockChain.MockChainSt

Description

This module exposes the internal state in which our direct simulation is run, and functions to update and querry it.

Documentation

data MockChainSt

The state used to run the simulation in Direct

Constructors

MockChainSt

Fields

mcstParams :: Params

mcstIndex :: UtxoIndex

mcstDatums :: Map DatumHash (TxSkelOutDatum, Integer)

mcstScripts :: Map ScriptHash (Versioned Script)

mcstCurrentSlot :: Slot

getIndex :: UtxoIndex -> Map TxOutRef TxOut

Converts a builtin UtxoIndex into our own usable map between utxos and associated outputs.

mcstToUtxoState :: MockChainSt -> UtxoState

Builds a UtxoState from a MockChainSt

mcstToSkelContext :: MockChainSt -> SkelContext

Generating a skeleton context from a mockchain state. This is dedicated to allowing the pretty printer to resolve skeleton parts.

mcstToEmulatedLedgerState :: MockChainSt -> EmulatedLedgerState

Generating an emulated state for the emulator from a mockchain state and some parameters, based on a standard initial state

addDatums :: [(DatumHash, TxSkelOutDatum)] -> MockChainSt -> MockChainSt

Adds a list of pairs (datumHash, datum) into a MockChainSt

removeDatums :: [DatumHash] -> MockChainSt -> MockChainSt

Removes a certain amound of datum hashes from a MockChainSt

addScript :: (ToScriptHash s, ToVersioned Script s) => s -> MockChainSt -> MockChainSt

Stores a script in a MockChainSt

Initial MockChainSt from an InitialDistribution

mockChainSt0From :: MonadBlockChainBalancing m => InitialDistribution -> m MockChainSt

Builds a MockChainSt from an InitialDistribution. This lives in MonadBlockChainBalancing because the creation of mcstIndex is impure

referenceScriptMap0From :: InitialDistribution -> Map ScriptHash (Versioned Script)

Collects the reference scripts present in an InitialDistribution

scriptMap0From :: InitialDistribution -> Map ScriptHash (Versioned Script)

Collects the scripts paid to in an InitialDistribution

datumMap0From :: InitialDistribution -> Map DatumHash (TxSkelOutDatum, Integer)

Collects the datums paid in an InitialDistribution

utxoIndex0From :: MonadBlockChainBalancing m => InitialDistribution -> m UtxoIndex

This creates the initial UtxoIndex from an initial distribution by submitting an initial transaction with the appropriate content:

inputs consist of a single dummy pseudo input

outputs are translated from the TxSkelOut list in the initial distribution

Two things to note:

We don't know what Magic means for the network ID (TODO)

The genesis key hash has been taken from https://github.com/input-output-hk/cardano-node/blob/543b267d75d3d448e1940f9ec04b42bd01bbb16b/cardano-api/test/Test/Cardano/Api/Genesis.hs#L60

utxoIndex0 :: MonadBlockChainBalancing m => m UtxoIndex

Same as utxoIndex0From with the default InitialDistribution