Cooked.MockChain.MockChainState

Description

This module exposes the internal state in which our direct simulation is run, and functions to update and query it.

Documentation

data MockChainState

The state used to run the simulation in Direct

Constructors

MockChainState

Fields

mcstParams :: Params

mcstLedgerState :: EmulatedLedgerState

mcstOutputs :: Map TxOutRef (TxSkelOut, Bool)

Associates to each TxOutRef the TxSkelOut that produced it, alongside a boolean to state whether this UTxO is still present in the index (True) or has already been consumed (False).

mcstConstitution :: Maybe (Versioned Script)

The constitution script to be used with proposals

mcstParamsL :: Lens' MockChainState Params

A lens to set or get the parameters of the MockChainState

mcstLedgerStateL :: Lens' MockChainState EmulatedLedgerState

A lens to set or get the ledger state of the MockChainState

mcstOutputsL :: Lens' MockChainState (Map TxOutRef (TxSkelOut, Bool))

A lens to set or get the outputs of the MockChainState

mcstConstitutionL :: Lens' MockChainState (Maybe (Versioned Script))

A lens to set or get the constitution script of the MockChainState

mcstToUtxoState :: MockChainState -> UtxoState

Builds a UtxoState from a MockChainState

addOutput :: TxOutRef -> TxSkelOut -> MockChainState -> MockChainState

Stores an output in a MockChainState

removeOutput :: TxOutRef -> MockChainState -> MockChainState

Removes an output from the MockChainState