Cooked.MockChain.Monad

BlockChain Monad

class MonadFail m => MonadBlockChain m where

Base methods for interacting with a UTxO graph through Cooked.Tx.Constraints. The operations within MonadBlockChain have an implementation over the Contract monad and can be used to build the first layer of off-chain code that needs to interact with code in the Contract monad.

It is strongly recomended that you write your code polymorphic for some monad m:

f :: (MonadBlockChain m) => m ()

This enables you to choose different interpretations at run time. The two immediately interesting interpretations are running your code within the Contract monad to generate and submit transactions and to test the code by running it in the MockChain monad, provided in this library.

Methods

validateTxSkel :: TxSkel -> m TxId

Generates and balances a transaction from a skeleton, then attemps to validate such transaction. A balanced transaction is such that inputs + mints == outputs + fees. To balance a transaction, we need access to the current UTxO state to choose which inputs to add in case the output-side of the balancing equation is bigger.

The TxSkel receives a TxOpts record with a number of options to customize how validation works.

utxosSuchThat :: FromData a => Address -> UtxoPredicate a -> m [(SpendableOut, Maybe a)]

Returns a list of spendable outputs that belong to a given address and satisfy a given predicate; Additionally, return the datum present in there if it happened to be a script output. It is important to use -XTypeApplications and pass a value for type variable a below.

txOutByRef :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

ownPaymentPubKeyHash :: m PubKeyHash

Returns the hash of our own public key. When running in the Plutus.Contract.Contract monad, this is a proxy to ownPubKey; when running in mock mode, the return value can be controlled with signingWith: the head of the non-empty list will be considered as the "ownPubkey".

currentSlot :: m Slot

Returns the current slot number

currentTime :: m POSIXTime

Returns the current time

awaitSlot :: Slot -> m Slot

Either waits until the given slot or returns the current slot. Note that that it might not wait for anything if the current slot is larger than the argument.

awaitTime :: POSIXTime -> m POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

Example: if starting time is 0 and slot length is 3s, then `awaitTime 4` waits until slot 2 and returns the value `POSIXTime 5`.

validateTxConstrOpts :: (MonadBlockChain m, ConstraintsSpec constraints) => TxOpts -> constraints -> m TxId

Calls validateTxSkel with a skeleton that is set with some specific options.

validateTxConstr :: (MonadBlockChain m, ConstraintsSpec constraints) => constraints -> m TxId

Calls validateTx with the default set of options and no label.

validateTxConstrLbl :: (LabelConstrs lbl, MonadBlockChain m, ConstraintsSpec constraints) => lbl -> constraints -> m TxId

Calls validateTxSkel with the default set of options but passes an arbitrary showable label to it.

spendableRef :: MonadBlockChain m => TxOutRef -> m SpendableOut

pkUtxosSuchThat :: forall a m. (MonadBlockChain m, FromData a) => PubKeyHash -> UtxoPredicate a -> m [(SpendableOut, Maybe a)]

Select public-key UTxOs that might contain some datum but no staking address. This is just a simpler variant of utxosSuchThat. If you care about staking credentials you must use utxosSuchThat directly.

pkUtxosSuchThatValue :: MonadBlockChain m => PubKeyHash -> (Value -> Bool) -> m [SpendableOut]

Select public-key UTxOs that do not contain some datum nor staking address. This is just a simpler variant of pkUtxosSuchThat.

scriptUtxosSuchThat

Arguments

:: (MonadBlockChain m, FromData (DatumType tv))

=> TypedValidator tv

-> (DatumType tv -> Value -> Bool)
Slightly different from UtxoPredicate: here we're guaranteed to have a datum present.

-> m [(SpendableOut, DatumType tv)]

Script UTxOs always have a datum, hence, can be selected easily with a simpler variant of utxosSuchThat. It is important to pass a value for type variable a with an explicit type application to make sure the conversion to and from Datum happens correctly.

outFromOutRef :: MonadBlockChain m => TxOutRef -> m TxOut

Returns the output associated with a given reference

pkUtxos :: MonadBlockChain m => PubKeyHash -> m [SpendableOut]

Return all UTxOs belonging to a pubkey

pkUtxos' :: MonadBlockChain m => PubKeyHash -> m [(TxOutRef, TxOut)]

Return all UTxOs belonging to a pubkey, but keep them as TxOut. This is for internal use.

Slot and Time Management

Slots are integers that monotonically increase and model the passage of time. By looking at the current slot, a validator gets to know that it is being executed within a certain window of wall-clock time. Things can get annoying pretty fast when trying to mock traces and trying to exercise certain branches of certain validators; make sure you also read the docs on autoSlotIncrease to be able to simulate sending transactions in parallel.

waitNSlots :: MonadBlockChain m => Integer -> m Slot

waitNMilliSeconds :: MonadBlockChain m => DiffMilliSeconds -> m POSIXTime

Monad Mock Chain: multiple identities

class MonadBlockChain m => MonadMockChain m where

The Contract monad has certain design decisions that make it suboptimal for testing. Therefore, we divided the abstract interface into those functions that have an interpretation in Contract and those that do not.

For example, a function returning a value of type Contract can be thought of as an interaction with a plutus contract from a single user point of view. When testing, however, we might want to interact with a contract from multiple different users. Changing the set of wallets that sign a transaction has no interpretation in Contract and can only be used with the testing monads.

Methods

signingWith :: NonEmpty Wallet -> m a -> m a

Sets a list of wallets that will sign every transaction emitted in the respective block

askSigners :: m (NonEmpty Wallet)

Returns the current set of signing wallets.

slotConfig :: m SlotConfig

Returns the slot configuration of the mock chain.

as :: MonadMockChain m => m a -> Wallet -> m a

Runs a given block of computations signing transactions as w.

signs :: MonadMockChain m => Wallet -> m a -> m a

Flipped version of as

Deriving further MonadBlockChain instances

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTrans.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the boilerplate of defining all the methods of the class yourself.

Constructors

AsTrans

Fields

getTrans :: t m a

unliftOn :: (MonadTransControl t, Monad m, Monad (t m)) => (m (StT t a) -> m (StT t a)) -> t m a -> t m a