Cooked.MockChain.Monad.Direct

Direct Emulation

The MockChainT monad provides a direct emulator; that is, it gives us a simple way to call validator scripts directly, without the need for all the complexity the Contract monad introduces.

Running a MockChain produces a UtxoState, which is a map from Address to (Pl.Value, Maybe Pl.Datum), and corresponds to the utxo mental model most people have. Internally, however, we keep a UtxoIndex in our state and feeding it to validateTx. For convenience, we also keep a map of Address to Datum, giving is a simple way of managing the current utxo state.

mcstToUtxoState :: MockChainSt -> UtxoState

data MockChainSt

Slightly more concrete version of UtxoState, used to actually run the simulation. We keep a map from datum hash to datum, then a map from txOutRef to datumhash Additionally, we also keep a map from datum hash to the underlying value's "show" result, in order to display the contents of the state to the user.

Constructors

MockChainSt

Fields

mcstIndex :: UtxoIndex

mcstDatums :: Map DatumHash Datum

mcstStrDatums :: Map DatumHash String

mcstCurrentSlot :: Slot

data MockChainError

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError ValidationErrorInPhase

MCETxError MkTxError

MCEUnbalanceable BalanceStage Tx BalanceTxRes

MCENoSuitableCollateral

FailWith String

data BalanceStage

Describes us which stage of the balancing process are we at. This is needed to distinguish the successive calls to balancing while computing fees from the final call to balancing

Constructors

BalCalcFee

BalFinalizing

data MockChainEnv

Constructors

MockChainEnv

Fields

mceParams :: Params

mceSigners :: NonEmpty Wallet

newtype MockChainT m a

The actual MockChainT is a trivial combination of StateT and ExceptT

Constructors

MockChainT

Fields

unMockChain :: ReaderT MockChainEnv (StateT MockChainSt (ExceptT MockChainError m)) a

type MockChain = MockChainT Identity

Non-transformer variant

combineMockChainT :: Monad m => (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

mapMockChainT :: (m (Either MockChainError (a, MockChainSt)) -> n (Either MockChainError (b, MockChainSt))) -> MockChainT m a -> MockChainT n b

runMockChainTRaw :: Monad m => MockChainEnv -> MockChainSt -> MockChainT m a -> m (Either MockChainError (a, MockChainSt))

Executes a MockChainT from some initial state and environment; does not convert the MockChainSt into a UtxoState.

runMockChainTFrom :: Monad m => InitialDistribution -> MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from an initial state set up with the given initial value distribution. Similar to runMockChainT, uses the default environment. Returns a UtxoState instead of a MockChainSt. If you need the later, use runMockChainTRaw

runMockChainT :: Monad m => MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from the canonical initial state and environment. The canonical environment uses the default SlotConfig and [Cooked.MockChain.Wallet.wallet 1] as the sole wallet signing transactions.

runMockChainRaw :: MockChainEnv -> MockChainSt -> MockChain a -> Either MockChainError (a, MockChainSt)

See runMockChainTRaw

runMockChainFrom :: InitialDistribution -> MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainTFrom

runMockChain :: MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainT

utxoState0 :: UtxoState

mockChainSt0 :: MockChainSt

mockChainSt0From :: InitialDistribution -> MockChainSt

utxoIndex0From :: InitialDistribution -> UtxoIndex

utxoIndex0 :: UtxoIndex

Direct Interpretation of Operations

runTransactionValidation

Arguments

:: Slot

-> Params

-> UtxoIndex

-> [PaymentPubKeyHash]
List of required signers

-> [Wallet]
List of signers

-> Tx

-> (UtxoIndex, Maybe ValidationErrorInPhase, [ScriptValidationEvent])

This validates a given Tx in its proper context; this is a very tricky thing to do. We're basing ourselves off from how plutus-apps is doing it.

TL;DR: we need to use Ledger.Index to compute the new UtxoIndex, but we neet to rely on Ledger.Validation to run the validation akin to how it happens on-chain, with proper checks on transactions fees and signatures.

For more details, check the following relevant pointers:

https://github.com/tweag/plutus-libs/issues/92

https://github.com/input-output-hk/plutus-apps/blob/03ba6b7e8b9371adf352ffd53df8170633b6dffa/plutus-ledger/src/Ledger/Tx.hs#L126

https://github.com/input-output-hk/plutus-apps/blob/03ba6b7e8b9371adf352ffd53df8170633b6dffa/plutus-contract/src/Wallet/Emulator/Chain.hs#L209

https://github.com/input-output-hk/plutus-apps/blob/03ba6b7e8b9371adf352ffd53df8170633b6dffa/plutus-contract/src/Wallet/Emulator/Wallet.hs#L314

Finally; because fromPlutusTx doesn't preserve signatures, we need the list of signers around to re-sign the transaction.

validateTx' :: Monad m => [PaymentPubKeyHash] -> Tx -> MockChainT m TxId

Check validateTx for details; we pass the list of required signatories since that is only truly available from the unbalanced tx, so we bubble that up all the way here.

utxosSuchThisAndThat' :: forall a m. (Monad m, FromData a) => (Address -> Bool) -> (Maybe a -> Value -> Bool) -> MockChainT m [(SpendableOut, Maybe a)]

utxosSuchThat' :: forall a m. (Monad m, FromData a) => Address -> (Maybe a -> Value -> Bool) -> MockChainT m [(SpendableOut, Maybe a)]

Check utxosSuchThat for details

generateUnbalTx :: TxSkel -> Either MockChainError UnbalancedTx

Generates an unbalanced transaction from a skeleton; A transaction is unbalanced whenever inputs + mints != outputs + fees. In order to submit a transaction, it must be balanced, otherwise we will see a ValueNotPreserved error.

See Cooked.Tx.Balance for balancing capabilities or stick to generateTx, which generates and balances a transaction.

myAdjustUnbalTx :: Params -> UnbalancedTx -> UnbalancedTx

generateTx' :: Monad m => TxSkel -> MockChainT m ([PaymentPubKeyHash], Tx)

Check generateTx for details

setFeeAndValidRange :: Monad m => BalanceOutputPolicy -> Wallet -> UnbalancedTx -> MockChainT m Tx

Sets the txFee and txValidRange according to our environment. The transaction fee gets set realistically, based on a fixpoint calculation taken from plutus-apps, see https://github.com/input-output-hk/plutus-apps/blob/03ba6b7e8b9371adf352ffd53df8170633b6dffa/plutus-contract/src/Wallet/Emulator/Wallet.hs#L314

balanceTxFrom :: Monad m => BalanceOutputPolicy -> Bool -> Collateral -> Wallet -> UnbalancedTx -> MockChainT m ([PaymentPubKeyHash], Tx)

calcCollateral :: Monad m => Wallet -> Collateral -> MockChainT m (Set TxIn)

Calculates the collateral for a some transaction

balanceTxFromAux :: Monad m => BalanceOutputPolicy -> BalanceStage -> Wallet -> Tx -> MockChainT m Tx

data BalanceTxRes

Constructors

BalanceTxRes

Fields

newInputs :: [TxOutRef]

returnValue :: Value

remainderUtxos :: [(TxOutRef, TxOut)]

calcBalanceTx :: Monad m => Wallet -> Tx -> MockChainT m BalanceTxRes

Calculate the changes needed to balance a transaction with money from a given wallet. Every transaction that is sent to the chain must be balanced, that is: inputs + mint == outputs + fee.

applyBalanceTx :: BalanceOutputPolicy -> Wallet -> BalanceTxRes -> Tx -> Maybe Tx

Once we calculated what is needed to balance a transaction tx, we still need to apply those changes to tx. Because of the minAdaTxOut constraint, this might not be possible: imagine the leftover is less than minAdaTxOut, but the transaction has no output addressed to the sending wallet. If we just create a new ouput for w and place the leftover there the resulting tx will fail to validate with LessThanMinAdaPerUTxO error. Instead, we need to consume yet another UTxO belonging to w to then create the output with the proper leftover. If w has no UTxO, then there's no way to balance this transaction.

Utilities

onlyAdaPkTxOut :: TxOut -> Maybe PubKeyHash

returns public key hash when txout contains only ada tokens and that no datum hash is specified.

addressIsPK :: Address -> Maybe PubKeyHash

rstr :: Monad m => (a, m b) -> m (a, b)

assocl :: (a, (b, c)) -> ((a, b), c)

Orphan instances

Default Slot

Instance details

Methods

def :: Slot