Cooked.MockChain.Monad.Staged

Interpreting and Running StagedMockChain

interpretAndRunWith :: (forall m. Monad m => MockChainT m a -> m res) -> StagedMockChain a -> [(res, TraceDescr)]

Interprets the staged mockchain then runs the resulting computation with a custom function. This can be used, for example, to supply a custom InitialDistribution by providing runMockChainTFrom.

interpretAndRun :: StagedMockChain a -> [(Either MockChainError (a, UtxoState), TraceDescr)]

Interprets and runs the mockchain computation from the default InitialDistribution

Interpreting StagedMockChain

type InterpMockChain = MockChainT (WriterT TraceDescr [])

The semantic domain in which StagedMockChain gets interpreted; see the interpret function for more.

interpret :: StagedMockChain a -> InterpMockChain a

The interpret function gives semantics to our traces. One StagedMockChain computation yields a potential list of MockChainT computations, which emmit a description of their operation. Recall a MockChainT is a state and except monad composed:

    MockChainT (WriterT TraceDescr []) a
=~= st -> (WriterT TraceDescr []) (Either err (a, st))
=~= st -> [(Either err (a, st) , TraceDescr)]

data MockChainOp a where

This is an initial encoding of the MockChain operations, it provides a simple way of altering the AST of a trace before actually executing it. On top of the operations from MonadBlockChain we also have Fail to make sure the resulting monad will be an instance of MonadFail.

Constructors

ValidateTxSkel :: TxSkel -> MockChainOp TxId

TxOutByRef :: TxOutRef -> MockChainOp (Maybe TxOut)

GetCurrentSlot :: MockChainOp Slot

AwaitSlot :: Slot -> MockChainOp Slot

GetCurrentTime :: MockChainOp POSIXTime

AwaitTime :: POSIXTime -> MockChainOp POSIXTime

UtxosSuchThat :: FromData a => Address -> (Maybe a -> Value -> Bool) -> MockChainOp [(SpendableOut, Maybe a)]

OwnPubKey :: MockChainOp PubKeyHash

SigningWith :: NonEmpty Wallet -> StagedMockChain a -> MockChainOp a

AskSigners :: MockChainOp (NonEmpty Wallet)

GetSlotConfig :: MockChainOp SlotConfig

Modify :: Modality TxSkel -> StagedMockChain a -> MockChainOp a

Fail :: String -> MockChainOp a

MPlus :: StagedMockChain a -> StagedMockChain a -> MockChainOp a

MZero :: MockChainOp a

data Staged (op :: * -> *) :: * -> * where

Constructors

Return :: a -> Staged op a

Instr :: op a -> (a -> Staged op b) -> Staged op b

type StagedMockChain = Staged MockChainOp

singleton :: MockChainOp a -> StagedMockChain a

interpretOp :: MonadPlus m => MockChainOp a -> InterpMockChainMod m a

Interprets a single operation in the direct MockChainT monad. We need to pass around the modalities since some operations refer back to StagedMockChain, and that needs to be interpreted taking into account the existing modalities.

type InterpMockChainMod m = MockChainT (WriterT TraceDescr (StateT [Modality TxSkel] m))

This is an internal type that keeps track of the list of modalities that are present in each branch, it's isomorphic to:

St -> ListMods -> m ((Either Err (a, St) , TraceDescr), ListMods)

interpretNonDet :: forall m a. MonadPlus m => StagedMockChain a -> InterpMockChainMod m a

Interprets a StagedMockChain that must be modified with a certain list of modalities.

interpAndTellOp :: forall m a. MonadPlus m => MockChainOp a -> InterpMockChainMod m a

Auxiliar function to interpret a bind.

MonadBlockChain Instance

Modalities

data Modality a

Modalities apply a function to a trace;

Constructors

Somewhere (a -> Maybe a)

Everywhere (a -> Maybe a)

interpModalities :: [Modality a] -> a -> [(a, [Modality a])]

Performs one step of interpreting a composition of modalities to an input. For example,

interpModalities [Somewhere h, Everywhere g, Somewhere f] x

Returns:

[ ( h (g (f x)) , [Everywhere g] )
, ( h (g    x)) , [Everywhere g, Somewhere f] )
, (   (g (f x)) , [Somewhere h, Everywhere g] )
, (   (g    x)) , [Somewhere h, Everywhere g, Somewhere f])

Where each element of the list corresponds to a universe where x suffered the effect of a modality, and which modalities to consider when continuing to evaluate such universe.

Human Readable Traces

prettyMockChainOp :: NonEmpty Wallet -> MockChainOp a -> TraceDescr

Generates a TraceDescription for the given operation; we're mostly interested in seeing the transactions that were validated, so many operations have no description.

newtype TraceDescr

A TraceDescr is a list of Doc encoded as a difference list for two reasons (check ShowS if you're confused about how this works, its the same idea). 1) Naturally, these make for efficient concatenation 2) More importantly, this makes it easy to define the empty TraceDescr as TraceDescr id instead of relying on emptyDoc, which generates empty lines when used with vsep. This avoids generating these empty lines

Constructors

TraceDescr

Fields

trApp :: [Doc ()] -> [Doc ()]

trSingleton :: Doc ann -> TraceDescr