Cooked.MockChain.RawUPLC

Documentation

unsafeTypedValidatorFromUPLC :: forall a. Program NamedDeBruijn DefaultUni DefaultFun () -> TypedValidator a

Returns a TypedValidator from a UPLC program. The "unsafe" refers to the use of unsafeCoerce to cast a TypedValidator Any, resulting from typedValidatorFromUPLC, to a TypedValidator a. This enables us to avoid using fromBuiltinData and toBuiltinData all the time.

This function is meant to be used with -XTypeApplications and the programmer is responsible for ensuring that the type variable a, below, gets instantiated to the correct type.

typedValidatorFromUPLC :: Program NamedDeBruijn DefaultUni DefaultFun () -> TypedValidator Any

Returns a TypedValidator from a UPLC program. The resulting typed validator is instantiated to Any, which means all of its arguments receive a value of type BuiltinData.

unsafeTypedValidatorFromBS :: forall a. ByteString -> Either String (TypedValidator a)

Loads a typed validator from a bytestring that was produced by flat the outputs of getPlc applied to a mkTypedValidator. If the compiled validator was not wrapped, stick to typedValidatorFromBS.

typedValidatorFromBS :: ByteString -> Either String (TypedValidator Any)

Loads a typed validator from a bytestring that was produced by flat the outputs of getPlc.