Cooked.MockChain.Runnable

Description

This module exposes the infrastructure to execute mockchain runs. In particular:

The notion of initial distribution (a list of payments)

The return types of the runs (raw and refined)

The initial configuration with which to execute a run

The notion of RunnableMockChain to actually execute computations

Initial distribution of funds

type InitialDistribution = [TxSkelOut]

Describes the initial distribution of UTxOs per user.

The following specifies a starting state where wallet 1 owns two UTxOs, one with 42 Ada and one with 2 Ada and one TOK token; wallet 2 owns a single UTxO with 10 Ada and wallet 3 has 10 Ada and a permanent value

i0 = distributionFromList $
       [ (wallet 1 , [ ada 42 , ada 2 <> quickValue "TOK" 1 ]
       , (wallet 2 , [ ada 10 ])
       , (wallet 3 , [ ada 10 <> permanentValue "XYZ" 10])
       ]

Note that payment issued through an initial distribution will be attached enough ADA to sustain themselves unless a fixed value is explicitly required.

initialDistributionTemplate :: InitialDistribution

4 UTxOs with 100 Ada each, for each of the first 4 knownWallets

distributionFromList :: IsTxSkelOutAllowedOwner owner => [(owner, [Value])] -> InitialDistribution

Creating a initial distribution with simple values assigned to owners

type RawMockChainReturn a = (MockChainJournal, (MockChainState, Either MockChainError a))

Raw return type of running a mockchain

data MockChainReturn a where

The returned type when running a mockchain. This is both a reorganizing and filtering of the natural returned type RawMockChainReturn.

Constructors

MockChainReturn

Fields

:: { mcrValue :: Either MockChainError a

The value returned by the computation, or an error

, mcrOutputs :: Map TxOutRef (TxSkelOut, Bool)

The outputs at the end of the run

, mcrUtxoState :: UtxoState

The UtxoState at the end of the run

, mcrJournal :: MockChainJournal

The final journal emitted during the run

} -> MockChainReturn a

type FunOnMockChainResult a b = RawMockChainReturn a -> b

The type of functions transforming an element of type RawMockChainReturn a into an element of type b

unRawMockChainReturn :: FunOnMockChainResult a (MockChainReturn a)

Building a MockChainReturn from a RawMockChainReturn

data MockChainConf a b where

Configuration from which to run a mockchain

Constructors

MockChainConf

Fields

:: { mccInitialState :: MockChainState

The initial state from which to run the mockchain

, mccInitialDistribution :: InitialDistribution

The initial payments to issue in the run

, mccFunOnResult :: FunOnMockChainResult a b

The function to apply on the results of the run

} -> MockChainConf a b

mockChainConfTemplate :: MockChainConf a (MockChainReturn a)

The default MockChainConf, which uses the default initial state and initial distribution, and returns a refined MockChainReturn

class RunnableMockChain effs where

The class of effects that represent a mockchain run

Methods

runMockChain :: MockChainState -> Sem effs a -> [RawMockChainReturn a]

Runs a computation from an initial MockChainState, while returning a list of RawMockChainReturn

runMockChainFromConf :: (RunnableMockChain effs, Member MockChainWrite effs) => MockChainConf a b -> Sem effs a -> [b]

Runs a RunnableMockChain from an initial MockChainConf

runMockChainFromInitDist :: (RunnableMockChain effs, Member MockChainWrite effs) => InitialDistribution -> Sem effs a -> [MockChainReturn a]

Runs a RunnableMockChain from an initial distribution

runMockChainFromInitDistTemplate :: (RunnableMockChain effs, Member MockChainWrite effs) => Sem effs a -> [MockChainReturn a]

Same as runMockChainFromInitDist using the initialDistributionTemplate

runMockChainDef :: (RunnableMockChain effs, Member MockChainWrite effs) => Sem effs a -> [MockChainReturn a]

Runs a RunnableMockChain from a default configuration