Cooked.MockChain.Staged

Interpreting and running StagedMockChain

interpretAndRunWith :: (forall m. Monad m => MockChainT m a -> m res) -> StagedMockChain a -> [(res, MockChainLog)]

Interprets the staged mockchain then runs the resulting computation with a custom function. This can be used, for example, to supply a custom InitialDistribution by providing runMockChainTFrom.

interpretAndRun :: StagedMockChain a -> [(Either MockChainError (a, UtxoState), MockChainLog)]

data MockChainLogEntry

Constructors

MCLogSubmittedTxSkel SkelContext TxSkel

MCLogNewTx TxId

MCLogFail String

type MockChainLog = [MockChainLogEntry]

type InterpMockChain = MockChainT (WriterT MockChainLog [])

The semantic domain in which StagedMockChain gets interpreted; see the interpret function for more.

interpret :: StagedMockChain a -> InterpMockChain a

The interpret function gives semantics to our traces. One StagedMockChain computation yields a potential list of MockChainT computations, which emit a description of their operation. Recall a MockChainT is a state and except monad composed:

    MockChainT (WriterT TraceDescr []) a
=~= st -> (WriterT TraceDescr []) (Either err (a, st))
=~= st -> [(Either err (a, st) , TraceDescr)]

StagedMockChain: An AST for MonadMockChain computations

data MockChainBuiltin a where

Constructors

ValidateTxSkel :: TxSkel -> MockChainBuiltin CardanoTx

TxOutByRef :: TxOutRef -> MockChainBuiltin (Maybe TxOut)

GetCurrentSlot :: MockChainBuiltin Slot

AwaitSlot :: Slot -> MockChainBuiltin Slot

GetCurrentTime :: MockChainBuiltin POSIXTime

AwaitTime :: POSIXTime -> MockChainBuiltin POSIXTime

DatumFromHash :: DatumHash -> MockChainBuiltin (Maybe (Datum, DocCooked))

OwnPubKey :: MockChainBuiltin PubKeyHash

AllUtxos :: MockChainBuiltin [(TxOutRef, TxOut)]

Empty :: MockChainBuiltin a
The empty set of traces

Alt :: StagedMockChain a -> StagedMockChain a -> MockChainBuiltin a
The union of two sets of traces

Fail :: String -> MockChainBuiltin a
The failing operation

type MockChainOp = LtlOp (UntypedTweak InterpMockChain) MockChainBuiltin

type StagedMockChain = Staged MockChainOp

InterpLtl instance

Helpers to run tweaks for use in tests for tweaks

runTweak :: Tweak InterpMockChain a -> TxSkel -> [Either MockChainError (a, TxSkel)]

runTweakFrom :: MockChainEnv -> MockChainSt -> Tweak InterpMockChain a -> TxSkel -> [Either MockChainError (a, TxSkel)]

Modalities

type MonadModalBlockChain m = (MonadBlockChain m, MonadModal m, Modification m ~ UntypedTweak InterpMockChain)

A modal mock chain is a mock chain that allows us to use LTL modifications with Tweaks

somewhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to some transaction in the given Trace. The tweak must apply at least once.

everywhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to every transaction in a given trace. This is also successful if there are no transactions at all.

withTweak :: MonadModalBlockChain m => m x -> Tweak InterpMockChain a -> m x

Apply a Tweak to the next transaction in the given trace. The order of arguments is reversed compared to somewhere and everywhere, because that enables an idiom like

do ...
   endpoint arguments `withTweak` someModification
   ...

where endpoint builds and validates a single transaction depending on the given arguments. Then withTweak says "I want to modify the transaction returned by this endpoint in the following way".

MonadBlockChain and MonadMockChain instances

singletonBuiltin :: builtin a -> Staged (LtlOp modification builtin) a

Orphan instances

MonadPlus m => MonadPlus (MockChainT m)

Instance details

Methods

mzero :: MockChainT m a

mplus :: MockChainT m a -> MockChainT m a -> MockChainT m a