Cooked.MockChain.State

Description

This module exposes the internal state in which our direct simulation is run (MockChainState), as well as a restricted and simplified version (UtxoState). The latter only consists of Utxos with a focus on who owns those Utxos. You can see this as having some sort of an "account" view of the ledger state, which typically does not exist in Cardano. This is useful for two reasons:

For printing purposes, where it is much more convient to see the available assets as "who owns what" rather than as a set of mixed Utxos.

For testings purposes, when querying the final state of a run is needed. For instance, properties such as "does Alice indeed owns 3 XXX tokens at the end of this run?" become much easier to express.

MockChainState and associated optics

data MockChainState where

The state used to run the simulation in Direct

Constructors

MockChainState

Fields

:: { mcstParams :: Params

The parametors of the emulated blockchain

, mcstLedgerState :: EmulatedLedgerState

The ledger state of the emulated blockchain

, mcstOutputs :: Map TxOutRef (TxSkelOut, Bool)

Associates to each TxOutRef the TxSkelOut that produced it, alongside a boolean to state whether this UTxO is still present in the index (True) or has already been consumed (False).

, mcstConstitution :: Maybe VScript

The constitution script to be used with proposals

} -> MockChainState

mcstParamsL :: Lens' MockChainState Params

A lens to set or get the parameters of the MockChainState

mcstLedgerStateL :: Lens' MockChainState EmulatedLedgerState

A lens to set or get the ledger state of the MockChainState

mcstOutputsL :: Lens' MockChainState (Map TxOutRef (TxSkelOut, Bool))

A lens to set or get the outputs of the MockChainState

mcstConstitutionL :: Lens' MockChainState (Maybe VScript)

A lens to set or get the constitution script of the MockChainState

mcstMOutputL :: TxOutRef -> Lens' MockChainState (Maybe TxSkelOut)

Accesses a given available Utxo from a MockChainState

Helpers to add or remove outputs from a MockChainState

addOutput :: TxOutRef -> TxSkelOut -> MockChainState -> MockChainState

Stores an output in a MockChainState

removeOutput :: TxOutRef -> MockChainState -> MockChainState

Removes an output from the MockChainState. This does not actually remove it from the map, but instead marks its availability to False

UtxoState: A simplified, address-focused view on a MockChainState

data UtxoPayloadDatum where

A simplified version of a TxSkelOutDatum which only stores the actual datum and whether it is hashed (True) or inline (False). The only difference is that whether the datum was resolved in the transaction creating it on the ledger is absent, which makes sense after the fact.

Constructors

NoUtxoPayloadDatum :: UtxoPayloadDatum

SomeUtxoPayloadDatum :: DatumConstrs dat => dat -> Bool -> UtxoPayloadDatum

utxoPayloadDatumKindAT :: AffineTraversal' UtxoPayloadDatum Bool

Focuses on whether on not this UtxoPayloadDatum isHashed

utxoPayloadDatumTypedAT :: (DatumConstrs a, DatumConstrs b) => AffineTraversal UtxoPayloadDatum UtxoPayloadDatum a b

Extracts, or sets, the typed datum of a UtxoPayloadDatum following the same rules as txSkelOutDatumTypedAT

data UtxoPayload where

A convenient wrapping of the interesting information of a UTxO.

Constructors

UtxoPayload

Fields

:: { utxoPayloadTxOutRef :: TxOutRef

The reference of this UTxO

, utxoPayloadValue :: Value

The value stored in this UTxO

, utxoPayloadDatum :: UtxoPayloadDatum

The optional datum stored in this UTxO

, utxoPayloadReferenceScriptHash :: Maybe ScriptHash

The hash of the optional reference script stored in this UTxO

} -> UtxoPayload

utxoPayloadTxOutRefL :: Lens' UtxoPayload TxOutRef

A lens to set or get the UTxO reference from this UtxoPayload

utxoPayloadValueL :: Lens' UtxoPayload Value

A lens to set or get the value from this UtxoPayload

utxoPayloadDatumL :: Lens' UtxoPayload UtxoPayloadDatum

A lens to set or get the datum from this UtxoPayload

utxoPayloadMReferenceScriptHashL :: Lens' UtxoPayload (Maybe ScriptHash)

A lens to set or get the optional reference script hash from this UtxoPayload

utxoPayloadReferenceScriptHashAT :: AffineTraversal' UtxoPayload ScriptHash

Focusing on the optional reference script hash of a UtxoPayload

newtype UtxoPayloadSet

Represents a set of payloads.

Constructors

UtxoPayloadSet

Fields

utxoPayloadSet :: [UtxoPayload]

List of UTxOs contained in this UtxoPayloadSet

utxoPayloadSetListI :: Iso' UtxoPayloadSet [UtxoPayload]

Going back and forth between a list of UtxoPayload and a UtxoPayloadSet

data UtxoState where

A description of who owns what in a blockchain. Owners are addresses and they each own a UtxoPayloadSet.

Constructors

UtxoState

Fields

:: { availableUtxos :: Map Address UtxoPayloadSet

Utxos available to be consumed

, consumedUtxos :: Map Address UtxoPayloadSet

Utxos already consumed

} -> UtxoState

availableUtxosL :: Lens' UtxoState (Map Address UtxoPayloadSet)

A lens to set or get the available UTxOs from a UtxoState

consumedUtxosL :: Lens' UtxoState (Map Address UtxoPayloadSet)

A lens to set or get the consumed UTxOs from a UtxoState

Querying the assets owned by a given address

holdsInState :: ToAddress a => a -> UtxoState -> Value

Total value accessible to what's pointed by the address.

Transforming a MockChainState into an UtxoState

mcstToUtxoState :: MockChainState -> UtxoState

Builds a UtxoState from a MockChainState