Cooked.MockChain.UtxoPredicate

Predicates over Utxos

type UtxoPredicate a = Maybe a -> Value -> Bool

A UtxoPredicate is a predicate that is used to filter and select utxos in utxosSuchThat. We declare them through a type synonym to create a convenient shallow DSL of predicates.

Basic building blocks

(.&&) :: UtxoPredicate a -> UtxoPredicate a -> UtxoPredicate a infixr 3

Conjunction of UtxoPredicates

(.||) :: UtxoPredicate a -> UtxoPredicate a -> UtxoPredicate a infixr 2

Disjunction of UtxoPredicates

valueSat :: (Value -> Bool) -> UtxoPredicate a

Lifts a predicate over values to a UtxoPredicate by ignoring the datum

datumSat :: (a -> Bool) -> UtxoPredicate a

Lifts a predicate over datums to a UtxoPredicate by forcing the datum to be present and satisfy the predicate.

noDatum :: UtxoPredicate a

Returns true when no datum is present

Predicatesover Values

hasCurrencySymbol :: CurrencySymbol -> Value -> Bool

Returns whether or not a given value has a currency symbol present.

hasAssetClass :: AssetClass -> Value -> Bool

Returns whether or not a given value has an asset class present

hasAssetClasses :: [AssetClass] -> Value -> Bool

Returns whether or not a given value has multiple asset classes present

hasQuickValue :: String -> Value -> Bool

Returns whether a value contains a "quick value". Quick values are tokens belonging to "quick currencies". Check quickValue for more info.

hasOnlyAda :: Value -> Bool

Returns whther or not a value has no token except for Ada