Cooked.MockChain.UtxoState

Description

This module provides a depiction of the state we return when running a MockChain.

Documentation

data UtxoState

A description of who owns what in a blockchain. Owners are addresses and they each own a UtxoPayloadSet.

Constructors

UtxoState

Fields

availableUtxos :: Map Address UtxoPayloadSet

consumedUtxos :: Map Address UtxoPayloadSet

newtype UtxoPayloadSet

Represents a set of payloads.

Constructors

UtxoPayloadSet

Fields

utxoPayloadSet :: [UtxoPayload]

List of UTxOs contained in this UtxoPayloadSet

data UtxoPayload

A convenient wrapping of the interesting information of a UTxO.

Constructors

UtxoPayload

Fields

utxoPayloadTxOutRef :: TxOutRef

The reference of this UTxO

utxoPayloadValue :: Value

The value stored in this UTxO

utxoPayloadDatum :: UtxoPayloadDatum

The optional datum stored in this UTxO

utxoPayloadReferenceScript :: Maybe ScriptHash

The optional reference script stored in this UTxO

data UtxoPayloadDatum where

A simplified version of a TxSkelOutDatum which only stores the actual datum and whether it is hashed or inline.

Constructors

NoUtxoPayloadDatum :: UtxoPayloadDatum

SomeUtxoPayloadDatum :: DatumConstrs dat => dat -> Bool -> UtxoPayloadDatum

holdsInState :: ToAddress a => a -> UtxoState -> Value

Total value accessible to what's pointed by the address.