Cooked.MockChain.UtxoState.Testing

Relations between states

equalModuloAda :: IsProp prop => UtxoState -> UtxoState -> prop

Returns whether two states are equal except for arbitrary differences in Ada. This implies that the set of addresses in each state is the same, hence, the "equal" prefix. If you don't care about that, check equivModuloAtMost.

equivModuloAtMost :: IsProp prop => Value -> UtxoState -> UtxoState -> prop

Returns whether the difference in value between two states is at most a given bound. Note that the difference in value between states can be negative, hence, the order of parameters matters here. See utxoStateDiffTotal for more.