Cooked.MockChain.Wallet

MockChain Wallets

Because the mock wallets from the plutus-apps changes somewhat often, we will provide our own wrapper on top of them to make sure that we can easily deal changes from Plutus.

type Wallet = MockWallet

type PrivateKey = XPrv

knownWallets :: [Wallet]

wallet :: Int -> Wallet

walletPKHashToId :: PubKeyHash -> Maybe Int

walletPK :: Wallet -> PubKey

walletStakingPK :: Wallet -> Maybe PubKey

walletPKHash :: Wallet -> PubKeyHash

walletStakingPKHash :: Wallet -> Maybe PubKeyHash

walletAddress :: Wallet -> Address

walletSK :: MockWallet -> PrivateKey

newtype HACK

Constructors

HACK

Fields

please :: XPrv

hackUnMockPrivateKey :: a -> XPrv

Don't use this; its a hack and will be deprecated once we have time to make a PR into plutus exporting the things we need. If you use this anyway, make sure that you only apply it to MockPrivateKey; the function is polymorphic because MockPrivateKey is not exported either; having a dedicated function makes it easy to test that this works: check the Cooked.MockChain.WalletSpec test module.

walletStakingSK :: Wallet -> Maybe PrivateKey

toPKHMap :: [Wallet] -> Map PubKeyHash Wallet

Signs a transaction

txAddSignature :: Wallet -> Tx -> Tx

txAddSignatureAPI :: Wallet -> Tx AlonzoEra -> Tx AlonzoEra

Initial distribution of funds

Are nothing but is a map from Wallet to Value; we'll just proxy the underlying plutus definitions to make it easer when we have to plug our own, if we ever have the need

newtype InitialDistribution

Describes the initial distribution of UTxOs per wallet. This is important since transaction validation must specify a collateral, hence, wallets must posses more than one UTxO to begin with in order to execute a transaction and have some collateral option. The txCollateral is transfered to the node operator in case the transaction fails to validate.

An initial distribution defined by:

i0 = InitialDistribution $ M.fromList
       [ (wallet 1 , [ Pl.lovelaveValueOf 42000000
                     , Pl.lovelaceValueOf 2000000 <> quickValue "TOK" 1
                     ]
       , (wallet 2 , [Pl.lovelaveValueOf 10000000])
       , (wallet 3 , [Pl.lovelaceValueOf 10000000 <> permanentValue "XYZ" 10])
       ]

Specifies a starting state where wallet 1 contains two UTxOs, one with 42 Ada and one with 2 Ada and one TOK token; wallet 2 contains a single UTxO with 10 Ada and wallet 3 has 10 Ada and a permanent value. Check #quickvalues for more on quick and permanent values. (Remember: 1 Ada = 1000000 Lovelace)

Check the corresponding Default InitialDistribution instance for the default value.

Constructors

InitialDistribution

Fields

distribution :: Map Wallet [Value]

validInitialDistribution :: InitialDistribution -> Bool

An initial distribution is valid if all utxos being created contain at least a minimum amount of Ada: minAda.

minAda :: Value

Proxy to minAdaTxOut as a Value

distributionFromList :: [(Wallet, [Value])] -> InitialDistribution

initialDistribution' :: [(Wallet, [Value])] -> InitialDistribution

Extension of the default initial distribution with additional value in some wallets.

initialTxFor :: InitialDistribution -> Tx

Orphan instances

Eq Wallet

Instance details

Methods

(==) :: Wallet -> Wallet -> Bool

(/=) :: Wallet -> Wallet -> Bool

Ord Wallet

Instance details

Methods

compare :: Wallet -> Wallet -> Ordering

(<) :: Wallet -> Wallet -> Bool

(<=) :: Wallet -> Wallet -> Bool

(>) :: Wallet -> Wallet -> Bool

(>=) :: Wallet -> Wallet -> Bool

max :: Wallet -> Wallet -> Wallet

min :: Wallet -> Wallet -> Wallet