Cooked.Output

Description

This module provide abstractions around the notion of outputs. The idea is to use types to carry additional information on which data is carried by various kinds of concrete outputs used in cooked specifically or in Plutus as a whole, such as TxSkelOut or TxOut.

Documentation

class IsAbstractOutput o

A generalisation of TxOut: With the four type families, we can lift some information about

who owns the output (a public key, a script...?)

what kind of datum is there (do we have an inline datum, a datum hash, nothing...?)

what kind of value does the output hold (pure Ada ...?)

what information do we have on the reference script (only a hash, a complete script, a typed validator...?)

to the type level.

Minimal complete definition

outputOwnerL, outputStakingCredentialL, outputDatumL, outputValueL, outputReferenceScriptL

type family OwnerType o

type family DatumType o

type family ValueType o

type family ReferenceScriptType o

outputOwnerL :: IsAbstractOutput o => Lens' o (OwnerType o)

outputStakingCredentialL :: IsAbstractOutput o => Lens' o (Maybe StakingCredential)

outputDatumL :: IsAbstractOutput o => Lens' o (DatumType o)

outputValueL :: IsAbstractOutput o => Lens' o (ValueType o)

outputReferenceScriptL :: IsAbstractOutput o => Lens' o (Maybe (ReferenceScriptType o))

class ToCredential a where

Methods

toCredential :: a -> Credential

class ToOutputDatum a where

Methods

toOutputDatum :: a -> OutputDatum

class ToValue a where

Methods

toValue :: a -> Value

class ToScript a where

Methods

toScript :: a -> Versioned Script

class ToScriptHash a where

Methods

toScriptHash :: a -> ScriptHash

type IsTxInfoOutput o = (IsAbstractOutput o, ToCredential (OwnerType o), ToOutputDatum (DatumType o), ToValue (ValueType o), ToScriptHash (ReferenceScriptType o))

An output that can be translated into its script-perspective (as seen on the TxInfo) representation

outputAddress :: (IsAbstractOutput o, ToCredential (OwnerType o)) => o -> Address

outputOutputDatum :: (IsAbstractOutput o, ToOutputDatum (DatumType o)) => o -> OutputDatum

outputValue :: (IsAbstractOutput o, ToValue (ValueType o)) => o -> Value

outputReferenceScriptHash :: (IsAbstractOutput o, ToScriptHash (ReferenceScriptType o)) => o -> Maybe ScriptHash

outputTxOut :: IsTxInfoOutput o => o -> TxOut

Return the output as it is seen by a validator on the TxInfo.

data ConcreteOutput ownerType datumType valueType referenceScriptType where

A type constructed to be the most general instance of IsAbstractOutput.

Constructors

ConcreteOutput :: ownerType -> Maybe StakingCredential -> datumType -> valueType -> Maybe referenceScriptType -> ConcreteOutput ownerType datumType valueType referenceScriptType

toOutputWithReferenceScriptHash :: IsTxInfoOutput out => out -> ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) ScriptHash

Convert the reference script type on the output to ScriptHash.

isOutputWithoutDatum :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) () (ValueType out) (ReferenceScriptType out))

Test if there is no datum on an output.

isOutputWithInlineDatum :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out))

Test if the output carries some inlined datum.

isOutputWithDatumHash :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) DatumHash (ValueType out) (ReferenceScriptType out))

Test if the output carries some datum hash.

isOutputWithInlineDatumOfType :: (FromData a, IsTxInfoOutput out) => out -> Maybe (ConcreteOutput (OwnerType out) a (ValueType out) (ReferenceScriptType out))

Test if the output carries some inlined datum that can be parsed from builtin data on to something of a specific type.

isScriptOutputFrom :: IsTxInfoOutput out => TypedValidator a -> out -> Maybe (ConcreteOutput (TypedValidator a) (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the owner of an output is a specific typed validator. If it is, return an output with the validator type as its OwnerType.

isPKOutputFrom :: IsTxInfoOutput out => PubKeyHash -> out -> Maybe (ConcreteOutput PubKeyHash (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the owner of an output is a specific public key. If it is, return an output of the same DatumType, but with PubKeyHash as its OwnerType.

isOnlyAdaOutput :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) (DatumType out) Ada (ReferenceScriptType out))

Test if the value on an output contains only Ada.

fromAbstractOutput :: IsAbstractOutput out => out -> ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (ReferenceScriptType out)

Creates a generic concrete instance of any kind of abstract output