Cooked.Pretty

Description

This module centralizes functions to pretty-print transaction skeletons, utxo states, addresses, pubkey hashes, values, etc.

We provide the PrettyCooked class and instances for common Plutus types. We don't rely on Pretty from Prettyprinter in order to define better printers for Plutus types which already have instances of Pretty. Also, PrettyCooked makes it possible to optionally modify pretty printing settings PrettyCookedOpts (e.g. length of printed hashes).

Requirements on datum and redeemers

Datums and redeemers are required to have a PrettyCooked instance.

For trivial datatypes, you can rely on Show by using viaShow from Prettyprinter: 'prettyCooked = Prettyprinter.viaShow'.

For more complex datatypes, you can rely on existing PrettyCooked instances. Prefer implementing the prettyCookedOpt function and relay the PrettyCookedOpts settings to other printers.

    data Foo = Bar Pl.Value | Baz Pl.PubkeyHash Pl.Value

    instance PrettyCooked Foo where
      prettyCookedOpt pcOpts (Bar value) =
        Bar + prettyCookedOpt pcOpts value
      prettyCookedOpt pcOpts (Baz pkh value) =
        prettyItemize
          Baz
          "-"
          [ "user:" + prettyCookedOpt pcOpts pkh,
            "deposit:" + prettyCookedOpt pcOpts value ]

The prettyItemize function is useful to nicely lay down nested lists of elements. Since we manipulate regular Doc values, any function from Prettyprinter can be used to implement your printers.

How to pretty print?

Pretty printing of transaction skeletons and UTxO states is done automatically by the end-user functions provided in Cooked.MockChain.Testing.

To do it manually, use prettyTxSkel or prettyUtxoState.

Documentation

prettyItemize :: DocCooked -> DocCooked -> [DocCooked] -> DocCooked

prettyItemizeNonEmpty :: DocCooked -> DocCooked -> [DocCooked] -> Maybe DocCooked

prettyEnumerate :: DocCooked -> DocCooked -> [DocCooked] -> DocCooked

renderString :: (a -> DocCooked) -> a -> String

Use this to convert a pretty-printer to a regular show function using default layout options. This is used in Testing because Tasty uses strings.

prettyEndState :: Show a => PrettyCookedOpts -> (a, UtxoState) -> DocCooked

prettyMockChainLog :: PrettyCookedOpts -> MockChainLog -> DocCooked

This pretty prints a mock chain log that usually consists of the list of validated or submitted transactions. In the log, we know a transaction has been validated if the MCLogSubmittedTxSkel is followed by a MCLogNewTx.

prettyTxSkel :: PrettyCookedOpts -> SkelContext -> TxSkel -> DocCooked

prettyBalancingWallet :: PrettyCookedOpts -> Wallet -> DocCooked

Same as prettyPubKeyHash with a suffix mentionning this is the balancing wallet

prettySigners :: PrettyCookedOpts -> TxOpts -> NonEmpty Wallet -> [DocCooked]

Prints a list of pubkeys with a flag next to the balancing wallet

prettyMints :: PrettyCookedOpts -> (Versioned MintingPolicy, MintsRedeemer, TokenName, NonZero Integer) -> DocCooked

prettyTxSkelOut :: PrettyCookedOpts -> TxSkelOut -> DocCooked

prettyTxSkelIn :: PrettyCookedOpts -> SkelContext -> (TxOutRef, TxSkelRedeemer) -> Maybe DocCooked

prettyTxSkelInReference :: PrettyCookedOpts -> SkelContext -> TxOutRef -> Maybe DocCooked

getReferenceScriptDoc :: (IsAbstractOutput output, ToScriptHash (ReferenceScriptType output)) => PrettyCookedOpts -> output -> Maybe DocCooked

lookupOutputWithDatumDoc :: SkelContext -> TxOutRef -> Maybe (TxOut, Maybe DocCooked)

prettyHash :: Show a => Int -> a -> DocCooked

mPrettyTxOpts :: PrettyCookedOpts -> TxOpts -> Maybe DocCooked

Pretty-print a list of transaction skeleton options, only printing an option if its value is non-default. If no non-default options are in the list, return nothing.

Pretty-printing

prettyUtxoState :: PrettyCookedOpts -> UtxoState -> DocCooked

Pretty prints a UtxoState. Print the known wallets first, then unknown pks, then scripts.

prettyAddressState :: PrettyCookedOpts -> Address -> [(Value, Maybe UtxoDatum)] -> DocCooked

Pretty prints the state of an address, that is the list of utxos (including value and datum), grouped

prettyPayloadGrouped :: PrettyCookedOpts -> [(Value, Maybe UtxoDatum)] -> Maybe DocCooked

Pretty prints payloads (datum and value corresponding to 1 utxo) that have been grouped together when they belong to the same utxo

prettyPayload :: PrettyCookedOpts -> Value -> Maybe UtxoDatum -> Maybe DocCooked

Orphan instances

Show UtxoState

Instance details

Methods

showsPrec :: Int -> UtxoState -> ShowS

show :: UtxoState -> String

showList :: [UtxoState] -> ShowS

PrettyCooked Address

Instance details

Methods

prettyCooked :: Address -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> Address -> DocCooked

PrettyCooked Credential

Instance details

Methods

prettyCooked :: Credential -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> Credential -> DocCooked

PrettyCooked PubKeyHash

Instance details

Methods

prettyCooked :: PubKeyHash -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> PubKeyHash -> DocCooked

PrettyCooked MintingPolicy

Instance details

Methods

prettyCooked :: MintingPolicy -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> MintingPolicy -> DocCooked

PrettyCooked TxId

Instance details

Methods

prettyCooked :: TxId -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> TxId -> DocCooked

PrettyCooked TxOutRef

Instance details

Methods

prettyCooked :: TxOutRef -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> TxOutRef -> DocCooked

PrettyCooked Value

Instance details

Methods

prettyCooked :: Value -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> Value -> DocCooked

PrettyCooked MockChainError

Instance details

Methods

prettyCooked :: MockChainError -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> MockChainError -> DocCooked