Cooked.Pretty.Class

Description

This module provides common functions to help implement pretty-printers in cooked-validators

Documentation

type DocCooked = Doc ()

A standard Doc without any annotation

class PrettyCooked a where

Type class of things that can be pretty printed as a single document. You need to implement either prettyCookedOpt or prettyCooked manually, otherwise calling either of them will resulting in a infinite loop.

Minimal complete definition

Nothing

Methods

prettyCookedOpt :: PrettyCookedOpts -> a -> DocCooked

Pretty prints an element based on some PrettyCookedOpts

prettyCooked :: a -> DocCooked

Pretty prints an element directly

class PrettyCookedList a where

Type class of things that can be pretty printed as a list of documents. Similarly to PrettyCooked, at least of the functions from this class needs to be manually implemented to avoid infinite loops.

Minimal complete definition

Nothing

Methods

prettyCookedOptList :: PrettyCookedOpts -> a -> [DocCooked]

Pretty prints an element as a list on some PrettyCookedOpts

prettyCookedOptListMaybe :: PrettyCookedOpts -> a -> [Maybe DocCooked]

Pretty prints an element as a list of optional documents

prettyCookedList :: a -> [DocCooked]

Pretty prints an elements as a list

class PrettyCookedMaybe a where

Type class of things that can be optionally pretty printed as a document

Minimal complete definition

Nothing

Methods

prettyCookedOptMaybe :: PrettyCookedOpts -> a -> Maybe DocCooked

Pretty prints an optional document on some PrettyCookedOpts

prettyCookedMaybe :: a -> Maybe DocCooked

Pretty prints an option document

printCookedOpt :: PrettyCooked a => PrettyCookedOpts -> a -> IO ()

Use this in the REPL as an alternative to the default print function when dealing with pretty-printable cooked values.

For example, printCookedOpt def runMockChain i0 foo

printCooked :: PrettyCooked a => a -> IO ()

Version of printCookedOpt that uses default pretty printing options.

renderString :: (a -> DocCooked) -> a -> String

Use this to convert a pretty-printer to a regular show function using default layout options. This is used in Testing because Tasty uses strings.

prettyHash :: ToHash a => PrettyCookedOpts -> a -> DocCooked

Pretty prints hashable elements based on pcOptHashes in the PrettyCookedOpts. This cannot be made an instance as it would be undecidable (the hope was (ToHash a) => PrettyCooked a)

prettyItemize :: PrettyCookedList a => PrettyCookedOpts -> DocCooked -> DocCooked -> a -> DocCooked

Print an item list with a title

>>> prettyCookedOpts opts "Foo" "-" ["bar1", "bar2", "bar3"]
Foo
  - bar1
  - bar2
  - bar3

prettyItemizeNoTitle :: PrettyCookedList a => PrettyCookedOpts -> DocCooked -> a -> DocCooked

Print an item list without a title

prettyItemizeNonEmpty :: PrettyCookedList a => PrettyCookedOpts -> DocCooked -> DocCooked -> a -> Maybe DocCooked

Print an item list with a title, but only when the list is non-empty