Cooked.Pretty.Common

Description

This module provides common functions to help implement pretty-printers in cooked-validators

Documentation

type DocCooked = Doc ()

renderString :: (a -> DocCooked) -> a -> String

Use this to convert a pretty-printer to a regular show function using default layout options. This is used in Testing because Tasty uses strings.

prettyItemize :: DocCooked -> DocCooked -> [DocCooked] -> DocCooked

Print an item list with a title.

>>> prettyItemize "Foo" "-" ["bar1", "bar2", "bar3"]
Foo
  - bar1
  - bar2
  - bar3

prettyItemizeNoTitle :: DocCooked -> [DocCooked] -> DocCooked

prettyItemizeNonEmpty :: DocCooked -> DocCooked -> [DocCooked] -> Maybe DocCooked

prettyEnumerate :: DocCooked -> DocCooked -> [DocCooked] -> DocCooked

prettyHash :: PrettyCookedHashOpts -> BuiltinByteString -> DocCooked

Pretty print a prefix of a hash with a given length.