Cooked.Pretty.Cooked

Description

This module implements pretty-printing for Cooked structures such as skeletons and chain state.

It contains orphaned instances of PrettyCooked for Cooked datatypes. They cannot be provided in Cooked.Pretty.Class because of dependency cycles and, for ease of maintainability, we chose to centralize all pretty-printing related code in submodules of Cooked.Pretty instead of having PrettyCooked instances scattered around.

Some structure require additional arguments to be pretty-printed and have therefore no instances PrettyCooked (for example TxSkel needs some TxSkelContext).

Documentation

prettyMockChainLog :: PrettyCookedOpts -> MockChainLog -> DocCooked

This pretty prints a mock chain log that usually consists of the list of validated or submitted transactions. In the log, we know a transaction has been validated if the MCLogSubmittedTxSkel is followed by a MCLogNewTx.

prettyTxSkel :: PrettyCookedOpts -> SkelContext -> TxSkel -> DocCooked

prettyBalancingWallet :: PrettyCookedOpts -> Wallet -> DocCooked

Same as prettyPubKeyHash with a suffix mentionning this is the balancing wallet

prettySigners :: PrettyCookedOpts -> TxOpts -> [Wallet] -> [DocCooked]

Prints a list of pubkeys with a flag next to the balancing wallet

prettyMints :: PrettyCookedOpts -> (Versioned MintingPolicy, MintsRedeemer, TokenName, NonZero Integer) -> DocCooked

prettyTxSkelOut :: PrettyCookedOpts -> TxSkelOut -> DocCooked

prettyTxSkelOutDatumMaybe :: PrettyCookedOpts -> TxSkelOutDatum -> Maybe DocCooked

prettyTxSkelIn :: PrettyCookedOpts -> SkelContext -> (TxOutRef, TxSkelRedeemer) -> Maybe DocCooked

prettyTxSkelInReference :: PrettyCookedOpts -> SkelContext -> TxOutRef -> Maybe DocCooked

getReferenceScriptDoc :: (IsAbstractOutput output, ToScriptHash (ReferenceScriptType output)) => PrettyCookedOpts -> output -> Maybe DocCooked

lookupOutput :: SkelContext -> TxOutRef -> Maybe (TxOut, TxSkelOutDatum)

mPrettyTxOpts :: PrettyCookedOpts -> TxOpts -> Maybe DocCooked

Pretty-print a list of transaction skeleton options, only printing an option if its value is non-default. If no non-default options are in the list, return nothing.

Pretty-printing

prettyUtxoState :: PrettyCookedOpts -> UtxoState -> DocCooked

Pretty prints a UtxoState. Print the known wallets first, then unknown pks, then scripts.

prettyAddressState :: PrettyCookedOpts -> Address -> UtxoPayloadSet -> DocCooked

Pretty prints the state of an address, that is the list of utxos (including value and datum), grouped

prettyPayloadGrouped :: PrettyCookedOpts -> [UtxoPayload] -> Maybe DocCooked

Pretty prints payloads (datum and value corresponding to 1 utxo) that have been grouped together when they belong to the same utxo

prettyPayload :: PrettyCookedOpts -> Bool -> UtxoPayload -> Maybe DocCooked

prettyReferenceScriptHash :: PrettyCookedOpts -> ScriptHash -> DocCooked

Orphan instances

PrettyCooked TxSkelOutDatum
The PrettyCooked instance for TxSkelOutDatum relays the pretty-printing of the datum it contains.

Instance details

Methods

prettyCooked :: TxSkelOutDatum -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> TxSkelOutDatum -> DocCooked

PrettyCooked MockChainError

Instance details

Methods

prettyCooked :: MockChainError -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> MockChainError -> DocCooked

Show a => PrettyCooked (Either MockChainError (a, UtxoState))

Instance details

Methods

prettyCooked :: Either MockChainError (a, UtxoState) -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> Either MockChainError (a, UtxoState) -> DocCooked

Show a => PrettyCooked (a, UtxoState)

Instance details

Methods

prettyCooked :: (a, UtxoState) -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> (a, UtxoState) -> DocCooked