Cooked.Pretty.Hashable

Description

This module provides an interface for plutus elements that can be hashed. This is used to provide aliases for hashes when pretty printing those elements.

Documentation

class ToHash a where

Hashable elements can be transformed to BuiltinByteString

Methods

toHash :: a -> BuiltinByteString