Cooked.Pretty.MockChain

Description

This module implements PrettyCooked, PrettyCookedList and PrettyCookedMaybe instances for data types returned by a MockChain run.

Orphan instances

PrettyCooked MockChainError

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> MockChainError -> DocCooked

prettyCooked :: MockChainError -> DocCooked

PrettyCooked MockChainLogEntry
This prints a MockChainLogEntry. In the log, we know a transaction has been validated if the MCLogSubmittedTxSkel is followed by a MCLogNewTx.

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> MockChainLogEntry -> DocCooked

prettyCooked :: MockChainLogEntry -> DocCooked

PrettyCooked UtxoState
Pretty print a UtxoState. Print the known wallets first, then unknown pubkeys, then scripts.

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> UtxoState -> DocCooked

prettyCooked :: UtxoState -> DocCooked

PrettyCookedList UtxoPayloadSet
Pretty prints the state of an address, that is the list of UTxOs (including value and datum), grouped

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> UtxoPayloadSet -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> UtxoPayloadSet -> [Maybe DocCooked]

prettyCookedList :: UtxoPayloadSet -> [DocCooked]

PrettyCooked [MockChainLogEntry]

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> [MockChainLogEntry] -> DocCooked

prettyCooked :: [MockChainLogEntry] -> DocCooked

Show a => PrettyCooked (MockChainReturn a UtxoState)

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> MockChainReturn a UtxoState -> DocCooked

prettyCooked :: MockChainReturn a UtxoState -> DocCooked

Show a => PrettyCooked (a, UtxoState)

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> (a, UtxoState) -> DocCooked

prettyCooked :: (a, UtxoState) -> DocCooked