Cooked.Pretty.Options

Description

This module defines pretty-printing options for prettyCookedOpt and their default values.

Documentation

data PrettyCookedOpts

Constructors

PrettyCookedOpts

Fields

pcOptPrintTxHashes :: Bool

Whether to print transaction ids of validated transactions. By default: False

pcOptPrintTxOutRefs :: PCOptTxOutRefs

Whether to print transaction outputs references. By default: hidden

pcOptPrintDefaultTxOpts :: Bool

Whether to print tx options that have not been modified from their default. By default: False

pcOptNumericUnderscores :: Bool

Whether to print big integers with numeric underscores. For example 53_000_000 instead of 53000000. By default: True

pcOptHashes :: PrettyCookedHashOpts

Options related to printing hashes

pcOptPrintLog :: Bool

Whether to display the log

data PrettyCookedHashOpts

Constructors

PrettyCookedHashOpts

Fields

pcOptHashLength :: Int

Length of printed hash prefix. By default: 7

pcOptHashNames :: Map BuiltinByteString String

Association between hashes and given names to ease readability. For example Map.singleton (walletPKHash (wallet 1)) Alice By default: "defaultHashNames" which assigns Lovelace, Quick, and Permanent as names for the associated currency symbols

pcOptHashVerbose :: Bool

When a given name exists for a hash, this flag also prints the original hash after the name. By default: False

data PCOptTxOutRefs

Whether to print transaction outputs references.

Constructors

PCOptTxOutRefsHidden
Hide them

PCOptTxOutRefsFull
Always show them.

Warning: this will disable printing similar UTxOs as a group (for instance (×10) Lovelace: 100_000_000)

PCOptTxOutRefsPartial
Show them for UTxOs which are not grouped with similar others. This avoids the downside of PCOptTxOutRefsFull which disables printing UTxOs as a group.

hashNamesFromList :: ToHash a => [(a, String)] -> Map BuiltinByteString String

Smart constructor for maps to be used in the "pcOptHashNames" pretty-printing option.

defaultHashNames :: Map BuiltinByteString String

Default hash to names map that assigns Lovelace, Quick, and Permanent to the associated currency symbols. This is used as the default for the pretty-printing option and is recommended to use as a basis to extend with custom names.