Cooked.Pretty.Skeleton

Description

This module implements PrettyCooked, PrettyCookedList and PrettyCookedMaybe instances for TxSkels and its components.

Documentation

data Contextualized a

Some elements of a skeleton can only be printed when they are associated with a context. This is typically the case for elements that need some TxOutRefs and datums to be resolved.

Constructors

Contextualized

Fields

_ctxOutputs :: Map TxOutRef (TxSkelOut, Bool)

ctxContent :: a

Orphan instances

PrettyCooked TxSkelCertificate

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> TxSkelCertificate -> DocCooked

prettyCooked :: TxSkelCertificate -> DocCooked

PrettyCooked Mint
Prints a minting specification

Example: > #abcdef - Redeemer: red - Reference script at: txOutRef - Foo: 500 - Bar: 1000

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> Mint -> DocCooked

prettyCooked :: Mint -> DocCooked

PrettyCooked TxSkelOut

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> TxSkelOut -> DocCooked

prettyCooked :: TxSkelOut -> DocCooked

PrettyCooked ParameterChange

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> ParameterChange -> DocCooked

prettyCooked :: ParameterChange -> DocCooked

PrettyCooked TxSkelSignatory

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> TxSkelSignatory -> DocCooked

prettyCooked :: TxSkelSignatory -> DocCooked

PrettyCooked Withdrawal

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> Withdrawal -> DocCooked

prettyCooked :: Withdrawal -> DocCooked

PrettyCooked DRep

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> DRep -> DocCooked

prettyCooked :: DRep -> DocCooked

PrettyCooked Delegatee

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> Delegatee -> DocCooked

prettyCooked :: Delegatee -> DocCooked

PrettyCookedList TxSkelOpts
Pretty-print a list of transaction skeleton options, only printing an option if its value is non-default.

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelOpts -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelOpts -> [Maybe DocCooked]

prettyCookedList :: TxSkelOpts -> [DocCooked]

PrettyCookedList TxSkelOut

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelOut -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelOut -> [Maybe DocCooked]

prettyCookedList :: TxSkelOut -> [DocCooked]

PrettyCookedList TxSkelProposal

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelProposal -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelProposal -> [Maybe DocCooked]

prettyCookedList :: TxSkelProposal -> [DocCooked]

PrettyCookedList TxSkelRedeemer

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelRedeemer -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelRedeemer -> [Maybe DocCooked]

prettyCookedList :: TxSkelRedeemer -> [DocCooked]

PrettyCookedMaybe TxSkelOutDatum
Prints a TxSkelOutDatum when different from NoTxSkelOutDatum

Instance details

Methods

prettyCookedOptMaybe :: PrettyCookedOpts -> TxSkelOutDatum -> Maybe DocCooked

prettyCookedMaybe :: TxSkelOutDatum -> Maybe DocCooked

PrettyCooked (CertificateAction req)

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> CertificateAction req -> DocCooked

prettyCooked :: CertificateAction req -> DocCooked

PrettyCooked (GovernanceAction a)

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> GovernanceAction a -> DocCooked

prettyCooked :: GovernanceAction a -> DocCooked

PrettyCookedList (User req mode)

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> User req mode -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> User req mode -> [Maybe DocCooked]

prettyCookedList :: User req mode -> [DocCooked]

PrettyCookedList (TxSkelOpts, [TxSkelSignatory])
Prints a list of pubkeys with a flag next to the balancing user

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> (TxSkelOpts, [TxSkelSignatory]) -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> (TxSkelOpts, [TxSkelSignatory]) -> [Maybe DocCooked]

prettyCookedList :: (TxSkelOpts, [TxSkelSignatory]) -> [DocCooked]