Cooked.ShowBS

Description

Print all the types that occur on the TxInfo to BuiltinString. This is useful for debugging of validators. You probably do not want to use this in production code, as many of the functions in this module are wildly inefficient due to limitations of the BuiltinString type.

If the functions from this module make script execution on your transactions go over budget, consider using txOptEmulatorParamsModification to temporarily loosen the limits (at the cost of breaking compatibility with mainnet)

Documentation

class ShowBS a where

analogue of Haskell's Show class for use in Plutus scripts.

Minimal complete definition

Nothing

Methods

showBS :: a -> BuiltinString

analogue of show

showBSsPrec :: Integer -> a -> BuiltinString -> BuiltinString

analogue of showsPrec

showBSs :: ShowBS a => a -> BuiltinString -> BuiltinString

analogue of shows

app_prec :: Integer

Precedence of function applications