Cooked.Skeleton.Anchor

Description

This module exposes the notion of Anchor used in proposals and certificates.

Data types

type TxSkelAnchor = Maybe (String, Maybe ByteString)

A TxSkelAnchor optionally bundles an URL as a String alongside an optional resolved page as a ByteString.

Optics

txSkelAnchorMResolvedPageAT :: AffineTraversal' TxSkelAnchor (Maybe ByteString)

Focusing on the optional resolved page of a TxSkelAnchor

txSkelAnchorResolvedPageAT :: AffineTraversal' TxSkelAnchor ByteString

Focusing on the existing resolved page of a TxSkelAnchor

txSkelAnchorURLAT :: AffineTraversal' TxSkelAnchor String

Focusing on the URL of a TxSkelAnchor

Smart constructors

simpleURLAnchor :: String -> TxSkelAnchor

Builds a simple Anchor with an URL