Cooked.Skeleton.Certificate

Description

This module exposes the certificate constructs used in a TxSkel and their associated utilities. To emit certificates in a skeleton, the usual way is to invoke txSkelCertificates = [pubKeyCertificate pk action, scriptCertificate script redeemer action ...]

Data types

data CertificateAction :: UserKind -> Type where

The depiction of the possible actions in a certificate. Each actions exposes, in its types, the possible owners it can have.

Constructors

StakingRegister :: CertificateAction IsEither

StakingUnRegister :: CertificateAction IsEither

StakingDelegate :: Delegatee -> CertificateAction IsEither

StakingRegisterDelegate :: Delegatee -> CertificateAction IsEither

DRepRegister :: CertificateAction IsEither

DRepUpdate :: CertificateAction IsEither

DRepUnRegister :: CertificateAction IsEither

PoolRegister :: PubKeyHash -> CertificateAction IsPubKey

PoolRetire :: Slot -> CertificateAction IsPubKey

CommitteeRegisterHot :: Credential -> CertificateAction IsEither

CommitteeResign :: CertificateAction IsEither

data TxSkelCertificate where

Certificates used in TxSkel. The types ensure that each certificate action is associated with a proper owner.

Constructors

TxSkelCertificate

Fields

:: Typeable kind

=> { txSkelCertificateOwner :: User kind Redemption

All owners of certificates must be in Redemption mode

, txSkelCertificateAction :: CertificateAction kind

The certificate itself does impose a UserKind

} -> TxSkelCertificate

Optics

txSkelCertificateOwnerAT :: Typeable user => AffineTraversal' TxSkelCertificate (User user Redemption)

Focuses on the owner of a TxSkelCertificate

txSkelCertificateActionAT :: Typeable user => AffineTraversal' TxSkelCertificate (CertificateAction user)

Focuses on the action of a TxSkelCertificate

Smart constructors

pubKeyCertificate :: (ToPubKeyHash pkh, Typeable pkh, Typeable a, a ∈ '[IsPubKey, IsEither]) => pkh -> CertificateAction a -> TxSkelCertificate

Smart constructor for a pubkey certificate

scriptCertificate :: (ToVScript script, Typeable script, RedeemerConstrs red) => script -> red -> CertificateAction IsEither -> TxSkelCertificate

Smart constructor for a script certificate