Cooked.Skeleton.Datum

Description

This module exposes the notion of datums as they are handled within a TxSkel

Documentation

type TxSkelOutDatumConstrs a = (Show a, PrettyCooked a, ToData a, Eq a, Typeable a)

These are the type constraints that must be satisfied by the datum content

data TxSkelOutDatum where

On transaction outputs, we have the options to use

no datum

only a datum hash

a "normal" datum

an inline datum

These four options are also what the type TxSkelOutDatum records. The following table explains their differences.

datum stored in in the simulated chain state	datum resolved on the txInfoData	OutputDatum constructor seen by the validator

TxSkelOutNoDatum	no	no	NoOutputDatum

TxSkelOutDatumHash	yes	no	OutputDatumHash

TxSkelOutDatum	yes	yes	OutputDatumHash

TxSkelOutInlineDatum	yes	no	OutputDatum

That is:

Whenever there is a datum, we'll store it in the state of our simulated chain. This will make it possible to retrieve it later, using functions such as datumFromHash.

Both of the TxSkelOutDatumHash and TxSkelOutDatum constructors will create an output that scripts see on the txInfo as having a datum hash. The difference is whether that hash will be resolvable using validator functions like findDatum.

Constructors

TxSkelOutNoDatum :: TxSkelOutDatum
use no datum

TxSkelOutDatumHash :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
only include the hash on the transaction

TxSkelOutDatum :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
use a OutputDatumHash on the transaction output, but generate the transaction in such a way that the complete datum is included in the txInfoData seen by validators

TxSkelOutInlineDatum :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
use an inline datum

txSkelOutTypedDatum :: Typeable a => TxSkelOutDatum -> Maybe a

Attempts to cast the content of this TxSkelOutDatum to a given type

txSkelOutUntypedDatum :: TxSkelOutDatum -> Maybe Datum

Converts a TxSkelOutDatum into a possible Plutus datum