Cooked.Skeleton.Label

Description

This module exposes the labels that can be used to stamp a TxSkel with additional arbitrary pieces of information.

Type constraints

type LabelConstrs x = (PrettyCooked x, Show x, Typeable x, Eq x, Ord x)

These are type constraints that must be satisfied by labels

Data types

data TxSkelLabel where

Labels are arbitrary information that can be added to skeleton. They are meant to be pretty-printed. The common use case we currently have is to tag skeletons that have been modified by tweaks and automated attacks.

The IsString instance will add a Text label, which can be used with labelled to apply tweaks to arbitrary transactions annotated with a label.

Constructors

TxSkelLabel :: LabelConstrs x => x -> TxSkelLabel

Optics

txSkelLabelTypedP :: LabelConstrs a => Prism' TxSkelLabel a

A prism to create a label and retrieve a typed content

label :: Text -> TxSkelLabel

Helper for defining TxSkelLabel values.