Cooked.Skeleton.Output

Description

This module exposes outputs as they can be defined in a TxSkel with various utilities around them.

Documentation

data TxSkelOut where

Transaction outputs. The Pays constructor is really general, and you'll probably want to use the receives smart constructor in most cases.

Constructors

Pays :: (Show o, Typeable o, IsTxInfoOutput o, OwnerConstraints (OwnerType o), DatumType o ~ TxSkelOutDatum, ValueType o ~ TxSkelOutValue, ReferenceScriptConstraints (ReferenceScriptType o)) => o -> TxSkelOut

receives :: (Show owner, Typeable owner, IsTxSkelOutAllowedOwner owner, ToCredential owner) => owner -> Payable els -> TxSkelOut

Smart constructor to build a TxSkelOut from an owner and payment. This should be the main way of building outputs.

txSkelOutValueL :: Lens' TxSkelOut TxSkelOutValue

A lens to get or set a TxSkelOutValue in a TxSkelOut

txSkelOutDatumL :: Lens' TxSkelOut TxSkelOutDatum

A lens to get or set a TxSkelOutDatum in a TxSkelOut

txSkelOutValue :: TxSkelOut -> Value

Returns the value contained in a TxSkelOut

txSkelOutValidator :: TxSkelOut -> Maybe (Versioned Validator)

Returns the optional validator owning a given TxSkelOut

txSkelOutOwnerTypeP :: forall ownerType. OwnerConstraints ownerType => Prism' TxSkelOut (ConcreteOutput ownerType TxSkelOutDatum TxSkelOutValue (Versioned Script))

Decides if a transaction output has a certain owner and datum type.

txSkelOutputDatumTypeAT :: (FromData a, Typeable a) => AffineTraversal' TxSkelOut a

A traversal over datums of type a in a TxSkelOut

class IsTxSkelOutAllowedOwner a where

Depicts the entities that are allowed to own a TxSkelOut

Methods

toPKHOrValidator :: a -> Either PubKeyHash (Versioned Validator)

txSkelOutReferenceScript :: TxSkelOut -> Maybe (Versioned Script)

Returns the optional reference script in a TxSkelOut

type OwnerConstraints owner = (IsTxSkelOutAllowedOwner owner, ToCredential owner, Typeable owner, Show owner)

Type constraints over the owner of a TxSkelOut

type ReferenceScriptConstraints refScript = (ToVersioned Script refScript, Show refScript, Typeable refScript)

Type constraints over the reference script in a TxSkelOut