Cooked.Skeleton.Payable

Documentation

data Payable :: [Symbol] -> Type where

Payable elements. Created from concrete elements or composed. Notice that there is no way of building an element of Type Payable '[] so when using an element of Type Payable els we are sure that something was in fact paid.

Constructors

VisibleHashedDatum :: TxSkelOutDatumConstrs a => a -> Payable '["Datum"]
Hashed datums visible in the transaction are payable

InlineDatum :: TxSkelOutDatumConstrs a => a -> Payable '["Datum"]
Inline datums are payable

HiddenHashedDatum :: TxSkelOutDatumConstrs a => a -> Payable '["Datum"]
Hashed datums hidden from the transaction are payable

ReferenceScript :: ToVersionedScript s => s -> Payable '["Reference Script"]
Reference scripts are payable

Value :: ToValue a => a -> Payable '["Value"]
Values are payable and are subject to min ada adjustment

FixedValue :: ToValue a => a -> Payable '["Value"]
Fixed Values are payable but are NOT subject to min ada adjustment

StakingCredential :: ToMaybeStakingCredential cred => cred -> Payable '["Staking Credential"]
Staking credentials are payable

PayableAnd :: els ⩀ els' => Payable els -> Payable els' -> Payable (els ∪ els')
Payables can be combined as long as their list of tags are disjoint

(<&&>) :: els ⩀ els' => Payable els -> Payable els' -> Payable (els ∪ els')

Basically re-exporting PayableAnd as a builtin operator