Cooked.Skeleton.Proposal

Description

This module exposes the proposals constructs used in a TxSkel and their associated utilities. To issue proposals in a skeleton, the usual way is to invoke txSkelProposals = [simpleProposal script govAction1, simpleProposal pk govAction2, ... ]

Data types

data ParameterChange where

These are all the protocol parameters. They are taken from https://github.com/IntersectMBO/cardano-ledger/blob/c4fbc05999866fea7c0cb1b211fd5288f286b95d/eras/conway/impl/cddl-files/conway.cddl#L381-L412 and will most likely change in future eras.

Constructors

FeePerByte :: Integer -> ParameterChange
The linear factor for the minimum fee calculation

FeeFixed :: Integer -> ParameterChange
The constant factor for the minimum fee calculation

MaxBlockBodySize :: Integer -> ParameterChange
Maximal block body size

MaxTxSize :: Integer -> ParameterChange
Maximal transaction size

MaxBlockHeaderSize :: Integer -> ParameterChange
Maximal block header size

KeyDeposit :: Integer -> ParameterChange
The amount of a key registration deposit

PoolDeposit :: Integer -> ParameterChange
The amount of a pool registration deposit

PoolRetirementMaxEpoch :: Integer -> ParameterChange
Maximum number of epochs in the future a pool retirement is allowed to be scheduled future for.

PoolNumber :: Integer -> ParameterChange
Desired number of pools

PoolInfluence :: Rational -> ParameterChange
Pool influence

MonetaryExpansion :: Rational -> ParameterChange
Monetary expansion

TreasuryCut :: Rational -> ParameterChange
Treasury expansion

MinPoolCost :: Integer -> ParameterChange
Minimum Stake Pool Cost

CoinsPerUTxOByte :: Integer -> ParameterChange
Cost in lovelace per byte of UTxO storage

CostModels
Cost models for non-native script languages

Fields

:: { cmPlutusV1Costs :: [Integer]

, cmPlutusV2Costs :: [Integer]

, cmPlutusV3Costs :: [Integer]

} -> ParameterChange

Prices
Prices of execution units

Fields

:: { pMemoryCost :: Rational

, pStepCost :: Rational

} -> ParameterChange

MaxTxExUnits
Max total script execution resources units allowed per tx

Fields

:: { mteuMemory :: Integer

, mteuSteps :: Integer

} -> ParameterChange

MaxBlockExUnits
Max total script execution resources units allowed per block

Fields

:: { mbeuMemory :: Integer

, mbeuSteps :: Integer

} -> ParameterChange

MaxValSize :: Integer -> ParameterChange
Max size of a Value in an output

CollateralPercentage :: Integer -> ParameterChange
Percentage of the txfee which must be provided as collateral when including non-native scripts.

MaxCollateralInputs :: Integer -> ParameterChange
Maximum number of collateral inputs allowed in a transaction

PoolVotingThresholds
Thresholds for pool votes

Fields

:: { pvtMotionNoConfidence :: Rational

, pvtCommitteeNormal :: Rational

, pvtCommitteeNoConfidence :: Rational

, pvtHardFork :: Rational

, pvtSecurityGroup :: Rational

} -> ParameterChange

DRepVotingThresholds
Thresholds for DRep votes

Fields

:: { drvtMotionNoConfidence :: Rational

, drvtCommitteeNormal :: Rational

, drvtCommitteeNoConfidence :: Rational

, drvtUpdateConstitution :: Rational

, drvtHardForkInitialization :: Rational

, drvtNetworkGroup :: Rational

, drvtEconomicGroup :: Rational

, drvtTechnicalGroup :: Rational

, drvtGovernanceGroup :: Rational

, drvtTreasuryWithdrawal :: Rational

} -> ParameterChange

CommitteeMinSize :: Integer -> ParameterChange
Minimum size of the Constitutional Committee

CommitteeMaxTermLength :: Integer -> ParameterChange
The Constitutional Committee Term limit in number of Slots

GovActionLifetime :: Integer -> ParameterChange
Gov action lifetime in number of Epochs

GovActionDeposit :: Integer -> ParameterChange
The amount of the Gov Action deposit

DRepRegistrationDeposit :: Integer -> ParameterChange
The amount of a DRep registration deposit

DRepActivity :: Integer -> ParameterChange
The number of Epochs that a DRep can perform no activity without losing their Active status.

MinFeeRefScriptCostPerByte :: Rational -> ParameterChange
Reference scripts fee for the minimum fee calculation

data GovernanceAction :: UserKind -> Type where

This lists the various possible governance actions. Only two of these action need to be witnessed by the constitution script, which are annotated by IsScript.

Constructors

ParameterChange :: [ParameterChange] -> GovernanceAction IsScript

TreasuryWithdrawals :: Map Credential Lovelace -> GovernanceAction IsScript

HardForkInitiation :: ProtocolVersion -> GovernanceAction IsNone

NoConfidence :: GovernanceAction IsNone

UpdateCommittee :: [ColdCommitteeCredential] -> Map ColdCommitteeCredential Integer -> Rational -> GovernanceAction IsNone

NewConstitution :: Constitution -> GovernanceAction IsNone

data TxSkelProposal where

This bundles a governance action into an actual proposal

Constructors

TxSkelProposal

Fields

:: (Typeable kind, ToCredential cred)

=> { txSkelProposalReturnCredential :: cred

The credential that should be used for a return account

, txSkelProposalGovernanceAction :: GovernanceAction kind

The proposed action gov action, either witnessed or simple

, txSkelProposalConstitution :: Maybe (User kind Redemption)

The constitution witness of this proposal, when paired with a witnessed governance action. Is the governance action is simple, only Nothing can be provided there.

, txSkelProposalAnchor :: TxSkelAnchor

An optional anchor to be given as additional data. It should correspond to the URL of a web page

} -> TxSkelProposal

Optics

txSkelProposalAnchorL :: Lens' TxSkelProposal TxSkelAnchor

A lens to get or set the anchor of a TxSkelProposal

txSkelProposalMConstitutionAT :: forall kind. Typeable kind => AffineTraversal' TxSkelProposal (Maybe (User kind Redemption))

Focuses on the optional constitution of a TxSkelProposal

txSkelProposalReturnCredentialL :: Lens' TxSkelProposal Credential

Focuses on the return credential from a TxSkelProposal

txSkelProposalGovernanceActionAT :: forall req. Typeable req => AffineTraversal' TxSkelProposal (GovernanceAction req)

Focuses on the governance action of a TxSkelProposal

txSkelProposalConstitutionAT :: AffineTraversal' TxSkelProposal (User IsScript Redemption)

Focuses on the constitution of a TxSkelProposal

Smart constructors

simpleProposal :: (ToCredential cred, Typeable kind) => cred -> GovernanceAction kind -> TxSkelProposal

Builds a TxSkelProposal from a credential and a gov action. Does not set any constitution (when applicable) nor anchor.

Utilities

fillConstitution :: (ToVScript script, Typeable script) => script -> TxSkelProposal -> TxSkelProposal

Sets the constitution script with an empty redeemer when empty. This will not tamper with an existing constitution script and redeemer.