Cooked.Skeleton.Redeemer

Description

This module exposes the notion of redeemers used whenever a script in invoked in a TxSkel.

Documentation

data TxSkelRedeemer where

A bundle around a redeemer which allows to provide a reference input in which the script associated with the redeemer can be found

Constructors

TxSkelRedeemer

Fields

:: RedeemerConstrs redeemer

=> { txSkelRedeemerContent :: redeemer

The redeemer data with which the script will be executed

, txSkelRedeemerReferenceInput :: Maybe TxOutRef

An optional reference input containing the script to execute

, txSkelRedeemerAutoFill :: Bool

Whether the reference input can be automatically assigned. This will only trigger if txSkelRedeemerReferenceInput is Nothing

} -> TxSkelRedeemer

type RedeemerConstrs redeemer = (ToData redeemer, Show redeemer, PrettyCooked redeemer, Eq redeemer, Typeable redeemer)

These are the constraints that must be satisfied by the inner content of a redeemer, that is the actual data that will be passed to the script as its redeemer during during validation

withReferenceInput :: TxSkelRedeemer -> TxOutRef -> TxSkelRedeemer

Attaches a reference input to a given TxSkelRedeemer. This should usually be of no use if option txOptAutoReferenceScripts is turned on, which is the case by default.

someTxSkelRedeemer :: RedeemerConstrs redeemer => redeemer -> TxSkelRedeemer

Creates a TxSkelRedeemer from an inner content with no reference input

emptyTxSkelRedeemer :: TxSkelRedeemer

Creates a TxSkelRedeemer without an inner content nor a reference input

getTypedRedeemer :: Typeable a => TxSkelRedeemer -> Maybe a

Attempts to retrieve the content of a TxSkelRedeemer and cast it to a given type

setTypedRedeemer :: RedeemerConstrs redeemer => redeemer -> TxSkelRedeemer -> TxSkelRedeemer

Changes the inner content of this TxSkelRedeemer, leaving the reference input unchanged. This operation is type-changing.

someTxSkelRedeemerNoAutoFill :: RedeemerConstrs redeemer => redeemer -> TxSkelRedeemer

Creates a TxSkelRedeemer from an inner content with no reference input, while not allowing it to be automatically assigned

emptyTxSkelRedeemerNoAutoFill :: TxSkelRedeemer

Creates a TxSkelRedeemer with no inner content and no reference input, while dissallowing it to be automatically assinged