Cooked.Skeleton.ReferenceScript

Description

This module exposes the notion of reference scripts used in our TxSkel

Documentation

type ReferenceScriptConstrs refScript = (ToVersioned Script refScript, Show refScript, Typeable refScript)

Type constraints over the reference script in a TxSkelOut

data TxSkelOutReferenceScript where

Reference scripts in TxSkelOut

Constructors

TxSkelOutNoReferenceScript :: TxSkelOutReferenceScript

TxSkelOutSomeReferenceScript :: ReferenceScriptConstrs a => a -> TxSkelOutReferenceScript

txSkelOutTypedRefScriptAT :: ReferenceScriptConstrs a => AffineTraversal' TxSkelOutReferenceScript a

Retrieving, or setting, a typed reference script

txSkelOutRefScriptVersioned :: TxSkelOutReferenceScript -> Maybe (Versioned Script)

Retrieving the versioned reference script

txSkelOutRefScriptHash :: TxSkelOutReferenceScript -> Maybe ScriptHash

Retrieving the hash of the reference script