Cooked.Skeleton.Signatory

Description

This module exposes the notion of signatory for out TxSkel

Data types

data TxSkelSignatory where

Signatories in skeletons

Constructors

TxSkelSignatory

Fields

:: (ToPubKeyHash pkh, Show pkh)

=> { txSkelSignatoryPubKeyHash :: pkh

Identifying the signatory with their pubkey hash

, txSkelSignatoryPrivateKey :: Maybe XPrv

The private key with which this signatory should sign. If set to Nothing the signature won't be added (but will be needed later on).oiu clairement

} -> TxSkelSignatory

Optics

txSkelSignatoryMPrivateKeyL :: Lens' TxSkelSignatory (Maybe XPrv)

A lens focusing on the option private key for this signatory

txSkelSignatoryPrivateKeyAT :: AffineTraversal' TxSkelSignatory XPrv

An affine traversal focusing on the existing private key for this signatory

txSkelSignatoryPubKeyHashL :: Lens' TxSkelSignatory PubKeyHash

A lens focusing on the public key hash of this signatory

Smart constructors

signatoryWallet :: Wallet -> TxSkelSignatory

Builds a signatory from a wallet, which will be able to actually sign the transaction.

signatoryPubKey :: (ToPubKeyHash pkh, Show pkh) => pkh -> TxSkelSignatory

Builds a signatory from a pubkey, which will no be able to actually sign the transaction, but will act as a requirement.

txSkelSignatoriesFromList :: [Wallet] -> [TxSkelSignatory]

Builds a list of signatories from a list of wallets