Cooked.Skeleton.User

Description

This module exposes the notion of user used everywhere in our TxSkel. A user can either be a script or a pubkey, and might be used in a redemption or an allocation setting. All of this is grouped under a single data type. Typically, users should rarely be created by hands, and instead, smart constructors adapted to the user's location in the skeleton should be used.

Aliases

type VScript = Versioned Script

VScript is a convenient alias as we have versioned scripts everywhere.

type ToVScript = ToVersioned Script

The ToVScript alias will come in handy when dealing with constrains.

toVScript :: ToVScript script => script -> VScript

The toVScript alias will come in handy to default the type parameter of toVersioned to Script.

Data types

data UserMode

The UserMode corresponds to the way the user will be used in our TxSkel which can either be for allocation (allocation a certain entity to a user) or for redemption (using this user as a witness in a transaction).

Constructors

Allocation

Redemption

data UserKind

The UserKind corresponds to the requirement on the type of users. Some elements will require specifically a script and some others a pubkey.

Constructors

IsScript

IsPubKey

IsEither

IsNone

data User :: UserKind -> UserMode -> Type where

Building users. The type exposes the mode for which the user has been built, and the requirements on the kind of the user.

Constructors

UserPubKey :: forall pkh kind mode. (kind ∈ '[IsPubKey, IsEither], ToPubKeyHash pkh, Typeable pkh) => pkh -> User kind mode
A pubkey user. This can be used whenever a pubkey is needed, and for either of the possible modes.

UserScript :: forall script kind. (kind ∈ '[IsScript, IsEither], ToVScript script, Typeable script) => script -> User kind Allocation
A script user. This can be used whenever a script is needed, but only for the allocation mode.

UserRedeemedScript :: forall script kind. (kind ∈ [IsScript, IsEither], ToVScript script, Typeable script) => script -> TxSkelRedeemer -> User kind Redemption
A script user with an associated redeemer. This can be used whenever a script is needed for redemption mode.

type Peer = User IsPubKey Allocation

An alias for 'User IsPubKey Allocation'

Optics

userHashG :: forall kind mode. Getter (User kind mode) BuiltinByteString

Retrieves the hash of the script or pubkey represented by this user

userCredentialG :: Getter (User kind mode) Credential

Extracts the Credential from a User

userTxSkelRedeemerAT :: AffineTraversal' (User kind mode) TxSkelRedeemer

Focusing on the possible TxSkelRedeemer of this User

userVScriptAT :: AffineTraversal' (User kind mode) VScript

Focusing on the possible VScript of this User

userScriptHashAF :: AffineFold (User kind mode) ScriptHash

Focusing on the possible ScriptHash of this User

userPubKeyHashAT :: AffineTraversal' (User kind mode) PubKeyHash

Focusing on the possible PubKeyHash of this User

userPubKeyHashI :: Iso' (User IsPubKey mode) PubKeyHash

An isomorphism between users required to be pubkeys and PubKeyHash

userVScriptL :: Lens' (User IsScript mode) VScript

Focusing on the VScript from a script

userScriptHashG :: Getter (User IsScript mode) ScriptHash

Focusing on the ScriptHash from a script

userTxSkelRedeemerL :: Lens' (User IsScript Redemption) TxSkelRedeemer

Focus on the TxSkelRedeemer from a script being redeemed

userEitherScriptP :: Prism' (User IsEither mode) (User IsScript mode)

Builds a User IsEither from a User IsScript

userEitherPubKeyP :: Prism' (User IsEither mode) (User IsPubKey mode)

Builds a User IsEither from a User IsPubKey

userTypedAF :: forall user kind mode. Typeable user => AffineFold (User kind mode) user

Retrieves a possible typed user from a User

userTypedScriptAT :: forall userScript mode. (ToVScript userScript, Typeable userScript) => AffineTraversal' (User IsScript mode) userScript

Focuses on a possible typed script in this User

userTypedPubKeyAT :: forall userPK mode. (ToPubKeyHash userPK, Typeable userPK) => AffineTraversal' (User IsPubKey mode) userPK

Focuses on a possible typed pubkey in this User