Cooked.Skeleton.ValidityRange

Description

This module provides a way of accessing lower and upper bounds of transaction validity through optics using at and ix.

Documentation

data ValidityBound

A type used to index optics within a SlotRange. This allows the usage of the following optics: at Lower, at Upper, ix Lower and ix Upper to modify parts of a slot range. When set to Nothing, the associated bound of the interval is considered infinite, and otherwise it is considered finite, with closed closure.

Constructors

Lower

Upper

Orphan instances

At SlotRange

Instance details

Methods

at :: Index SlotRange -> Lens' SlotRange (Maybe (IxValue SlotRange))

Ixed SlotRange

Instance details

Associated Types

type IxKind SlotRange

Methods

ix :: Index SlotRange -> Optic' (IxKind SlotRange) NoIx SlotRange (IxValue SlotRange)