Cooked.Skeleton.Value

Description

This module exposes constructs around the value as it is stored in a TxSkel.

Documentation

data TxSkelOutValue where

A bundle arond an Value to be stored in a TxSkel. This bundles offers the possibility to mark a value as adjustable, in case the ADA amount it contains is insufficient to sustain the storage cost of the UTxO containing it.

Constructors

TxSkelOutValue

Fields

:: { txSkelOutValueContent :: Value

Value to be paid

, txSkelOutValueAutoAdjust :: Bool

Whether this value can be subject to automated adjustment

} -> TxSkelOutValue

txSkelOutValueContentL :: Lens' TxSkelOutValue Value

A lens to get or set the inner value of a TxSkelOutValue

txSkelOutValueAutoAdjustL :: Lens' TxSkelOutValue Bool

A lens to get or set if this value should be auto-adjusted if needed