Cooked.Skeleton.Withdrawal

Description

This module exposes the withdrawing constructs used in a TxSkel and their associated utilities. To issue withdrawals in a skeleton, the usual way is to invoke txSkelWithdrawals = txSkelWithdrawalsFromList [pubKeyWithdrawal pk amount, scriptWithdrawal script redeemer amount, ...]

Data types

data Withdrawal where

A single Withdrawal, owned by a pubkey or redeemed script

Constructors

Withdrawal

Fields

:: { withdrawalUser :: User IsEither Redemption

The user making a withdrawals of their rewards

, withdrawalAmount :: Maybe Lovelace

The amount of lovelace to withdraw. If set to Nothing, cooked will attempt to fill this out with the current rewards for the user, which is the only acceptable amount ledger-wise. Manually setting this value is only left as a possibility for testing purposes.

} -> Withdrawal

data TxSkelWithdrawals

Withdrawals associate either a script or a private key with a redeemer and a certain amount of ada. They are uniquely identified by the hash of either.

Optics

withdrawalUserL :: Lens' Withdrawal (User 'IsEither 'Redemption)

Focuses on the user owning a Withdrawal

withdrawalMAmountL :: Lens' Withdrawal (Maybe Lovelace)

Focuses on the optional amount in a Withdrawal

withdrawalAmountAT :: AffineTraversal' Withdrawal Lovelace

Focuses on the amount in a Withdrawal

txSkelWithdrawalsListI :: Iso' TxSkelWithdrawals [Withdrawal]

Transforms a [Withdrawal] to a 'TxSkelWithdrawals and vice versa. Accumulates amount of withdrawals with similar owners, and keep the latest found redeemer in the case of scripts, discarding the previous ones.

Smart constructors

pubKeyWithdrawal :: (ToPubKeyHash pkh, Typeable pkh) => pkh -> Withdrawal

Creates a Withdrawal from a private key hash

scriptWithdrawal :: (ToVScript script, Typeable script, RedeemerConstrs red) => script -> red -> Withdrawal

Creates a Withdrawal from a redeemed script and lovelace amount

txSkelWithdrawalsFromList :: [Withdrawal] -> TxSkelWithdrawals

Builds a TxSkelWithdrawals from a list of Withdrawal. This is equivalent to calling review txSkelWithdrawalsListI

Utilities

fillAmount :: Lovelace -> Withdrawal -> Withdrawal

Fills a given amount of lovelace to withdraw, if not already set