Cooked.Staged

Description

This module exposes a simple notion of a Staged computation (or a freer monad) to be used when modifying mockchain runs with Ltl formulas.

Documentation

data Staged (op :: Type -> Type) :: Type -> Type where

The freer monad on op. We think of this as the AST of a computation with operations of types op a.

Constructors

Return :: a -> Staged op a

Instr :: op a -> (a -> Staged op b) -> Staged op b

interpStaged :: Monad m => (forall a. op a -> m a) -> forall a. Staged op a -> m a

Interprets a staged computation given a interpreter of the builtins