Cooked.Tweak

Documentation

class (MonadPlus m, MonadBlockChainWithoutValidation m) => MonadTweak m where

Methods

getTxSkel :: m TxSkel

putTxSkel :: TxSkel -> m ()

failingTweak :: MonadTweak m => m a

The never-applicable tweak.

doNothingTweak :: MonadTweak m => m ()

The tweak that always applies and leaves the transaction unchanged.

viewTweak :: (MonadTweak m, Is k A_Getter) => Optic' k is TxSkel a -> m a

The "tweak" that obtains some value from the TxSkel. This does *not* modify the transaction.

viewAllTweak :: (MonadTweak m, Is k A_Fold) => Optic' k is TxSkel a -> m [a]

Like the viewTweak, but returns a list of all foci

setTweak :: (MonadTweak m, Is k A_Setter) => Optic' k is TxSkel a -> a -> m ()

The tweak that sets a certain value in the TxSkel.

overTweak :: (MonadTweak m, Is k A_Setter) => Optic' k is TxSkel a -> (a -> a) -> m ()

The tweak that modifies a certain value in the TxSkel.

overMaybeTweak :: (MonadTweak m, Is k A_Traversal) => Optic' k is TxSkel a -> (a -> Maybe a) -> m [a]

Like overTweak, but only modifies foci on which the argument function returns Just the new focus. Returns a list of the foci that were modified, as they were before the tweak, and in the order in which they occurred on the original transaction.

overMaybeSelectingTweak :: forall a m k is. (MonadTweak m, Is k A_Traversal) => Optic' k is TxSkel a -> (a -> Maybe a) -> (Integer -> Bool) -> m [a]

Sometimes overMaybeTweak modifies too many foci. This might be the case if there are several identical foci, but you only want to modify some of them. This is where this Tweak becomes useful: The (Integer -> Bool) argument can be used to select which of the modifiable foci should be actually modified.

selectP :: (a -> Bool) -> Prism' a a

overMaybeTweak requires a modification that can fail (targeting Maybe). Sometimes, it can prove more convenient to explicitly state which property the foci shoud satisfy to be eligible for a modification that cannot fail instead. selectP provides a prism to make such a selection. The intended use case is 'overTweak (optic % selectP prop) mod' where optic gives the candidate foci, prop is the predicate to be satisfied by the foci, and mod is the modification to be applied to the selected foci.

data PermutOutTweakMode

Constructors

KeepIdentity (Maybe Int)

OmitIdentity (Maybe Int)

allOutPermutsTweak :: MonadTweak m => PermutOutTweakMode -> m ()

Modify transactions by changing the ordering of output constraints. If the PermutTweakMode is

KeepIdentity (Just n), the unmodified transaction is included in the list of modified transactions and only the first n outputs are permuted,

KeepIdentity Nothing, the unmodified transaction is included and all outputs are permuted. Use this with care; there might be a lot of permutations!

OmitIdentity (Just n), the unmodified transaction is not included in the list of modified transactions and only the first n outputs are permuted,

OmitIdentity Nothing, the unmodified transaction is not included and all outputs are permuted. Use this with care; there might be a lot of permutations!

(In particular, this is clever enough to generate only the distinct permutations, even if some outputs are identical.)

singleOutPermutTweak :: MonadTweak m => Int -> m ()

This randomly permutes the outputs of a transaction with a given seed Can be used to assess if a certain validator is order-dependant

addLabelTweak :: (MonadTweak m, LabelConstrs x) => x -> m ()

Add a label to a TxSkel.

removeLabelTweak :: (MonadTweak m, LabelConstrs x) => x -> m ()

Removes a label from a TxSkel

hasLabelTweak :: (MonadTweak m, LabelConstrs x) => x -> m Bool

Checks if a given label is present in the TxSkel

data TamperDatumLbl

Constructors

TamperDatumLbl

tamperDatumTweak

Arguments

:: forall a m. (MonadTweak m, Show a, PrettyCooked a, ToData a, FromData a, Typeable a)

=> (a -> Maybe a)
Use this function to return Just the changed datum, if you want to perform a change, and Nothing, if you want to leave it as-is. All datums on outputs not paying to a validator of type a are never touched.

-> m [a]

A tweak that tries to change the datum on outputs that go to scripts with a prescribed tampering function, that only applies to datums of a certain type.

The tweak returns a list of the modified datums, as they were *before* the modification was applied to them.

ensureInputTweak :: MonadTweak m => TxOutRef -> TxSkelRedeemer -> m (Maybe (TxOutRef, TxSkelRedeemer))

Ensure that a given TxOutRef is being spent with a given TxSkelRedeemer. The return value will be Just the added data, if anything changed.

addInputTweak :: MonadTweak m => TxOutRef -> TxSkelRedeemer -> m ()

Add an input to a transaction. If the given TxOutRef is already being consumed by the transaction, fail.

removeInputTweak :: MonadTweak m => (TxOutRef -> TxSkelRedeemer -> Bool) -> m [(TxOutRef, TxSkelRedeemer)]

Remove transaction inputs according to a given predicate. The returned list contains all removed inputs.

ensureOutputTweak :: MonadTweak m => TxSkelOut -> m (Maybe TxSkelOut)

Ensure that a certain output is produced by a transaction. The return value will be Just the added output, if there was any change.

addOutputTweak :: MonadTweak m => TxSkelOut -> m ()

Add a transaction output, at the end of the current list of outputs, thus retaining the order in which they have been specified.

removeOutputTweak :: MonadTweak m => (TxSkelOut -> Bool) -> m [TxSkelOut]

Remove transaction outputs according to some predicate. The returned list contains all the removed outputs.

addMintTweak :: MonadTweak m => (Versioned MintingPolicy, MintsRedeemer, TokenName, Integer) -> m ()

Add a new entry to the TxSkelMints of the transaction skeleton under modification. As this is implemented in terms of addToTxSkelMints, the same caveats apply as do to that function!

removeMintTweak :: MonadTweak m => ((Versioned MintingPolicy, MintsRedeemer, TokenName, Integer) -> Bool) -> m [(Versioned MintingPolicy, MintsRedeemer, TokenName, Integer)]

Remove some entries from the TxSkelMints of a transaction, according to some predicate. The returned list holds the removed entries.