Cooked.Tweak.Common

Description

This module defines Tweaks which are the building blocks of our DSL for attacks. They are skeleton modifications aware of the mockchain state.

Tweak effect

data Tweak :: Effect where

An effect that allows to store or retrieve a TxSkel from a context

Constructors

GetTxSkel :: Tweak m TxSkel
Retrieves the TxSkel from the context

PutTxSkel :: TxSkel -> Tweak m ()
Overrides the TxSkel in the context

getTxSkel :: forall r. Member Tweak r => Sem r TxSkel

Retrieves the TxSkel from the context

putTxSkel :: forall r. Member Tweak r => TxSkel -> Sem r ()

Overrides the TxSkel in the context

Running a tweak

runTweak :: TxSkel -> Sem (Tweak : effs) a -> Sem effs (TxSkel, a)

Running a Tweak is equivalent to running a state monad storing a TxSkel

evalTweak :: TxSkel -> Sem (Tweak : effs) a -> Sem effs a

Same as runTweak but discards the returned TxSkel

execTweak :: TxSkel -> Sem (Tweak : effs) a -> Sem effs TxSkel

Same as runTweak but discards the returned value