Cooked.Tweak.Mint

Description

Tweaks working on the minting part of a skeleton

Documentation

addMintTweak :: MonadTweak m => (Versioned MintingPolicy, TxSkelRedeemer, TokenName, Integer) -> m ()

Add a new entry to the TxSkelMints of the transaction skeleton under modification. As this is implemented in terms of addToTxSkelMints, the same caveats apply as do to that function!

removeMintTweak :: MonadTweak m => ((Versioned MintingPolicy, TxSkelRedeemer, TokenName, Integer) -> Bool) -> m [(Versioned MintingPolicy, TxSkelRedeemer, TokenName, Integer)]

Remove some entries from the TxSkelMints of a transaction, according to some predicate. The returned list holds the removed entries.