Cooked.Tweak.Query

Description

This module provides tweaks that query the stored TxSkel based on various kinds of optics.

Querying via a getter

viewTweak :: (Member Tweak effs, Is k A_Getter) => Optic' k is TxSkel a -> Sem effs a

Retrieves the focus from the TxSkel given a getter

iviewTweak :: (Member Tweak effs, Is k A_Getter) => Optic' k (WithIx i) TxSkel a -> Sem effs (i, a)

Like viewTweak, for indexed optics

Querying via a fold

toListOfTweak :: (Member Tweak effs, Is k A_Fold) => Optic' k is TxSkel a -> Sem effs [a]

Like viewTweak, but returns all the foci as a list

itoListOfTweak :: (Member Tweak effs, Is k A_Fold) => Optic' k (WithIx i) TxSkel a -> Sem effs [(i, a)]

Like toListOfTweak, for indexed optics

Querying via an affine fold

previewTweak :: (Member Tweak effs, Is k An_AffineFold) => Optic' k is TxSkel a -> Sem effs (Maybe a)

Like viewTweak, but the foci might not exist

ipreviewTweak :: (Member Tweak effs, Is k An_AffineFold) => Optic' k (WithIx i) TxSkel a -> Sem effs (Maybe (i, a))

Like previewTweak, for indexed optics