Cooked.Tweak.Remove

Description

This module provides tweaks able to remove elements from collections focused in a TxSkel.

Removing elements from lists

removeIfTweak :: (Member Tweak effs, Is k A_Lens) => Optic' k is TxSkel [a] -> (a -> Bool) -> Sem effs [a]

Removes elements satisfying a predicate from a list focused in a TxSkel, returning the elements that were removed.

removeAtPosTweak :: (Member Tweak effs, Is k A_Lens) => Optic' k is TxSkel [a] -> Int -> Sem effs (Maybe a)

Removes an element at a specific position in a list focused in a TxSkel, returning the removed element, if any.

Removing elements from maps

removeAtTweak :: (Member Tweak effs, Is k A_Lens, Ord a) => Optic' k is TxSkel (Map a b) -> a -> Sem effs (Maybe b)

Removes an element at a specific key in a map focused in a TxSkel, returning the removed value, if any.

Removing elements from sets

removeInTweak :: (Member Tweak effs, Is k A_Lens, Ord a) => Optic' k is TxSkel (Set a) -> a -> Sem effs (Maybe a)

Removes an element in a set focused in a TxSkel, returning the removed value, if any.