Cooked.Tweak.Signers

Description

This module defines Tweaks revolving around the signers of a transaction. They assume but do not ensure that the list of signers is free of duplicates.

Documentation

getSignersTweak :: MonadTweak m => m [Wallet]

Returns the current list of signers

modifySignersTweak :: MonadTweak m => ([Wallet] -> [Wallet]) -> m [Wallet]

Apply a function to the list of signers and return the old ones

setSignersTweak :: MonadTweak m => [Wallet] -> m [Wallet]

Change the current signers and return the old ones

signersSatisfyTweak :: MonadTweak m => ([Wallet] -> Bool) -> m Bool

Check if the signers satisfy a certain predicate

isSignerTweak :: MonadTweak m => Wallet -> m Bool

Check if a wallet signs a transaction

hasSignersTweak :: MonadTweak m => m Bool

Check if the transaction has at least a signer

addFirstSignerTweak :: MonadTweak m => Wallet -> m [Wallet]

Add a signer to the transaction, at the head of the list of signers, and return the old list of signers

addSignersTweak :: MonadTweak m => [Wallet] -> m [Wallet]

Add signers at the end of the list of signers, and return the old list of signers

addLastSignerTweak :: MonadTweak m => Wallet -> m [Wallet]

Add a signer to the transaction, at the end of the list of signers, and return the old list of signers

removeSignersTweak :: MonadTweak m => [Wallet] -> m [Wallet]

Remove signers from the transaction and return the old list of signers

removeSignerTweak :: MonadTweak m => Wallet -> m [Wallet]

Remove a signer from the transaction and return the old list of signers

replaceFirstSignerTweak :: MonadTweak m => Wallet -> m [Wallet]

Changes the first signer (adds it if there are no signers) and return the old list of signers.