Cooked.Tweak.TamperDatum

Documentation

tamperDatumTweak

Arguments

:: forall a m. (MonadTweak m, Show a, PrettyCooked a, ToData a, FromData a, Typeable a)

=> (a -> Maybe a)
Use this function to return Just the changed datum, if you want to perform a change, and Nothing, if you want to leave it as-is. All datums on outputs not paying to a validator of type a are never touched.

-> m [a]

A tweak that tries to change the datum on outputs that go to scripts with a prescribed tampering function, that only applies to datums of a certain type.

The tweak returns a list of the modified datums, as they were *before* the modification was applied to them.

data TamperDatumLbl

Constructors

TamperDatumLbl