Cooked.Tx.Balance

Documentation

class (Show out, Show (BOutRef out), Eq (BOutRef out)) => BalancableOut out where

Associated Types

type BOutRef out = (result :: Type) | result -> out

The BOutRefs are only used to keep track of what outputs were used during balancing. They might as well be identity (type) functions if out is cheap and comparable.

Methods

outValue :: out -> Value

outRef :: out -> BOutRef out

balanceWithUTxOs :: forall out. BalancableOut out => Value -> [out] -> ([BOutRef out], Value, ExcessiveValue)

This function will try to select a selection of outputs from the set of available outputs given to it in order to have enough to balance out the given value.

type ExcessiveValue = Value

spendValueFrom :: forall out. BalancableOut out => Value -> [out] -> ([BOutRef out], Value, ExcessiveValue)

A helper to make sure that input + mint = output + fees holds for a transaction.

Namely, val is output - input, and utxos is the list of the unspent outputs that can be used to balance the transaction.

If a token value in val is negative, it means that the transaction got more of that token as input than as output. We don't balance it in this case, since we don't want to (and can't, really) choose where that extra output goes.

This function returns the UTXOs that were spent (a subset of utxos), any leftover value (in case some UTXOs had more output than we actually need to balance the transaction) and the excess value which needs to be routed somewhere.

necessaryUtxosFor :: BalancableOut out => CurrencySymbol -> TokenName -> Integer -> [out] -> [BOutRef out] -> ([BOutRef out], Value)

singletonValue :: CurrencySymbol -> TokenName -> Integer -> Value