Cooked.Tx.Constraints

Documentation

module Cooked.Tx.Constraints.Type

module Cooked.Tx.Constraints.Pretty

type LedgerConstraint a = (ScriptLookups a, TxConstraints (RedeemerType a) (DatumType a))

extractDatumStr :: ToLedgerConstraint constraint => constraint -> Map DatumHash String

Map from datum hashes to string representation of all the datums carried. We use this in order to display data to the use when testing. Its often easier to read the original datatype that was placed into a UTxO instead of its respective toBuilinData image.

signedByWallets :: [Wallet] -> MiscConstraint

signedByWallets ws == SignedBy $ map walletPKHash ws

toLedgerConstraint :: ToLedgerConstraint constraint => constraint -> LedgerConstraint a

Converts our constraint into a LedgerConstraint, which later can be used to generate a transaction. The universally quantified type-variable is there on purpose, to enable us to easily spend from multiple scripts at the same time.

orderTxOutputs :: [OutConstraint] -> [TxOut] -> [TxOut]

Reorders the outputs of a transaction according to the ordered list of output constraints that generate them. Fails in case of mismatch. The reordered outputs are put at the beginning of the list. orderTxOutputs :: MonadFail m => [OutConstraint] -> [Pl.TxOut] -> m [Pl.TxOut] -- TODO Check staking credentials orderTxOutputs [] txOuts = return txOuts orderTxOutputs (oc : ocs) txOuts = case findGeneratedTxOutput oc txOuts of Nothing -> fail $ "Could not locate output corresponding to constraint " <> show (prettyOutConstraint oc) Just txOut -> (txOut :) $ orderTxOutputs ocs (List.delete txOut txOuts)