Cooked.Tx.Constraints.Optics

Description

Various Optics on TxSkels, constraints, and all the other types defined in Type.

Picking apart TxSkels

data TxLabel where

Constructors

TxLabel :: LabelConstrs x => Maybe x -> TxLabel

txLabelL :: Lens' TxSkel TxLabel

txOptsL :: Lens' TxSkel TxOpts

txConstraintsL :: Lens' TxSkel Constraints

constraintPairI :: Iso' Constraints ([MiscConstraint], [OutConstraint])

outConstraintsL :: Lens' TxSkel [OutConstraint]

miscConstraintsL :: Lens' TxSkel [MiscConstraint]

Picking apart MiscConstraints

data MintsConstraint where

Constructors

MintsConstraint :: MintsConstrs a => Maybe a -> [MintingPolicy] -> Value -> MintsConstraint

mintsConstraintP :: Prism' MiscConstraint MintsConstraint

mintsConstraintsT :: Traversal' TxSkel MintsConstraint

data SpendsScriptConstraint where

Constructors

SpendsScriptConstraint :: SpendsConstrs a => TypedValidator a -> RedeemerType a -> SpendableOut -> SpendsScriptConstraint

spendsScriptConstraintP :: Prism' MiscConstraint SpendsScriptConstraint

spendsScriptConstraintsT :: Traversal' TxSkel SpendsScriptConstraint

spendableOutL :: Lens' SpendsScriptConstraint SpendableOut

spendsPKConstraintP :: Prism' MiscConstraint SpendableOut

Picking apart OutConstraints

data PaysScriptConstraint where

Constructors

PaysScriptConstraint :: PaysScriptConstrs a => TypedValidator a -> Maybe StakingCredential -> DatumType a -> Value -> PaysScriptConstraint

paysScriptConstraintP :: Prism' OutConstraint PaysScriptConstraint

paysScriptConstraintsT :: Traversal' TxSkel PaysScriptConstraint

paysScriptConstraintTypeP :: forall a. (Typeable a, PaysScriptConstrs a) => Prism' PaysScriptConstraint (TypedValidator a, Maybe StakingCredential, DatumType a, Value)

data PaysPKWithDatumConstraint where

Constructors

PaysPKWithDatumConstraint :: (ToData a, Eq a, Show a, Typeable a) => PubKeyHash -> Maybe StakePubKeyHash -> Maybe a -> Value -> PaysPKWithDatumConstraint

paysPKWithDatumConstraintP :: Prism' OutConstraint PaysPKWithDatumConstraint

paysPKWithDatumConstraintsT :: Traversal' TxSkel PaysPKWithDatumConstraint

Extracting Values from different types

class HasValue a where

Methods

valueL :: Lens' a Value

valueAT :: AffineTraversal' MiscConstraint Value

txSkelInValue :: TxSkel -> Value

The combined value contained in all MiscConstraints of a TxSkel.

txSkelOutValue :: TxSkel -> Value

The combined value contained in all OutConstraints of a TxSkel.

Picking apart Values

flattenValueI :: Iso' Value [(AssetClass, Integer)]

nonAdaValue :: Value -> Value

The portion of a Value that is not Ada.