Cooked.Tx.Constraints.Pretty

Documentation

prettyEnum :: Doc ann -> Doc ann -> [Doc ann] -> Doc ann

prettyTxSkel :: [Wallet] -> TxSkel -> Doc ann

prettyWallet :: PubKeyHash -> Doc ann

prettyOutConstraint :: OutConstraint -> Doc ann

prettyMiscConstraint :: MiscConstraint -> Doc ann

prettyHash :: Show a => a -> Doc ann

prettyMintingPolicy :: MintingPolicy -> Doc ann

prettyOutputDatum :: Show (DatumType a) => TypedValidator a -> (SpendableOut, DatumType a) -> Doc ann

prettyTxOut :: TxOut -> (Doc ann, Maybe (Doc ann))

prettyTypedValidator :: TypedValidator a -> Doc ann

prettyDatumVal :: Show (DatumType a) => TypedValidator a -> DatumType a -> Value -> Doc ann

prettyOpts :: TxOpts -> Maybe (Doc ann)

Prettifies a TxOpts; returns Nothing if we're looking at default options.

data Field record where

Constructors

Field :: (Show x, Eq x) => String -> (record -> x) -> Field record

Shortening Hashes from Show Instances

prettyDatum :: Show a => a -> Doc ann

The prettyDatum relies on Haskell's default show implementation, but it display shortened hashes. It is a big hack to detect hashes and it works in the vast majority of cases. It might shorten a string that is not a hash iff such string has more than 24 hexadecimal characters, which arguably, is pretty unlikely for something that is not a hash.

data WordPunct

Constructors

Word String

Punct String