Cooked.Tx.Constraints.Type

Documentation

type SpendableOut = (TxOutRef, ChainIndexTxOut)

A SpendableOut is an outref that is ready to be spend; with its underlying ChainIndexTxOut.

sOutValue :: SpendableOut -> Value

Accesses the Value within a SpendableOut

sOutAddress :: SpendableOut -> Address

Accesses the Address within a SpendableOut

sOutDatumHash :: SpendableOut -> Maybe DatumHash

Accesses a potential DatumHash within a SpendableOut; note that the existence (or not) of a datum hash DOES NOT indicate the SpendableOut belongs to a script or a public key; you must pattern match on the result of sOutAddress or use one of sBelongsToPubKey or sBelongsToScript to distinguish that.

sBelongsToPubKey :: SpendableOut -> Maybe PubKeyHash

If a SpendableOut belongs to a public key, return its hash.

sBelongsToScript :: SpendableOut -> Maybe ValidatorHash

If a SpendableOut belongs to a validator, return its hash.

type SpendsConstrs a = (ToData (DatumType a), ToData (RedeemerType a), Show (DatumType a), Show (RedeemerType a), Eq (DatumType a), Eq (RedeemerType a), Typeable a)

type PaysScriptConstrs a = (ToData (DatumType a), Show (DatumType a), Eq (DatumType a), Typeable a)

type MintsConstrs a = (ToData a, Eq a, Show a, Typeable a)

data Constraints

The main constraints datatype. It combines output constraints (OutConstraint) in a specific order that is respected by the generated transaction, and miscellaneous constraints MiscConstraint including input constraints, time constraints, or signature requirements.

Constructors

[MiscConstraint] :=>: [OutConstraint]

data MiscConstraint where

Constraints which do not specify new transaction outputs

Constructors

SpendsScript :: SpendsConstrs a => TypedValidator a -> RedeemerType a -> (SpendableOut, DatumType a) -> MiscConstraint
Ensure that the given TypedValidator spends a specific UTxO (which must belong to the validator). That is: Unlock the UTxO described by the given pair of SpendableOut and DatumType, passing the given redeemer to the validator script.

SpendsPK :: SpendableOut -> MiscConstraint
Ensure that a PubKeyHash spends a specific UTxO. The hash is not an argument since it can be read off the given SpendableOut.

Mints :: MintsConstrs a => Maybe a -> [MintingPolicy] -> Value -> MiscConstraint
Ensure that the transaction mints the given Value. For each CurrencySymbol in the value that is to be minted, the given list of MintingPolicys has to include the policy governing that currency symbol. The Maybe a-argument is the optional redeemer passed to the policies: If you pass Nothing, the constraint is translated using mustMintValue, which uses no redeemer; if you pass Just r, mustMintValueWithRedeemer is used to pass r as a redeemer to the policies.

Before :: POSIXTime -> MiscConstraint
Ensure that the transaction happens before the given time.

After :: POSIXTime -> MiscConstraint
Ensure that the transaction happens after the given time.

ValidateIn :: POSIXTimeRange -> MiscConstraint
Ensure that the transaction happens in the given time range.

SignedBy :: [PubKeyHash] -> MiscConstraint
Ensure that the transaction is signed by the given PubKeyHashes.

(~*~?) :: forall ty a1 a2. (Typeable a1, Typeable a2) => ty a1 -> ty a2 -> Maybe (a1 :~~: a2)

data OutConstraint where

Constraints which specify new transaction outputs

Constructors

PaysScript :: PaysScriptConstrs a => TypedValidator a -> DatumType a -> Value -> OutConstraint
Creates an UTxO to the given validator, with the given datum and the given value. That is, lets the scirpt lock the given value.

PaysPKWithDatum :: (ToData a, Eq a, Show a, Typeable a) => PubKeyHash -> Maybe StakePubKeyHash -> Maybe a -> Value -> OutConstraint
Creates a UTxO to a specific PubKeyHash with a potential StakePubKeyHash. and datum. If the stake pk is present, it will call ownStakePubKeyHash to update the script lookups, hence, generating a transaction with two PaysPKWithDatum with two different staking keys will cause the first staking key to be overriden by calling ownStakePubKeyHash a second time. If this is an issue for you, please do submit an issue with an explanation on GitHub.

class (Eq a, Typeable a) => ConstraintsSpec a where

This typeclass provides user-friendly convenience to overload the txConstraints field of TxSkel. For instance, this enables us to ommit the (:=>:) constructor whenever we're using only one kind of constraints. It also opens up future alternative of constraint specification.

Methods

toConstraints :: a -> Constraints

paysPK :: PubKeyHash -> Value -> OutConstraint

mints :: [MintingPolicy] -> Value -> MiscConstraint

type LabelConstrs x = (Show x, Typeable x, Eq x)

data TxSkel where

A Transaction skeleton is a set of our constraints, and an optional showable label, which is useful when displaying traces. The label was encoded as an existential to enable us to create ad-hoc datatypes and pass their values around, for instance:

data ProposePaymentLbl = ProposePayment Value deriving Show
f = ... (txSkelLbl (ProposePayment val) ...) ...

This way we can (A) modify the show behavior if we want and (B) not worry about constructing consistent strings when constructing transactions

A TxSkel does NOT include a Wallet since wallets only exist in mock mode.

Constructors

TxSkel

Fields

:: LabelConstrs x

=> { txLabel :: Maybe x

, txOpts :: TxOpts

Set of options to use when generating this transaction.

, txConstraints :: Constraints

} -> TxSkel

txSkel :: ConstraintsSpec constraints => constraints -> TxSkel

Constructs a skeleton without a default label and with default TxOpts

txSkelOpts :: ConstraintsSpec constraints => TxOpts -> constraints -> TxSkel

Constructs a skeleton without a default label, but with custom options

txSkelLbl :: (LabelConstrs x, ConstraintsSpec constraints) => x -> constraints -> TxSkel

Constructs a skeleton with a label

txSkelLblOpts :: (LabelConstrs x, ConstraintsSpec constraints) => x -> TxOpts -> constraints -> TxSkel

Constructs a skeleton with the given labtl, options, and constraints

data TxOpts

Set of options to modify the behavior of generating and validating some transaction. Some of these options only have an effect when running in the Contract, some only have an effect when running in MockChainT. If nothing is explicitely stated, the option has an effect independently of the running context.

Constructors

TxOpts

Fields

adjustUnbalTx :: Bool

Performs an adjustment to unbalanced txs, making sure every UTxO that is produced has the necessary minimum amount of Ada.

By default, this is set to False, given this is the default behavior in Plutus: https://github.com/input-output-hk/plutus-apps/issues/143#issuecomment-1013012744

awaitTxConfirmed :: Bool

When submitting a transaction for real (i.e., running in the Contract monad), it is common to call awaitTxConfirmed after submitTxConstraints. If you do NOT wish to do so, please set this to False.

This has NO effect when running outside of Contract. By default, this is set to True.

autoSlotIncrease :: Bool

Whether to increase the slot counter automatically on this submission. This is useful for modelling transactions that could be submitted in parallel in reality, so there should be no explicit ordering of what comes first. One good example is in the Crowdfunding use case contract.

This has NO effect when running in Contract. By default, this is set to True.

forceOutputOrdering :: Bool

Reorders the transaction outputs to fit the ordering of output constraints. Those outputs are put at the very beginning of the list. This has NO effect when running in Contract. By default, this is set to True.

unsafeModTx :: RawModTx

Applies an arbitrary modification to a transaction after it has been pottentially adjusted (adjustUnbalTx) and balanced. This is prefixed with unsafe to draw attention that modifying a transaction at that stage might make it invalid. Still, this offers a hook for being able to alter a transaction in unforeseen ways. It is mostly used to test contracts that have been written for custom PABs.

One interesting use of this function is to observe a transaction just before it is being sent for validation, with unsafeModTx = RawModTx Debug.Trace.traceShowId.

This has NO effect when running in Contract. By default, this is set to Id.

balance :: Bool

Whether to balance the transaction or not. Balancing is the process of ensuring that input + mint = output + fees, if you decide to set balance = false you will have trouble satisfying that equation by hand because fees are variable. You will likely see a ValueNotPreserved error and should adjust the fees accordingly. For now, there is no option to skip the fee computation because without it, validation through Ledger.Validation would fail with InsufficientFees.

This has NO effect when running in Contract. By default, this is set to True.

collateral :: Collateral

Which collateral utxo to use for this transaction. A collateral UTxO must be an Ada-only utxo and can be specified manually, or it can be chosen automatically, if any is available.

This has NO effect when running in Contract. By default, this is set to CollateralAuto.

balanceOutputPolicy :: BalanceOutputPolicy

The BalanceOutputPolicy to apply when balancing the transaction.

This has NO effect when running in Contract. By default, this is set to AdjustExistingOutput.

data BalanceOutputPolicy

Whether to adjust existing public key outputs during transaction balancing.

Constructors

AdjustExistingOutput
Try to adjust an existing public key output with the change. If no suitable output can be found, create a new change output.

DontAdjustExistingOutput
Do not change the existing outputs, always create a new change output.

data RawModTx

Wraps a function that can be applied to a transaction right before submitting it. We have a distinguished datatype to be able to provide a little more info on the show instance.

Constructors

Id

RawModTx (Tx -> Tx)

applyRawModTx :: RawModTx -> Tx -> Tx

data Collateral

Specifies how to select the collateral input

Constructors

CollateralAuto
Will select the first Ada-only UTxO we find belonging to ownPaymentPubKeyHash

CollateralUtxos [TxOutRef]
Will use the TxOutRefs given in the list. This list can be empty, in which case no collateral will be used whatsoever.