Cooked.Validators

Description

This module introduces standard dummy validators to be used in attacks, traces or tests. More precisely, it introduces the always True and always False validators, which will respectively always succeed or always fail.

Documentation

alwaysTrueValidator :: TypedValidator a

The trivial validator that always succeds; this is in particular a sufficient target for the datum hijacking attack since we only want to show feasibility of the attack.

alwaysFalseValidator :: TypedValidator a

The trivial validator that always fails

data MockContract

A Mock contract type to instantiate validators with