Cooked.ValueUtils

Description

Utilities to work with Value

Documentation

flattenValueI :: Iso' Value [(AssetClass, Integer)]

positivePart :: Value -> Value

The positive part of a value. For every asset class in the given value, this asset class and its amount are included in the output iff the amount is strictly positive. It holds

x == positivePart x <> Pl.negate negativePart x

negativePart :: Value -> Value

The negative part of a value. For every asset class in the given value, this asset class and its negated amount are included in the output iff the amount is strictly negative. It holds

x == positivePart x <> Pl.negate negativePart x

adaL :: Lens' Value Ada

Focus the Ada part in a value.