Cooked.Wallet

Description

This module defines convenient wrappers for mock chain wallets (around Plutus mock wallets) and initial distributions (that is the initial state associating a list of UTxOs with some initial values to each known wallet). It also exposes a convenient API to construct wallets and distributions, manipulate them, and fetch information (such as public/private keys and staking keys).

Documentation

knownWallets :: [Wallet]

All the wallets corresponding to known Plutus mock wallets. This is a list of 10 wallets which will

receive funds in the standard initial distribution of cooked-validators, and

be pretty-printed as part the final state after running a few transactions.

wallet :: Int -> Wallet

Wallet corresponding to a given wallet number (or wallet ID)

walletPKHashToId :: PubKeyHash -> Maybe Int

Retrieves the id of the known wallet that corresponds to a public key hash, if any.

walletPKHashToId (walletPKHash (wallet 3)) == Just 3

walletPK :: Wallet -> PubKey

Retrieves a wallet public key (PK)

walletStakingPK :: Wallet -> Maybe PubKey

Retrieves a wallet's public staking key (PK), if any

walletPKHash :: Wallet -> PubKeyHash

Retrieves a wallet's public key hash

walletStakingPKHash :: Wallet -> Maybe PubKeyHash

Retrieves a wallet's public staking key hash, if any

walletAddress :: Wallet -> Address

Retrieves a wallet's address

walletSK :: MockWallet -> PrivateKey

Retrieves a wallet private key (secret key SK)

walletStakingSK :: Wallet -> Maybe PrivateKey

Retrieves a wallet's private staking key (secret key SK), if any

initialDistribution :: [(Wallet, [Value])] -> InitialDistribution

newtype InitialDistribution

Describes the initial distribution of UTxOs per wallet. This is important since transaction validation must specify a collateral. Hence, wallets must have more than one UTxO to begin with in order to execute a transaction and have some collateral option. The txCollateral is transferred to the node operator in case the transaction fails to validate.

The following specifies a starting state where wallet 1 contains two UTxOs, one with 42 Ada and one with 2 Ada and one TOK token; wallet 2 contains a single UTxO with 10 Ada and wallet 3 has 10 Ada and a permanent value. See Cooked.Currencies for more information on quick and permanent values.

i0 = InitialDistribution $ M.fromList
       [ (wallet 1 , [ Pl.lovelaveValueOf 42000000
                     , Pl.lovelaceValueOf 2000000 <> quickValue "TOK" 1
                     ]
       , (wallet 2 , [Pl.lovelaveValueOf 10000000])
       , (wallet 3 , [Pl.lovelaceValueOf 10000000 <> permanentValue "XYZ" 10])
       ]

Constructors

InitialDistribution

Fields

unInitialDistribution :: Map Wallet [Value]

type Wallet = MockWallet

type PrivateKey = XPrv

Orphan instances

Eq Wallet

Instance details

Methods

(==) :: Wallet -> Wallet -> Bool

(/=) :: Wallet -> Wallet -> Bool

Ord Wallet

Instance details

Methods

compare :: Wallet -> Wallet -> Ordering

(<) :: Wallet -> Wallet -> Bool

(<=) :: Wallet -> Wallet -> Bool

(>) :: Wallet -> Wallet -> Bool

(>=) :: Wallet -> Wallet -> Bool

max :: Wallet -> Wallet -> Wallet

min :: Wallet -> Wallet -> Wallet