Example

MockChain Example

example :: Either MockChainError ((), UtxoState)

Start from the initial UtxoIndex, where each known wallets have the same amount of Ada, then transfers 4200 lovelace from wallet 1 to wallet 2. This transfers from wallet 1 because that's the default signer of transactions if nothing else is specified.