Cooked.Attack

The type of attacks, and functions to turn optics into attacks

type Attack = TxSkel -> Maybe TxSkel

The type of attacks that operate on a single transaction. The idea is to try to modify a transaction, or return Nothing if the modification does not apply; use this with the somewhere and everywhere modalities from Cooked.MockChain.Monad.

mkAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify.

-> (a -> Maybe a)
The modification to apply; return Nothing if you want to leave the given focus as it is.

-> Attack

The simplest way to make an attack from an optic: Try to apply a given function of type a -> Maybe a to all foci of an optic, modifying all foci that return Just. This attack returns Just if at least one modification was successful, i.e. if at least one of the foci evaluates to Just; otherwise it returns Nothing.

mkSelectAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify.

-> (a -> Maybe a)
The modification to apply; return Nothing if you want to leave the given focus as it is.

-> (Integer -> Bool)
Maybe the modification applies to (i.e. returns Just on) more than one focus. Use this function to select the foci to modify by the order in which they are traversed. If you set this function to const True, you should probably use mkAttack, because that the semantic of that function.

-> Attack

A more fine-grained attack: Traverse all foci of the optic from the left to the right, while counting them, and modify only those foci that return Just _and_ whose index satisfies a given predicate. This is useful if you have many indentical outputs but you only want to modify a few of them.

mkAccumLAttack

Arguments

:: Is k A_Traversal

=> Optic' k is TxSkel a
Optic focussing potentially interesting points to modify

-> (acc -> a -> (a, acc))
function that describes the modification of the accumulator and the current focus.

-> acc
initial accumulator

-> (TxSkel -> acc -> Maybe TxSkel)
function to decide whether the traversal modified the TxSkel in the desired way. This will typically look like this: > skel acc -> if someTest skel acc > then Just (someFinalModification skel) > else Nothing

-> Attack

A very general attack: Traverse all foci of the optic from the left to the right, collecting an accumulator while also (optionally) modifying the TxSkel. At the end, look at the modified TxSkel and the accumulator to decide whether the traversal was a success, or whether we should return Nothing.

General helpers

addLabel :: LabelConstrs x => x -> TxSkel -> TxSkel

Add a label to a TxSkel. If there is already a pre-existing label, the given label will be added, forming a pair (newlabel, oldlabel).

Token duplication attack

dupTokenAttack

Arguments

:: (AssetClass -> Integer -> Maybe Integer)
A function describing how the amount of minted tokens of an asset class should be changed by the attack. The given function f should probably satisfy > f ac i == Just j -> i < j for all ac and i, i.e. it should increase in the minted amount. If it does *not* increase the minted amount, or if f ac i == Nothing, the minted amount will be left unchanged.

-> Wallet
The wallet of the attacker. Any extra tokens that were minted by the transaction are paid to this wallet.

-> Attack

A token duplication attack which modifies every Mints-constraint of a TxSkel that satisfies some conditions. This adds a DupTokenLbl to the labels of the transaction using addLabel.

data DupTokenLbl

Constructors

DupTokenLbl

Datum hijacking attack

datumHijackingAttack

Arguments

:: forall a. (Typeable a, UnsafeFromData (DatumType a), UnsafeFromData (RedeemerType a))

=> (TypedValidator a -> DatumType a -> Value -> Bool)
Predicate to select outputs to steal, depending on the intended recipient, the datum, and the value.

-> (Integer -> Bool)
If the selection predicate matches more than one output, select the n-th output(s) with this predicate.

-> Attack

A datum hijacking attack, simplified: This attack tries to substitute a different recipient on PaysScript constraints, but leaves the datum as it is. That is, it tests for careless uses of something like txInfoOutputs in places where something like getContinuingOutputs should be used. If this attack goes through, however, a "proper" datum hijacking attack that modifies the datum in a way that the (relevant part of) the toBuiltinData-translation stays the same will also work. A DatumHijackingLbl with the hash of the "thief" validator is added to the labels of the TxSkel using addLabel.

newtype DatumHijackingLbl

Constructors

DatumHijackingLbl ValidatorHash

datumHijackingTarget :: TypedValidator a

The trivial validator that always succeds; this is a sufficient target for the datum hijacking attack since we only want to show feasibility of the attack.