Cooked.Attack.AddToken

Description

This module provides an automated attack to mint and give extra tokens to a certain wallet.

Documentation

addTokenAttack

Arguments

:: (MonadTweak m, OwnerConstrs o)

=> (Versioned MintingPolicy -> [(TokenName, Integer)])
For each policy that occurs in some Mint constraint, return a list of token names together with how many tokens with that name should be minted.

-> o
The wallet of the attacker where extra tokens will be paid to

-> m Value

This attack adds extra tokens, depending on the minting policy. It is different from the dupTokenAttack in that it does not merely try to increase the amount of tokens minted: It tries to mint tokens of asset classes that were not necessarily present on the unmodified transaction.

This attack adds an AddTokenLbl label.

data AddTokenLbl

A label that is added to a TxSkel that has successfully been modified by addTokenAttack

Constructors

AddTokenLbl