Cooked.Attack.DatumHijacking

Description

This module provides an automated attack to try and redirect outputs to a certain target with a similar datum type.

Documentation

redirectOutputTweakAny :: forall owner owner' m. (MonadTweak m, OwnerConstrs owner, OwnerConstrs owner') => (TxSkelOut -> Maybe owner') -> (Integer -> Bool) -> m TxSkelOut

A version of redirectOutputTweakAll where, instead of modifying all the outputs targeted by the input predicates in the same transaction, we modify one of them at a time, relying on the MonadPlus instance of m.

datumHijackingAttackAny

Arguments

:: forall owner owner' m. (MonadTweak m, OwnerConstrs owner, OwnerConstrs owner')

=> (TxSkelOut -> Bool)
Predicate to select outputs to steal, depending on the intended recipient, the datum, and the value.

-> (Integer -> Bool)
The selection predicate may match more than one output. Use this predicate to restrict to the i-th of the outputs (counting from the left, starting at zero) chosen by the selection predicate with this predicate.

-> owner'
The thief

-> m TxSkelOut

A version of datumHijackingAttackAll relying on the rules of redirectOutputTweakAny.

datumHijackingAttack :: forall owner owner' m. (MonadTweak m, OwnerConstrs owner, OwnerConstrs owner') => owner' -> m TxSkelOut

The default datum hijacking attack. It tries to redirect any output for which the owner is of type owner and branches at each attempt.

redirectOutputTweakAll

Arguments

:: forall owner owner' m. (MonadTweak m, OwnerConstrs owner, OwnerConstrs owner')

=> (TxSkelOut -> Maybe owner')
Return Just the new owner, or Nothing if you want to leave this output unchanged.

-> (Integer -> Bool)
The redirection described by the previous argument might apply to more than one of the outputs of the transaction. Use this predicate to select which of the redirectable outputs to actually redirect. We count the redirectable outputs from the left to the right, starting with zero.

-> m [TxSkelOut]
Returns the list of outputs it redirected (as they were before the modification), in the order in which they occurred on the original transaction.

Redirects some outputs from one owner to another owner, which can be of different types.

datumHijackingAttackAll

Arguments

:: forall owner owner' m. (MonadTweak m, OwnerConstrs owner, OwnerConstrs owner')

=> (TxSkelOut -> Bool)
Predicate to select outputs to steal, depending on the intended recipient, the datum, and the value.

-> (Integer -> Bool)
The selection predicate may match more than one output. Use this predicate to restrict to the i-th of the outputs (counting from the left, starting at zero) chosen by the selection predicate with this predicate.

-> owner'
The thief

-> m [TxSkelOut]

A datum hijacking attack, simplified: This attack tries to substitute a different recipient on outputs belonging to scripts, but leaves the datum as it is. That is, it tests for careless uses of something like txInfoOutputs in places where something like getContinuingOutputs should be used. If this attack goes through, however, a "proper" datum hijacking attack that modifies the datum in a way that (the relevant part of) the toBuiltinData-translation stays the same will also work.

A DatumHijackingLbl with the hash of the "thief" validator is added to the labels of the TxSkel using addLabelTweak.

This attack returns the list of outputs it redirected, in the order in which they occurred on the original transaction. If no output is redirected, this attack fails.

newtype DatumHijackingLbl

A label that is added to a TxSkel that has successfully been modified by any of the datum hijacking attacks

Constructors

DatumHijackingLbl Credential