Cooked.Attack.DatumHijacking

Documentation

datumHijackingAttack

Arguments

:: forall a. (Typeable a, UnsafeFromData (DatumType a), UnsafeFromData (RedeemerType a))

=> (TypedValidator a -> DatumType a -> Value -> Bool)
Predicate to select outputs to steal, depending on the intended recipient, the datum, and the value.

-> (Integer -> Bool)
If the selection predicate matches more than one output, select the n-th output(s) with this predicate.

-> Attack

A datum hijacking attack, simplified: This attack tries to substitute a different recipient on PaysScript constraints, but leaves the datum as it is. That is, it tests for careless uses of something like txInfoOutputs in places where something like getContinuingOutputs should be used. If this attack goes through, however, a "proper" datum hijacking attack that modifies the datum in a way that the (relevant part of) the toBuiltinData-translation stays the same will also work. A DatumHijackingLbl with the hash of the "thief" validator is added to the labels of the TxSkel using addLabel.

newtype DatumHijackingLbl

Constructors

DatumHijackingLbl Address

datumHijackingTarget :: TypedValidator a

The trivial validator that always succeds; this is a sufficient target for the datum hijacking attack since we only want to show feasibility of the attack.