Cooked.Attack.DoubleSat

Documentation

data DoubleSatParams a

Parameters for a double satisfaction attack. You probably don't want to initialise these by hand; better use one of the smart constructors defined below.

Constructors

DoubleSatParams

Fields

dsOptic :: Traversal' TxSkel a

Each focus of this traversal is one potential reason to add extra Constraints.

dsExtraConstraints :: MockChainSt -> a -> [Constraints]

For every a in the transaction under modification, calculate a list of extra Constraints to add to the transaction in order to try a double satisfaction attack.

Extra Constraints are added to pre-existing Constraints with the function addConstraints, which makes sure that no additional constraint tries spending an UTxO that's already being spent and other such checks. See the documentation comments for that function below.

We add OutConstraints at the end of the list of transaction outputs, because some contracts rely on the ordering of (the first few) outputs. If you want to try permutations of transaction outputs, look at the function tryOutPermutations, which is an attack-agnostic way to accomplish that.

dsAttacker :: Wallet

The wallet to which any surplus will be paid.

dsSplitStrategy :: SplitStrategy

How to combine extra Constraints:

There are at the moment the following two strategies for how to combine additional Constraints belonging to _different_ SpendsScript constraints of the original transaction.

With dsSplitStrategy = OneChange, each extra Constraints is added onto a separate modified TxSkel: Each of the output modified TxSkels will receive exactly one of the extra Constraints.

With dsSplitStrategy = AllCombinations, al TxSkels obtained by adding _at least one_ extra Constraints belonging to an original a are tried: If there are n as in the original transaction, the modified transactions will incorporate up to n extra Constraints. For a given a on the original transaction, each of its extra Constraints is tried together with all options for all other as. This means that all cases that the OneChange-case checks are also explored here, plus some (potentially very many) more.

If there is a conflict between some of the Constraints that are to be added to an output transaction (for example, two Constraints might require the same UTxO to be spent with different redeemers), that combination is not tried.

See the comments for mkSplittingAttack for more explanation on SplitStrategy.

Smart constructors for DoubleSatParams

dsAddOneSscFromOwner

Arguments

:: (SpendsConstrs b, FromData (DatumType b))

=> Traversal' TxSkel a

-> TypedValidator b
The validator to take extra inputs from.

-> (a -> SpendableOut -> [RedeemerType b])
For all as of the original transaction, decide whether to add an extra UTxO currently belonging to the extraInputOwner, and if so, which redeemers to try. Each redeemer is tried on a separate output transaction.

-> Wallet

-> DoubleSatParams a

Parameters for a double satisfaction attack that adds one SpendsScriptConstraint for a UTxO belonging to a given validator. Each of the modified transactions will contain exactly one extra SpendsScript constraint.

dsAddOneSscToPsc :: (SpendsConstrs b, FromData (DatumType b)) => TypedValidator b -> (PaysScriptConstraint -> SpendableOut -> [RedeemerType b]) -> Wallet -> DoubleSatParams PaysScriptConstraint

Parameters for a double satisfaction attack that adds one extra SpendsScript constraint for a Utxo belonging to a given validator to transactions, depending on PaysScript constraints already present on the transaction. Each of the modified transactions will contain exactly one extra SpendsScript constraint. See the comments at dsAddOneSscFromOwner for explanation of the arguments.

dsAddOneSscToSsc :: (SpendsConstrs b, FromData (DatumType b)) => TypedValidator b -> (SpendsScriptConstraint -> SpendableOut -> [RedeemerType b]) -> Wallet -> DoubleSatParams SpendsScriptConstraint

Parameters for a double satisfaction attack that adds one extra SpendsScript constraint for a Utxo belonging to a given validator to transactions, depending on SpendsScript constraints already present on the transaction. Each of the modified transactions will contain exactly one extra SpendsScript constraint. See the comments at dsAddOneSscFromOwner for explanation of the arguments.

dsAddOneSscToMc :: (SpendsConstrs b, FromData (DatumType b)) => TypedValidator b -> (MintsConstraint -> SpendableOut -> [RedeemerType b]) -> Wallet -> DoubleSatParams MintsConstraint

Parameters for a double satisfaction attack that adds one extra SpendsScript constraint for a Utxo belonging to a given validator to transactions, depending on Mints constraints already present on the transaction. Each of the modified transactions will contain exactly one extra SpendsScript constraint. See the comments at dsAddOneSscFromOwner for explanation of the arguments.

dsAddMcToSsc

Arguments

:: MintsConstrs a

=> (SpendsScriptConstraint -> [(Maybe a, [MintingPolicy], Value)])
For all SpendsScript constraints of the original transaction, decide whether to mint some extra value, and which redeemer and minting policies to use to do so.

-> Wallet
Wallet of the attacker. Any extra minted values are paid to this wallet.

-> DoubleSatParams SpendsScriptConstraint

Parameters for a double satisfaction attack that adds one or more extra Mints constraints to transactions, depending on SpendsScript constraints already present on the transaction. Each of the modified transactions will contain at least one extra Mints constraint.

dsAddMcToMc

Arguments

:: MintsConstrs a

=> (MintsConstraint -> [(Maybe a, [MintingPolicy], Value)])
For all Mints constraints of the original transaction, decide whether to mint some extra value, and which redeemer and minting policies to use to do so.

-> Wallet
Wallet of the attacker. Any extra minted values are paid to this wallet.

-> DoubleSatParams MintsConstraint

Parameters for a double satisfaction attack that adds one or more extra Mints constraints to transactions, depending on Mints constraints already present on the transaction. Each of the modified transactions will contain at least one extra Mints constraint.

Note that this is more general than just a token duplication attack as implemented by dupTokenAttack: That attack is used to make _one_ minting policy mint more tokens, with this attack, more than one minting policy may be involved.

dsAddSsc :: Traversal' TxSkel a -> (MockChainSt -> a -> [SpendsScriptConstraint]) -> Wallet -> SplitStrategy -> DoubleSatParams a

Parameters for a double satisfaction attack that adds SpendsScript constraints.

dsAddMc :: Traversal' TxSkel a -> (MockChainSt -> a -> [MintsConstraint]) -> Wallet -> SplitStrategy -> DoubleSatParams a

Parameters for a doubleSatisfactionAttack that adds Mints constraints.

The double satisfaction attack

doubleSatAttack :: DoubleSatParams a -> Attack

Double satisfaction attack. See the comments for DoubleSatParams and its smart constructors for explanation.

Extending Constraints

addConstraints :: Constraints -> Constraints -> Constraints

Extend the constraints in the second argument with the ones from the first argument, but only if there are no conflicts:

No SpendsScript or SpendsPK constraints trying to spend the same UTxO

No incompatible time ranges

strictlyExtendsConstraints :: Constraints -> Constraints -> Bool

Return True iff the constraints in the first argument can be added without conflict to the ones in the second argument, and they are not alreay covered by the constraints in the second argument.

data CheckMiscExtend

Constructors

Incompatible

AlreadyThere

CompatibleExtension

checkMiscExtend :: MiscConstraint -> [MiscConstraint] -> CheckMiscExtend

Check how a given MiscConstraint relates to a set of other MiscConstraints:

If the set of constraints already includes the given constraint (either directly or as a consequence), return AlreadyThere

If the given constraint can be added to the set without conflict, return CompatibleExtension

Otherwise, return Incompatible

data DoubleSatLbl

Constructors

DoubleSatLbl