Cooked.Attack.DupToken

Documentation

dupTokenAttack

Arguments

:: (AssetClass -> Integer -> Maybe Integer)
A function describing how the amount of minted tokens of an asset class should be changed by the attack. The given function f should probably satisfy > f ac i == Just j -> i < j for all ac and i, i.e. it should increase in the minted amount. If it does *not* increase the minted amount, or if f ac i == Nothing, the minted amount will be left unchanged.

-> Wallet
The wallet of the attacker. Any additional tokens that are minted by the modified transaction but were not minted by the original transaction are paid to this wallet.

-> Attack

A token duplication attack which modifies every Mints-constraint of a TxSkel that satisfies some conditions. This adds a DupTokenLbl to the labels of the transaction using addLabel.

data DupTokenLbl

Constructors

DupTokenLbl