Cooked.Currencies

Quick Values

Quick values are convenient to manipulate assets that are supposed to exist when running a mock chain. For example, a market maker would exchange Ada against other assets. Yet, when writing traces for such contract we'd need to define a minting policy for those tokens, which is very repetitive. Moreover, most of the times we'd want wallets to start with some positive balance of tokens.

There are two classes of functions for using custom tokens on traces:

The quick prefixed functions provide access to tokens from the const (const True) minting policy. That is, these can be minted and burnt at will, at any point in time.

The perpetual prefixed functions provide access to tokens from the const (const False) minting policy. That is, these cannot ever be minted or burnt and must be present in an initial distribution to be useful.

See the docs for InitialDistribution for an example usage.

quickTokenName :: String -> TokenName

Token name of a quick asset class; prefixes the name with a q to make it easy to distinguish between quick and permanent tokens.

quickAssetClass :: String -> AssetClass

Quick asset class from a token name

quickValue :: String -> Integer -> Value

Constructor for quick values from token name and amount

permanentTokenName :: String -> TokenName

Token name of a permanent asset class

permanentAssetClass :: String -> AssetClass

Permanent asset class from a token name

permanentValue :: String -> Integer -> Value

Constructor for Permanent values from token name and amount

QuickValue Minting Policies

mkQuickCurrencyPolicy :: () -> ScriptContext -> Bool

quickCurrencyPolicy :: MintingPolicy

quickCurrencySymbol :: CurrencySymbol

mkPermanentCurrencyPolicy :: () -> ScriptContext -> Bool

permanentCurrencyPolicy :: MintingPolicy

permanentCurrencySymbol :: CurrencySymbol