Cooked.Ltl

Documentation

data Ltl a

Type of LTL formulas with atomic formulas of type a. Think of a as a type of "modifications", then a value of type Ltl a describes where to apply modifications. Since it does not make (obvious) sense to talk of a negated modification or of one modification (possibly in the future) to imply another modification, implication and negation are absent.

Constructors

LtlTruth
The "do nothing" modification that never fails

LtlFalsity
The modification that never applies (i.e. always fails)

LtlAtom a
The modification that applies a given atomic modification at the current time step

LtlOr (Ltl a) (Ltl a)
Disjunction will be interpreted in an "intuitionistic" way, i.e. as branching into the "timeline" where the left disjunct holds and the one where the right disjunct holds. In that sense, it is an exclusive or, as it does not introduce the branch where both disjuncts hold.

LtlAnd (Ltl a) (Ltl a)
Conjunction will be interpreted as "apply both modifications". Attention: The "apply both" operation will be user-defined for atomic modifications, so that conjunction may for example fail to be commutative if the operation on atomic modification is not commutative.

LtlNext (Ltl a)
Assert that the given formula holds at the next time step.

LtlUntil (Ltl a) (Ltl a)
Assert that the first formula holds at least until the second one begins to hold, which must happen eventually. The formulas > a LtlUntil b and > b LtlOr (a LtlAnd LtlNext (a LtlUntil b)) are equivalent.

LtlRelease (Ltl a) (Ltl a)
Assert that the second formula has to be true up to and including the point when the first one becomes true; if that never happens, the second formula has to remain true forever. View this as dual to LtlUntil. The formulas > a LtlRelease b and > b LtlAnd (a LtlOr LtlNext (a LtlRelease b)) are equivalent.

nowLater :: Monoid a => Ltl a -> [(a, Ltl a)]

Split an LTL formula that describes a modification of a computation into a list of (doNow, doLater) pairs, where

doNow is the modification to be applied to the current time step,

doLater is an LTL formula describing the modification that should be applied from the next time step onwards, and

The return value is a list because a formula might be satisfied in different ways. For example, the modification described by a LtlUntil b might be accomplished by applying the modification b right now, or by applying a right now and a LtlUntil b from the next step onwards; the returned list will contain these two options.

Modifications should form a Monoid, where mempty is the do-nothing modification, and <> is the composition of modifications. We interpret a <> b as the modification that first applies b and then a. Attention: Since we use <> to define conjunction, if <> is not commutative, conjunction will also fail to be commutative!

nowLaterList :: Monoid a => [Ltl a] -> [(a, [Ltl a])]

Say we're passing around more than one formula from each time step to the next, where the intended meaning of a list of formulas is the modification that applies the first formula in the list first, then the second formula, then the third and so on. We'd still like to compute a list of (doNow, doLater) pairs as in nowLater, only that the doLater should again be a list of formulas.

data LtlOp (modification :: *) (builtin :: * -> *) :: * -> * where

The idea is that a value of type Staged (LtlOp modification builtin) a describes a set of (monadic) computations that return an a such that

every step of the computations that returns a b is reified as a builtin b, and

every step can be modified by a modification.

Operations for computations that can be modified using LTL formulas.

Constructors

StartLtl :: Ltl modification -> LtlOp modification builtin ()
The operation that introduces a new LTL formula that should be used to modify the following computations. Think of this operation as coming between time steps and adding a new formula to be applied before all of the formulas that should already be applied to the next time step.

StopLtl :: LtlOp modification builtin ()
The operation that removes the last LTL formula that was introduced. If the formula is not yet finished, the current time line will fail.

Builtin :: builtin a -> LtlOp modification builtin a

data Staged (op :: * -> *) :: * -> * where

The freer monad on op. We think of this as the AST of a computation with operations of types op a.

Constructors

Return :: a -> Staged op a

Instr :: op a -> (a -> Staged op b) -> Staged op b

interpLtl :: InterpLtl modification builtin m => Staged (LtlOp modification builtin) a -> StateT [Ltl modification] m a

Interpret a Staged computation into a suitable domain, using the function interpBuiltin to interpret the builtins.

interpLtlAndPruneUnfinished :: InterpLtl modification builtin m => Staged (LtlOp modification builtin) a -> StateT [Ltl modification] m a

Interpret a Staged computation into a suitable domain, using the function interpBuiltin to interpret the builtins. At the end of the computation, prune branches that still have unfinished modifications applied to them. See the discussion on the regression test case for PRs 110 and 131 in hs for a discussion on why this function has to exist.

class MonadPlus m => InterpLtl modification builtin m where

To be a suitable semantic domain for computations modified by LTL formulas, a monad m has to

have the right builtin functions, which can be modified by the right modifications,

be a MonadPlus, because one LTL formula might yield different modified versions of the computation, and

This type class only requires from the user to specify how to interpret the (modified) builtins. In order to do so, it passes around the formulas that are to be applied to the next time step in a StateT. A common idiom to modify an operation should be this:

interpBuiltin op =
 get
   >>= msum
     . map (\(now, later) -> applyModification now op <* put later)
     . nowLaterList

(But to write this, modification has to be a Monoid to make nowLaterList work!) Look at the tests for this module and at Cooked.MockChain.Monad.Staged for examples of how to use this type class.

Methods

interpBuiltin :: builtin a -> StateT [Ltl modification] m a

class Monad m => MonadModal m where

Monads that allow modificaitons with LTL formulas.

Associated Types

type Modification m :: Type

Methods

modifyLtl :: Ltl (Modification m) -> m a -> m a