Cooked.MockChain

Documentation

utxoPayloadReferenceScript :: UtxoPayload -> Maybe ScriptHash

utxoPayloadSkelOutDatum :: UtxoPayload -> TxSkelOutDatum

utxoPayloadTxOutRef :: UtxoPayload -> TxOutRef

utxoPayloadValue :: UtxoPayload -> Value

newtype UtxoPayloadSet

Represents a set of payloads. Payloads are the name we give to the information we care about on a UTxO: value, datum, and reference script. We use a list instead of a set because Value doesn't implement Ord and because it is possible that we want to distinguish between utxo states that have additional utxos, even if these could have been merged together.

Constructors

UtxoPayloadSet

Fields

utxoPayloadSet :: [UtxoPayload]

holdingInState :: UtxoState -> Wallet -> Value

Helper function to compute what the given wallet owns in the given state

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTransControl.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the boilerplate of defining all the methods of the class yourself.

Constructors

AsTrans

Fields

getTrans :: t m a

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

The main abstraction of the blockchain.

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates, balances and validates a transaction from a skeleton. It returns the validated transaction and updates the state of the blockchain. In MockChainT, this means: - deletes the consumed outputs from mcstIndex - adds the produced outputs to msctIndex - deletes the consumed datums from mcstDatums - adds the produced datums to mcstDatums - adds the validators on outputs to the mcstValidators.

class MonadBlockChainBalancing m => MonadBlockChainWithoutValidation m where

Methods

allUtxosLedger :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs.

currentSlot :: m Slot

Returns the current slot number

awaitSlot :: Slot -> m Slot

Waits until the current slot becomes greater or equal to the given slot, and returns the current slot after waiting.

Note that it might not wait for anything if the current slot is large enough.

class (MonadFail m, MonadError MockChainError m) => MonadBlockChainBalancing m where

Contains methods needed for balancing.

Methods

getParams :: m Params

Returns the parameters of the chain.

utxosAtLedger :: Address -> m [(TxOutRef, TxOut)]

Returns a list of all UTxOs at a certain address.

datumFromHash :: DatumHash -> m (Maybe Datum)

Returns the datum with the given hash or Nothing if there is none

validatorFromHash :: ValidatorHash -> m (Maybe (Versioned Validator))

Returns the full validator corresponding to hash, if that validator owns something or if it is stored in the reference script field of some UTxO.

txOutByRefLedger :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

data MCEUnbalanceableError

Constructors

MCEUnbalNotEnoughFunds Wallet Value
The balancing wallet misses some value to pay what is needed to balance the transaction.

MCEUnbalNotEnoughReturning (Value, [TxOutRef]) (Value, [TxOutRef]) Value
There is value to return to the balancing wallet but not enough to fullfill the min ada requirement and there is not enough in additional inputs to make it possible.

data MockChainError where

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError :: ValidationErrorInPhase -> MockChainError

MCEUnbalanceable :: MCEUnbalanceableError -> TxSkel -> MockChainError

MCENoSuitableCollateral :: MockChainError
Thrown when the balancing wallet owns no output that is pure Ada and with no datum.

MCEGenerationError :: GenerateTxError -> MockChainError

MCECalcFee :: MockChainError -> MockChainError
Errors happening at fee calculation time.

MCEUnknownOutRefError :: String -> TxOutRef -> MockChainError
Thrown when an output reference should be in the state of the mockchain, but isn't.

MCEUnknownValidator :: String -> ValidatorHash -> MockChainError
Same as MCEUnknownOutRefError for validators.

MCEUnknownDatum :: String -> DatumHash -> MockChainError
Same as MCEUnknownOutRefError for datums.

FailWith :: String -> MockChainError
Used to provide MonadFail instances.

OtherMockChainError :: (Show err, Eq err) => err -> MockChainError

allUtxos :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

utxosAt :: MonadBlockChainBalancing m => Address -> m [(TxOutRef, TxOut)]

txOutByRef :: MonadBlockChainBalancing m => TxOutRef -> m (Maybe TxOut)

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

txOutV2FromLedger :: TxOut -> TxOut

resolveDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out)))

Try to resolve the datum on the output: If there's an inline datum, take that; if there's a datum hash, look the corresponding datum up (with datumFromHash), returning Nothing if it can't be found; if there's no datum or hash at all, return Nothing.

resolveValidator :: (IsAbstractOutput out, ToCredential (OwnerType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (Versioned Validator) (DatumType out) (ValueType out) (ReferenceScriptType out)))

Try to resolve the validator that owns an output: If the output is owned by a public key, or if the validator's hash is not known (i.e. if validatorFromHash returns Nothing) return Nothing.

resolveReferenceScript :: (IsAbstractOutput out, ToScriptHash (ReferenceScriptType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (Versioned Validator)))

Try to resolve the reference script on an output: If the output has no reference script, or if the reference script's hash is not known (i.e. if validatorFromHash returns Nothing), this function will return Nothing.

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

waitNSlots :: MonadBlockChainWithoutValidation m => Integer -> m Slot

Moves n slots fowards

currentTime :: MonadBlockChainWithoutValidation m => m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the current slot

slotToTimeInterval :: MonadBlockChainWithoutValidation m => Slot -> m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the slot with the given number. It holds that

slotToTimeInterval (getEnclosingSlot t) == (a, b)    ==>   a <= t <= b

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot a == n && getEnclosingSlot b == n

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot (a-1) == n-1 && getEnclosingSlot (b+1) == n+1

getEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Return the slot that contains the given time. See slotToTimeInterval for some equational properties this function satisfies.

awaitEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Waits until the current slot becomes greater or equal to the slot containing the given POSIX time. Note that that it might not wait for anything if the current slot is large enough.

slotRangeBefore :: MonadBlockChain m => POSIXTime -> m SlotRange

The infinite range of slots ending before or at the given POSIX time

slotRangeAfter :: MonadBlockChain m => POSIXTime -> m SlotRange

The infinite range of slots starting after or at the given POSIX time

runUtxoSearch :: Monad m => UtxoSearch m a -> m [(TxOutRef, a)]

Given a UTxO search, we can run it to obtain a list of UTxOs.

allUtxosSearch :: MonadBlockChain m => UtxoSearch m TxOut

Search all currently known TxOutRefs together with their corresponding TxInfo-TxOut.

allUtxosLedgerSearch :: MonadBlockChain m => UtxoSearch m TxOut

Like allUtxosSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

utxosAtSearch :: MonadBlockChainBalancing m => Address -> UtxoSearch m TxOut

Search all TxOutRefs at a certain address, together with their TxInfo-TxOut.

utxosAtLedgerSearch :: MonadBlockChainBalancing m => Address -> UtxoSearch m TxOut

Like utxosAtSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

utxosFromCardanoTxSearch :: Monad m => CardanoTx -> UtxoSearch m TxOut

Search all TxOutRefs of a transaction, together with their TxInfo-TxOut.

filterWith :: Monad m => UtxoSearch m a -> (a -> m (Maybe b)) -> UtxoSearch m b

Transform a UtxoSearch by applying a possibly failing monadic "lookup" on every output.

filterWithPure :: Monad m => UtxoSearch m a -> (a -> Maybe b) -> UtxoSearch m b

filterWithOptic :: (Is k An_AffineFold, Monad m) => UtxoSearch m a -> Optic' k is a b -> UtxoSearch m b

filterWithPred :: Monad m => UtxoSearch m a -> (a -> Bool) -> UtxoSearch m a

balancedTxSkel :: MonadBlockChainBalancing m => TxSkel -> m (TxSkel, Fee, Set TxOutRef)

balancedTx :: MonadBlockChainBalancing m => (TxSkel, Fee, Set TxOutRef) -> m (Tx BabbageEra)

Take the output of balancedTxSkel and turn it into an actual Cardano transaction.

txSkelInputData :: MonadBlockChainBalancing m => TxSkel -> m (Map DatumHash Datum)

Look up the data on UTxOs the transaction consumes.

type MockChain = MockChainT Identity

newtype MockChainT m a

Constructors

MockChainT

Fields

unMockChain :: ReaderT MockChainEnv (StateT MockChainSt (ExceptT MockChainError m)) a

newtype MockChainEnv

Constructors

MockChainEnv

Fields

mceParams :: Params

data MockChainSt

Slightly more concrete version of UtxoState, used to actually run the simulation. We keep a map from datum hash to datum in order to display the contents of the state to the user.

Constructors

MockChainSt

Fields

mcstIndex :: UtxoIndex

mcstDatums :: Map DatumHash TxSkelOutDatum

mcstValidators :: Map ValidatorHash (Versioned Validator)

mcstCurrentSlot :: Slot

mcstToUtxoState :: MockChainSt -> UtxoState

utxoIndexToTxOutMap :: UtxoIndex -> Map TxOutRef TxOut

The UtxoIndex contains TxOuts, but we want a map that contains TxOut.

combineMockChainT :: Monad m => (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

mapMockChainT :: (m (Either MockChainError (a, MockChainSt)) -> n (Either MockChainError (b, MockChainSt))) -> MockChainT m a -> MockChainT n b

runMockChainTRaw :: Monad m => MockChainEnv -> MockChainSt -> MockChainT m a -> m (Either MockChainError (a, MockChainSt))

Executes a MockChainT from some initial state and environment; does not convert the MockChainSt into a UtxoState.

runMockChainTFrom :: Monad m => InitialDistribution -> MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from an initial state set up with the given initial value distribution. Similar to runMockChainT, uses the default environment. Returns a UtxoState instead of a MockChainSt. If you need the later, use runMockChainTRaw

runMockChainT :: Monad m => MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from the canonical initial state and environment. The canonical environment uses the default SlotConfig and Cooked.Wallet.wallet 1 as the sole wallet signing transactions.

runMockChainRaw :: MockChainEnv -> MockChainSt -> MockChain a -> Either MockChainError (a, MockChainSt)

See runMockChainTRaw

runMockChainFrom :: InitialDistribution -> MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainTFrom

runMockChain :: MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainT

utxoState0 :: UtxoState

mockChainSt0 :: MockChainSt

mockChainSt0From :: InitialDistribution -> MockChainSt

utxoIndex0From :: InitialDistribution -> UtxoIndex

utxoIndex0 :: UtxoIndex

runTransactionValidation

Arguments

:: Monad m

=> Tx BabbageEra
The transaction to validate. It should already be balanced, and include appropriate fees and collateral.

-> [RawModTx]
Modifications to apply to the transaction right before it is submitted.

-> Map DatumHash Datum
The data consumed by the transaction

-> Map DatumHash TxSkelOutDatum
The data produced by the transaction

-> Map ValidatorHash (Versioned Validator)
The validators protecting transaction outputs, and the validators in the reference script field of transaction outputs. The MockChain will remember them.

-> MockChainT m SomeCardanoApiTx

type MonadModalBlockChain m = (MonadBlockChain m, MonadModal m, Modification m ~ UntypedTweak InterpMockChain)

A modal mock chain is a mock chain that allows us to use LTL modifications with Tweaks

unMockChainLog :: MockChainLog -> [MockChainLogEntry]

pattern MCLogFail :: String -> MockChainLogEntry

pattern MCLogSubmittedTxSkel :: SkelContext -> TxSkel -> MockChainLogEntry

pattern MCLogNewTx :: TxId -> MockChainLogEntry

somewhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to some transaction in the given Trace. The tweak must apply at least once.

everywhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to every transaction in a given trace. This is also successful if there are no transactions at all.

withTweak :: MonadModalBlockChain m => m x -> Tweak InterpMockChain a -> m x

Apply a Tweak to the next transaction in the given trace. The order of arguments is reversed compared to somewhere and everywhere, because that enables an idiom like

do ...
   endpoint arguments `withTweak` someModification
   ...

where endpoint builds and validates a single transaction depending on the given arguments. Then withTweak says "I want to modify the transaction returned by this endpoint in the following way".

class IsProp prop where

This module provides a common interface for HUnit and QuickCheck tests. We do so by abstracting uses of Assertion and Property for (IsProp prop) => prop, then provide instances for both HU.Asserton and QC.Property.

Minimal complete definition

testCounterexample, testConjoin, testDisjoin

Methods

testCounterexample :: String -> prop -> prop

Displays the string to the user in case of failure

testConjoin :: [prop] -> prop

Conjunction of a number of results

testDisjoin :: [prop] -> prop

Disjunction of a number of results

testFailure :: prop

Flags a failure

testSuccess :: prop

Flags a success

testFailureMsg :: String -> prop

Flags a failure with a message

testBool :: IsProp prop => Bool -> prop

testAll :: IsProp prop => (a -> prop) -> [a] -> prop

(.==.) :: (IsProp prop, Eq a) => a -> a -> prop infix 4

(.&&.) :: IsProp prop => prop -> prop -> prop infixr 3

(.||.) :: IsProp prop => prop -> prop -> prop infixr 2

testSucceeds :: IsProp prop => PrettyCookedOpts -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed, starting from the default initial distribution

testFails :: (IsProp prop, Show a) => PrettyCookedOpts -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from the default initial distribution

testSucceedsFrom :: IsProp prop => PrettyCookedOpts -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed starting from some initial distribution but doesn't impose any additional condition on success. Use testSucceedsFrom' for that.

testSucceedsFrom' :: IsProp prop => PrettyCookedOpts -> (a -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed starting from some initial distribution. Additionally impose a condition over the resulting state and value.

testFailsFrom :: (IsProp prop, Show a) => PrettyCookedOpts -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from some initial distribution

testFailsFrom' :: (IsProp prop, Show a) => PrettyCookedOpts -> (MockChainError -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from some initial distribution, moreover, ensures that a certain predicate over the error holds.

isCekEvaluationFailure :: IsProp prop => PrettyCookedOpts -> MockChainError -> prop

Is satisfied when the given MockChainError is wrapping a CekEvaluationFailure. This is particularly important when writing negative tests. For example, if we are simulating an attack and writing a test with testFailsFrom, we might have made a mistake in the attack, yielding a test that fails for reasons such as ValueLessThanMinAda or ValueNotPreserved, which does not rule out the attack being caught by the validator script. For these scenarios it is paramount to rely on testFailsFrom' isCekEvaluationFailure instead.

isCekEvaluationFailureWithMsg :: IsProp prop => PrettyCookedOpts -> (String -> Bool) -> MockChainError -> prop

Similar to isCekEvaluationFailure, but enables us to check for a specific error message in the error.

testAllSatisfiesFrom :: forall prop a. IsProp prop => PrettyCookedOpts -> (Either MockChainError (a, UtxoState) -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the set of traces encoded by the StagedMockChain satisfy the given predicate. If you wish to build custom predicates you can use testSatisfiesFrom' directly and see testBinaryRelatedBy as an example.

testBinaryRelatedBy :: IsProp prop => PrettyCookedOpts -> (UtxoState -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Asserts that the given StagedMockChain produces exactly two outcomes, both of which are successful and have their resulting states related by a given predicate. A typical usage would look like:

testBinaryRelatedBy equalModuloAda myInitDistr $ do
  x <- trPrepare
  execOption1 x <|> execOption2 x

testOneEquivClass :: IsProp prop => PrettyCookedOpts -> (UtxoState -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Generalizes testBinaryRelatedBy, asserting that the given StagedMockChain produces more than two outcomes, say [x,y,z,w], all of which are successful (i.e. are not a MockChainError) and these states are in the same equivalence class of (~); that is, they satisfy:

x ~ y && x ~ z && x ~ z && x ~ w

Because (~) should be symmetric and transitive we can estabilish that these states all belong to the same equivalence class. This function does not check each pointwise case.

testSatisfiesFrom' :: ([(Either MockChainError (a, UtxoState), MockChainLog)] -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Asserts that the results produced by running the given StagedMockChain from some speficied InitialDistribution satisfy a given assertion. In this case, the predicate gets the trace descriptions that led to each potential outcome and is responsible for calling testCounterexample communicate these to the user.

Although this function is mainly used internally, as a building block for the simpler predicates, it can be useful in building some custom predicates. Check testAllSatisfiesFrom or testBinaryRelatedBy for examples on using this.

assertionToMaybe :: Assertion -> IO (Maybe HUnitFailure)

forAll :: Show a => Gen a -> (a -> Property) -> Property

Here we provide our own universsal quantifier instead of forAll, so we can monomorphize it to returning a Property