Cooked.MockChain

Description

This module centralizes everything related to our mockchain, while hiding elements related to logs and inner state.

Documentation

type MockChain = MockChainT Identity

data UtxoState

A description of who owns what in a blockchain. Owners are addresses and they each own a UtxoPayloadSet.

type UtxoSearch m a = ListT m (TxOutRef, a)

If a UTxO is a TxOutRef with some additional information, this type captures a "stream" of UTxOs.

data MockChainError where

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError :: ValidationPhase -> ValidationError -> MockChainError

MCEUnbalanceable :: Wallet -> Value -> TxSkel -> MockChainError
Thrown when the balancing wallet does not have enough funds

MCENoSuitableCollateral :: Integer -> Integer -> Value -> MockChainError
Thrown when not enough collateral are provided. Built upon the fee, the percentage and the expected minimal collateral value.

MCEGenerationError :: GenerateTxError -> MockChainError

MCEUnknownOutRefError :: String -> TxOutRef -> MockChainError
Thrown when an output reference should be in the state of the mockchain, but isn't.

MCEUnknownValidator :: String -> ValidatorHash -> MockChainError
Same as MCEUnknownOutRefError for validators.

MCEUnknownDatum :: String -> DatumHash -> MockChainError
Same as MCEUnknownOutRefError for datums.

FailWith :: String -> MockChainError
Used to provide MonadFail instances.

class (MonadFail m, MonadError MockChainError m) => MonadBlockChainBalancing m where

Contains methods needed for balancing.

Methods

getParams :: m Params

Returns the parameters of the chain.

utxosAtLedger :: Address -> m [(TxOutRef, TxOut)]

Returns a list of all UTxOs at a certain address.

datumFromHash :: DatumHash -> m (Maybe Datum)

Returns the datum with the given hash if present.

validatorFromHash :: ValidatorHash -> m (Maybe (Versioned Validator))

Returns the full validator corresponding to hash, if that validator owns something or if it is stored in the reference script field of some UTxO.

txOutByRefLedger :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

class MonadBlockChainBalancing m => MonadBlockChainWithoutValidation m where

Methods

allUtxosLedger :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs.

setParams :: Params -> m ()

Updates parameters

currentSlot :: m Slot

Returns the current slot number

awaitSlot :: Slot -> m Slot

Waits until the current slot becomes greater or equal to the given slot, and returns the current slot after waiting.

Note that it might not wait for anything if the current slot is large enough.

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

The main abstraction of the blockchain.

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates, balances and validates a transaction from a skeleton. It returns the validated transaction and updates the state of the blockchain. In MockChainT, this means:

deletes the consumed outputs from mcstIndex

adds the produced outputs to msctIndex

deletes the consumed datums from mcstDatums

adds the produced datums to mcstDatums

adds the validators on outputs to the mcstValidators.

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTransControl.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the trouble of defining all the class methods yourself.

Constructors

AsTrans

Fields

getTrans :: t m a

type MonadModalBlockChain m = (MonadBlockChain m, MonadModal m, Modification m ~ UntypedTweak InterpMockChain)

A modal mock chain is a mock chain that allows us to use LTL modifications with Tweaks

class IsProp prop where

This module provides a common interface for HUnit and QuickCheck tests. We do so by abstracting uses of Assertion and Property for (IsProp prop) => prop, then provide instances for both HU.Asserton and QC.Property.

Minimal complete definition

testCounterexample, testConjoin, testDisjoin

Methods

testCounterexample :: String -> prop -> prop

Displays the string to the user in case of failure

testConjoin :: [prop] -> prop

Conjunction of a number of results

testDisjoin :: [prop] -> prop

Disjunction of a number of results

testFailure :: prop

Flags a failure

testSuccess :: prop

Flags a success

testFailureMsg :: String -> prop

Flags a failure with a message

newtype MockChainT m a

Constructors

MockChainT

Fields

unMockChain :: StateT MockChainSt (ExceptT MockChainError m) a

data MockChainSt

Slightly more concrete version of UtxoState, used to actually run the simulation.

Constructors

MockChainSt

Fields

mcstParams :: Params

mcstIndex :: UtxoIndex

mcstDatums :: Map DatumHash (TxSkelOutDatum, Integer)

mcstValidators :: Map ValidatorHash (Versioned Validator)

mcstCurrentSlot :: Slot

pattern MCLogSubmittedTxSkel :: SkelContext -> TxSkel -> MockChainLogEntry

pattern MCLogNewTx :: TxId -> MockChainLogEntry

pattern MCLogFail :: String -> MockChainLogEntry

forAll :: Show a => Gen a -> (a -> Property) -> Property

Here we provide our own universsal quantifier instead of forAll, so we can monomorphize it to returning a Property

(.&&.) :: IsProp prop => prop -> prop -> prop infixr 3

(.||.) :: IsProp prop => prop -> prop -> prop infixr 2

utxosAt :: MonadBlockChainBalancing m => Address -> m [(TxOutRef, TxOut)]

(.==.) :: (IsProp prop, Eq a) => a -> a -> prop infix 4

getIndex :: UtxoIndex -> Map TxOutRef TxOut

everywhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to every transaction in a given trace. This is also successful if there are no transactions at all.

somewhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to some transaction in the given Trace. The tweak must apply at least once.

balanceTxSkel :: MonadBlockChainBalancing m => TxSkel -> m (TxSkel, Fee, Collaterals, Wallet)

This is the main entry point of our balancing mechanism. This function takes a skeleton and returns a (possibly) balanced skeleton alongside the associated fee, collateral inputs and return collateral wallet. The options from the skeleton control whether it should be balanced, and how to compute its associated elements.

getMinAndMaxFee :: MonadBlockChainBalancing m => m (Fee, Fee)

This computes the minimum and maximum possible fee a transaction can cost based on the current protocol parameters

estimateTxSkelFee :: MonadBlockChainBalancing m => TxSkel -> Fee -> Collaterals -> Wallet -> m Fee

This function is essentially a copy of https://github.com/input-output-hk/plutus-apps/blob/d4255f05477fd8477ee9673e850ebb9ebb8c9657/plutus-ledger/src/Ledger/Fee.hs#L19

allUtxos :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

currentTime :: MonadBlockChainWithoutValidation m => m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the current slot

waitNSlots :: MonadBlockChainWithoutValidation m => Integer -> m Slot

Moves n slots fowards

txOutByRef :: MonadBlockChainBalancing m => TxOutRef -> m (Maybe TxOut)

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

txOutV2FromLedger :: TxOut -> TxOut

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

resolveDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out)))

Try to resolve the datum on the output: If there's an inline datum, take that; if there's a datum hash, look the corresponding datum up (with datumFromHash), returning Nothing if it can't be found; if there's no datum or hash at all, return Nothing.

resolveTypedDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m, FromData a) => out -> m (Maybe (ConcreteOutput (OwnerType out) a (ValueType out) (ReferenceScriptType out)))

Like resolveDatum, but also tries to use fromBuiltinData to extract a datum of the suitable type.

resolveValidator :: (IsAbstractOutput out, ToCredential (OwnerType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (Versioned Validator) (DatumType out) (ValueType out) (ReferenceScriptType out)))

Try to resolve the validator that owns an output: If the output is owned by a public key, or if the validator's hash is not known (i.e. if validatorFromHash returns Nothing) return Nothing.

resolveReferenceScript :: (IsAbstractOutput out, ToScriptHash (ReferenceScriptType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (Versioned Validator)))

Try to resolve the reference script on an output: If the output has no reference script, or if the reference script's hash is not known (i.e. if validatorFromHash returns Nothing), this function will return Nothing.

getEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Return the slot that contains the given time. See slotToTimeInterval for some satisfied equational properties.

awaitEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Waits until the current slot becomes greater or equal to the slot containing the given POSIX time. Note that that it might not wait for anything if the current slot is large enough.

slotRangeBefore :: MonadBlockChainWithoutValidation m => POSIXTime -> m SlotRange

The infinite range of slots ending before or at the given time

slotRangeAfter :: MonadBlockChainWithoutValidation m => POSIXTime -> m SlotRange

The infinite range of slots starting after or at the given time

slotToTimeInterval :: MonadBlockChainWithoutValidation m => Slot -> m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the slot with the given number. It holds that

slotToTimeInterval (getEnclosingSlot t) == (a, b)    ==>   a <= t <= b

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot a == n && getEnclosingSlot b == n

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot (a-1) == n-1 && getEnclosingSlot (b+1) == n+1

txSkelInputUtxosPl :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelInputUtxos :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelReferenceInputUtxosPl :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelReferenceInputUtxos :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelInputValidators :: MonadBlockChainBalancing m => TxSkel -> m (Map ValidatorHash (Versioned Validator))

All validators which protect transaction inputs

txSkelInputValue :: MonadBlockChainBalancing m => TxSkel -> m Value

look up the UTxOs the transaction consumes, and sum their values.

txSkelInputData :: MonadBlockChainBalancing m => TxSkel -> m (Map DatumHash Datum)

Look up the data on UTxOs the transaction consumes.

lookupUtxos :: MonadBlockChainBalancing m => [TxOutRef] -> m (Map TxOutRef TxOut)

lookupUtxosPl :: MonadBlockChainBalancing m => [TxOutRef] -> m (Map TxOutRef TxOut)

Go through all of the TxOutRefs in the list and look them up in the state of the blockchain, throwing an error if one of them cannot be resolved.

validateTxSkel' :: MonadBlockChain m => TxSkel -> m [TxOutRef]

Validates a skeleton, and retuns the ordered list of produced output references

txSkelProposalsDeposit :: MonadBlockChainBalancing m => TxSkel -> m Lovelace

Retrieves the total amount of lovelace deposited in proposals in this skeleton (equal to govActionDeposit times the number of proposals).

govActionDeposit :: MonadBlockChainBalancing m => m Lovelace

Retrieves the required deposit amount for issuing governance actions.

mcstToUtxoState :: MockChainSt -> UtxoState

toTxSkelOutWithMinAda :: Params -> TxSkelOut -> Either GenerateTxError TxSkelOut

This transforms an output into another output which necessarily contains at least the minimal required ada. If the previous quantity of ada was sufficient, it remains unchanged. This requires an iterative process, as adding ada into an output can potentially increase its size and thus make it require more minimal ada (although this remains to be witnessed in practice). This approach was inspired by https://github.com/input-output-hk/plutus-apps/blob/8706e6c7c525b4973a7b6d2ed7c9d0ef9cd4ef46/plutus-ledger/src/Ledger/Index.hs#L124

toTxSkelWithMinAda :: MonadBlockChainBalancing m => TxSkel -> m TxSkel

This transforms a skeleton by replacing all its TxSkelOut by their updated variants with their minimal amount of required ada. Any error raised in the transformation process is transformed into an MCEGenerationError

getTxSkelOutMinAda :: Params -> TxSkelOut -> Either GenerateTxError Integer

This provides the minimum amount of ada required in a given TxSkelOut. As we need to transform our output into a Cardano output to compute this value, this function can fail.

interpretAndRunWith :: (forall m. Monad m => MockChainT m a -> m res) -> StagedMockChain a -> [(res, MockChainLog)]

Interprets the staged mockchain then runs the resulting computation with a custom function. This can be used, for example, to supply a custom InitialDistribution by providing runMockChainTFrom.

interpretAndRun :: StagedMockChain a -> [(Either MockChainError (a, UtxoState), MockChainLog)]

runTweakFrom :: MockChainSt -> Tweak InterpMockChain a -> TxSkel -> [Either MockChainError (a, TxSkel)]

runTweak :: Tweak InterpMockChain a -> TxSkel -> [Either MockChainError (a, TxSkel)]

withTweak :: MonadModalBlockChain m => m x -> Tweak InterpMockChain a -> m x

Apply a Tweak to the next transaction in the given trace. The order of arguments is reversed compared to somewhere and everywhere, because that enables an idiom like

do ...
   endpoint arguments `withTweak` someModification
   ...

where endpoint builds and validates a single transaction depending on the given arguments. Then withTweak says "I want to modify the transaction returned by this endpoint in the following way".

there :: MonadModalBlockChain m => Integer -> Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to the (0-indexed) nth transaction in a given trace. Successful when this transaction exists and can be modified.

runUtxoSearch :: Monad m => UtxoSearch m a -> m [(TxOutRef, a)]

Given a UTxO search, we can run it to obtain a list of UTxOs.

allUtxosSearch :: MonadBlockChain m => UtxoSearch m TxOut

Search all currently known TxOutRefs together with their corresponding TxInfo-TxOut.

allUtxosLedgerSearch :: MonadBlockChain m => UtxoSearch m TxOut

Like allUtxosSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

utxosAtSearch :: (MonadBlockChainBalancing m, ToAddress addr) => addr -> UtxoSearch m TxOut

Search all TxOutRefs at a certain address, together with their TxInfo-TxOut.

utxosAtLedgerSearch :: (MonadBlockChainBalancing m, ToAddress addr) => addr -> UtxoSearch m TxOut

Like utxosAtSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

utxosFromCardanoTxSearch :: Monad m => CardanoTx -> UtxoSearch m TxOut

Search all TxOutRefs of a transaction, together with their TxInfo-TxOut.

txOutByRefSearch :: MonadBlockChainBalancing m => [TxOutRef] -> UtxoSearch m TxOut

Search all TxInfo-TxOuts corresponding to given the list of TxOutRefs. Any TxOutRef that doesn't correspond to a known output will be filtered out.

filterWith :: Monad m => UtxoSearch m a -> (a -> m (Maybe b)) -> UtxoSearch m b

Transform a UtxoSearch by applying a possibly failing monadic "lookup" on every output.

filterWithPure :: Monad m => UtxoSearch m a -> (a -> Maybe b) -> UtxoSearch m b

filterWithOptic :: (Is k An_AffineFold, Monad m) => UtxoSearch m a -> Optic' k is a b -> UtxoSearch m b

filterWithPred :: Monad m => UtxoSearch m a -> (a -> Bool) -> UtxoSearch m a

filterWithValuePred :: Monad m => UtxoSearch m TxOut -> (Value -> Bool) -> UtxoSearch m Value

filterWithOnlyAda :: Monad m => UtxoSearch m TxOut -> UtxoSearch m Value

filterWithNotOnlyAda :: Monad m => UtxoSearch m TxOut -> UtxoSearch m Value

onlyValueOutputsAtSearch :: (MonadBlockChainBalancing m, ToAddress addr) => addr -> UtxoSearch m (ConcreteOutput Credential () Value ScriptHash)

Search for UTxOs which only carry address and value information (no datum, staking credential, or reference script).

vanillaOutputsAtSearch :: (MonadBlockChainBalancing m, ToAddress addr) => addr -> UtxoSearch m (ConcreteOutput Credential () Ada ScriptHash)

A vanilla output only possesses an ada-only value and does not have a staking credential, a datum or a reference script. A vanilla UTxO is a perfect candidate to be used for fee, balancing or collateral.

filterWithAlways :: Monad m => UtxoSearch m a -> (a -> b) -> UtxoSearch m b

scriptOutputsSearch :: (MonadBlockChain m, ToScriptHash s) => s -> UtxoSearch m (ConcreteOutput s OutputDatum Value ScriptHash)

referenceScriptOutputsSearch :: (MonadBlockChain m, ToScriptHash s) => s -> UtxoSearch m (ConcreteOutput Credential OutputDatum Value ScriptHash)

holdsInState :: Address -> UtxoState -> Value

Total value accessible to what's pointed by the address.

utxoPayloadSet :: UtxoPayloadSet -> [UtxoPayload]

utxoPayloadTxOutRef :: UtxoPayload -> TxOutRef

utxoPayloadValue :: UtxoPayload -> Value

utxoPayloadSkelOutDatum :: UtxoPayload -> TxSkelOutDatum

utxoPayloadReferenceScript :: UtxoPayload -> Maybe ScriptHash

mcstToEmulatedLedgerState :: MockChainSt -> EmulatedLedgerState

Generating an emulated state for the emulator from a mockchain state and some parameters, based on a standard initial state

combineMockChainT :: Monad m => (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

mapMockChainT :: (m (Either MockChainError (a, MockChainSt)) -> n (Either MockChainError (b, MockChainSt))) -> MockChainT m a -> MockChainT n b

runMockChainTRaw :: Monad m => MockChainSt -> MockChainT m a -> m (Either MockChainError (a, MockChainSt))

Executes a MockChainT from some initial state; does not convert the MockChainSt into a UtxoState.

runMockChainTFrom :: Monad m => InitialDistribution -> MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from an initial state set up with the given initial value distribution. Similar to runMockChainT, uses the default environment. Returns a UtxoState instead of a MockChainSt. If you need the later, use runMockChainTRaw

mockChainSt0From :: InitialDistribution -> MockChainSt

runMockChainT :: Monad m => MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from the canonical initial state and environment. The canonical environment uses the default SlotConfig and Cooked.Wallet.wallet 1 as the sole wallet signing transactions.

runMockChainRaw :: MockChain a -> Either MockChainError (a, MockChainSt)

See runMockChainTRaw

runMockChainFrom :: InitialDistribution -> MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainTFrom

runMockChain :: MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainT

utxoState0 :: UtxoState

mockChainSt0 :: MockChainSt

utxoIndex0 :: UtxoIndex

utxoIndex0From :: InitialDistribution -> UtxoIndex

This creates the initial UtxoIndex from an initial distribution by submitting an initial transaction with the appropriate content:

inputs consist of a single dummy pseudo input

all non-ada assets in outputs are considered minted

outputs are translated from the TxSkelOut list in the initial distribution

Two things to note:

We don't know what Magic means for the network ID (TODO)

The genesis key hash has been taken from https://github.com/input-output-hk/cardano-node/blob/543b267d75d3d448e1940f9ec04b42bd01bbb16b/cardano-api/test/Test/Cardano/Api/Genesis.hs#L60

datumMap0From :: InitialDistribution -> Map DatumHash (TxSkelOutDatum, Integer)

Datums from initial distributions should be accounted for in the MockChainSt which is done using this function.

referenceScriptMap0From :: InitialDistribution -> Map ValidatorHash (Versioned Validator)

Reference scripts from initial distributions should be accounted for in the MockChainSt which is done using this function.

unMockChainLog :: MockChainLog -> [MockChainLogEntry]

testBool :: IsProp prop => Bool -> prop

testAll :: IsProp prop => (a -> prop) -> [a] -> prop

testSucceeds :: IsProp prop => PrettyCookedOpts -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed, starting from the default initial distribution

testSucceedsFrom :: IsProp prop => PrettyCookedOpts -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed starting from some initial distribution but doesn't impose any additional condition on success. Use testSucceedsFrom' for that.

testFails :: (IsProp prop, Show a) => PrettyCookedOpts -> (MockChainError -> prop) -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail and that a predicate holds over the error.

To test that validation fails, use > testFails def (isCekEvaluationFailure def) e

testFailsFrom :: (IsProp prop, Show a) => PrettyCookedOpts -> (MockChainError -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from some initial distribution.

testSucceedsFrom' :: IsProp prop => PrettyCookedOpts -> (a -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed starting from some initial distribution. Additionally impose a condition over the resulting state and value.

testAllSatisfiesFrom :: forall prop a. IsProp prop => PrettyCookedOpts -> (Either MockChainError (a, UtxoState) -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the set of traces encoded by the StagedMockChain satisfy the given predicate. If you wish to build custom predicates you can use testSatisfiesFrom' directly and see testBinaryRelatedBy as an example.

isCekEvaluationFailure :: IsProp prop => PrettyCookedOpts -> MockChainError -> prop

Is satisfied when the given MockChainError is wrapping a CekEvaluationFailure. This is particularly important when writing negative tests. For example, if we are simulating an attack and writing a test with testFailsFrom, we might have made a mistake in the attack, yielding a test that fails for reasons such as ValueLessThanMinAda or ValueNotPreserved, which does not rule out the attack being caught by the validator script. For these scenarios it is paramount to rely on testFailsFrom' isCekEvaluationFailure instead.

isCekEvaluationFailureWithMsg :: IsProp prop => PrettyCookedOpts -> (String -> Bool) -> MockChainError -> prop

Similar to isCekEvaluationFailure, but enables us to check for a specific error message in the error.

testSatisfiesFrom' :: ([(Either MockChainError (a, UtxoState), MockChainLog)] -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Asserts that the results produced by running the given StagedMockChain from some speficied InitialDistribution satisfy a given assertion. In this case, the predicate gets the trace descriptions that led to each potential outcome and is responsible for calling testCounterexample communicate these to the user.

Although this function is mainly used internally, as a building block for the simpler predicates, it can be useful in building some custom predicates. Check testAllSatisfiesFrom or testBinaryRelatedBy for examples on using this.

testBinaryRelatedBy :: IsProp prop => PrettyCookedOpts -> (UtxoState -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Asserts that the given StagedMockChain produces exactly two outcomes, both of which are successful and have their resulting states related by a given predicate. A typical usage would look like:

testBinaryRelatedBy equalModuloAda myInitDistr $ do
  x <- trPrepare
  execOption1 x <|> execOption2 x

testOneEquivClass :: IsProp prop => PrettyCookedOpts -> (UtxoState -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Generalizes testBinaryRelatedBy, asserting that the given StagedMockChain produces more than two outcomes, say [x,y,z,w], all of which are successful (i.e. are not a MockChainError) and these states are in the same equivalence class of (~); that is, they satisfy:

x ~ y && x ~ z && x ~ z && x ~ w

Because (~) should be symmetric and transitive we can estabilish that these states all belong to the same equivalence class. This function does not check each pointwise case.

assertionToMaybe :: Assertion -> IO (Maybe HUnitFailure)

assertSubset :: (Show a, Eq a) => [a] -> [a] -> Assertion

assertSameSets :: (Show a, Eq a) => [a] -> [a] -> Assertion