Cooked.MockChain

Description

Exports all the machinery necessary for using MonadBlockChain and MonadMockChain for writing and testing contracts. You might actually want to import just Cooked which re-exports this module and a few others too

Documentation

module Cooked.MockChain.Constraints

module Cooked.MockChain.Time

module Cooked.MockChain.UtxoPredicate

module Cooked.MockChain.UtxoState

module Cooked.MockChain.UtxoState.Testing

module Cooked.MockChain.Wallet

module Cooked.MockChain.Monad.Staged

module Cooked.MockChain.Monad.Direct

module Cooked.MockChain.Monad

module Cooked.MockChain.RawUPLC

module Cooked.MockChain.Testing

type SpendableOut = (TxOutRef, ChainIndexTxOut)

A SpendableOut is an outref that is ready to be spend; with its underlying ChainIndexTxOut.