Cooked.MockChain.Balancing

Description

This module handles auto-balancing of transaction skeleton. This includes computation of fees and collaterals because their computation cannot be separated from the balancing.

Documentation

balanceTxSkel :: Members '[MockChainRead, MockChainLog, Error MockChainError, Error ToCardanoError, Fail] effs => TxSkel -> Sem effs (TxSkel, Fee, Collaterals)

This is the main entry point of our balancing mechanism. This function takes a skeleton and returns a (possibly) balanced skeleton alongside the associated fee, collateral inputs and return collateral user, which might be empty when no script is involved in the transaction. The options from the skeleton control whether it should be balanced, and how to compute its associated elements.

getMinAndMaxFee :: Member MockChainRead effs => Fee -> Sem effs (Fee, Fee)

This computes the minimum and maximum possible fee a transaction can cost based on the current protocol parameters and its number of scripts. In the Dijsktra era, this will be modified with new protocol parameters. See https://github.com/IntersectMBO/cardano-ledger/blob/master/docs/adr/2024-08-14_009-refscripts-fee-change.md for more information

estimateTxSkelFee :: Members '[MockChainRead, Error MockChainError, Error ToCardanoError, Fail] effs => TxSkel -> Fee -> Collaterals -> Sem effs Fee

This function was originally inspired by https://github.com/input-output-hk/plutus-apps/blob/d4255f05477fd8477ee9673e850ebb9ebb8c9657/plutus-ledger/src/Ledger/Fee.hs#L19