Cooked.MockChain.Balancing

Description

This module handles auto-balancing of transaction skeleton. This includes computation of fees and collaterals as well as ensuring outputs have the required minimum ada. Most of the balancing process can be customized through TxSkelOpts

Documentation

balanceTxSkel :: MonadBlockChainBalancing m => TxSkel -> m (TxSkel, Fee, Set TxOutRef)