Cooked.MockChain.Balancing

Documentation

balancedTxSkel :: MonadBlockChainBalancing m => TxSkel -> m (TxSkel, Fee, Set TxOutRef)

balancedTx :: MonadBlockChainBalancing m => (TxSkel, Fee, Set TxOutRef) -> m (Tx BabbageEra)

Take the output of balancedTxSkel and turn it into an actual Cardano transaction.

ensureTxSkelOutsMinAda :: MonadBlockChainBalancing m => TxSkel -> m TxSkel

Ensure that the transaction outputs have the necessary minimum amount of Ada on them. This will only be applied if the txOptEnsureMinAda is set to True.

txSkelInputUtxosPV2 :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelInputUtxos :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelReferenceInputUtxosPV2 :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelReferenceInputUtxos :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelInputValidators :: MonadBlockChainBalancing m => TxSkel -> m (Map ValidatorHash (Versioned Validator))

All validators which protect transaction inputs

lookupUtxosPV2 :: MonadBlockChainBalancing m => [TxOutRef] -> m (Map TxOutRef TxOut)

lookupUtxos :: MonadBlockChainBalancing m => [TxOutRef] -> m (Map TxOutRef TxOut)

txSkelInputValue :: MonadBlockChainBalancing m => TxSkel -> m Value

look up the UTxOs the transaction consumes, and sum the value contained in them.

txSkelInputData :: MonadBlockChainBalancing m => TxSkel -> m (Map DatumHash Datum)

Look up the data on UTxOs the transaction consumes.

setFeeAndBalance :: MonadBlockChainBalancing m => Wallet -> TxSkel -> m (TxSkel, Fee)

estimateTxSkelFee :: Params -> UTxO EmulatorEra -> Map DatumHash Datum -> Map TxOutRef TxOut -> Map ValidatorHash (Versioned Validator) -> TxSkel -> Fee -> Either MockChainError Fee

This funcion is essentially a copy of https://github.com/input-output-hk/plutus-apps/blob/d4255f05477fd8477ee9673e850ebb9ebb8c9657/plutus-ledger/src/Ledger/Fee.hs#L19

calcCollateral :: MonadBlockChainBalancing m => Wallet -> m (Set TxOutRef)

Calculates the collateral for a transaction

balanceTxFromAux :: MonadBlockChainBalancing m => BalanceStage -> Wallet -> TxSkel -> Fee -> m TxSkel

data BalanceTxRes

Constructors

BalanceTxRes

Fields

newInputs :: [(TxOutRef, TxOut)]

Inputs that need to be added in order to cover the value in the transaction outputs

returnValue :: Value

The newInputs will add _at least_ the missing value to cover the outputs, this is the difference of the input value together with the newInputs and the output value. This value must be nonnegative in every asset class.

availableUtxos :: [(TxOutRef, TxOut)]

Some additional UTxOs that could be used as extra inputs. These all belong to the same wallet that was passed to calcBalanceTx as an argument, and are sorted in decreasing order of their Ada value.

calcBalanceTx :: MonadBlockChainBalancing m => BalanceStage -> Wallet -> TxSkel -> Fee -> m BalanceTxRes

Calculate the changes needed to balance a transaction with money from a given wallet. Every transaction that is sent to the chain must be balanced, that is: inputs + mints == outputs + fee + burns.

applyBalanceTx :: PubKeyHash -> BalanceTxRes -> TxSkel -> Maybe TxSkel

Once we calculated what is needed to balance a transaction skel, we still need to apply those changes to skel. Because of the minAdaTxOut constraint, this might not be possible: imagine the leftover is less than minAdaTxOut, but the transaction has no output addressed to the balancing wallet. If we just create a new ouput for the balancing wallet and place the leftover there, the resulting transaction will fail to validate with LessThanMinAdaPerUTxO error. Instead, we need to consume yet another UTxO belonging to the wallet to then create the output with the proper leftover. If the wallet has no UTxO, then there's no way to balance this transaction.