Cooked.MockChain.Balancing

Description

This module handles auto-balancing of transaction skeleton. This includes computation of fees and collaterals because their computation cannot be separated from the balancing.

Documentation

balanceTxSkel :: MonadBlockChainBalancing m => TxSkel -> m (TxSkel, Fee, Collaterals)

This is the main entry point of our balancing mechanism. This function takes a skeleton and returns a (possibly) balanced skeleton alongside the associated fee, collateral inputs and return collateral user, which might be empty when no script is involved in the transaction. The options from the skeleton control whether it should be balanced, and how to compute its associated elements.

getMinAndMaxFee :: MonadBlockChainBalancing m => Fee -> m (Fee, Fee)

This computes the minimum and maximum possible fee a transaction can cost based on the current protocol parameters and its number of scripts.

estimateTxSkelFee :: MonadBlockChainBalancing m => TxSkel -> Fee -> Collaterals -> m Fee

This function was originally inspired by https://github.com/input-output-hk/plutus-apps/blob/d4255f05477fd8477ee9673e850ebb9ebb8c9657/plutus-ledger/src/Ledger/Fee.hs#L19