Cooked.MockChain.BlockChain

Documentation

data MockChainError where

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError :: ValidationErrorInPhase -> MockChainError

MCEUnbalanceable :: MCEUnbalanceableError -> TxSkel -> MockChainError

MCENoSuitableCollateral :: MockChainError
Thrown when the balancing wallet owns no output that is pure Ada and with no datum.

MCEGenerationError :: GenerateTxError -> MockChainError

MCECalcFee :: MockChainError -> MockChainError
Errors happening at fee calculation time.

MCEUnknownOutRefError :: String -> TxOutRef -> MockChainError
Thrown when an output reference should be in the state of the mockchain, but isn't.

MCEUnknownValidator :: String -> ValidatorHash -> MockChainError
Same as MCEUnknownOutRefError for validators.

MCEUnknownDatum :: String -> DatumHash -> MockChainError
Same as MCEUnknownOutRefError for datums.

FailWith :: String -> MockChainError
Used to provide MonadFail instances.

OtherMockChainError :: (Show err, Eq err) => err -> MockChainError

data MCEUnbalanceableError

Constructors

MCEUnbalNotEnoughFunds Wallet Value
The balancing wallet misses some value to pay what is needed to balance the transaction.

MCEUnbalNotEnoughReturning (Value, [TxOutRef]) (Value, [TxOutRef]) Value
There is value to return to the balancing wallet but not enough to fullfill the min ada requirement and there is not enough in additional inputs to make it possible.

class (MonadFail m, MonadError MockChainError m) => MonadBlockChainBalancing m where

Contains methods needed for balancing.

Methods

getParams :: m Params

Returns the parameters of the chain.

utxosAtLedger :: Address -> m [(TxOutRef, TxOut)]

Returns a list of all UTxOs at a certain address.

datumFromHash :: DatumHash -> m (Maybe Datum)

Returns the datum with the given hash or Nothing if there is none

validatorFromHash :: ValidatorHash -> m (Maybe (Versioned Validator))

Returns the full validator corresponding to hash, if that validator owns something or if it is stored in the reference script field of some UTxO.

txOutByRefLedger :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

class MonadBlockChainBalancing m => MonadBlockChainWithoutValidation m where

Methods

allUtxosLedger :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs.

currentSlot :: m Slot

Returns the current slot number

awaitSlot :: Slot -> m Slot

Waits until the current slot becomes greater or equal to the given slot, and returns the current slot after waiting.

Note that it might not wait for anything if the current slot is large enough.

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

The main abstraction of the blockchain.

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates, balances and validates a transaction from a skeleton. It returns the validated transaction and updates the state of the blockchain. In MockChainT, this means: - deletes the consumed outputs from mcstIndex - adds the produced outputs to msctIndex - deletes the consumed datums from mcstDatums - adds the produced datums to mcstDatums - adds the validators on outputs to the mcstValidators.

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTransControl.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the boilerplate of defining all the methods of the class yourself.

Constructors

AsTrans

Fields

getTrans :: t m a

allUtxos :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

currentTime :: MonadBlockChainWithoutValidation m => m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the current slot

waitNSlots :: MonadBlockChainWithoutValidation m => Integer -> m Slot

Moves n slots fowards

utxosAt :: MonadBlockChainBalancing m => Address -> m [(TxOutRef, TxOut)]

txOutByRef :: MonadBlockChainBalancing m => TxOutRef -> m (Maybe TxOut)

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

txOutV2FromLedger :: TxOut -> TxOut

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

resolveDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out)))

Try to resolve the datum on the output: If there's an inline datum, take that; if there's a datum hash, look the corresponding datum up (with datumFromHash), returning Nothing if it can't be found; if there's no datum or hash at all, return Nothing.

resolveTypedDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m, FromData a) => out -> m (Maybe (ConcreteOutput (OwnerType out) a (ValueType out) (ReferenceScriptType out)))

Like resolveDatum, but also tries to use fromBuiltinData to extract a datum of the suitable type.

resolveValidator :: (IsAbstractOutput out, ToCredential (OwnerType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (Versioned Validator) (DatumType out) (ValueType out) (ReferenceScriptType out)))

Try to resolve the validator that owns an output: If the output is owned by a public key, or if the validator's hash is not known (i.e. if validatorFromHash returns Nothing) return Nothing.

resolveReferenceScript :: (IsAbstractOutput out, ToScriptHash (ReferenceScriptType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (Versioned Validator)))

Try to resolve the reference script on an output: If the output has no reference script, or if the reference script's hash is not known (i.e. if validatorFromHash returns Nothing), this function will return Nothing.

getEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Return the slot that contains the given time. See slotToTimeInterval for some equational properties this function satisfies.

awaitEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Waits until the current slot becomes greater or equal to the slot containing the given POSIX time. Note that that it might not wait for anything if the current slot is large enough.

slotRangeBefore :: MonadBlockChainWithoutValidation m => POSIXTime -> m SlotRange

The infinite range of slots ending before or at the given POSIX time

slotRangeAfter :: MonadBlockChainWithoutValidation m => POSIXTime -> m SlotRange

The infinite range of slots starting after or at the given POSIX time

slotToTimeInterval :: MonadBlockChainWithoutValidation m => Slot -> m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the slot with the given number. It holds that

slotToTimeInterval (getEnclosingSlot t) == (a, b)    ==>   a <= t <= b

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot a == n && getEnclosingSlot b == n

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot (a-1) == n-1 && getEnclosingSlot (b+1) == n+1