Cooked.MockChain.BlockChain

BlockChain Monad

data MockChainError where

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError :: ValidationErrorInPhase -> MockChainError

MCEUnbalanceable :: String -> BalanceStage -> TxSkel -> MockChainError

MCENoSuitableCollateral :: MockChainError

MCEGenerationError :: GenerateTxError -> MockChainError

MCECalcFee :: MockChainError -> MockChainError

MCEUnknownOutRefError :: String -> TxOutRef -> MockChainError

MCEUnknownValidator :: String -> ValidatorHash -> MockChainError

MCEUnknownDatum :: String -> DatumHash -> MockChainError

FailWith :: String -> MockChainError

OtherMockChainError :: (Show err, Eq err) => err -> MockChainError

data BalanceStage

Describes us which stage of the balancing process are we at. This is needed to distinguish the successive calls to balancing while computing fees from the final call to balancing

Constructors

BalCalcFee

BalFinalizing

class (MonadFail m, MonadError MockChainError m) => MonadBlockChainBalancing m where

Methods

getParams :: m Params

Returns the paramters of the chain.

utxosAtLedger :: Address -> m [(TxOutRef, TxOut)]

Return a list of all UTxOs at a certain address.

datumFromHash :: DatumHash -> m (Maybe Datum)

Returns the datum with the given hash or Nothing if there is none

validatorFromHash :: ValidatorHash -> m (Maybe (Versioned Validator))

Returns the full validator corresponding to hash, if that validator is currently the owner of some UTxO. (If it is not, there's no guarantee that this will return Just.)

txOutByRefLedger :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

class MonadBlockChainBalancing m => MonadBlockChainWithoutValidation m where

Methods

allUtxosLedger :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs.

currentSlot :: m Slot

Returns the current slot number

currentTime :: m POSIXTime

Returns the current time

awaitSlot :: Slot -> m Slot

Either waits until the given slot or returns the current slot. Note that that it might not wait for anything if the current slot is larger than the argument.

awaitTime :: POSIXTime -> m POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

Example: if starting time is 0 and slot length is 3s, then `awaitTime 4` waits until slot 2 and returns the value `POSIXTime 5`.

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates and balances a transaction from a skeleton, then attemps to validate such transaction. A balanced transaction is such that inputs + mints == outputs + fees. To balance a transaction, we need access to the current UTxO state to choose which inputs to add in case the output-side of the balancing equation is bigger.

The TxSkel receives a TxOpts record with a number of options to customize how validation works.

allUtxos :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

utxosAt :: MonadBlockChainBalancing m => Address -> m [(TxOutRef, TxOut)]

txOutByRef :: MonadBlockChainBalancing m => TxOutRef -> m (Maybe TxOut)

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

txOutV2FromLedger :: TxOut -> TxOut

filterUtxos :: (o1 -> Maybe o2) -> [(TxOutRef, o1)] -> [(TxOutRef, o2)]

Helper function to filter the output of allUtxos and utxosFromCardanoTx

allUtxosWithDatums :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

Like allUtxos, but on every OutputDatumHash, try to resolve the complete datum from the state

resolveDatums :: MonadBlockChainBalancing m => [(TxOutRef, TxOut)] -> m [(TxOutRef, TxOut)]

filteredUtxos :: MonadBlockChainWithoutValidation m => (TxOut -> Maybe output) -> m [(TxOutRef, output)]

filteredUtxosWithDatums :: MonadBlockChainWithoutValidation m => (TxOut -> Maybe output) -> m [(TxOutRef, output)]

Like filteredUtxos, but will all resolvable datum hashes resolved. This means that the outputs are presented differently from how a script would see them; the information on whether there are inline datums or datum hashes is lost.

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

Deriving further MonadBlockChain instances

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTransControl.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the boilerplate of defining all the methods of the class yourself.

Constructors

AsTrans

Fields

getTrans :: t m a