Cooked.MockChain.BlockChain

Documentation

data MockChainError where

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError :: ValidationErrorInPhase -> MockChainError

MCEUnbalanceable :: MCEUnbalanceableError -> TxSkel -> MockChainError

MCENoSuitableCollateral :: MockChainError

MCEGenerationError :: GenerateTxError -> MockChainError

MCECalcFee :: MockChainError -> MockChainError

MCEUnknownOutRefError :: String -> TxOutRef -> MockChainError

MCEUnknownValidator :: String -> ValidatorHash -> MockChainError

MCEUnknownDatum :: String -> DatumHash -> MockChainError

FailWith :: String -> MockChainError

OtherMockChainError :: (Show err, Eq err) => err -> MockChainError

data MCEUnbalanceableError

Constructors

MCEUnbalNotEnoughFunds Wallet Value
The balancing wallet misses some value to pay what is needed to balance the transaction.

MCEUnbalNotEnoughReturning (Value, [TxOutRef]) (Value, [TxOutRef]) Value
There is value to return to the balancing wallet but not enough to fullfill the min ada requirement and there is not enough in additional inputs to make it possible.

class (MonadFail m, MonadError MockChainError m) => MonadBlockChainBalancing m where

Methods

getParams :: m Params

Returns the paramters of the chain.

utxosAtLedger :: Address -> m [(TxOutRef, TxOut)]

Return a list of all UTxOs at a certain address.

datumFromHash :: DatumHash -> m (Maybe Datum)

Returns the datum with the given hash or Nothing if there is none

validatorFromHash :: ValidatorHash -> m (Maybe (Versioned Validator))

Returns the full validator corresponding to hash, if that validator is currently the owner of some UTxO. (If it is not, there's no guarantee that this will return Just.)

txOutByRefLedger :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

class MonadBlockChainBalancing m => MonadBlockChainWithoutValidation m where

Methods

allUtxosLedger :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs.

currentSlot :: m Slot

Returns the current slot number

currentTime :: m POSIXTime

Returns the current time

awaitSlot :: Slot -> m Slot

Either waits until the given slot or returns the current slot. Note that that it might not wait for anything if the current slot is larger than the argument.

awaitTime :: POSIXTime -> m POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

Example: if starting time is 0 and slot length is 3s, then `awaitTime 4` waits until slot 2 and returns the value `POSIXTime 5`.

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates and balances a transaction from a skeleton, then attemps to validate such transaction. A balanced transaction is such that inputs + mints == outputs + fees. To balance a transaction, we need access to the current UTxO state to choose which inputs to add in case the output-side of the balancing equation is bigger.

The TxSkel receives a TxOpts record with a number of options to customize how validation works.

allUtxos :: MonadBlockChainWithoutValidation m => m [(TxOutRef, TxOut)]

utxosAt :: MonadBlockChainBalancing m => Address -> m [(TxOutRef, TxOut)]

txOutByRef :: MonadBlockChainBalancing m => TxOutRef -> m (Maybe TxOut)

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

txOutV2FromLedger :: TxOut -> TxOut

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

resolveDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out)))

Try to resolve the datum on the output: If there's an inline datum, take that; if there's a datum hash, look the corresponding datum up (with datumFromHash), returning Nothing if it can't be found; if there's no datum or hash at all, return Nothing.

resolveValidator :: (IsAbstractOutput out, ToCredential (OwnerType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (Versioned Validator) (DatumType out) (ValueType out) (ReferenceScriptType out)))

Try to resolve the validator that owns an output: If the output is owned by a public key, or if the validator's hash is not known (i.e. if validatorFromHash returns Nothing) return Nothing.

resolveReferenceScript :: (IsAbstractOutput out, ToScriptHash (ReferenceScriptType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (Versioned Validator)))

Try to resolve the reference script on an output: If the output has no reference script, or if the reference script's hash is not known (i.e. if validatorFromHash returns Nothing), this function will return Nothing.