Cooked.MockChain.BlockChain

Description

This modules provides a specification for our blockchain monads, in three layers:

MonadBlockChainBalancing provides what's needing for balancing purposes

MonadBlockChainWithoutValidation adds up remaining primitives without transaction validation

MonadBlockChain concludes with the addition of transaction validation, thus modifying the internal index of outputs

In addition, you will find here many helpers functions which can be derived from the core definition of our blockchain.

Documentation

data MockChainError

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError ValidationPhase ValidationError
Validation errors, either in Phase 1 or Phase 2

MCEUnbalanceable Wallet Value TxSkel
Thrown when the balancing wallet does not have enough funds

MCENoSuitableCollateral Integer Integer Value
Thrown when not enough collateral are provided. Built upon the fee, the percentage and the expected minimal collateral value.

MCEGenerationError GenerateTxError
Thrown when an error occured during transaction generation

MCEUnknownOutRefError String TxOutRef
Thrown when an output reference is missing from the mockchain state

MCEUnknownValidator String ValidatorHash
Same as MCEUnknownOutRefError for validators.

MCEUnknownDatum String DatumHash
Same as MCEUnknownOutRefError for datums.

FailWith String
Used to provide MonadFail instances.

data MockChainLogEntry

This represents the specific events that should be logged when processing transactions. If a new kind of event arises, then a new constructor should be provided here.

Constructors

MCLogSubmittedTxSkel SkelContext TxSkel
Logging a Skeleton as it is submitted by the user.

MCLogAdjustedTxSkel SkelContext TxSkel Integer (Maybe (Set TxOutRef, Wallet))
Logging a Skeleton as it has been adjusted by the balancing mechanism, alongside fee, and possible collateral utxos and return collateral wallet.

MCLogNewTx TxId
Logging the appearance of a new transaction, after a skeleton has been successfully sent for validation.

MCLogDiscardedUtxos Integer String
Logging the fact that utxos provided by the user for balancing have to be discarded for a specific reason.

MCLogUnusedCollaterals (Either Wallet (Set TxOutRef))
Logging the fact that utxos provided as collaterals will not be used because the transaction does not involve scripts. There are 2 cases, depending on whether the user has provided an explicit wallet or a set of utxos to be used as collaterals.

class (MonadFail m, MonadError MockChainError m) => MonadBlockChainBalancing m where

Contains methods needed for balancing.

Methods

getParams :: m Params

Returns the emulator parameters, including protocol parameters

utxosAt :: Address -> m [(TxOutRef, TxOut)]

Returns a list of all UTxOs at a certain address.

datumFromHash :: DatumHash -> m (Maybe Datum)

Returns the datum with the given hash if present.

validatorFromHash :: ValidatorHash -> m (Maybe (Versioned Validator))

Returns the full validator corresponding to hash, if that validator owns something or if it is stored in the reference script field of some UTxO.

txOutByRef :: TxOutRef -> m (Maybe TxOut)

Returns an output given a reference to it

logEvent :: MockChainLogEntry -> m ()

Logs an event that occured during a BlockChain run

class MonadBlockChainBalancing m => MonadBlockChainWithoutValidation m where

Methods

allUtxos :: m [(TxOutRef, TxOut)]

Returns a list of all currently known outputs.

setParams :: Params -> m ()

Updates parameters

currentSlot :: m Slot

Returns the current slot number

awaitSlot :: Slot -> m Slot

Waits until the current slot becomes greater or equal to the given slot, and returns the current slot after waiting.

Note that it might not wait for anything if the current slot is large enough.

class MonadBlockChainWithoutValidation m => MonadBlockChain m where

The main abstraction of the blockchain.

Methods

validateTxSkel :: TxSkel -> m CardanoTx

Generates, balances and validates a transaction from a skeleton. It returns the validated transaction and updates the state of the blockchain. In MockChainT, this means:

deletes the consumed outputs from mcstIndex

adds the produced outputs to msctIndex

deletes the consumed datums from mcstDatums

adds the produced datums to mcstDatums

adds the validators on outputs to the mcstValidators.

newtype AsTrans t (m :: Type -> Type) a

A newtype wrapper to be used with '-XDerivingVia' to derive instances of MonadBlockChain for any MonadTransControl.

For example, to derive 'MonadBlockChain m => MonadBlockChain (ReaderT r m)', you'd write

deriving via (AsTrans (ReaderT r) m) instance MonadBlockChain m => MonadBlockChain (ReaderT r m)

and avoid the trouble of defining all the class methods yourself.

Constructors

AsTrans

Fields

getTrans :: t m a

currentTime :: MonadBlockChainWithoutValidation m => m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the current slot

waitNSlots :: MonadBlockChainWithoutValidation m => Integer -> m Slot

Moves n slots fowards

utxosFromCardanoTx :: CardanoTx -> [(TxOutRef, TxOut)]

Retrieve the ordered list of outputs of the given CardanoTx.

This is useful when writing endpoints and/or traces to fetch utxos of interest right from the start and avoid querying the chain for them afterwards using allUtxos or similar functions.

typedDatumFromTxOutRef :: (FromData a, MonadBlockChainWithoutValidation m) => TxOutRef -> m (Maybe a)

valueFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Value)

outputDatumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe OutputDatum)

datumFromTxOutRef :: MonadBlockChainWithoutValidation m => TxOutRef -> m (Maybe Datum)

resolveDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out)))

Try to resolve the datum on the output: If there's an inline datum, take that; if there's a datum hash, look the corresponding datum up (with datumFromHash), returning Nothing if it can't be found; if there's no datum or hash at all, return Nothing.

resolveTypedDatum :: (IsAbstractOutput out, ToOutputDatum (DatumType out), MonadBlockChainBalancing m, FromData a) => out -> m (Maybe (ConcreteOutput (OwnerType out) a (ValueType out) (ReferenceScriptType out)))

Like resolveDatum, but also tries to use fromBuiltinData to extract a datum of the suitable type.

resolveValidator :: (IsAbstractOutput out, ToCredential (OwnerType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (Versioned Validator) (DatumType out) (ValueType out) (ReferenceScriptType out)))

Try to resolve the validator that owns an output: If the output is owned by a public key, or if the validator's hash is not known (i.e. if validatorFromHash returns Nothing) return Nothing.

resolveReferenceScript :: (IsAbstractOutput out, ToScriptHash (ReferenceScriptType out), MonadBlockChainBalancing m) => out -> m (Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (Versioned Validator)))

Try to resolve the reference script on an output: If the output has no reference script, or if the reference script's hash is not known (i.e. if validatorFromHash returns Nothing), this function will return Nothing.

getEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Return the slot that contains the given time. See slotToTimeInterval for some satisfied equational properties.

awaitEnclosingSlot :: MonadBlockChainWithoutValidation m => POSIXTime -> m Slot

Waits until the current slot becomes greater or equal to the slot containing the given POSIX time. Note that that it might not wait for anything if the current slot is large enough.

awaitDurationFromLowerBound :: MonadBlockChainWithoutValidation m => Integer -> m Slot

Wait a given number of ms from the lower bound of the current slot and returns the current slot after waiting.

awaitDurationFromUpperBound :: MonadBlockChainWithoutValidation m => Integer -> m Slot

Wait a given number of ms from the upper bound of the current slot and returns the current slot after waiting.

slotRangeBefore :: MonadBlockChainWithoutValidation m => POSIXTime -> m SlotRange

The infinite range of slots ending before or at the given time

slotRangeAfter :: MonadBlockChainWithoutValidation m => POSIXTime -> m SlotRange

The infinite range of slots starting after or at the given time

slotToTimeInterval :: MonadBlockChainWithoutValidation m => Slot -> m (POSIXTime, POSIXTime)

Returns the closed ms interval corresponding to the slot with the given number. It holds that

slotToTimeInterval (getEnclosingSlot t) == (a, b)    ==>   a <= t <= b

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot a == n && getEnclosingSlot b == n

and

slotToTimeInterval n == (a, b)   ==>   getEnclosingSlot (a-1) == n-1 && getEnclosingSlot (b+1) == n+1

txSkelInputUtxos :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelReferenceInputUtxos :: MonadBlockChainBalancing m => TxSkel -> m (Map TxOutRef TxOut)

txSkelInputValidators :: MonadBlockChainBalancing m => TxSkel -> m (Map ValidatorHash (Versioned Validator))

All validators which protect transaction inputs

txSkelInputValue :: MonadBlockChainBalancing m => TxSkel -> m Value

look up the UTxOs the transaction consumes, and sum their values.

txSkelHashedData :: MonadBlockChainBalancing m => TxSkel -> m (Map DatumHash Datum)

Looks up and resolves the hashed datums on UTxOs the transaction consumes or references, which will be needed by the transaction body.

txSkelInputDataAsHashes :: MonadBlockChainBalancing m => TxSkel -> m [DatumHash]

Looks up the data on UTxOs the transaction consumes and returns their hashes. This corresponds to the keys of what should be removed from the stored datums in our mockchain. There can be duplicates, which is expected.

lookupUtxos :: MonadBlockChainBalancing m => [TxOutRef] -> m (Map TxOutRef TxOut)

Go through all of the TxOutRefs in the list and look them up in the state of the blockchain, throwing an error if one of them cannot be resolved.

validateTxSkel' :: MonadBlockChain m => TxSkel -> m [TxOutRef]

Validates a skeleton, and retuns the ordered list of produced output references

validateTxSkel_ :: MonadBlockChain m => TxSkel -> m ()

Validates a skeleton, and erases the outputs

txSkelProposalsDeposit :: MonadBlockChainBalancing m => TxSkel -> m Lovelace

Retrieves the total amount of lovelace deposited in proposals in this skeleton (equal to govActionDeposit times the number of proposals).

govActionDeposit :: MonadBlockChainBalancing m => m Lovelace

Retrieves the required deposit amount for issuing governance actions.