Cooked.MockChain.Direct

Direct Emulation

The MockChainT monad provides a direct emulator; that is, it gives us a simple way to call validator scripts directly, without the need for all the complexity the Contract monad introduces.

Running a MockChain produces a UtxoState, which is a map from Address to (Pl.Value, Maybe Pl.Datum), and corresponds to the utxo mental model most people have. Internally, however, we keep a UtxoIndex in our state and feeding it to validateTx. For convenience, we also keep a map of Address to Datum, giving is a simple way of managing the current utxo state.

mcstToUtxoState :: MockChainSt -> UtxoState

data MockChainSt

Slightly more concrete version of UtxoState, used to actually run the simulation. We keep a map from datum hash to datum, then a map from txOutRef to datumhash Additionally, we also keep a map from datum hash to the underlying value's "show" result, in order to display the contents of the state to the user.

Constructors

MockChainSt

Fields

mcstIndex :: UtxoIndex

mcstDatums :: Map DatumHash (Datum, Doc ())

mcstValidators :: Map ValidatorHash (Versioned Validator)

mcstCurrentSlot :: Slot

utxoIndexToTxOutMap :: UtxoIndex -> Map TxOutRef TxOut

The UtxoIndex contains TxOuts, but we want a map that contains TxOut.

data MockChainError

The errors that can be produced by the MockChainT monad

Constructors

MCEValidationError ValidationErrorInPhase

MCEUnbalanceable String BalanceStage TxSkel

MCENoSuitableCollateral

MCEGenerationError GenerateTxError

MCECalcFee MockChainError

MCEUnknownOutRefError String TxOutRef

FailWith String

data BalanceStage

Describes us which stage of the balancing process are we at. This is needed to distinguish the successive calls to balancing while computing fees from the final call to balancing

Constructors

BalCalcFee

BalFinalizing

data MockChainEnv

Constructors

MockChainEnv

Fields

mceParams :: Params

mceSigners :: NonEmpty Wallet

newtype MockChainT m a

The actual MockChainT is a trivial combination of StateT and ExceptT

Constructors

MockChainT

Fields

unMockChain :: ReaderT MockChainEnv (StateT MockChainSt (ExceptT MockChainError m)) a

type MockChain = MockChainT Identity

Non-transformer variant

combineMockChainT :: Monad m => (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

mapMockChainT :: (m (Either MockChainError (a, MockChainSt)) -> n (Either MockChainError (b, MockChainSt))) -> MockChainT m a -> MockChainT n b

runMockChainTRaw :: Monad m => MockChainEnv -> MockChainSt -> MockChainT m a -> m (Either MockChainError (a, MockChainSt))

Executes a MockChainT from some initial state and environment; does not convert the MockChainSt into a UtxoState.

runMockChainTFrom :: Monad m => InitialDistribution -> MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from an initial state set up with the given initial value distribution. Similar to runMockChainT, uses the default environment. Returns a UtxoState instead of a MockChainSt. If you need the later, use runMockChainTRaw

runMockChainT :: Monad m => MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from the canonical initial state and environment. The canonical environment uses the default SlotConfig and [Cooked.Wallet.wallet 1] as the sole wallet signing transactions.

runMockChainRaw :: MockChainEnv -> MockChainSt -> MockChain a -> Either MockChainError (a, MockChainSt)

See runMockChainTRaw

runMockChainFrom :: InitialDistribution -> MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainTFrom

runMockChain :: MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainT

utxoState0 :: UtxoState

mockChainSt0 :: MockChainSt

mockChainSt0From :: InitialDistribution -> MockChainSt

utxoIndex0From :: InitialDistribution -> UtxoIndex

utxoIndex0 :: UtxoIndex

Direct Interpretation of Operations

runTransactionValidation

Arguments

:: Monad m

=> Slot
The current slot

-> Params
The parameters of the MockChain

-> UtxoIndex
The currently known UTxOs

-> [Wallet]
List of signers that were on the TxSkel. This will at least have to include all wallets that have to sign in order for the transaction to be phase 1 - valid.

-> TxBodyContent BuildTx BabbageEra
The transaction to validate. It should already be balanced and have appropriate collateral.

The witnesses in script inputs must use the InlineScriptDatum constructor. We've not really understood why this is required by validateCardanoTx, but if the ScriptDatumForTxIn constructor with an explicit datum is used, we see an error about NonOutputSupplimentaryDatums.

-> Map DatumHash Datum
The data on transaction inputs. If the transaction is successful, these will be deleted from the mcstDatums.

-> Map DatumHash (Datum, Doc ())
The data on transaction outputs. If the transaction is successful, these will be added to the mcstDatums.

-> Map ValidatorHash (Versioned Validator)
The validators on transaction outputs.

-> [RawModTx]
Modifications to apply to the transaction right before it is submitted.

-> MockChainT m SomeCardanoApiTx

ensureTxSkelOutsMinAda :: Monad m => TxSkel -> MockChainT m TxSkel

Ensure that the transaction outputs have the necessary minimum amount of Ada on them. This will only be applied if the txOptEnsureMinAda is set to True.

txSkelInputUtxos :: Monad m => TxSkel -> MockChainT m (Map TxOutRef TxOut)

Get all UTxOs that the TxSkel consumes from the MockChainState.

txSkelReferenceInputUtxos :: Monad m => TxSkel -> MockChainT m (Map TxOutRef TxOut)

Get all UTxOs that the TxSkel references from the MockChainState.

lookupUtxos :: Monad m => [TxOutRef] -> MockChainT m (Map TxOutRef TxOut)

txSkelInputValue :: Monad m => TxSkel -> MockChainT m Value

Look up the UTxOs the transaction consumes, and sum the value contained in them.

txSkelInputDatums :: Monad m => TxSkel -> MockChainT m (Map DatumHash Datum)

Look up the data on UTxOs the transaction consumes.

setFeeAndBalance :: Monad m => PubKeyHash -> TxSkel -> MockChainT m (TxSkel, Fee)

Balance the TxSkel and compute the the fee. Balancing is the process of ensuring that the equation

value in inputs + minted value = value in outputs + burned value + fee

holds. The fee depends on the transaction size, which might change during the process of balancing, because additional inputs belonging to the balancePK might be added to ensure that transaction inputs can cover all of the outputs. This means that fee calculation and balancing are tied together. We follow plutus-apps in breaking this mutual dependency with a fixpoint iteration, which should compute realistic fees.

This function also adjusts the transaction outputs to contain at least the minimum Ada amount, if the 'txOptEnsureMinAda option is True.

estimateTxSkelFee :: Params -> UTxO EmulatorEra -> Map DatumHash (Datum, Doc ()) -> Map TxOutRef TxOut -> Map ValidatorHash (Versioned Validator) -> TxSkel -> Fee -> Either MockChainError Fee

This funcion is essentially a copy of https://github.com/input-output-hk/plutus-apps/blob/d4255f05477fd8477ee9673e850ebb9ebb8c9657/plutus-ledger/src/Ledger/Fee.hs#L19

calcCollateral :: Monad m => Wallet -> MockChainT m (Set TxOutRef)

Calculates the collateral for a transaction

balanceTxFromAux :: Monad m => BalanceOutputPolicy -> BalanceStage -> PubKeyHash -> TxSkel -> Fee -> MockChainT m TxSkel

data BalanceTxRes

Constructors

BalanceTxRes

Fields

newInputs :: [(TxOutRef, PKOutputMaybeDatum)]

Inputs that need to be added in order to cover the value in the transaction outputs

returnValue :: Value

The newInputs will add _at least_ the missing value to cover the outputs, this is the difference of the input value together with the newInputs and the output value. This value must be nonnegative in every asset class.

availableUtxos :: [(TxOutRef, PKOutputMaybeDatum)]

Some additional UTxOs that could be used as extra inputs. These all belong to the same wallet that was passed to calcBalanceTx as an argument, and are sorted in decreasing order of their Ada value.

calcBalanceTx :: Monad m => BalanceStage -> PubKeyHash -> TxSkel -> Fee -> MockChainT m BalanceTxRes

Calculate the changes needed to balance a transaction with money from a given wallet. Every transaction that is sent to the chain must be balanced, that is: inputs + mints == outputs + fee + burns.

applyBalanceTx :: PubKeyHash -> BalanceTxRes -> TxSkel -> Maybe TxSkel

Once we calculated what is needed to balance a transaction skel, we still need to apply those changes to skel. Because of the minAdaTxOut constraint, this might not be possible: imagine the leftover is less than minAdaTxOut, but the transaction has no output addressed to the balancing wallet. If we just create a new ouput for the balancing wallet and place the leftover there, the resulting transaction will fail to validate with LessThanMinAdaPerUTxO error. Instead, we need to consume yet another UTxO belonging to the wallet to then create the output with the proper leftover. If the wallet has no UTxO, then there's no way to balance this transaction.

Orphan instances

Default Slot

Instance details

Methods

def :: Slot