Cooked.MockChain.Direct

Direct Emulation

The MockChainT monad provides a direct emulator; that is, it gives us a simple way to call validator scripts directly, without the need for all the complexity the Contract monad introduces.

Running a MockChain produces a UtxoState, a simplified view on UtxoIndex, which we also keep in our state.

mcstToUtxoState :: MockChainSt -> UtxoState

data MockChainSt

Slightly more concrete version of UtxoState, used to actually run the simulation.

Constructors

MockChainSt

Fields

mcstIndex :: UtxoIndex

mcstDatums :: Map DatumHash (TxSkelOutDatum, Integer)

mcstValidators :: Map ValidatorHash (Versioned Validator)

mcstCurrentSlot :: Slot

newtype MockChainEnv

Constructors

MockChainEnv

Fields

mceParams :: Params

newtype MockChainT m a

Constructors

MockChainT

Fields

unMockChain :: ReaderT MockChainEnv (StateT MockChainSt (ExceptT MockChainError m)) a

type MockChain = MockChainT Identity

combineMockChainT :: Monad m => (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

mapMockChainT :: (m (Either MockChainError (a, MockChainSt)) -> n (Either MockChainError (b, MockChainSt))) -> MockChainT m a -> MockChainT n b

runMockChainTRaw :: Monad m => MockChainEnv -> MockChainSt -> MockChainT m a -> m (Either MockChainError (a, MockChainSt))

Executes a MockChainT from some initial state and environment; does not convert the MockChainSt into a UtxoState.

runMockChainTFrom :: Monad m => InitialDistribution -> MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from an initial state set up with the given initial value distribution. Similar to runMockChainT, uses the default environment. Returns a UtxoState instead of a MockChainSt. If you need the later, use runMockChainTRaw

runMockChainT :: Monad m => MockChainT m a -> m (Either MockChainError (a, UtxoState))

Executes a MockChainT from the canonical initial state and environment. The canonical environment uses the default SlotConfig and Cooked.Wallet.wallet 1 as the sole wallet signing transactions.

runMockChainRaw :: MockChainEnv -> MockChainSt -> MockChain a -> Either MockChainError (a, MockChainSt)

See runMockChainTRaw

runMockChainFrom :: InitialDistribution -> MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainTFrom

runMockChain :: MockChain a -> Either MockChainError (a, UtxoState)

See runMockChainT

Canonical initial values

utxoState0 :: UtxoState

mockChainSt0 :: MockChainSt

mockChainSt0From :: InitialDistribution -> MockChainSt

utxoIndex0From :: InitialDistribution -> UtxoIndex

utxoIndex0 :: UtxoIndex

Direct Interpretation of Operations

getIndex :: UtxoIndex -> Map TxOutRef TxOut

runTransactionValidation

Arguments

:: Monad m

=> Params
The emulator parameters to use. They might have been changed by the txOptEmulatorParamsModification.

-> Tx BabbageEra
The transaction to validate. It should already be balanced, and include appropriate fees and collateral.

-> [RawModTx]
Modifications to apply to the transaction right before it is submitted.

-> Map DatumHash Datum
The data consumed by the transaction

-> Map DatumHash TxSkelOutDatum
The data produced by the transaction

-> Map ValidatorHash (Versioned Validator)
The validators protecting transaction outputs, and the validators in the reference script field of transaction outputs. The MockChain will remember them.

-> MockChainT m CardanoTx

Orphan instances

Default Slot

Instance details

Methods

def :: Slot