Cooked.MockChain.Direct

Description

This module provides a direct (as opposed to Staged) implementation of the MonadBlockChain specification. This rely on the emulator from cardano-node-emulator for transaction validation, although we have our own internal state. This choice might be revised in the future.

Direct Emulation

The MockChainT monad provides a direct emulator; that is, it gives us a simple way to run a full validation process directly, without relying on a deployed node. While simulated, the validation is performed by the cardano-ledger code, thus ensuring similar results on the real chain.

A MockChain:

stores and updates a MockChainSt

returns a UtxoState when run

emits entries in a MockChainBook

data MockChainBook

This represents elements that can be emitted throughout a MockChain run. These elements are either log entries corresponding to internal events worth logging, or aliases for hashables corresponding to elements users wishes to be properly displayed when printed with PrettyCooked

Constructors

MockChainBook

Fields

mcbJournal :: [MockChainLogEntry]

Log entries generated by cooked-validators

mcbAliases :: Map BuiltinByteString String

Aliases stored by the user

newtype MockChainT m a

A MockChainT builds up a stack of monads on top of a given monad m to reflect the requirements of the simulation.

Constructors

MockChainT

Fields

unMockChain :: StateT MockChainSt (ExceptT MockChainError (WriterT MockChainBook m)) a

type MockChain = MockChainT Identity

Our MockChain naturally instantiate the inner monad with Identity

combineMockChainT :: (forall a. m a -> m a -> m a) -> MockChainT m x -> MockChainT m x -> MockChainT m x

Combines two MockChainT together

type MockChainReturn a b = (Either MockChainError (a, b), MockChainBook)

A generic return type for a MockChain run

mapMockChainT :: (m (MockChainReturn a MockChainSt) -> n (MockChainReturn b MockChainSt)) -> MockChainT m a -> MockChainT n b

Transforms a MockChainT into another one

runMockChainTRaw :: MockChainT m a -> m (MockChainReturn a MockChainSt)

Runs a MockChainT from a default MockChainSt

runMockChainTFrom :: Monad m => InitialDistribution -> MockChainT m a -> m (MockChainReturn a UtxoState)

Runs a MockChainT from an initial MockChainSt built from a given InitialDistribution. Returns a UtxoState.

runMockChainT :: Monad m => MockChainT m a -> m (MockChainReturn a UtxoState)

Executes a MockChainT from the canonical initial state and environment.

runMockChainFrom :: InitialDistribution -> MockChain a -> MockChainReturn a UtxoState

See runMockChainTFrom

runMockChain :: MockChain a -> MockChainReturn a UtxoState

See runMockChainT

Direct Interpretation of Operations