Cooked.MockChain.GenerateTx

Description

This module provides primitives to translate elements from our skeleton to actual transaction elements, including the transaction itself. Ideally, this module should only export generateTx but we need to make visible a few other primitives that will be used in balancing.

Documentation

data GenerateTxError

Constructors

ToCardanoError String ToCardanoError

TxBodyError String TxBodyError

GenerateTxErrorGeneral String

generateBodyContent :: Fee -> Set TxOutRef -> Params -> Map DatumHash Datum -> Map TxOutRef TxOut -> Map ValidatorHash (Versioned Validator) -> TxSkel -> Either GenerateTxError (TxBodyContent BuildTx ConwayEra)

generateTxOut :: NetworkId -> TxSkelOut -> Either GenerateTxError (TxOut CtxTx ConwayEra)

generateTx :: Fee -> Set TxOutRef -> Params -> Map DatumHash Datum -> Map TxOutRef TxOut -> Map ValidatorHash (Versioned Validator) -> TxSkel -> Either GenerateTxError (Tx ConwayEra)