Cooked.MockChain.GenerateTx

Documentation

data GenerateTxError

Constructors

ToCardanoError String ToCardanoError

TxBodyError String TxBodyError

GenerateTxErrorGeneral String

data GenTxParams

The internal (do-not-modify unless you know what you're doing) parameters for generateTxBodyContent.

generateTxBodyContent

Arguments

:: GenTxParams
Parameters controlling transaction generation.

-> Params
Some parameters, coming from the MockChain.

-> Map DatumHash Datum
All of the currently known data on transaction inputs, also coming from the MockChain.

-> Map TxOutRef TxOut
All of the currently known UTxOs which will be used as transaction inputs or referenced, also coming from the MockChain.

-> Map ValidatorHash (Versioned Validator)
All of the currently known validators which protect transaction inputs, also coming from the MockChain.

-> TxSkel
The transaction skeleton to translate.

-> Either GenerateTxError (TxBodyContent BuildTx BabbageEra)

txSkelOutToCardanoTxOut :: Params -> TxSkelOut -> Either GenerateTxError (TxOut CtxTx BabbageEra)

generateTx

Arguments

:: GenTxParams
Parameters controlling transaction generation.

-> Params
Some parameters, coming from the MockChain.

-> Map DatumHash Datum
All of the currently known data on transaction inputs, also coming from the MockChain.

-> Map TxOutRef TxOut
All of the currently known UTxOs which will be used as transaction inputs or referenced, also coming from the MockChain.

-> Map ValidatorHash (Versioned Validator)
All of the currently known validators which protect transaction inputs, also coming from the MockChain.

-> TxSkel
The transaction skeleton to translate.

-> Either GenerateTxError (Tx BabbageEra)