Cooked.MockChain.GenerateTx.Collateral

Documentation

data CollateralContext where

Constructors

CollateralContext

Fields

:: { managedTxOuts :: Map TxOutRef TxOut

, fee :: Integer

, mCollaterals :: Maybe (Set TxOutRef, Wallet)

, params :: Params

} -> CollateralContext

type CollateralGen a = TxGen CollateralContext a

toCollateralTriplet :: CollateralGen (TxInsCollateral ConwayEra, TxTotalCollateral ConwayEra, TxReturnCollateral CtxTx ConwayEra)

Computes the collateral triplet from the fees and the collateral inputs in the context. What we call a collateral triplet is composed of: * The set of collateral inputs * The total collateral paid by the transaction in case of phase 2 failure * An output returning excess collateral value when collaterals are used These quantity should satisfy the equation (in terms of their values): collateral inputs = total collateral + return collateral