Cooked.MockChain.Monad.Staged

Interpreting and running StagedMockChain

interpretAndRunWith :: (forall m. Monad m => MockChainT m a -> m res) -> StagedMockChain a -> [(res, TraceDescr)]

Interprets the staged mockchain then runs the resulting computation with a custom function. This can be used, for example, to supply a custom InitialDistribution by providing runMockChainTFrom.

interpretAndRun :: StagedMockChain a -> [(Either MockChainError (a, UtxoState), TraceDescr)]

type InterpMockChain = MockChainT (WriterT TraceDescr [])

The semantic domain in which StagedMockChain gets interpreted; see the interpret function for more.

interpret :: StagedMockChain a -> InterpMockChain a

The interpret function gives semantics to our traces. One StagedMockChain computation yields a potential list of MockChainT computations, which emit a description of their operation. Recall a MockChainT is a state and except monad composed:

    MockChainT (WriterT TraceDescr []) a
=~= st -> (WriterT TraceDescr []) (Either err (a, st))
=~= st -> [(Either err (a, st) , TraceDescr)]

StagedMockChain: An AST for MonadMockChain computations

data MockChainBuiltin a where

Constructors

ValidateTxSkel :: TxSkel -> MockChainBuiltin CardanoTx

TxOutByRef :: TxOutRef -> MockChainBuiltin (Maybe TxOut)

GetCurrentSlot :: MockChainBuiltin Slot

AwaitSlot :: Slot -> MockChainBuiltin Slot

GetCurrentTime :: MockChainBuiltin POSIXTime

AwaitTime :: POSIXTime -> MockChainBuiltin POSIXTime

UtxosSuchThat :: FromData a => Address -> (Maybe a -> Value -> Bool) -> MockChainBuiltin [(SpendableOut, Maybe a)]

UtxosSuchThisAndThat :: FromData a => (Address -> Bool) -> (Maybe a -> Value -> Bool) -> MockChainBuiltin [(SpendableOut, Maybe a)]

DatumFromTxOut :: ChainIndexTxOut -> MockChainBuiltin (Maybe Datum)

OwnPubKey :: MockChainBuiltin PubKeyHash

SigningWith :: NonEmpty Wallet -> StagedMockChain a -> MockChainBuiltin a

AskSigners :: MockChainBuiltin (NonEmpty Wallet)

GetParams :: MockChainBuiltin Params

LocalParams :: (Params -> Params) -> StagedMockChain a -> MockChainBuiltin a

Empty :: MockChainBuiltin a
The empty set of traces

Alt :: StagedMockChain a -> StagedMockChain a -> MockChainBuiltin a
The union of two sets of traces

Fail :: String -> MockChainBuiltin a
The failing operation

type MockChainOp = LtlOp Attack MockChainBuiltin

type StagedMockChain = Staged MockChainOp

InterpLtl instance

Modalities

type MonadModalMockChain m = (MonadMockChain m, MonadModal m, Modification m ~ Attack)

A modal mock chain is a mock chain that allows us to use LTL modifications with Attacks

MonadBlockChain and MonadMockChain instances

singletonBuiltin :: builtin a -> Staged (LtlOp modification builtin) a

Human Readable Traces

prettyMockChainOp :: NonEmpty Wallet -> MockChainOp a -> TraceDescr

Generates a TraceDescription for the given operation; we're mostly interested in seeing the transactions that were validated, so many operations have no description.

newtype TraceDescr

A TraceDescr is a list of Doc encoded as a difference list for two reasons (check ShowS if you're confused about how this works, its the same idea). 1) Naturally, these make for efficient concatenation 2) More importantly, this makes it easy to define the empty TraceDescr as TraceDescr id instead of relying on emptyDoc, which generates empty lines when used with vsep. This avoids generating these empty lines

Constructors

TraceDescr

Fields

trApp :: [Doc ()] -> [Doc ()]

trSingleton :: Doc ann -> TraceDescr

Orphan instances

Semigroup Attack

Instance details

Methods

(<>) :: Attack -> Attack -> Attack

sconcat :: NonEmpty Attack -> Attack

stimes :: Integral b => b -> Attack -> Attack

Monoid Attack

Instance details

Methods

mempty :: Attack

mappend :: Attack -> Attack -> Attack

mconcat :: [Attack] -> Attack

MonadPlus m => MonadPlus (MockChainT m)

Instance details

Methods

mzero :: MockChainT m a

mplus :: MockChainT m a -> MockChainT m a -> MockChainT m a