Cooked.MockChain.Staged

Documentation

interpretAndRunWith :: (forall m. Monad m => MockChainT m a -> m res) -> StagedMockChain a -> [(res, MockChainLog)]

Interprets the staged mockchain then runs the resulting computation with a custom function. This can be used, for example, to supply a custom InitialDistribution by providing runMockChainTFrom.

interpretAndRun :: StagedMockChain a -> [(Either MockChainError (a, UtxoState), MockChainLog)]

data MockChainLogEntry

Constructors

MCLogSubmittedTxSkel SkelContext TxSkel

MCLogNewTx TxId

MCLogFail String

newtype MockChainLog

Constructors

MockChainLog

Fields

unMockChainLog :: [MockChainLogEntry]

type StagedMockChain = Staged MockChainOp

runTweak :: Tweak InterpMockChain a -> TxSkel -> [Either MockChainError (a, TxSkel)]

runTweakFrom :: MockChainEnv -> MockChainSt -> Tweak InterpMockChain a -> TxSkel -> [Either MockChainError (a, TxSkel)]

User API

type MonadModalBlockChain m = (MonadBlockChain m, MonadModal m, Modification m ~ UntypedTweak InterpMockChain)

A modal mock chain is a mock chain that allows us to use LTL modifications with Tweaks

somewhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to some transaction in the given Trace. The tweak must apply at least once.

everywhere :: MonadModalBlockChain m => Tweak InterpMockChain b -> m a -> m a

Apply a Tweak to every transaction in a given trace. This is also successful if there are no transactions at all.

withTweak :: MonadModalBlockChain m => m x -> Tweak InterpMockChain a -> m x

Apply a Tweak to the next transaction in the given trace. The order of arguments is reversed compared to somewhere and everywhere, because that enables an idiom like

do ...
   endpoint arguments `withTweak` someModification
   ...

where endpoint builds and validates a single transaction depending on the given arguments. Then withTweak says "I want to modify the transaction returned by this endpoint in the following way".

Orphan instances

MonadPlus m => MonadPlus (MockChainT m)

Instance details

Methods

mzero :: MockChainT m a

mplus :: MockChainT m a -> MockChainT m a -> MockChainT m a