Cooked.MockChain.Testing

Description

This modules provides primitives to run tests over mockchain executions and to give expectation on the result of these runs.

Common interface between HUnit and QuickCheck

class IsProp prop where

IsProp is a common interface for HUnit and QuickCheck tests. It abstracts uses of Assertion and Property for (IsProp prop) => prop, then provide instances for both HU.Asserton and QC.Property.

Minimal complete definition

testCounterexample, testConjoin, testDisjoin

Methods

testCounterexample :: String -> prop -> prop

Displays the string to the user in case of failure

testConjoin :: [prop] -> prop

Conjunction of a number of results

testDisjoin :: [prop] -> prop

Disjunction of a number of results

testFailure :: prop

Flags a failure

testSuccess :: prop

Flags a success

testFailureMsg :: String -> prop

Flags a failure with a message

testBool :: IsProp prop => Bool -> prop

Turns a boolean into a prop

testAll :: IsProp prop => (a -> prop) -> [a] -> prop

Ensures all elements of a list satisfy a given prop

(.==.) :: (IsProp prop, Eq a) => a -> a -> prop infix 4

Lifts an equality test to a prop

(.&&.) :: IsProp prop => prop -> prop -> prop infixr 3

Conjunction of two props

(.||.) :: IsProp prop => prop -> prop -> prop infixr 2

Disjunction of two props

assertionToMaybe :: Assertion -> IO (Maybe HUnitFailure)

Catches a HUnit test failure, if the test fails.

forAll :: Show a => Gen a -> (a -> Property) -> Property

Here we provide our own universsal quantifier instead of forAll, so we can monomorphize it to returning a Property

Extra HUnit assertions

assertSubset :: (Show a, Eq a) => [a] -> [a] -> Assertion

Asserts whether a set is a subset of another one, both given as lists.

assertSameSets :: (Show a, Eq a) => [a] -> [a] -> Assertion

Asserts whether 2 sets are equal, both given as lists.

Testing mockchain traces

data Test a prop

Data structure to test a mockchain trace

Constructors

Test

Fields

testTrace :: StagedMockChain a

The mockchain trace to test

testInitDist :: InitialDistribution

The initial distribution from which the trace should be run

testSizeProp :: Integer -> prop

The requirement on the number of results

testErrorProp :: PrettyCookedOpts -> MockChainError -> [MockChainLogEntry] -> prop

The property should hold in case of failure

testResultProp :: PrettyCookedOpts -> a -> UtxoState -> [MockChainLogEntry] -> prop

The property that should hold in case of success

testPrettyOpts :: PrettyCookedOpts

The printing option that should be use to render test results

mustSucceedTest :: IsProp prop => StagedMockChain a -> Test a prop

A test template which expects a success from a trace

mustFailTest :: (IsProp prop, Show a) => StagedMockChain a -> Test a prop

A test template which expects a failure from a trace

emptyTest :: IsProp prop => StagedMockChain a -> Test a prop

A test template with no particular requirement on the trace

withInitDist :: Test a prop -> InitialDistribution -> Test a prop

Appending an initial distribution to a test

withPrettyOpts :: Test a prop -> PrettyCookedOpts -> Test a prop

Appending printing options to a test

withJournalPred :: IsProp prop => Test a prop -> ([MockChainLogEntry] -> prop) -> Test a prop

Appending a predicate over the log to a test. This will be used both in case of success or failure of the trace.

withValueAndStatePred :: IsProp prop => Test a prop -> (a -> UtxoState -> prop) -> Test a prop

Appending a predicate over the return value and state, which will be used in case of success of the trace.

withExactSize :: IsProp prop => Test a prop -> Integer -> Test a prop

Ensuring the number of results of a test is exactly as expected

withValuePred :: IsProp prop => Test a prop -> (a -> prop) -> Test a prop

Appending a predicate over the return value, which will be used in case of success of the trace.

withStatePred :: IsProp prop => Test a prop -> (UtxoState -> prop) -> Test a prop

Appending a predicate over the return state, which will be used in case of success of the trace.

withPrettyAndErrorPred :: IsProp prop => Test a prop -> (PrettyCookedOpts -> MockChainError -> prop) -> Test a prop

Appending a predicate over an error which uses the printing options, which will be used in case of failure of the trace.

withErrorPred :: IsProp prop => Test a prop -> (MockChainError -> prop) -> Test a prop

Appends a predicate over an error to a Test

testToProp :: IsProp prop => Test a prop -> prop

This takes a test and transforms it into an actual test case in prop.

testSucceeds :: IsProp prop => StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed, starting from the default initial distribution

testFails :: (IsProp prop, Show a) => StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail

isPhase1Failure :: IsProp prop => PrettyCookedOpts -> MockChainError -> prop

A property to ensure a phase 1 failure

testFailsInPhase1 :: (IsProp prop, Show a) => StagedMockChain a -> prop

A test that succeeds when the trace results in a phase 1 failure

isPhase2Failure :: IsProp prop => PrettyCookedOpts -> MockChainError -> prop

A property to ensure a phase 2 failure

testFailsInPhase2 :: (IsProp prop, Show a) => StagedMockChain a -> prop

A test that succeeds when the trace results in a phase 2 failure

isPhase1FailureWithMsg :: IsProp prop => (String -> Bool) -> PrettyCookedOpts -> MockChainError -> prop

Same as isPhase1Failure with an added predicate on the text error

testFailsInPhase1WithMsg :: (IsProp prop, Show a) => (String -> Bool) -> StagedMockChain a -> prop

Same as testFailsInPhase1 with an added predicate on the text error

isPhase2FailureWithMsg :: IsProp prop => (String -> Bool) -> PrettyCookedOpts -> MockChainError -> prop

Same as isPhase2Failure with an added predicate over the text error

testFailsInPhase2WithMsg :: (IsProp prop, Show a) => (String -> Bool) -> StagedMockChain a -> prop

Same as testFailsInPhase2 with an added predicate over the text error