Cooked.MockChain.Testing

Documentation

class IsProp prop where

This module provides a common interface for HUnit and QuickCheck tests. We do so by abstracting uses of Assertion and Property for (IsProp prop) => prop, then provide instances for both HU.Asserton and QC.Property.

Minimal complete definition

testCounterexample, testConjoin, testDisjoin

Methods

testCounterexample :: String -> prop -> prop

Displays the string to the user in case of failure

testConjoin :: [prop] -> prop

Conjunction of a number of results

testDisjoin :: [prop] -> prop

Disjunction of a number of results

testFailure :: prop

Flags a failure

testSuccess :: prop

Flags a success

testFailureMsg :: String -> prop

Flags a failure with a message

testBool :: IsProp prop => Bool -> prop

testAll :: IsProp prop => (a -> prop) -> [a] -> prop

(.==.) :: (IsProp prop, Eq a) => a -> a -> prop infix 4

(.&&.) :: IsProp prop => prop -> prop -> prop infixr 3

(.||.) :: IsProp prop => prop -> prop -> prop infixr 2

testSucceeds :: IsProp prop => StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed, starting from the default initial distribution

testFails :: (IsProp prop, Show a) => StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from the default initial distribution

testSucceedsFrom :: IsProp prop => InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed starting from some initial distribution but doesn't impose any additional condition on success. Use testSucceedsFrom' for that.

testSucceedsFrom' :: IsProp prop => (a -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain succeed starting from some initial distribution. Additionally impose a condition over the resulting state and value.

testFailsFrom :: (IsProp prop, Show a) => InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from some initial distribution

testFailsFrom' :: (IsProp prop, Show a) => (MockChainError -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the staged mockchain fail starting from some initial distribution, moreover, ensures that a certain predicate over the error holds.

isCekEvaluationFailure :: IsProp prop => MockChainError -> prop

Is satisfied when the given MockChainError is wrapping a CekEvaluationFailure. This is particularly important when writing negative tests. For example, if we are simulating an attack and writing a test with testFailsFrom, we might have made a mistake in the attack, yielding a test that fails for reasons such as ValueLessThanMinAda or ValueNotPreserved, which does not rule out the attack being caught by the validator script. For these scenarios it is paramount to rely on testFailsFrom' isCekEvaluationFailure instead.

isCekEvaluationFailureWithMsg :: IsProp prop => (String -> Bool) -> MockChainError -> prop

Similar to isCekEvaluationFailure, but enables us to check for a specific error message in the error.

testAllSatisfiesFrom :: forall prop a. IsProp prop => (Either MockChainError (a, UtxoState) -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Ensure that all results produced by the set of traces encoded by the StagedMockChain satisfy the given predicate. If you wish to build custom predicates you can use testSatisfiesFrom' directly and see testBinaryRelatedBy as an example.

testBinaryRelatedBy :: IsProp prop => (UtxoState -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Asserts that the given StagedMockChain produces exactly two outcomes, both of which are successful and have their resulting states related by a given predicate. A typical usage would look like:

testBinaryRelatedBy equalModuloAda myInitDistr $ do
  x <- trPrepare
  execOption1 x <|> execOption2 x

testOneEquivClass :: IsProp prop => (UtxoState -> UtxoState -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Generalizes testBinaryRelatedBy, asserting that the given StagedMockChain produces more than two outcomes, say [x,y,z,w], all of which are successful (i.e. are not a MockChainError) and these states are in the same equivalence class of (~); that is, they satisfy:

x ~ y && x ~ z && x ~ z && x ~ w

Because (~) should be symmetric and transitive we can estabilish that these states all belong to the same equivalence class. This function does not check each pointwise case.

testSatisfiesFrom' :: ([(Either MockChainError (a, UtxoState), TraceDescr)] -> prop) -> InitialDistribution -> StagedMockChain a -> prop

Asserts that the results produced by running the given StagedMockChain from some speficied InitialDistribution satisfy a given assertion. In this case, the predicate gets the trace descriptions that led to each potential outcome and is responsible for calling testCounterexample communicate these to the user.

Although this function is mainly used internally, as a building block for the simpler predicates, it can be useful in building some custom predicates. Check testAllSatisfiesFrom or testBinaryRelatedBy for examples on using this.

TestResult instances

assertionToMaybe :: Assertion -> IO (Maybe HUnitFailure)

forAll :: Show a => Gen a -> (a -> Property) -> Property

Here we provide our own universsal quantifier instead of forAll, so we can monomorphize it to returning a Property