Cooked.MockChain.UtxoSearch

Description

This module provides a convenient framework to look through UTxOs and search relevant ones based on predicates. For instance, it makes it very convenient to gather all UTxOs at a certain address.

The type of UTxO searches * --

type UtxoSearch m a = MaybeT (ListT m) (TxOutRef, a)

If a UTxO is a TxOutRef with some additional information, this type captures a "stream" of potential UTxOs.

runUtxoSearch :: Monad m => UtxoSearch m a -> m [(TxOutRef, a)]

Given a UTxO search, we can run it to obtain a list of actual UTxOs.

Filters to trim down utxos * --

type Filter m a b = a -> m (Maybe b)

A filter changes a into a potential b within a certain context m. These filters are more powerful than the usual filters as they are type-changing. Thus they will only be composable in a certain type-compatible fashion

idFilter :: (a -> m (Maybe b)) -> Filter m a b

Creates a filter from a filtering function

pureFilter :: Monad m => (a -> Maybe b) -> Filter m a b

Creates a filter from a pure filtering function

pureAlwaysFilter :: Monad m => (a -> b) -> Filter m a b

Creates a filter from a pure always succeeding filter

alwaysFilter :: Monad m => (a -> m b) -> Filter m a b

Creates a filter from a monadic always succeeding filter

pureBoolFilter :: Monad m => (a -> Bool) -> Filter m a a

Creates a filter from a pure selection function

boolFilter :: Monad m => (a -> m Bool) -> Filter m a a

Creates a filter from a monadic selection function

opticFilter :: (Monad m, Is k An_AffineFold) => Optic' k is a b -> Filter m a b

Creates a filter from an optic

negateF :: Monad m => Filter m a b -> Filter m a a

Negates a filter, that is keep elements which should be filtered out. Does not produce any casting.

Searching and filtering utxos * --

(*+*) :: Monad m => UtxoSearch m a -> Filter m a b -> UtxoSearch m b

Refines a search using a filter

(*~*) :: Monad m => UtxoSearch m a -> Filter m a b -> UtxoSearch m a

Refines a seach using a negated filter

Initial searches * --

utxosAtSearch :: (MonadBlockChainBalancing m, HasAddress addr) => addr -> UtxoSearch m TxOut

Search all TxOutRefs at a certain address, together with their TxInfo-TxOut.

utxosAtLedgerSearch :: (MonadBlockChainBalancing m, HasAddress addr) => addr -> UtxoSearch m TxOut

Like utxosAtSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

allUtxosSearch :: MonadBlockChain m => UtxoSearch m TxOut

Search all currently known TxOutRefs together with their corresponding TxInfo-TxOut.

allUtxosLedgerSearch :: MonadBlockChain m => UtxoSearch m TxOut

Like allUtxosSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

concreteOutputSearch :: MonadBlockChain m => UtxoSearch m (ConcreteOutput Credential OutputDatum Value ScriptHash)

Initializes a search with all known Concrete outputs

utxosFromCardanoTxSearch :: Monad m => CardanoTx -> UtxoSearch m TxOut

Initializes a search with utxos from a specific transaction

txOutByRefSearch :: MonadBlockChainBalancing m => [TxOutRef] -> UtxoSearch m TxOut

Initializes a search with resolved utxos from a given list

onlyAdaSearch :: MonadBlockChain m => UtxoSearch m (ConcreteOutput Credential OutputDatum Ada ScriptHash)

notAdaOnlySearch :: MonadBlockChain m => UtxoSearch m (ConcreteOutput Credential OutputDatum Value ScriptHash)

vanillaUtxosSearch :: MonadBlockChain m => UtxoSearch m (ConcreteOutput Credential () Ada ScriptHash)

vanillaUtxosAtSearch :: (MonadBlockChainBalancing m, HasAddress addr) => addr -> UtxoSearch m (ConcreteOutput Credential () Ada ScriptHash)

scriptOutputSearch :: MonadBlockChain m => TypedValidator a -> UtxoSearch m (ConcreteOutput (TypedValidator a) OutputDatum Value ScriptHash)