Cooked.MockChain.UtxoSearch

Documentation

runUtxoSearch :: Monad m => UtxoSearch m a -> m [(TxOutRef, a)]

Given a UTxO search, we can run it to obtain a list of UTxOs.

allUtxosSearch :: MonadBlockChain m => UtxoSearch m TxOut

Search all currently known TxOutRefs together with their corresponding TxInfo-TxOut.

allUtxosLedgerSearch :: MonadBlockChain m => UtxoSearch m TxOut

Like allUtxosSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

utxosAtSearch :: MonadBlockChainBalancing m => Address -> UtxoSearch m TxOut

Search all TxOutRefs at a certain address, together with their TxInfo-TxOut.

utxosAtLedgerSearch :: MonadBlockChainBalancing m => Address -> UtxoSearch m TxOut

Like utxosAtSearch, but returns a Ledger-level representation of the transaction outputs, which might contain more information.

utxosFromCardanoTxSearch :: Monad m => CardanoTx -> UtxoSearch m TxOut

Search all TxOutRefs of a transaction, together with their TxInfo-TxOut.

filterWith :: Monad m => UtxoSearch m a -> (a -> m (Maybe b)) -> UtxoSearch m b

Transform a UtxoSearch by applying a possibly failing monadic "lookup" on every output.

filterWithPure :: Monad m => UtxoSearch m a -> (a -> Maybe b) -> UtxoSearch m b

filterWithOptic :: (Is k An_AffineFold, Monad m) => UtxoSearch m a -> Optic' k is a b -> UtxoSearch m b

filterWithPred :: Monad m => UtxoSearch m a -> (a -> Bool) -> UtxoSearch m a