Cooked.MockChain.UtxoState

Documentation

newtype UtxoState

A UtxoState provides us with the mental picture of the state of the UTxO graph: Each address has a set of UTxOs that consist in a value and some potential datum.

Constructors

UtxoState

Fields

utxoState :: Map Address UtxoValueSet

newtype UtxoValueSet

Represents a set of values, yet, we use a list instead of a set because Value doesn't implement Ord and because it is possible that we want to distinguish between utxo states that have additional utxos, even if these could have been merged together.

Constructors

UtxoValueSet

Fields

utxoValueSet :: [(Value, TxSkelOutDatum)]

utxoValueSetTotal :: UtxoValueSet -> Value

Computes the total value in a set

utxoStateTotal :: UtxoState -> Value

Computes the total value in the entire state