Cooked.MockChain.UtxoState

Documentation

newtype UtxoState

A description of who owns what in a blockchain. Owners are addresses and they each own a UtxoPayloadSet.

Constructors

UtxoState

Fields

utxoState :: Map Address UtxoPayloadSet

newtype UtxoPayloadSet

Represents a set of payloads.

Constructors

UtxoPayloadSet

Fields

utxoPayloadSet :: [UtxoPayload]

data UtxoPayload

A convenient wrapping of the interesting information of a UTxO.

Constructors

UtxoPayload

Fields

utxoPayloadTxOutRef :: TxOutRef

utxoPayloadValue :: Value

utxoPayloadSkelOutDatum :: TxSkelOutDatum

utxoPayloadReferenceScript :: Maybe ScriptHash

holdsInState :: Address -> UtxoState -> Value

Total value accessible to what's pointed by the address.