Cooked.MockChain.UtxoState

Description

This module provides a depiction of the state we return when running a MockChain.

Documentation

newtype UtxoState

A description of who owns what in a blockchain. Owners are addresses and they each own a UtxoPayloadSet.

Constructors

UtxoState

Fields

utxoState :: Map Address UtxoPayloadSet

newtype UtxoPayloadSet

Represents a set of payloads.

Constructors

UtxoPayloadSet

Fields

utxoPayloadSet :: [UtxoPayload]

List of UTxOs contained in this UtxoPayloadSet

data UtxoPayload

A convenient wrapping of the interesting information of a UTxO.

Constructors

UtxoPayload

Fields

utxoPayloadTxOutRef :: TxOutRef

The reference of this UTxO

utxoPayloadValue :: Value

The value stored in this UTxO

utxoPayloadSkelOutDatum :: TxSkelOutDatum

The datum stored in this UTxO

utxoPayloadReferenceScript :: Maybe ScriptHash

The optional reference script stored in this UTxO

holdsInState :: ToAddress a => a -> UtxoState -> Value

Total value accessible to what's pointed by the address.