Cooked.MockChain.UtxoState

Documentation

newtype UtxoState

A UtxoState provides us with the mental picture of the state of the UTxO graph: Each address has a set of UTxOs that consist in a value and some potential datum.

Constructors

UtxoState

Fields

utxoState :: Map Address UtxoValueSet

newtype UtxoValueSet

Represents a set of values, yet, we use a list instead of a set because Value doesn't implement Ord and because it is possible that we want to distinguish between utxo states that have additional utxos, even if these could have been merged together.

Constructors

UtxoValueSet

Fields

utxoValueSet :: [(Value, Maybe UtxoDatum)]

utxoValueSetTotal :: UtxoValueSet -> Value

Computes the total value in a set

utxoStateTotal :: UtxoState -> Value

Computes the total value in the entire state

data UtxoDatum

A UtxoDatum contains a datum which is Datum $ Pl.toBuiltinData x for some x :: X, but we also include show x to be able to print this value in a more user friendly fashion.

Constructors

UtxoDatum

Fields

utxoDatum :: Datum

utxoShow :: String

Differences between two UtxoState

type UtxoStateDiff = Map Address UtxoValueSetDiff

The differences between two states a and b is a map whose keys are M.keys a union M.keys b and values dictate what happened to that particular value set with a UtxoValueSetDiff

data UtxoValueSetDiff

A UtxoValueSet can have been inserted, deleted or modified.

Constructors

Inserted UtxoValueSet

Deleted UtxoValueSet

Modified
In case it has been modified, the order is irrelevant; hence we just need to keep the deltas around.

Fields

modDeleted :: UtxoValueSet

modInserted :: UtxoValueSet

modUnchanged :: UtxoValueSet

utxoValueSetDiffTotal :: UtxoValueSetDiff -> Value

Computes the numeric difference between the new total and the old total of the source and destinations of this diff of set of values

utxoStateDiffTotal :: UtxoStateDiff -> Value

Computes the total difference in value between a new state and an old state: utxoStateDiff new old. If this difference is positive for some token t, it means some t it was burnt. If it is negative, some t was minted.

utxoStateDiff :: UtxoState -> UtxoState -> UtxoStateDiff

Computes the difference between two states.

utxoStateDiffSrc :: UtxoStateDiff -> UtxoState

Returns the source of a difference. Satisfies: utxoStateDiffSrc (utxoStateDiff a b) == a

utxoStateDiffTgt :: UtxoStateDiff -> UtxoState

Returns the target of a difference. Satisfies: utxoStateDiffTgt (utxoStateDiff a b) == b

Pretty-printing

prettyUtxoState :: UtxoState -> Doc ann

Pretty prints a UtxoState. The entire point of producing a UtxoState instead of a UtxoIndex is to provide the user a picture that is closer to their mental model of what is going on. Hence, we must be able to display our UtxoStates in a human readable fashion.

prettyAddress :: Address -> [(Value, Maybe UtxoDatum)] -> Doc ann

prettyPayloadGroup :: [(Value, Maybe UtxoDatum)] -> Maybe (Doc ann)

prettyPayload :: Value -> Maybe UtxoDatum -> Maybe (Doc ann)

mPrettyValue :: Value -> Maybe (Doc ann)

prettyCurrencyAndAmount :: CurrencySymbol -> Map TokenName Integer -> Doc ann

prettyAddressTypeAndHash :: Address -> Doc ann