Cooked.Output

Documentation

class IsAbstractOutput o where

A generalisation of TxOut: With the four type families, we can lift some information about who owns the output (a public key, a script...?) and the datum (do we have an inline datum, a datum hash, nothing...?) to the type level.

Associated Types

type OwnerType o

type DatumType o

type ValueType o

type ReferenceScriptType o

Methods

outputOwnerL :: Lens' o (OwnerType o)

outputStakingCredentialL :: Lens' o (Maybe StakingCredential)

outputDatumL :: Lens' o (DatumType o)

outputValueL :: Lens' o (ValueType o)

outputReferenceScriptL :: Lens' o (Maybe (ReferenceScriptType o))

class ToCredential a where

Methods

toCredential :: a -> Credential

class ToOutputDatum a where

Methods

toOutputDatum :: a -> OutputDatum

class ToValue a where

Methods

toValue :: a -> Value

class ToScript a where

Methods

toScript :: a -> Versioned Script

class ToScriptHash a where

Methods

toScriptHash :: a -> ScriptHash

type IsOnchainOutput o = (IsAbstractOutput o, ToCredential (OwnerType o), ToOutputDatum (DatumType o), ToValue (ValueType o), ToScriptHash (ReferenceScriptType o))

An output that can be translated into its script-perspective (as seen on the TxInfo) representation

outputAddress :: (IsAbstractOutput o, ToCredential (OwnerType o)) => o -> Address

outputOutputDatum :: (IsAbstractOutput o, ToOutputDatum (DatumType o)) => o -> OutputDatum

outputValue :: (IsAbstractOutput o, ToValue (ValueType o)) => o -> Value

outputReferenceScript :: (IsAbstractOutput o, ToScriptHash (ReferenceScriptType o)) => o -> Maybe ScriptHash

outputTxOut :: IsOnchainOutput o => o -> TxOut

Return the output as it is seen by a validator. In particular the correctness of this specification will depend on the IsOnchainOutput instance, so make sure you get these instances (in particular the functions outputAddress, outputOutputDatum, and outputValue) right!

TxOuts are outputs

A concrete type for outputs

data ConcreteOutput ownerType datumType valueType referenceScriptType where

Constructors

ConcreteOutput

Fields

:: { concreteOutputOwner :: ownerType

, concreteOutputStakingCredential :: Maybe StakingCredential

, concreteOutputValue :: valueType

, concreteOutputDatum :: datumType

, concreteOutputReferenceScript :: Maybe referenceScriptType

} -> ConcreteOutput ownerType datumType valueType referenceScriptType

A few special concrete outputs

type PKAdaOnlyOutput = ConcreteOutput PubKeyHash () Ada ScriptHash

A public key output that only has Ada and no datum

type PKOutputMaybeDatum = ConcreteOutput PubKeyHash OutputDatum Value ScriptHash

A public key output where we don't know anything about the datum: It is a general OutputDatum

type ScriptOutputWithInlineDatum a = ConcreteOutput (TypedValidator a) (DatumType a) Value ScriptHash

An output that belongs to a typed validator and has an inline datum of the appropriate type.

Functions to translate between different output types

toOutputWithReferenceScriptHash :: (IsAbstractOutput output, ToScriptHash (ReferenceScriptType output)) => output -> ConcreteOutput (OwnerType output) (DatumType output) (ValueType output) ScriptHash

isOutputWithoutDatum :: IsOnchainOutput output => output -> Maybe (ConcreteOutput (OwnerType output) () (ValueType output) (ReferenceScriptType output))

Test if there is no datum on an output. If there is no datum, return an output with the same OwnerType, but with () as its DatumType.

Functions to translate between different output types

isOutputWithInlineDatumUntyped :: IsOnchainOutput output => output -> Maybe (ConcreteOutput (OwnerType output) Datum (ValueType output) (ReferenceScriptType output))

Test if the output carries some inlined datum (lose the type information about the datum in favour of Plutus' Datum type).

isOutputWithInlineDatum :: IsOnchainOutput output => output -> Maybe output

Test if the output carries some inlined datum.

isOutputWithDatumHash :: IsOnchainOutput output => output -> Maybe (ConcreteOutput (OwnerType output) DatumHash (ValueType output) (ReferenceScriptType output))

Test if the output carries some datum hash.

isOutputWithValueSuchThat :: IsAbstractOutput output => (ValueType output -> Bool) -> output -> Maybe output

Test if the value carried by an output verifies a given predicate.

isOutputWithDatumSuchThat :: IsAbstractOutput output => (DatumType output -> Bool) -> output -> Maybe output

Test if the datum carried by an output verifies a given predicate.

newtype ResolvedOrInlineDatum a

Wrapper type to make clear that the datum is a resolved or inline datum

Constructors

ResolvedOrInlineDatum a

isOutputWithDatumOfType :: forall a output. (FromData a, IsOnchainOutput output) => output -> Maybe (ConcreteOutput (OwnerType output) (ResolvedOrInlineDatum a) (ValueType output) (ReferenceScriptType output))

Test if an output has an inline datum of a certain type. In most applications, it will make sense to use this with filteredUtxosWithDatums, and not with filteredUtxos.

isScriptOutputFrom :: IsOnchainOutput output => TypedValidator a -> output -> Maybe (ConcreteOutput (TypedValidator a) (DatumType output) (ValueType output) (ReferenceScriptType output))

Test if the owner an output is a specific script. If it is, return an output with the validator type as its OwnerType.

isScriptOutputFrom' :: forall a output. (IsOnchainOutput output, FromData (DatumType a), ToOutputDatum (DatumType output), Show (DatumType output), ToValue (ValueType output), Show (ValueType output)) => TypedValidator a -> output -> Maybe (ConcreteOutput (TypedValidator a) (ResolvedOrInlineDatum (DatumType a)) (ValueType output) (ReferenceScriptType output))

like isScriptOutputFrom, but also makes sure that the output has an inline datum of the correct type. In most applications, it will make sense to use this with filteredUtxosWithDatums, and not with filteredUtxos.

isPKOutputFrom :: IsOnchainOutput output => PubKeyHash -> output -> Maybe (ConcreteOutput PubKeyHash (DatumType output) (ValueType output) (ReferenceScriptType output))

Test if the owner an output is a specific public key. If it is, return an output of the same DatumType, but with PubKeyHash as its OwnerType.

isOnlyAdaOutput :: IsOnchainOutput output => output -> Maybe (ConcreteOutput (OwnerType output) (DatumType output) Ada (ReferenceScriptType output))

Test if the value on an output contains only Ada, and adapt the return type accordingly if it is so.