Cooked.Output

Description

This module provide abstractions around the notion of outputs. The idea is to use types to carry additional information on which data is carried by various kinds of concrete outputs used in cooked specifically or in Plutus as a whole, such as TxSkelOut or TxOut.

Documentation

class IsAbstractOutput o

A generalisation of TxOut: With the four type families, we can lift some information about its content

Minimal complete definition

outputOwnerL, outputStakingCredentialL, outputDatumL, outputValueL, outputReferenceScriptL

type family OwnerType o

The owner of the output. Is it a public key, as script...?

type family DatumType o

The datum stored in the output. Is a nothing, an OutputDatum, a TxSkelOutDatum...?

type family ValueType o

The value in the output. Does it contain pure ADA? Other assets?

type family ReferenceScriptType o

The reference script in the output. Is it a hash, a typed validator...?

outputOwnerL :: IsAbstractOutput o => Lens' o (OwnerType o)

Getting or setting the owner of this output

outputStakingCredentialL :: IsAbstractOutput o => Lens' o (Maybe StakingCredential)

Getting or setting the staking credential of this output

outputDatumL :: IsAbstractOutput o => Lens' o (DatumType o)

Getting or setting the datum of this output

outputValueL :: IsAbstractOutput o => Lens' o (ValueType o)

Getting or setting the value of this output

outputReferenceScriptL :: IsAbstractOutput o => Lens' o (Maybe (ReferenceScriptType o))

Getting or setting the reference script of this output

type IsTxInfoOutput o = (IsAbstractOutput o, ToCredential (OwnerType o), ToOutputDatum (DatumType o), ToValue (ValueType o), ToScriptHash (ReferenceScriptType o))

An output that can be translated into its script-perspective (as seen on the TxInfo) representation

outputAddress :: (IsAbstractOutput o, ToCredential (OwnerType o)) => o -> Address

Retrieve the Address of a script-translatable output

outputOutputDatum :: (IsAbstractOutput o, ToOutputDatum (DatumType o)) => o -> OutputDatum

Retrieves the OutputDatum of a script-translatable output

outputValue :: (IsAbstractOutput o, ToValue (ValueType o)) => o -> Value

Retrieves the Value of a script-translatable output

outputReferenceScriptHash :: (IsAbstractOutput o, ToScriptHash (ReferenceScriptType o)) => o -> Maybe ScriptHash

Retrieves the optional ScriptHash of a script-translatable output

outputTxOut :: IsTxInfoOutput o => o -> TxOut

Translates a script-translatable output into an TxOut

data ConcreteOutput ownerType datumType valueType referenceScriptType where

A type constructed to be the most general instance of IsAbstractOutput.

Constructors

ConcreteOutput

Fields

:: { concreteOutputOwner :: ownerType

, concreteOutputStakingCredential :: Maybe StakingCredential

, concreteOutputDatum :: datumType

, concreteOutputValue :: valueType

, concreteOutputReferenceScript :: Maybe referenceScriptType

} -> ConcreteOutput ownerType datumType valueType referenceScriptType

toOutputWithReferenceScriptHash :: IsTxInfoOutput out => out -> ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) ScriptHash

Convert the reference script type on the output to ScriptHash.

isOutputWithoutDatum :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) () (ValueType out) (ReferenceScriptType out))

Test if there is no datum on an output.

isOutputWithInlineDatum :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) Datum (ValueType out) (ReferenceScriptType out))

Test if the output carries some inlined datum.

isOutputWithDatumHash :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) DatumHash (ValueType out) (ReferenceScriptType out))

Test if the output carries some datum hash.

isOutputWithInlineDatumOfType :: (FromData a, IsTxInfoOutput out) => out -> Maybe (ConcreteOutput (OwnerType out) a (ValueType out) (ReferenceScriptType out))

Test if the output carries some inlined datum that can be parsed from builtin data on to something of a specific type.

isScriptOutputFrom :: (IsTxInfoOutput out, ToScriptHash s) => s -> out -> Maybe (ConcreteOutput s (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the owner of an output is a specific script

isPKOutputFrom :: IsTxInfoOutput out => PubKeyHash -> out -> Maybe (ConcreteOutput PubKeyHash (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the owner of an output is a specific public key

isOnlyAdaOutput :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) (DatumType out) Lovelace (ReferenceScriptType out))

Test if the value on an output contains only Ada.

fromAbstractOutput :: IsAbstractOutput out => out -> ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (ReferenceScriptType out)

Creates a generic concrete instance of any kind of abstract output

isReferenceScriptOutputFrom :: (IsTxInfoOutput out, ToScriptHash s) => s -> out -> Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) ScriptHash)

Test if the reference script in an output is a specific script

isStakingCredentialOutputFrom :: (IsTxInfoOutput out, ToCredential cred) => cred -> out -> Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the given output possesses a certain staking credential

isEmptyStakingCredentialOutput :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the give output does not possess any staking credential

isPKOutput :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput PubKeyHash (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the owner of an output is a public key

isScriptOutput :: IsTxInfoOutput out => out -> Maybe (ConcreteOutput ScriptHash (DatumType out) (ValueType out) (ReferenceScriptType out))

Test if the owner of an output is a script

setDatum :: IsAbstractOutput out => out -> dat -> ConcreteOutput (OwnerType out) dat (ValueType out) (ReferenceScriptType out)

Type-changing setter for an abstract output datum

setOwner :: IsAbstractOutput out => out -> owner -> ConcreteOutput owner (DatumType out) (ValueType out) (ReferenceScriptType out)

Type-changing setter for an abstract output owner

setReferenceScript :: IsAbstractOutput out => out -> ref -> ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) ref

Type-changing setter for an abstract output reference script

setValue :: IsAbstractOutput out => out -> val -> ConcreteOutput (OwnerType out) (DatumType out) val (ReferenceScriptType out)

Type-changing setter for an abstract output value

setStakingCredential :: IsAbstractOutput out => out -> StakingCredential -> ConcreteOutput (OwnerType out) (DatumType out) (ValueType out) (ReferenceScriptType out)

Type-changing setter for an abstract output staking scredential