Cooked.Output

Documentation

class IsAbstractOutput o

A generalisation of TxOut: With the four type families, we can lift some information about

who owns the output (a public key, a script...?)

what kind of datum is there (do we have an inline datum, a datum hash, nothing...?)

what kind of value does the output hold (pure Ada ...?)

what information do we have on the reference script (only a hash, a complete script, a typed validator...?)

to the type level.

Minimal complete definition

outputOwnerL, outputStakingCredentialL, outputDatumL, outputValueL, outputReferenceScriptL

type family OwnerType o

type family DatumType o

type family ValueType o

type family ReferenceScriptType o

outputOwnerL :: IsAbstractOutput o => Lens' o (OwnerType o)

outputStakingCredentialL :: IsAbstractOutput o => Lens' o (Maybe StakingCredential)

outputDatumL :: IsAbstractOutput o => Lens' o (DatumType o)

outputValueL :: IsAbstractOutput o => Lens' o (ValueType o)

outputReferenceScriptL :: IsAbstractOutput o => Lens' o (Maybe (ReferenceScriptType o))

class ToCredential a where

Methods

toCredential :: a -> Credential

class ToOutputDatum a where

Methods

toOutputDatum :: a -> OutputDatum

class ToValue a where

Methods

toValue :: a -> Value

class ToScript a where

Methods

toScript :: a -> Versioned Script

class ToScriptHash a where

Methods

toScriptHash :: a -> ScriptHash

type IsTxInfoOutput o = (IsAbstractOutput o, ToCredential (OwnerType o), ToOutputDatum (DatumType o), ToValue (ValueType o), ToScriptHash (ReferenceScriptType o))

An output that can be translated into its script-perspective (as seen on the TxInfo) representation

outputAddress :: (IsAbstractOutput o, ToCredential (OwnerType o)) => o -> Address

outputOutputDatum :: (IsAbstractOutput o, ToOutputDatum (DatumType o)) => o -> OutputDatum

outputValue :: (IsAbstractOutput o, ToValue (ValueType o)) => o -> Value

outputReferenceScriptHash :: (IsAbstractOutput o, ToScriptHash (ReferenceScriptType o)) => o -> Maybe ScriptHash

outputTxOut :: IsTxInfoOutput o => o -> TxOut

Return the output as it is seen by a validator on the TxInfo.

data ConcreteOutput ownerType datumType valueType referenceScriptType where

A type constructed to be the most general instance of IsAbstractOutput.

Constructors

ConcreteOutput :: ownerType -> Maybe StakingCredential -> valueType -> datumType -> Maybe referenceScriptType -> ConcreteOutput ownerType datumType valueType referenceScriptType

toOutputWithReferenceScriptHash :: (IsAbstractOutput output, ToScriptHash (ReferenceScriptType output)) => output -> ConcreteOutput (OwnerType output) (DatumType output) (ValueType output) ScriptHash

Convert the reference script type on the output to ScriptHash.

isOutputWithoutDatum :: IsTxInfoOutput output => output -> Maybe (ConcreteOutput (OwnerType output) () (ValueType output) (ReferenceScriptType output))

Test if there is no datum on an output.

isOutputWithInlineDatum :: IsTxInfoOutput output => output -> Maybe (ConcreteOutput (OwnerType output) Datum (ValueType output) (ReferenceScriptType output))

Test if the output carries some inlined datum.

isOutputWithDatumHash :: IsTxInfoOutput output => output -> Maybe (ConcreteOutput (OwnerType output) DatumHash (ValueType output) (ReferenceScriptType output))

Test if the output carries some datum hash.

isOutputWithInlineDatumOfType :: (FromData a, IsTxInfoOutput output) => output -> Maybe (ConcreteOutput (OwnerType output) a (ValueType output) (ReferenceScriptType output))

Test if the output carries some inlined datum that can be parsed from builtin data on to something of a specific type.

isScriptOutputFrom :: IsTxInfoOutput output => TypedValidator a -> output -> Maybe (ConcreteOutput (TypedValidator a) (DatumType output) (ValueType output) (ReferenceScriptType output))

Test if the owner of an output is a specific typed validator. If it is, return an output with the validator type as its OwnerType.

isPKOutputFrom :: IsTxInfoOutput output => PubKeyHash -> output -> Maybe (ConcreteOutput PubKeyHash (DatumType output) (ValueType output) (ReferenceScriptType output))

Test if the owner of an output is a specific public key. If it is, return an output of the same DatumType, but with PubKeyHash as its OwnerType.

isOnlyAdaOutput :: IsTxInfoOutput output => output -> Maybe (ConcreteOutput (OwnerType output) (DatumType output) Ada (ReferenceScriptType output))

Test if the value on an output contains only Ada.