Cooked.Pretty

Description

This module exposes off-chain pretty-printing functions for transaction skeletons, utxo states, addresses, pubkey hashes, values, etc.

We provide the PrettyCooked class and instances for common Plutus types. We don't rely on Pretty from Prettyprinter in order to define better printers for Plutus types which already have instances of Pretty. Also, PrettyCooked makes it possible to optionally modify pretty printing settings PrettyCookedOpts (e.g. length of printed hashes).

Requirements on datum and redeemers

Datums and redeemers are required to have a PrettyCooked instance.

For trivial datatypes, you can rely on Show by using viaShow from Prettyprinter: prettyCooked = Prettyprinter.viaShow.

For more complex datatypes, you can rely on existing PrettyCooked instances. Prefer implementing the prettyCookedOpt function and relay the PrettyCookedOpts settings to other printers.

@ data Foo = Bar Api.Value | Baz Api.PubkeyHash Api.Value

instance PrettyCooked Foo where prettyCookedOpt pcOpts (Bar value) = Bar + prettyCookedOpt pcOpts value prettyCookedOpt pcOpts (Baz pkh value) = prettyItemize Baz "-" [ "user:" + prettyCookedOpt pcOpts pkh, "deposit:" + prettyCookedOpt pcOpts value ] @

The prettyItemize function is useful to nicely lay down nested lists of elements. Since we manipulate regular Doc values, any function from Prettyprinter can be used to implement your printers.

How to pretty print?

Pretty printing of transaction skeletons and UTxO states is done automatically by the end-user functions provided in Cooked.MockChain.Testing.

To do it manually, use instances of PrettyCooked, PrettyCookedList or PrettyCookedMaybe defined in Skeleton or MockChain such as the one for MockChainReturn a UtxoState.

Documentation

type DocCooked = Doc ()

A standard Doc without any annotation

class PrettyCooked a where

Type class of things that can be pretty printed as a single document. You need to implement either prettyCookedOpt or prettyCooked manually, otherwise calling either of them will resulting in a infinite loop.

Minimal complete definition

Nothing

Methods

prettyCookedOpt :: PrettyCookedOpts -> a -> DocCooked

Pretty prints an element based on some PrettyCookedOpts

prettyCooked :: a -> DocCooked

Pretty prints an element directly

class PrettyCookedList a where

Type class of things that can be pretty printed as a list of documents. Similarly to PrettyCooked, at least of the functions from this class needs to be manually implemented to avoid infinite loops.

Minimal complete definition

Nothing

Methods

prettyCookedOptList :: PrettyCookedOpts -> a -> [DocCooked]

Pretty prints an element as a list on some PrettyCookedOpts

prettyCookedOptListMaybe :: PrettyCookedOpts -> a -> [Maybe DocCooked]

Pretty prints an element as a list of optional documents

prettyCookedList :: a -> [DocCooked]

Pretty prints an elements as a list

class PrettyCookedMaybe a where

Type class of things that can be optionally pretty printed as a document

Minimal complete definition

Nothing

Methods

prettyCookedOptMaybe :: PrettyCookedOpts -> a -> Maybe DocCooked

Pretty prints an optional document on some PrettyCookedOpts

prettyCookedMaybe :: a -> Maybe DocCooked

Pretty prints an option document

class ToHash a where

Hashable elements can be transformed to BuiltinByteString

Methods

toHash :: a -> BuiltinByteString

data PrettyCookedOpts

A set of option to pilot pretty printing in cooked-validators

Constructors

PrettyCookedOpts

Fields

pcOptPrintTxHashes :: Bool

Whether to print transaction ids of validated transactions. By default: False

pcOptPrintTxOutRefs :: PCOptTxOutRefs

Whether to print transaction outputs references. By default: hidden

pcOptPrintDefaultTxSkelOpts :: Bool

Whether to print tx options that have not been modified from their default. By default: False

pcOptNumericUnderscores :: Bool

Whether to print big integers with numeric underscores. For example 53_000_000 instead of 53000000. By default: True

pcOptHashes :: PrettyCookedHashOpts

Options related to printing hashes

pcOptPrintLog :: Bool

Whether to display the log

pcOptPrintConsumedUTxOs :: Bool

Whether to display consumed UTxOs in the end state. Default: False

data PrettyCookedHashOpts

A set of options to pilot how hashes are pretty printed

Constructors

PrettyCookedHashOpts

Fields

pcOptHashLength :: Int

Length of printed hash prefix. By default: 7

pcOptHashNames :: Map BuiltinByteString String

Association between hashes and given names to ease readability. For example Map.singleton (walletPKHash (wallet 1)) Alice By default: "defaultHashNames" which assigns Lovelace, Quick, and Permanent as names for the associated currency symbols

pcOptHashVerbose :: Bool

When a given name exists for a hash, this flag also prints the original hash after the name. By default: False

data PCOptTxOutRefs

Whether to print transaction outputs references.

Constructors

PCOptTxOutRefsHidden
Hide them

PCOptTxOutRefsFull
Always show them.

Warning: this will disable printing similar UTxOs as a group (for instance (×10) Lovelace: 100_000_000)

PCOptTxOutRefsPartial
Show them for UTxOs which are not grouped with similar others. This avoids the downside of PCOptTxOutRefsFull which disables printing UTxOs as a group.

renderString :: (a -> DocCooked) -> a -> String

Use this to convert a pretty-printer to a regular show function using default layout options. This is used in Testing because Tasty uses strings.

printCookedOpt :: PrettyCooked a => PrettyCookedOpts -> a -> IO ()

Use this in the REPL as an alternative to the default print function when dealing with pretty-printable cooked values.

For example, printCookedOpt def runMockChain i0 foo

printCooked :: PrettyCooked a => a -> IO ()

Version of printCookedOpt that uses default pretty printing options.

prettyHash :: ToHash a => PrettyCookedOpts -> a -> DocCooked

Pretty prints hashable elements based on pcOptHashes in the PrettyCookedOpts. This cannot be made an instance as it would be undecidable (the hope was (ToHash a) => PrettyCooked a)

prettyItemize :: PrettyCookedList a => PrettyCookedOpts -> DocCooked -> DocCooked -> a -> DocCooked

Print an item list with a title

>>> prettyCookedOpts opts "Foo" "-" ["bar1", "bar2", "bar3"]
Foo
  - bar1
  - bar2
  - bar3

prettyItemizeNoTitle :: PrettyCookedList a => PrettyCookedOpts -> DocCooked -> a -> DocCooked

Print an item list without a title

prettyItemizeNonEmpty :: PrettyCookedList a => PrettyCookedOpts -> DocCooked -> DocCooked -> a -> Maybe DocCooked

Print an item list with a title, but only when the list is non-empty

hashNamesFromList :: ToHash a => [(a, String)] -> Map BuiltinByteString String

Smart constructor for maps to be used in the "pcOptHashNames" pretty-printing option.

defaultHashNames :: Map BuiltinByteString String

Default hash to names map that assigns Lovelace, Quick, and Permanent to the associated currency symbols. This is used as the default for the pretty-printing option and is recommended to use as a basis to extend with custom names.

addHashNames :: Map BuiltinByteString String -> PrettyCookedOpts -> PrettyCookedOpts

Adds some additional names to these pretty cooked options. This has two practical use cases:

Users can use it in conjuction to hashNamesFromList without having to remember to manually invoke defaultHashNames

We use it internally to account for names that have been registered during mockchain runs, such as for names that depend on on-chain data, typically a TxOutRef.