Cooked.Pretty.Class

Description

We provide the PrettyCooked class and instances for common Plutus types. We don't rely on Pretty from Prettyprinter in order to define better printers for Plutus types which already have instances of Pretty. Also, PrettyCooked makes it possible to optionally modify pretty printing settings PrettyCookedOpts (e.g. length of printed hashes). You just need to implement either prettyCooked or prettyCookedOpt.

Documentation

type DocCooked = Doc ()

data PrettyCookedOpts

Constructors

PrettyCookedOpts

Fields

pcOptPrintTxHashes :: Bool

Whether to print transaction ids of validated transactions. By default: False

pcOptPrintDefaultTxOpts :: Bool

Whether to print tx options that have not been modified from their default. By default: False

pcOptPrintedHashLength :: Int

Length of printed hashes (e.g. addresses, tx ids) By default: 7

class PrettyCooked a where

Minimal complete definition

Nothing

Methods

prettyCooked :: a -> DocCooked

prettyCookedOpt :: PrettyCookedOpts -> a -> DocCooked