Cooked.Pretty.Skeleton

Description

This module implements PrettyCooked, PrettyCookedList and PrettyCookedMaybe instances for TxSkels and its components.

Documentation

data Contextualized a

Some elements of a skeleton can only be printed when they are associated with a context. This is typically the case for elements that need some TxOutRefs and datums to be resolved.

Constructors

Contextualized

Fields

_ctxOutputs :: Map TxOutRef (TxSkelOut, Bool)

ctxContent :: a

Orphan instances

PrettyCooked DatumContent

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> DatumContent -> DocCooked

prettyCooked :: DatumContent -> DocCooked

PrettyCooked Mint
Prints a minting specification

Example: > #abcdef - Redeemer: red - Reference script at: txOutRef - Foo: 500 - Bar: 1000

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> Mint -> DocCooked

prettyCooked :: Mint -> DocCooked

PrettyCooked TxSkelOut

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> TxSkelOut -> DocCooked

prettyCooked :: TxSkelOut -> DocCooked

PrettyCooked TxGovAction

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> TxGovAction -> DocCooked

prettyCooked :: TxGovAction -> DocCooked

PrettyCooked TxParameterChange

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> TxParameterChange -> DocCooked

prettyCooked :: TxParameterChange -> DocCooked

PrettyCooked Wallet

Instance details

Methods

prettyCookedOpt :: PrettyCookedOpts -> Wallet -> DocCooked

prettyCooked :: Wallet -> DocCooked

PrettyCookedList TxOpts
Pretty-print a list of transaction skeleton options, only printing an option if its value is non-default.

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxOpts -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxOpts -> [Maybe DocCooked]

prettyCookedList :: TxOpts -> [DocCooked]

PrettyCookedList TxSkelOut

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelOut -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelOut -> [Maybe DocCooked]

prettyCookedList :: TxSkelOut -> [DocCooked]

PrettyCookedList TxSkelProposal

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelProposal -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelProposal -> [Maybe DocCooked]

prettyCookedList :: TxSkelProposal -> [DocCooked]

PrettyCookedList TxSkelRedeemer
Prints a list of docs corresponding to an instance of TxSkelRedeemer

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> TxSkelRedeemer -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> TxSkelRedeemer -> [Maybe DocCooked]

prettyCookedList :: TxSkelRedeemer -> [DocCooked]

PrettyCookedMaybe TxSkelOutDatum
Prints a TxSkelOutDatum when different from TxSkelOutNoDatum

Instance details

Methods

prettyCookedOptMaybe :: PrettyCookedOpts -> TxSkelOutDatum -> Maybe DocCooked

prettyCookedMaybe :: TxSkelOutDatum -> Maybe DocCooked

PrettyCookedList (TxOpts, [Wallet])
Prints a list of pubkeys with a flag next to the balancing wallet

Instance details

Methods

prettyCookedOptList :: PrettyCookedOpts -> (TxOpts, [Wallet]) -> [DocCooked]

prettyCookedOptListMaybe :: PrettyCookedOpts -> (TxOpts, [Wallet]) -> [Maybe DocCooked]

prettyCookedList :: (TxOpts, [Wallet]) -> [DocCooked]