Cooked.ShowBS

Description

This module exposes on-chain pretty-printing function for all the types that occur on the ScriptContext to BuiltinString. This is useful for debugging of validators. You probably do not want to use this in production code, as many of the functions in this module are wildly inefficient due to limitations of the BuiltinString type.

If the script execution on your transactions go over budget by using this module, consider using txOptEmulatorParamsModification to temporarily loosen the limits (at the cost of breaking compatibility with mainnet)

Documentation

class ShowBS a where

analogue of Haskell's Show class to be use in Plutus scripts.

Methods

showBS :: a -> BuiltinString

analogue of show