Cooked.Skeleton

Description

This module provides the description of a transaction skeleton. We have our own representation of a transaction for three main reasons:

our transaction skeletons are typed (datums, validators, outputs...)

with our own wrapper, we are less affected by plutus updates

we can have default or automated behavior for the parts of the transactions that are less relevant to testing, such as collaterals or fees

Documentation

type LabelConstrs x = (PrettyCooked x, Show x, Typeable x, Eq x, Ord x)

data TxLabel where

Constructors

TxLabel :: LabelConstrs x => x -> TxLabel

data BalanceOutputPolicy

Whether to adjust a potentially existing output to the balancing wallet with the change during transaction balancing.

Constructors

AdjustExistingOutput
Try to adjust an existing public key output with the change. If no suitable output can be found, create a new change output.

DontAdjustExistingOutput
Do not change the existing outputs, always create a new change output.

data FeePolicy

What fee policy to use in the transaction.

Constructors

AutoFeeComputation
Use automatic fee computation. If balancing is activated, an optimal fee will be computed based on the transaction and existing utxos in the balancing wallet. Otherwise, the maximum transaction fee will be applied.

ManualFee Integer
Provide a fee to the transaction. If the autobalancing is activated, it will be attempted around this fee, which might lead to failure if it is too low, otherwise, this fee will be given to transaction generation.

data BalancingPolicy

Whether to balance the transaction or not, and which wallet to use to provide outputs for balancing. Either the first signer or an explicit wallet. In the second case, this wallet must be a signer of the transaction.

Constructors

BalanceWithFirstSigner

BalanceWith Wallet

DoNotBalance

data BalancingUtxos

Which UTxOs to use when balancing. Note that utxos that are already known by the skeleton being balanced (in the sense of txSkelKnownTxOutRefs, i.e. inputs and reference inputs) will be filtered out during balancing.

Constructors

BalancingUtxosFromBalancingWallet
Use all UTxOs containing only a Value (no datum, no staking credential, and no reference script) belonging to the balancing wallet.

BalancingUtxosFromSet (Set TxOutRef)
Use the provided UTxOs. UTxOs belonging to scripts will be filtered out

newtype RawModTx

Wraps a function that will be applied to a transaction right before submission, and after balancing.

Constructors

RawModTxAfterBalancing (Tx ConwayEra -> Tx ConwayEra)

newtype EmulatorParamsModification

Wraps a function that will temporarily change the emulator parameters for the transaction's balancing and submission.

Constructors

EmulatorParamsModification (Params -> Params)

data CollateralUtxos

Describe which UTxOs to use as collaterals

Constructors

CollateralUtxosFromBalancingWallet
Rely on automated computation with only-value UTxOs from the balancing wallet. Return collaterals will be sent to this wallet.

CollateralUtxosFromWallet Wallet
Rely on automated computation with only-value UTxOs from a given wallet. Return collaterals will be sent to this wallet.

CollateralUtxosFromSet (Set TxOutRef) Wallet
Manually provide a set of candidate UTxOs to be used as collaterals alongside a wallet to send return collaterals back to.

applyEmulatorParamsModification :: Maybe EmulatorParamsModification -> Params -> Params

applyRawModOnBalancedTx :: [RawModTx] -> Tx ConwayEra -> Tx ConwayEra

Applies a list of modifications right before the transaction is submitted. The leftmost function in the argument list is applied first.

data TxOpts

Set of options to modify the behavior of generating and validating some transaction.

Constructors

TxOpts

Fields

txOptEnsureMinAda :: Bool

Performs an adjustment to unbalanced transactions, making sure every UTxO that is produced has the necessary minimum amount of Ada.

Default is False.

txOptAutoSlotIncrease :: Bool

Whether to increase the slot counter automatically on transaction submission. This is useful for modelling transactions that could be submitted in parallel in reality, so there should be no explicit ordering of what comes first.

Default is True.

txOptUnsafeModTx :: [RawModTx]

Applies an arbitrary modification to a transaction after it has been potentially adjusted (txOptEnsureMinAda) and balanced. The name of this option contains unsafe to draw attention to the fact that modifying a transaction at that stage might make it invalid. Still, this offers a hook for being able to alter a transaction in unforeseen ways. It is mostly used to test contracts that have been written for custom PABs.

One interesting use of this function is to observe a transaction just before it is being sent for validation, with

txOptUnsafeModTx = [RawModTxAfterBalancing Debug.Trace.traceShowId]

The leftmost function in the list is applied first.

Default is [].

txOptBalancingPolicy :: BalancingPolicy

Whether to balance the transaction or not, and which wallet should provide/reclaim the missing and surplus value. Balancing ensures that

input + mints == output + fees + burns

If you decide to set txOptBalance = DoNotBalance you will have trouble satisfying that equation by hand unless you use ManualFee. You will likely see a error about value preservation.

Default is BalanceWithFirstSigner

txOptFeePolicy :: FeePolicy

The fee to use when balancing the transaction

Default is AutomaticFeeComputation

txOptBalanceOutputPolicy :: BalanceOutputPolicy

The BalanceOutputPolicy to apply when balancing the transaction.

Default is AdjustExistingOutput.

txOptBalancingUtxos :: BalancingUtxos

Which UTxOs to use during balancing. This can either be a precise list, or rely on automatic searches for utxos with values only belonging to the balancing wallet.

Default is BalancingUtxosFromBalancingWallet.

txOptEmulatorParamsModification :: Maybe EmulatorParamsModification

Apply an arbitrary modification to the protocol parameters that are used to balance and submit the transaction. This is obviously a very unsafe thing to do if you want to preserve compatibility with the actual chain. It is useful mainly for testing purposes, when you might want to use extremely big transactions or transactions that exhaust the maximum execution budget. Such a thing could be accomplished with

txOptEmulatorParamsModification = Just $ EmulatorParamsModification increaseTransactionLimits

for example.

Default is Nothing.

txOptCollateralUtxos :: CollateralUtxos

Which utxos to use as collaterals. They can be given manually, or computed automatically from a given, or the balancing, wallet.

Default is CollateralUtxosFromBalancingWallet

txOptEnsureMinAdaL :: Lens' TxOpts Bool

txOptUnsafeModTxL :: Lens' TxOpts [RawModTx]

txOptAutoSlotIncreaseL :: Lens' TxOpts Bool

txOptBalancingPolicyL :: Lens' TxOpts BalancingPolicy

txOptBalanceOutputPolicyL :: Lens' TxOpts BalanceOutputPolicy

txOptFeePolicyL :: Lens' TxOpts FeePolicy

txOptBalancingUtxosL :: Lens' TxOpts BalancingUtxos

txOptEmulatorParamsModificationL :: Lens' TxOpts (Maybe EmulatorParamsModification)

txOptCollateralUtxosL :: Lens' TxOpts CollateralUtxos

type MintsConstrs redeemer = (ToData redeemer, Show redeemer, PrettyCooked redeemer, Typeable redeemer)

data MintsRedeemer where

Which redeemer to use for minting. Note that using NoMintsRedeemer corresponds to the redeemer () on-chain.

Constructors

NoMintsRedeemer :: MintsRedeemer

SomeMintsRedeemer :: MintsConstrs redeemer => redeemer -> MintsRedeemer

type TxSkelMints = Map (Versioned MintingPolicy) (MintsRedeemer, NEMap TokenName (NonZero Integer))

A description of what a transaction mints. For every policy, there can only be one MintsRedeemer, and if there is, there must be some token names, each with a non-zero amount of tokens.

You'll probably not construct this by hand, but use txSkelMintsFromList.

addToTxSkelMints :: (Versioned MintingPolicy, MintsRedeemer, TokenName, Integer) -> TxSkelMints -> TxSkelMints

Add a new entry to a TxSkelMints. There are a few wrinkles:

If for a given policy, redeemer, and token name, there are n tokens in the argument TxSkelMints, and you add -n tokens, the corresponding entry in the "inner map" of the policy will disappear (obviously, because all of its values have to be non-zero). If that also means that the inner map becomes empty, the policy will disappear from the TxSkelMints altogether.

If a policy is already present on the argument TxSkelMints with a redeemer a, and you add a mint with a different redeemer b, the old redeemer is thrown away. The values associated with the token names of that policy are added as described above, though. This means that any pre-existing values will be minted with a new redeemer.

If, for some reason, you really want to generate a TxSkelMints that has both a negative and a positive entry of the same asset class and redeemer, you'll have to do so manually. Note, however, that even if you do so, NO VALIDATOR OR MINTING POLICY WILL EVER GET TO SEE A TRANSACTION WITH SUCH CONFLICTING INFORMATION. This is not a design decision/limitation of cooked-validators: The Cardano API TxBodyContent type, that we're translating everything into eventually, stores minting information as a minted value together with a map from policy IDs to witnesses (which represent the used redeemers). That means that we can only store _one_ redeemer per minting policy, and no conflicting mints of the same asset class, since they'll just cancel.

txSkelMintsToList :: TxSkelMints -> [(Versioned MintingPolicy, MintsRedeemer, TokenName, Integer)]

Convert from TxSkelMints to a list of tuples describing eveything that's being minted.

txSkelMintsFromList :: [(Versioned MintingPolicy, MintsRedeemer, TokenName, Integer)] -> TxSkelMints

Smart constructor for TxSkelMints. This function relies on addToTxSkelMints. So, some non-empty lists (where all amounts for a given asset class an redeemer add up to zero) might be translated into the empty TxSkelMints.

txSkelMintsValue :: TxSkelMints -> Value

The value described by a TxSkelMints

txSkelOutValueL :: Lens' TxSkelOut Value

txSkelOutDatumL :: Lens' TxSkelOut TxSkelOutDatum

txSkelOutValue :: TxSkelOut -> Value

txSkelOutValidator :: TxSkelOut -> Maybe (Versioned Validator)

type TxSkelOutDatumConstrs a = (Show a, PrettyCooked a, ToData a, Eq a, Typeable a)

data TxSkelOutDatum where

On transaction outputs, we have the options to use

no datum

only a datum hash

a "normal" datum

an inline datum

These four options are also what the type TxSkelOutDatum records. The following table explains their differences.

datum stored in in the simulated chain state	 datum resolved on the txInfoData	OutputDatum constructor seen by the validator

TxSkelOutNoDatum	no	no	NoOutputDatum

TxSkelOutDatumHash	yes	no	OutputDatumHash

TxSkelOutDatum	yes	yes	OutputDatumHash

TxSkelOutInlineDatum	yes	no	OutputDatum

That is:

Whenever there is a datum, we'll store it in the state of our simulated chain. This will make it possible to retrieve it later, using functions such as datumFromHash.

Both of the TxSkelOutDatumHash and TxSkelOutDatum constructors will create an output that scripts see on the txInfo as having a datum hash. The difference is whether that hash will be resolvable using validator functions like findDatum.

Constructors

TxSkelOutNoDatum :: TxSkelOutDatum
use no datum

TxSkelOutDatumHash :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
only include the hash on the transaction

TxSkelOutDatum :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
use a OutputDatumHash on the transaction output, but generate the transaction in such a way that the complete datum is included in the txInfoData seen by validators

TxSkelOutInlineDatum :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
use an inline datum

data TxSkelOut where

Transaction outputs. The Pays constructor is really general, and you'll probably want to use one of the smart constructors like paysScript or paysPK in most cases.

Constructors

Pays

Fields

:: (Show o, Typeable o, IsTxInfoOutput o, IsTxSkelOutAllowedOwner (OwnerType o), Typeable (OwnerType o), ToCredential (OwnerType o), DatumType o ~ TxSkelOutDatum, ValueType o ~ Value, ToScript (ReferenceScriptType o), Show (OwnerType o), Show (ReferenceScriptType o), Typeable (ReferenceScriptType o))

=> { producedOutput :: o

} -> TxSkelOut

txSkelOutTypedDatum :: FromData a => TxSkelOutDatum -> Maybe a

txSkelOutUntypedDatum :: TxSkelOutDatum -> Maybe Datum

paysPK :: ToPubKeyHash a => a -> Value -> TxSkelOut

Pay a certain value to a public key.

paysScript :: (ToData (DatumType a), Show (DatumType a), Typeable (DatumType a), Eq (DatumType a), PrettyCooked (DatumType a), Typeable a) => TypedValidator a -> DatumType a -> Value -> TxSkelOut

Pays a script a certain value with a certain datum, using the TxSkelOutDatum constructor. (See the documentation of TxSkelOutDatum.)

paysScriptInlineDatum :: (ToData (DatumType a), Show (DatumType a), Typeable (DatumType a), Eq (DatumType a), PrettyCooked (DatumType a), Typeable a) => TypedValidator a -> DatumType a -> Value -> TxSkelOut

Pays a script a certain value with a certain inlined datum.

paysScriptDatumHash :: (ToData (DatumType a), Show (DatumType a), Typeable (DatumType a), Eq (DatumType a), PrettyCooked (DatumType a), Typeable a) => TypedValidator a -> DatumType a -> Value -> TxSkelOut

Pays a script a certain value with a certain hashed (not resolved in transaction) datum.

paysScriptNoDatum :: Typeable a => TypedValidator a -> Value -> TxSkelOut

Pays a script a certain value without any datum. Intended to be used with withDatum, withDatumHash, or withInlineDatum to try a datum whose type does not match the validator's.

withDatum :: (ToData a, Show a, Typeable a, Eq a, PrettyCooked a) => TxSkelOut -> a -> TxSkelOut

Set the datum in a payment to the given datum (whose type may not fit the typed validator in case of a script).

withInlineDatum :: (ToData a, Show a, Typeable a, Eq a, PrettyCooked a) => TxSkelOut -> a -> TxSkelOut

Set the datum in a payment to the given inlined datum (whose type may not fit the typed validator in case of a script).

withDatumHash :: (ToData a, Show a, Typeable a, Eq a, PrettyCooked a) => TxSkelOut -> a -> TxSkelOut

Set the datum in a payment to the given hashed (not resolved in the transaction) datum (whose type may not fit the typed validator in case of a script).

withReferenceScript :: (Show script, ToScript script, Typeable script, ToScriptHash script) => TxSkelOut -> script -> TxSkelOut

Add a reference script to a transaction output (or replace it if there is already one)

withStakingCredential :: TxSkelOut -> StakingCredential -> TxSkelOut

Add a staking credential to a transaction output (or replace it if there is already one)

data TxSkelRedeemer where

Constructors

TxSkelNoRedeemerForPK :: TxSkelRedeemer

TxSkelRedeemerForScript :: SpendsScriptConstrs redeemer => redeemer -> TxSkelRedeemer

TxSkelRedeemerForReferencedScript :: SpendsScriptConstrs redeemer => TxOutRef -> redeemer -> TxSkelRedeemer
The first argument is a reference to the output where the referenced script is stored.

txSkelTypedRedeemer :: FromData (RedeemerType a) => TxSkelRedeemer -> Maybe (RedeemerType a)

data TxSkel where

Constructors

TxSkel

Fields

:: { txSkelLabel :: Set TxLabel

Labels do not influence the transaction generation at all; they are pretty-printed whenever cooked-validators prints a transaction, and can therefore make the output more informative (and greppable).

, txSkelOpts :: TxOpts

Some options that control transaction generation.

, txSkelMints :: TxSkelMints

Any value minted or burned by the transaction. You'll probably want to use txSkelMintsFromList to construct this.

, txSkelSigners :: [Wallet]

The wallets signing the transaction. This list must contain at least one element. By default, the first signer will pay for fees and balancing. You can change that with txOptBalanceWallet.

, txSkelValidityRange :: SlotRange

, txSkelIns :: Map TxOutRef TxSkelRedeemer

To each TxOutRef the transaction should consume, add a redeemer specifying how to spend it. You must make sure that

On TxOutRefs referencing UTxOs belonging to public keys, you use the TxSkelNoRedeemerForPK constructor.

On TxOutRefs referencing UTxOs belonging to scripts, you must make sure that the type of the redeemer is appropriate for the script.

, txSkelInsReference :: Set TxOutRef

All outputs referenced by the transaction.

, txSkelOuts :: [TxSkelOut]

The outputs of the transaction. These will occur in exactly this order on the transaction.

} -> TxSkel

txSkelLabelL :: Lens' TxSkel (Set TxLabel)

txSkelOptsL :: Lens' TxSkel TxOpts

txSkelMintsL :: Lens' TxSkel TxSkelMints

txSkelValidityRangeL :: Lens' TxSkel SlotRange

txSkelSignersL :: Lens' TxSkel [Wallet]

txSkelInsL :: Lens' TxSkel (Map TxOutRef TxSkelRedeemer)

txSkelInsReferenceL :: Lens' TxSkel (Set TxOutRef)

txSkelOutsL :: Lens' TxSkel [TxSkelOut]

txSkelTemplate :: TxSkel

A convenience template of an empty transaction skeleton.

txSkelDataInOutputs :: TxSkel -> Map DatumHash TxSkelOutDatum

Return all data on transaction outputs.

txSkelValidatorsInOutputs :: TxSkel -> Map ValidatorHash (Versioned Validator)

All validators which will receive transaction outputs

txSkelOutOwnerTypeP :: forall ownerType. (ToCredential ownerType, Show ownerType, IsTxSkelOutAllowedOwner ownerType, Typeable ownerType) => Prism' TxSkelOut (ConcreteOutput ownerType TxSkelOutDatum Value (Versioned Script))

Decide if a transaction output has a certain owner and datum type.

txSkelOutputDatumTypeAT :: (FromData a, Typeable a) => AffineTraversal' TxSkelOut a

data SkelContext

The missing information on a TxSkel that can only be resolved by querying the state of the blockchain.

Constructors

SkelContext

Fields

skelContextTxOuts :: Map TxOutRef TxOut

skelContextTxSkelOutDatums :: Map DatumHash TxSkelOutDatum

txSkelReferenceScript :: TxSkelRedeemer -> Maybe TxOutRef

txSkelKnownTxOutRefs :: TxSkel -> [TxOutRef]

All TxOutRefs known by a given transaction skeleton. This includes TxOutRef`s used as inputs of the skeleton and TxOutRefs used as reference inputs of the skeleton. This does not include additional possible TxOutRefs used for balancing and additional TxOutRefs used as collateral inputs, as they are not part of the skeleton.

txSkelValueInOutputs :: TxSkel -> Value

Returns the full value contained in the skeleton outputs

txSkelReferenceScripts :: TxSkel -> Map ValidatorHash (Versioned Validator)

All validators in the reference script field of transaction outputs