Cooked.Skeleton

Documentation

type LabelConstrs x = (Show x, Typeable x, Eq x, Ord x)

data TxLabel where

Constructors

TxLabel :: LabelConstrs x => x -> TxLabel

data BalanceOutputPolicy

Whether to adjust existing public key outputs during transaction balancing. TODO: Why do we need these two options? Are they just historical baggage?

Constructors

AdjustExistingOutput
Try to adjust an existing public key output with the change. If no suitable output can be found, create a new change output.

DontAdjustExistingOutput
Do not change the existing outputs, always create a new change output.

data BalancingWallet

Which wallet to use to provide outputs for balancing and collaterals. Either the first signer or an explicit wallet. In the second case, this wallet must be a signer of the transaction.

Constructors

BalanceWithFirstSigner

BalanceWith Wallet

newtype RawModTx

Wraps a function that will be applied to a transaction right before submitting it.

Constructors

RawModTxAfterBalancing (Tx BabbageEra -> Tx BabbageEra)
Apply modification on transaction after balancing, fee calculation, and final signing areperformed

applyRawModOnBalancedTx :: [RawModTx] -> Tx BabbageEra -> Tx BabbageEra

Applies a list of modifications right before the transaction is submitted. The leftmost function in the argument list is applied first.

data TxOpts

Set of options to modify the behavior of generating and validating some transaction. Some of these options only have an effect when running in the Contract, some only have an effect when running in MockChainT. If nothing is explicitely stated, the option has an effect independently of the running context.

TODO Refactor field names to avoid clashes on common terms such as "collateral" or "balance"

Constructors

TxOpts

Fields

txOptEnsureMinAda :: Bool

Performs an adjustment to unbalanced transactions, making sure every UTxO that is produced has the necessary minimum amount of Ada.

By default, this is set to False, given this is the default behavior in Plutus: https://github.com/input-output-hk/plutus-apps/issues/143#issuecomment-1013012744

txOptAwaitTxConfirmed :: Bool

When submitting a transaction for real (i.e., running in the Contract monad), it is common to call awaitTxConfirmed after submitTxConstraints. If you do NOT wish to do so, please set this to False.

This has NO effect when running outside of Contract. By default, this is set to True.

txOptAutoSlotIncrease :: Bool

Whether to increase the slot counter automatically on this submission. This is useful for modelling transactions that could be submitted in parallel in reality, so there should be no explicit ordering of what comes first. One good example is in the Crowdfunding use case contract.

This has NO effect when running in Contract. By default, this is set to True.

txOptUnsafeModTx :: [RawModTx]

Applies an arbitrary modification to a transaction after it has been potentially adjusted ('txOptEnsureMinAda) and balanced. This is prefixed with unsafe to draw attention to the fact that modifying a transaction at that stage might make it invalid. Still, this offers a hook for being able to alter a transaction in unforeseen ways. It is mostly used to test contracts that have been written for custom PABs.

One interesting use of this function is to observe a transaction just before it is being sent for validation, with unsafeModTx = [RawModTxAfterBalancing Debug.Trace.traceShowId].

This has NO effect when running in Contract. By default, this is set to the empty list.

The leftmost function in the list is applied first.

txOptBalance :: Bool

Whether to balance the transaction or not. Balancing ensures that input + mint = output + fees + burns, if you decide to set balance = false you will have trouble satisfying that equation by hand because fees are variable. You will likely see a ValueNotPreserved error and should adjust the fees accordingly.

This has NO effect when running in Contract. By default, this is set to True.

txOptBalanceOutputPolicy :: BalanceOutputPolicy

The BalanceOutputPolicy to apply when balancing the transaction.

This has NO effect when running in Contract. By default, this is set to AdjustExistingOutput.

txOptBalanceWallet :: BalancingWallet

Which wallet to use to provide outputs for balancing and collaterals. Either the first signer by default, or an explicit wallet. In the second case, this wallet must be a signer of the transaction. This option WILL NOT ensure that it is added in case it is not already present in the list of signers.

type MintsConstrs redeemer = (ToData redeemer, Show redeemer, PrettyCooked redeemer, Typeable redeemer)

data MintsRedeemer where

Constructors

NoMintsRedeemer :: MintsRedeemer

SomeMintsRedeemer :: MintsConstrs redeemer => redeemer -> MintsRedeemer

type TxSkelMints = Map (Versioned MintingPolicy) (MintsRedeemer, NEMap TokenName (NonZero Integer))

A description of what a transaction mints: For every policy, there can only be one redeemer, and if there is a redeemer, there mus be some token names, each with a non-zero amount of tokens.

addToTxSkelMints :: (Versioned MintingPolicy, MintsRedeemer, TokenName, NonZero Integer) -> TxSkelMints -> TxSkelMints

Add a new entry to a TxSkelMints. There are a few wrinkles:

If for a given policy, redeemer, and token name, there are n tokens in the argument TxSkelMints, and you add -n tokens, the corresponding entry in the "inner map" of the policy will disappear (obviously, because all of its values have to be non-zero). If that also means that the inner map becomes empty, the policy will disappear from the TxSkelMints altogether.

If a policy is already present on the argument TxSkelMints with a redeemer a, and you add a mint with a different redeemer b, the old redeemer is thrown away. This is because, on the plutus-apps Tx type, every minting policy can have only one redeemer per transaction. The values associated with the token names of that policy are added as described above, though. This means that any pre-exixting values will be minted with a new redeemer.

If, for some reason, you really want to generate a TxSkelMints that has both a negative and a positive entry of the same asset class and redeemer, you'll have to do so manually. Note, however, that even if you do so, NO VALIDATOR OR MINTING POLICY WILL EVER GET TO SEE A TRANSACTION WITH SUCH CONFLICTING INFORMATION. This is not a design decision/limitation of cooked-validators: The Cardano API TxBodyContent type, that we're translating everything into eventually, stores minting information as a minted value together with a map from policy IDs to witnesses (which represent the used redeemers). That means that we can only store _one_ redeemer per minting policy, and no conflicting mints of the same asset class, since they'll just cancel.

txSkelMintsToList :: TxSkelMints -> [(Versioned MintingPolicy, MintsRedeemer, TokenName, NonZero Integer)]

txSkelMintsFromList :: [(Versioned MintingPolicy, MintsRedeemer, TokenName, NonZero Integer)] -> TxSkelMints

This function relies on the Monoid instance of TxSkelMints. So, some non-empty lists (where all amounts for a given asset class an redeemer add up to zero) might be translated into the empty TxSkelMints. (See the comment at the Semigroup instance definition of TxSkelMints, and at the definition of addToTxSkelMints.)

mintsListIso :: Iso' TxSkelMints [(Versioned MintingPolicy, MintsRedeemer, TokenName, NonZero Integer)]

Convert between TxSkelMints and a list of tuples describing eveything that's being minted. This is implemented in terms of txSkelMintsFromList (see the comment at that function). The upshot is that

review mintsListIso . view mintsListIso

is the identity on TxSkelMints, but

view mintsListIso . review mintsListIso

is NOT THE IDENTITY on [(Pl.MintingPolicy, MintsRedeemer, Pl.TokenName, NonZero Integer)].

txSkelMintsValue :: TxSkelMints -> Value

The value described by a TxSkelMints

txSkelOutValueL :: Lens' TxSkelOut Value

txSkelOutDatumL :: Lens' TxSkelOut TxSkelOutDatum

txSkelOutValue :: TxSkelOut -> Value

txSkelOutValidator :: TxSkelOut -> Maybe (Versioned Validator)

type TxSkelOutDatumConstrs a = (Show a, PrettyCooked a, ToData a, Eq a, Typeable a)

data TxSkelOutDatum where

See the [note on TxSkelOut data]

Constructors

TxSkelOutNoDatum :: TxSkelOutDatum
use no datum

TxSkelOutDatumHash :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
only include the hash on the transaction

TxSkelOutDatum :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
use a OutputDatumHash on the transaction output, but generate the transaction in such a way that the complete datum is included in the txInfoData seen by validators

TxSkelOutInlineDatum :: TxSkelOutDatumConstrs a => a -> TxSkelOutDatum
use an inline datum

data TxSkelOut where

Constructors

Pays

Fields

:: (Show o, IsOnchainOutput o, IsTxSkelOutAllowedOwner (OwnerType o), Typeable (OwnerType o), ToCredential (OwnerType o), DatumType o ~ TxSkelOutDatum, ValueType o ~ Value, ToScript (ReferenceScriptType o))

=> { producedOutput :: o

} -> TxSkelOut

paysPK :: PubKeyHash -> Value -> TxSkelOut

Pay a certain value to a public key, without using a datum.

paysPKWithReferenceScript :: PubKeyHash -> Value -> TypedValidator a -> TxSkelOut

Pay a certain value to a public key, without a datum, but including a reference script. This can be used to put reference scripts on chain.

txSkelOutTypedDatum :: FromData a => TxSkelOutDatum -> Maybe a

txSkelOutUntypedDatum :: TxSkelOutDatum -> Maybe Datum

paysScript :: (ToData (DatumType a), Show (DatumType a), Typeable (DatumType a), Eq (DatumType a), PrettyCooked (DatumType a), Typeable a) => TypedValidator a -> DatumType a -> Value -> TxSkelOut

Pays a script a certain value with a certain datum, which will be included as a datum hash on the transaction.

paysScriptInlineDatum :: (ToData (DatumType a), Show (DatumType a), Typeable (DatumType a), Eq (DatumType a), PrettyCooked (DatumType a), Typeable a) => TypedValidator a -> DatumType a -> Value -> TxSkelOut

Like paysScript, but using an inline datum.

paysScriptDatumHash :: (ToData (DatumType a), Show (DatumType a), Typeable (DatumType a), Eq (DatumType a), PrettyCooked (DatumType a), Typeable a) => TypedValidator a -> DatumType a -> Value -> TxSkelOut

Like paysScript, but won't include the complete datum on the transaction. This is only useful if there's no script that checks the output datum.

data TxSkelRedeemer where

Constructors

TxSkelNoRedeemerForPK :: TxSkelRedeemer

TxSkelRedeemerForScript :: SpendsScriptConstrs redeemer => redeemer -> TxSkelRedeemer

TxSkelRedeemerForReferencedScript :: SpendsScriptConstrs redeemer => redeemer -> TxSkelRedeemer

txSkelTypedRedeemer :: FromData (RedeemerType a) => TxSkelRedeemer -> Maybe (RedeemerType a)

data TxSkel where

Constructors

TxSkel

Fields

:: { txSkelLabel :: Set TxLabel

, txSkelOpts :: TxOpts

, txSkelMints :: TxSkelMints

, txSkelValidityRange :: POSIXTimeRange

, txSkelSigners :: [Wallet]

, txSkelIns :: Map TxOutRef TxSkelRedeemer

, txSkelInsReference :: Set TxOutRef

, txSkelOuts :: [TxSkelOut]

} -> TxSkel

txSkelLabelL :: Lens' TxSkel (Set TxLabel)

txSkelOptsL :: Lens' TxSkel TxOpts

txSkelMintsL :: Lens' TxSkel TxSkelMints

txSkelValidityRangeL :: Lens' TxSkel POSIXTimeRange

txSkelSignersL :: Lens' TxSkel [Wallet]

txSkelInsL :: Lens' TxSkel (Map TxOutRef TxSkelRedeemer)

txSkelInsReferenceL :: Lens' TxSkel (Set TxOutRef)

txSkelOutsL :: Lens' TxSkel [TxSkelOut]

txSkelTemplate :: TxSkel

A convenience template where wallet 1 is the default signer of an otherwise empty transaction skeleton.

txSkelOutputData :: TxSkel -> Map DatumHash TxSkelOutDatum

Return all data on transaction outputs.

newtype Fee

Constructors

Fee

Fields

feeLovelace :: Integer

txSkelOutputValue :: TxSkel -> Fee -> Value

The value in all transaction inputs, plus the negative parts of the minted value. This is the right hand side of the "balancing equation":

mints + inputs = fees + burns + outputs

flattenValueI :: Iso' Value [(AssetClass, Integer)]

txSkelOutValidators :: TxSkel -> Map ValidatorHash (Versioned Validator)

All validators protecting transaction outputs

positivePart :: Value -> Value

The positive part of a value. For every asset class in the given value, this asset class and its amount are included in the output iff the amount is strictly positive. It holds

x == positivePart x <> Pl.negate negativePart x

negativePart :: Value -> Value

The negative part of a value. For every asset class in the given value, this asset class and its negated amount are included in the output iff the amount is strictly negative. It holds

x == positivePart x <> Pl.negate negativePart x

adaL :: Lens' Value Ada

Focus the Ada part in a value. This is useful if you want to chcange only that part.

txSkelOutOwnerTypeP :: forall ownerType. (ToCredential ownerType, Show ownerType, IsTxSkelOutAllowedOwner ownerType, Typeable ownerType) => Prism' TxSkelOut (ConcreteOutput ownerType TxSkelOutDatum Value (Versioned Script))

Decide if a transaction output has a certain owner and datum type.

txSkelOutputDatumTypeAT :: (FromData a, Typeable a) => AffineTraversal' TxSkelOut a

data SkelContext

Constructors

SkelContext

Fields

skelContextTxOuts :: Map TxOutRef TxOut

skelContextTxSkelOutDatums :: Map DatumHash TxSkelOutDatum

txSkelOutReferenceScripts :: TxSkel -> Map ValidatorHash (Versioned Validator)

All validators in the reference script field of transaction outputs