Cooked.Skeleton

Description

This module provides the description of a transaction skeleton. We have our own representation of a transaction for many reasons. Here are some of them:

our transaction skeletons are typed (datums, validators, outputs...)

each transaction skeleton comes with its own set of generation options

our transaction skeleton is by default anchored in the latest Cardano era

each field in our transaction skeleton comes with a set of helpers and

smart constructor to ease the transaction creation

we can have default or automated behavior for the parts of the transactions that are less relevant to testing, such as collaterals or fees

Documentation

module Cooked.Skeleton.Withdrawal

module Cooked.Skeleton.Value

module Cooked.Skeleton.Redeemer

module Cooked.Skeleton.Proposal

module Cooked.Skeleton.Payable

module Cooked.Skeleton.Output

module Cooked.Skeleton.Option

module Cooked.Skeleton.Mint

module Cooked.Skeleton.Label

module Cooked.Skeleton.Datum

data TxSkel where

A transaction skeleton. This is cooked-validators's variant of transaction bodies, eventually translated to Cardano TxBody.

Constructors

TxSkel

Fields

:: { txSkelLabel :: Set TxLabel

Labels do not influence the transaction generation at all; they are pretty-printed whenever cooked-validators prints a transaction, and can therefore make the output more informative (and greppable).

, txSkelOpts :: TxOpts

Some options that control transaction generation.

, txSkelMints :: TxSkelMints

Any value minted or burned by the transaction. You'll probably want to use txSkelMintsFromList to construct this.

, txSkelSigners :: [Wallet]

The wallets signing the transaction. This list must contain at least one element. By default, the first signer will pay for fees and balancing. You can change that with txOptBalancingPolicy.

, txSkelValidityRange :: SlotRange

, txSkelIns :: Map TxOutRef TxSkelRedeemer

To each TxOutRef the transaction should consume, add a redeemer specifying how to spend it. You must make sure that

On TxOutRefs referencing UTxOs belonging to public keys, use the emptyTxSkelRedeemer smart constructor.

On TxOutRefs referencing UTxOs belonging to scripts, use the someTxSkelRedeemer smart constructor.

, txSkelInsReference :: Set TxOutRef

All outputs directly referenced by the transaction. Note that additional reference inputs can be found within the various redeemers of this skeleton. Function txSkelReferenceTxOutRefs collects them all.

, txSkelOuts :: [TxSkelOut]

The outputs of the transaction. These will occur in exactly this order on the transaction.

, txSkelProposals :: [TxSkelProposal]

Possible proposals issued in this transaction to be voted on and possible enacted later on.

, txSkelWithdrawals :: TxSkelWithdrawals

Withdrawals performed by the transaction

} -> TxSkel

txSkelLabelL :: Lens' TxSkel (Set TxLabel)

A lens to set of get labels from a TxSkel

txSkelOptsL :: Lens' TxSkel TxOpts

A lens to set of get options from a TxSkel

txSkelMintsL :: Lens' TxSkel TxSkelMints

A lens to set of get the minted value of a TxSkel

txSkelValidityRangeL :: Lens' TxSkel SlotRange

A lens to set of get the validity range of a TxSkel

txSkelProposalsL :: Lens' TxSkel [TxSkelProposal]

A lens to set of get proposals from a TxSkel

txSkelSignersL :: Lens' TxSkel [Wallet]

A lens to set of get signers from a TxSkel

txSkelInsL :: Lens' TxSkel (Map TxOutRef TxSkelRedeemer)

A lens to set of get inputs from a TxSkel

txSkelInsReferenceL :: Lens' TxSkel (Set TxOutRef)

A lens to set of get reference inputs from a TxSkel

txSkelOutsL :: Lens' TxSkel [TxSkelOut]

A lens to set of get outputs from a TxSkel

txSkelWithdrawalsL :: Lens' TxSkel TxSkelWithdrawals

A lens to set of get withdrawals from a TxSkel

txSkelTemplate :: TxSkel

A convenience template of an empty transaction skeleton.

txSkelDataInOutputs :: TxSkel -> [(DatumHash, TxSkelOutDatum)]

Returns all data on transaction outputs. This can contain duplicates, which is intended.

txSkelKnownTxOutRefs :: TxSkel -> [TxOutRef]

All TxOutRefs known by a given transaction skeleton. This includes TxOutRef`s used as inputs of the skeleton and TxOutRefs used as reference inputs of the skeleton. This does not include additional possible TxOutRefs used for balancing and additional TxOutRefs used as collateral inputs, as they are not part of the skeleton.

txSkelWithdrawnValue :: TxSkel -> Value

Returns the total value withdrawn in this TxSkel

txSkelWithdrawalsScripts :: TxSkel -> [Versioned Script]

Returns all the scripts involved in withdrawals in this TxSkel

txSkelValueInOutputs :: TxSkel -> Value

Returns the full value contained in the skeleton outputs

txSkelReferenceTxOutRefs :: TxSkel -> [TxOutRef]

All TxOutRefs in reference inputs