Cooked.Skeleton.Datum

Description

This module exposes the notion of datums as they are handled within a TxSkel

Documentation

type DatumConstrs a = (Show a, PrettyCooked a, ToData a, Eq a, Typeable a)

Type constraints that must be satisfied by the datum content

data DatumContent where

Data type of wrapped datums satisfying DatumConstrs

Constructors

DatumContent :: DatumConstrs a => a -> DatumContent
Wraps an element satisfying DatumConstrs

datumContentToDatum :: DatumContent -> Datum

Extracts the datum from a DatumContent

datumContentToDatumHash :: DatumContent -> DatumHash

Extracts the datum hash from a DatumContent

data DatumResolved

Whether the datum should be resolved in the transaction

Constructors

NotResolved
Do not resolve the datum (absent from txInfoData)

Resolved
Resolve the datum (present from txInfoData)

data DatumKind

Options on how to include the datum in the transaction

Constructors

Inline
Include the full datum in the UTxO

Hashed DatumResolved
Only include the datum hash in the UTxO. Resolve, or do not resolve, the full datum in the transaction body.

data TxSkelOutDatum where

Datums to be placed in TxSkel outputs, which are either empty, or composed of a datum content and its placement

Constructors

TxSkelOutNoDatum :: TxSkelOutDatum
use no datum

TxSkelOutSomeDatum :: DatumContent -> DatumKind -> TxSkelOutDatum
use some datum content and associated placement

txSkelOutDatumHash :: TxSkelOutDatum -> Maybe DatumHash

Converts a TxSkelOutDatum into a possible Plutus datum hash

txSkelOutUntypedDatum :: TxSkelOutDatum -> Maybe Datum

Converts a TxSkelOutDatum into a possible Plutus datum

datumContentTypedDatumAT :: DatumConstrs a => AffineTraversal' DatumContent a

Extracts a typed datum for a DatumContent when of the right type

txSkelOutDatumContentAT :: AffineTraversal' TxSkelOutDatum DatumContent

Extracts or changes the DatumContent of a TxSkelOutDatum

txSkelOutTypedDatumAT :: DatumConstrs a => AffineTraversal' TxSkelOutDatum a

Extracts or changes the inner typed datum of a TxSkelOutDatum