Cooked.Skeleton.Label

Description

Labels are arbitrary information that can be added to skeleton. They are meant to be pretty-printed. The common use case we currently have is to tag skeletons that have been modified by tweaks and automated attacks.

Documentation

type LabelConstrs x = (PrettyCooked x, Show x, Typeable x, Eq x, Ord x)

data TxLabel where

Constructors

TxLabel :: LabelConstrs x => x -> TxLabel