Cooked.Skeleton.Output

Documentation

data TxSkelOut where

Transaction outputs. The Pays constructor is really general, and you'll probably want to use the receives smart constructor in most cases.

Constructors

Pays :: (Show o, Typeable o, IsTxInfoOutput o, IsTxSkelOutAllowedOwner (OwnerType o), ToCredential (OwnerType o), Typeable (OwnerType o), DatumType o ~ TxSkelOutDatum, ValueType o ~ Value, ToVersionedScript (ReferenceScriptType o), Show (OwnerType o), Show (ReferenceScriptType o), Typeable (ReferenceScriptType o)) => o -> TxSkelOut

receives :: (Show owner, Typeable owner, IsTxSkelOutAllowedOwner owner, ToCredential owner) => owner -> Payable els -> TxSkelOut

Smart constructor to build TxSkelOut from an owner and payment. This should be the main way of building outputs.

txSkelOutValueL :: Lens' TxSkelOut Value

txSkelOutDatumL :: Lens' TxSkelOut TxSkelOutDatum

txSkelOutValue :: TxSkelOut -> Value

txSkelOutValidator :: TxSkelOut -> Maybe (Versioned Validator)

txSkelOutOwnerTypeP :: forall ownerType. (ToCredential ownerType, Show ownerType, IsTxSkelOutAllowedOwner ownerType, Typeable ownerType) => Prism' TxSkelOut (ConcreteOutput ownerType TxSkelOutDatum Value (Versioned Script))

Decide if a transaction output has a certain owner and datum type.

txSkelOutputDatumTypeAT :: (FromData a, Typeable a) => AffineTraversal' TxSkelOut a