Cooked.Skeleton.Output

Description

This module exposes the outputs constructs used in a TxSkel and their associated utilities. To build payments in a skeleton, the usual way is to invoke txSkelIns = [pk receives Value v, script receives (InlineDatum dat && ReferenceScript script)]

Type constraints

class IsTxSkelOutAllowedOwner a where

A conveniency typeclass to automated the creation of TxSkelOut owners, to be used alongside Payable with the smart constructor receives.

Methods

toPKHOrVScript :: a -> User IsEither Allocation

Data types

data PayableKind where

The kind of possible components of a TxSkelOut, other than the owner

Constructors

IsDatum :: PayableKind

IsReferenceScript :: PayableKind

IsValue :: PayableKind

IsStakingCredential :: PayableKind

data Payable :: [PayableKind] -> Type where

Payable elements. Created from concrete elements or composed. Notice that there is no way of building an element of Type Payable '[] so when using an element of Type Payable els we are sure that something was in fact paid. Also, there is no way of building an element of type Payable '[a,a] so we also know at most one occurrence of each type of payment is performed.

Constructors

VisibleHashedDatum :: DatumConstrs a => a -> Payable '[IsDatum]
Hashed datums visible in the transaction are payable

InlineDatum :: DatumConstrs a => a -> Payable '[IsDatum]
Inline datums are payable

HiddenHashedDatum :: DatumConstrs a => a -> Payable '[IsDatum]
Hashed datums hidden from the transaction are payable

ReferenceScript :: ToVScript s => s -> Payable '[IsReferenceScript]
Reference scripts are payable

Value :: ToValue a => a -> Payable '[IsValue]
Values are payable and are subject to min ada adjustment

FixedValue :: ToValue a => a -> Payable '[IsValue]
Fixed Values are payable but are NOT subject to min ada adjustment

StakingCredential :: ToMaybeStakingCredential cred => cred -> Payable '[IsStakingCredential]
Staking credentials are payable

PayableAnd :: els ⩀ els' => Payable els -> Payable els' -> Payable (els ∪ els')
Payables can be combined as long as their list of tags are disjoint

data TxSkelOut where

An output to be put into a TxSkel

Constructors

TxSkelOut

Fields

:: { txSkelOutOwner :: User IsEither Allocation

, txSkelOutStakingCredential :: Maybe StakingCredential

, txSkelOutDatum :: TxSkelOutDatum

, txSkelOutValue :: Value

, txSkelOutValueAutoAdjust :: Bool

, txSkelOutReferenceScript :: Maybe VScript

} -> TxSkelOut

Optics

txSkelOutValueL :: Lens' TxSkelOut Value

Focuses on the Value of this TxSkelOut

txSkelOutValueAutoAdjustL :: Lens' TxSkelOut Bool

Focuses on whether the Value contained in this TxSkelOut can be adjusted to min ADA during transaction generation

txSkelOutDatumL :: Lens' TxSkelOut TxSkelOutDatum

Optics focusing on the datum of a TxSkelOut

Focuses on the TxSkelOutDatum of this TxSkelOut

txSkelOutMReferenceScriptL :: Lens' TxSkelOut (Maybe VScript)

Focuses on the Maybe VScript corresponding to the possible reference script contained in this TxSkelOut

txSkelOutReferenceScriptAT :: AffineTraversal' TxSkelOut VScript

Focuses on the reference script of this TxSkelOut

txSkelOutMStakingCredentialL :: Lens' TxSkelOut (Maybe StakingCredential)

Focuses on the Maybe StakingCredential of this TxSkelOut

txSkelOutStakingCredentialAT :: AffineTraversal' TxSkelOut StakingCredential

Focuses on the staking credential of this TxSkelOut

txSkelOutCredentialG :: Getter TxSkelOut Credential

Returns the credential of this TxSkelOut

txSkelOutAddressG :: Getter TxSkelOut Address

Returns the address of this TxSkelOut

txSkelOutReferenceScriptHashAF :: AffineFold TxSkelOut ScriptHash

Returns the possible reference script has of this TxSkelOut

txSkelOutOwnerL :: Lens' TxSkelOut (User 'IsEither 'Allocation)

Focuses on the user of this TxSkelOut

Smart constructors

(<&&>) :: els ⩀ els' => Payable els -> Payable els' -> Payable (els ∪ els')

An infix-usable alias for PayableAnd

receives :: IsTxSkelOutAllowedOwner owner => owner -> Payable els -> TxSkelOut infix 1

Smart constructor to build a TxSkelOut from an owner and Payable. This should be the main way of building outputs.