Cooked.Skeleton.Output

Description

This module exposes outputs as they can be defined in a TxSkel with various utilities around them.

Documentation

data TxSkelOut where

A rich output to be put into a TxSkel

Constructors

TxSkelOut

Fields

:: OwnerConstrs owner

=> { txSkelOutOwner :: owner

, txSkelOutStakingCredential :: Maybe StakingCredential

, txSkelOutDatum :: TxSkelOutDatum

, txSkelOutValue :: Value

, txSkelOutValueAutoAdjust :: Bool

, txSkelOutReferenceScript :: TxSkelOutReferenceScript

} -> TxSkelOut

receives :: OwnerConstrs owner => owner -> Payable els -> TxSkelOut infix 1

Smart constructor to build a TxSkelOut from an owner and payment. This should be the main way of building outputs.

txSkelOutValueL :: Lens' TxSkelOut Value

A lens to get or set the Value from a TxSkelOut

txSkelOutValueAutoAdjustL :: Lens' TxSkelOut Bool

A lens to get or set if the value can be auto-adjusted if needed

txSkelOutDatumL :: Lens' TxSkelOut TxSkelOutDatum

A lens to get or set the TxSkelOutDatum from a TxSkelOut

txSkelOutReferenceScriptL :: Lens' TxSkelOut TxSkelOutReferenceScript

A lens to get or set the TxSkelOutReferenceScript from a TxSkelOut

txSkelOutStakingCredentialL :: Lens' TxSkelOut (Maybe StakingCredential)

A lens to get or set the 'Maybe Api.StakingCredential' from a TxSkelOut

txSkelOutValidatorAT :: AffineTraversal' TxSkelOut (Versioned Validator)

Returns the optional validator owning a given TxSkelOut

class IsTxSkelOutAllowedOwner a where

A TxSkelOut can either be owned by a pubkeyhash or a versioned validator

Methods

toPKHOrValidator :: a -> Either PubKeyHash (Versioned Validator)

type OwnerConstrs owner = (IsTxSkelOutAllowedOwner owner, Typeable owner, Show owner)

Type constraints over the owner of a TxSkelOut

txSkelOutCredentialG :: Getter TxSkelOut Credential

Returns the credential of this TxSkelOut

txSkelOutAddressG :: Getter TxSkelOut Address

Returns the address of this TxSkelOut

txSkelOutPKHashAT :: AffineTraversal' TxSkelOut PubKeyHash

Returns the optional private key owning a given TxSkelOut

txSkelOutTypedOwnerAT :: (OwnerConstrs a, OwnerConstrs b) => AffineTraversal TxSkelOut TxSkelOut a b

Attempts to retrieve or set a typed owner from this TxSkelOut

txSkelOutValidatorHashAF :: AffineFold TxSkelOut ValidatorHash

Returns the optional validator hash owning a given TxSkelOut

valueAssetClassAmountL :: ToMintingPolicyHash mp => mp -> TokenName -> Lens' Value Integer

A lens to get or set the amount of tokens of a certain AssetClass from a given Value. This removes the entry if the new amount is 0.

lovelaceIntegerI :: Iso' Lovelace Integer

Isomorphism between Lovelace and integers

valueLovelaceL :: Lens' Value Lovelace

Focus the Lovelace part in a value.

valueAssetClassAmountP :: ToMintingPolicyHash mp => mp -> TokenName -> Prism' Value Integer

A prism to build a value from an asset class and amount, or retrieves the amount from this asset class if it is not zero

valueLovelaceP :: Prism' Value Lovelace

An instance of valueAssetClassAmountP for Lovelace

ownerCredentialG :: IsTxSkelOutAllowedOwner owner => Getter owner Credential

Retrieves the credential of a TxSkelOut allowed owner