Cooked.Skeleton.Proposal

Description

This module exposes the notion of proposal within a TxSkel

Documentation

data TxParameterChange where

These are all the protocol parameters. They are taken from https://github.com/IntersectMBO/cardano-ledger/blob/c4fbc05999866fea7c0cb1b211fd5288f286b95d/eras/conway/impl/cddl-files/conway.cddl#L381-L412 and will most likely change in future eras.

Constructors

FeePerByte :: Integer -> TxParameterChange
The linear factor for the minimum fee calculation

FeeFixed :: Integer -> TxParameterChange
The constant factor for the minimum fee calculation

MaxBlockBodySize :: Integer -> TxParameterChange
Maximal block body size

MaxTxSize :: Integer -> TxParameterChange
Maximal transaction size

MaxBlockHeaderSize :: Integer -> TxParameterChange
Maximal block header size

KeyDeposit :: Integer -> TxParameterChange
The amount of a key registration deposit

PoolDeposit :: Integer -> TxParameterChange
The amount of a pool registration deposit

PoolRetirementMaxEpoch :: Integer -> TxParameterChange
Maximum number of epochs in the future a pool retirement is allowed to be scheduled future for.

PoolNumber :: Integer -> TxParameterChange
Desired number of pools

PoolInfluence :: Rational -> TxParameterChange
Pool influence

MonetaryExpansion :: Rational -> TxParameterChange
Monetary expansion

TreasuryCut :: Rational -> TxParameterChange
Treasury expansion

MinPoolCost :: Integer -> TxParameterChange
Minimum Stake Pool Cost

CoinsPerUTxOByte :: Integer -> TxParameterChange
Cost in lovelace per byte of UTxO storage

CostModels
Cost models for non-native script languages

Fields

:: { cmPlutusV1Costs :: [Integer]

, cmPlutusV2Costs :: [Integer]

, cmPlutusV3Costs :: [Integer]

} -> TxParameterChange

Prices
Prices of execution units

Fields

:: { pMemoryCost :: Rational

, pStepCost :: Rational

} -> TxParameterChange

MaxTxExUnits
Max total script execution resources units allowed per tx

Fields

:: { mteuMemory :: Integer

, mteuSteps :: Integer

} -> TxParameterChange

MaxBlockExUnits
Max total script execution resources units allowed per block

Fields

:: { mbeuMemory :: Integer

, mbeuSteps :: Integer

} -> TxParameterChange

MaxValSize :: Integer -> TxParameterChange
Max size of a Value in an output

CollateralPercentage :: Integer -> TxParameterChange
Percentage of the txfee which must be provided as collateral when including non-native scripts.

MaxCollateralInputs :: Integer -> TxParameterChange
Maximum number of collateral inputs allowed in a transaction

PoolVotingThresholds
Thresholds for pool votes

Fields

:: { pvtMotionNoConfidence :: Rational

, pvtCommitteeNormal :: Rational

, pvtCommitteeNoConfidence :: Rational

, pvtHardFork :: Rational

, pvtSecurityGroup :: Rational

} -> TxParameterChange

DRepVotingThresholds
Thresholds for DRep votes

Fields

:: { drvtMotionNoConfidence :: Rational

, drvtCommitteeNormal :: Rational

, drvtCommitteeNoConfidence :: Rational

, drvtUpdateConstitution :: Rational

, drvtHardForkInitialization :: Rational

, drvtNetworkGroup :: Rational

, drvtEconomicGroup :: Rational

, drvtTechnicalGroup :: Rational

, drvtGovernanceGroup :: Rational

, drvtTreasuryWithdrawal :: Rational

} -> TxParameterChange

CommitteeMinSize :: Integer -> TxParameterChange
Minimum size of the Constitutional Committee

CommitteeMaxTermLength :: Integer -> TxParameterChange
The Constitutional Committee Term limit in number of Slots

GovActionLifetime :: Integer -> TxParameterChange
Gov action lifetime in number of Epochs

GovActionDeposit :: Integer -> TxParameterChange
The amount of the Gov Action deposit

DRepRegistrationDeposit :: Integer -> TxParameterChange
The amount of a DRep registration deposit

DRepActivity :: Integer -> TxParameterChange
The number of Epochs that a DRep can perform no activity without losing their Active status.

data TxGovAction where

This lists the various possible governance actions

Constructors

TxGovActionParameterChange :: [TxParameterChange] -> TxGovAction

TxGovActionHardForkInitiation :: ProtocolVersion -> TxGovAction

TxGovActionTreasuryWithdrawals :: Map Credential Lovelace -> TxGovAction

TxGovActionNoConfidence :: TxGovAction

TxGovActionUpdateCommittee :: [ColdCommitteeCredential] -> Map ColdCommitteeCredential Integer -> Rational -> TxGovAction

TxGovActionNewConstitution :: Constitution -> TxGovAction

data TxSkelProposal where

This bundles a governance action into an actual proposal

Constructors

TxSkelProposal

Fields

:: { txSkelProposalAddress :: Address

Whatever credential will get back the deposit

, txSkelProposalAction :: TxGovAction

The proposed action

, txSkelProposalWitness :: Maybe (Versioned Script, TxSkelRedeemer)

An optional script (typically the constitution script) to witness the proposal and validate it. Only parameter changes and treasury withdrawals can be subject to such a validation and transactions will not pass validation phase 1 if other actions are given a witness.

, txSkelProposalAnchor :: Maybe String

An optional anchor to be given as additional data. It should correspond to the URL of a web page

} -> TxSkelProposal

txSkelProposalAddressL :: Lens' TxSkelProposal Address

A lens to get or set the address of a TxSkelProposal

txSkelProposalActionL :: Lens' TxSkelProposal TxGovAction

A lens to get or set the governance action of a TxSkelProposal

txSkelProposalWitnessL :: Lens' TxSkelProposal (Maybe (Versioned Script, TxSkelRedeemer))

A lens to get or set the witness of a TxSkelProposal

txSkelProposalAnchorL :: Lens' TxSkelProposal (Maybe String)

A lens to get or set the anchor of a TxSkelProposal

simpleTxSkelProposal :: ToAddress a => a -> TxGovAction -> TxSkelProposal

Builds a TxSkelProposal from an address and a TxGovAction

withWitness :: ToVersioned Script a => TxSkelProposal -> (a, TxSkelRedeemer) -> TxSkelProposal

Assigns a witness to a TxSkelProposal

withAnchor :: TxSkelProposal -> String -> TxSkelProposal

Assigns an anchor to a TxSkelProposal