Cooked.Skeleton.Withdrawal

Documentation

type TxSkelWithdrawals = Map (Either (Versioned Script) PubKeyHash) (TxSkelRedeemer, Ada)

Withdrawals associate either a script or a private key with a redeemer and a certain amount of ada. Note that the redeemer will be ignored in the case of a private key.

pkWithdrawal :: ToPubKeyHash pkh => pkh -> Ada -> TxSkelWithdrawals

scriptWithdrawal :: ToVersionedScript script => script -> TxSkelRedeemer -> Ada -> TxSkelWithdrawals