Cooked.Tweak.Common

Description

This module defines Tweaks which are the fundamental building blocks of our "domain specific language" for attacks.

The type of tweaks

class (MonadPlus m, MonadBlockChainWithoutValidation m) => MonadTweak m where

Methods

getTxSkel :: m TxSkel

putTxSkel :: TxSkel -> m ()

type Tweak m = StateT TxSkel (ListT m)

runTweakInChain :: (MonadBlockChainWithoutValidation m, MonadPlus m) => Tweak m a -> TxSkel -> m (a, TxSkel)

This is the function that gives a meaning to Tweaks: A Tweak is a computation that, depending on the state of the chain, looks at a transaction and returns zero or more modified transactions, together with some additional values.

Our intuition (and also the language of the comments pertaining to Tweaks) is that a Tweak t

fails if runTweakInChain t skel is mzero.

returns the value in the first component of the pair returned by this function (which is also the value it returns in the monad Tweak m).

modifies a TxSkel. Since it can use every method of MonadBlockChainWithoutValidateTxSkel to do so, this also includes stateful lookups or even things like waiting for a certain amount of time before submitting the transaction.

data UntypedTweak m where

This is a wrapper type used in the implementation of the Staged monad. You will probably never use it while you're building Tweaks.

Constructors

UntypedTweak :: Tweak m a -> UntypedTweak m

A few fundamental tweaks

failingTweak :: MonadTweak m => m a

The never-applicable tweak.

doNothingTweak :: MonadTweak m => m ()

The tweak that always applies and leaves the transaction unchanged.

Constructing Tweaks from Optics

viewTweak :: (MonadTweak m, Is k A_Getter) => Optic' k is TxSkel a -> m a

The "tweak" that obtains some value from the TxSkel. This does *not* modify the transaction.

viewAllTweak :: (MonadTweak m, Is k A_Fold) => Optic' k is TxSkel a -> m [a]

Like the viewTweak, but returns a list of all foci

setTweak :: (MonadTweak m, Is k A_Setter) => Optic' k is TxSkel a -> a -> m ()

The tweak that sets a certain value in the TxSkel.

overTweak :: (MonadTweak m, Is k A_Setter) => Optic' k is TxSkel a -> (a -> a) -> m ()

The tweak that modifies a certain value in the TxSkel.

overMaybeTweak :: (MonadTweak m, Is k A_Traversal) => Optic' k is TxSkel a -> (a -> Maybe a) -> m [a]

Like overTweak, but only modifies foci on which the argument function returns Just the new focus. Returns a list of the foci that were modified, as they were before the tweak, and in the order in which they occurred on the original transaction.

overMaybeSelectingTweak :: forall a m k is. (MonadTweak m, Is k A_Traversal) => Optic' k is TxSkel a -> (a -> Maybe a) -> (Integer -> Bool) -> m [a]

Sometimes overMaybeTweak modifies too many foci. This might be the case if there are several identical foci, but you only want to modify some of them. This is where this Tweak becomes useful: The (Integer -> Bool) argument can be used to select which of the modifiable foci should be actually modified.

selectP :: (a -> Bool) -> Prism' a a

overMaybeTweak requires a modification that can fail (targeting Maybe). Sometimes, it can prove more convenient to explicitly state which property the foci shoud satisfy to be eligible for a modification that cannot fail instead. selectP provides a prism to make such a selection. The intended use case is 'overTweak (optic % selectP prop) mod' where optic gives the candidate foci, prop is the predicate to be satisfied by the foci, and mod is the modification to be applied to the selected foci.