Cooked.Tweak.TamperDatum

Documentation

tamperDatumTweak

Arguments

:: forall a m. (MonadTweak m, Show a, PrettyCooked a, ToData a, FromData a, Typeable a)

=> (a -> Maybe a)
Use this function to return Just the changed datum, if you want to perform a change, and Nothing, if you want to leave it as-is. All datums on outputs not paying to a validator of type a are never touched.

-> m [a]

A tweak that tries to change the datum on outputs carrying datums of a certain type with a prescribed tampering function.

The tweak returns a list of the modified datums, as they were *before* the modification was applied to them.

data TamperDatumLbl

Constructors

TamperDatumLbl

malformDatumTweak :: forall a m. (MonadTweak m, ToData a, FromData a, Typeable a) => (a -> [BuiltinData]) -> m ()

A tweak that tries to change the datum on outputs carrying datums of a certain type with a prescribed tampering function. There are two main differences with tamperDatumTweak. First, the tampering function returns BuiltinData, allowing it to do pretty much anything with the datums. Second, for every output datum there are zero or more options for how to modify it, and all combinations of these modifications are tried.

That is, if there are n output datums, for which there are 'k_1,...,k_n' possible modifications, this tweak will try

  k_1 + ... + k_n
+ k_1 * k_2 + ... + k_{n-1} * k_n
+ k_1 * k_2 * k_3 + ... + k_{n-2} * k_{n-1} * k_n
+ ...
+ k_1 * k_2 * ... * k_{n-1} * k_n
== (k_1 + 1) * ... * (k_n + 1) - 1

modified transactions.

data MalformDatumLbl

Constructors

MalformDatumLbl