Cooked.Wallet

Description

This module defines convenient wrappers for mock chain wallets (around Plutus mock wallets) with an associate API to construct them, manipulate them, and fetch information (such as public/private and staking keys).

Documentation

knownWallets :: [Wallet]

All the wallets corresponding to known Plutus mock wallets. This is a list of 10 wallets which will

receive funds in the standard initial distribution of cooked-validators,

be pretty-printed as part the final state after running a few transactions.

wallet :: Integer -> Wallet

Wallet corresponding to a given wallet number (or wallet ID) with an offset of 1 to start at 1 instead of 0

walletPKHashToId :: PubKeyHash -> Maybe Int

Retrieves the id of the known wallet that corresponds to a public key hash

walletPKHashToId (walletPKHash (wallet 3)) == Just 3

walletPKHashToWallet :: PubKeyHash -> Maybe Wallet

Retrieves the known wallet that corresponds to a public key hash

walletPK :: Wallet -> PubKey

Retrieves a wallet public key (PK)

walletStakingPK :: Wallet -> Maybe PubKey

Retrieves a wallet's public staking key (PK), if any

walletPKHash :: Wallet -> PubKeyHash

Retrieves a wallet's public key hash

walletStakingPKHash :: Wallet -> Maybe PubKeyHash

Retrieves a wallet's public staking key hash, if any

walletSK :: Wallet -> XPrv

Retrieves a wallet private key (secret key SK)

walletStakingSK :: Wallet -> Maybe XPrv

Retrieves a wallet's private staking key (secret key SK), if any

walletStakingCredential :: Wallet -> Maybe StakingCredential

Retrieves a wallet's staking credential

type Wallet = MockWallet

A Wallet is a MockWallet from plutus-ledger

Orphan instances

Eq Wallet

Instance details

Methods

(==) :: Wallet -> Wallet -> Bool

(/=) :: Wallet -> Wallet -> Bool

Ord Wallet

Instance details

Methods

compare :: Wallet -> Wallet -> Ordering

(<) :: Wallet -> Wallet -> Bool

(<=) :: Wallet -> Wallet -> Bool

(>) :: Wallet -> Wallet -> Bool

(>=) :: Wallet -> Wallet -> Bool

max :: Wallet -> Wallet -> Wallet

min :: Wallet -> Wallet -> Wallet

ToAddress Wallet

Instance details

Methods

toAddress :: Wallet -> Address

ToCredential Wallet

Instance details

Methods

toCredential :: Wallet -> Credential